PLNmodels: Poisson Lognormal Models

The Poisson-lognormal model and variants (Chiquet, Mariadassou and Robin, 2021 doi:10.3389/fevo.2021.588292 ) can be used for a variety of multivariate problems when count data are at play, including principal component analysis for count data, discriminant analysis, model-based clustering and network inference. Implements variational algorithms to fit such models accompanied with a set of functions for visualization and diagnostic.

See also

Author

Maintainer: Julien Chiquet julien.chiquet@inrae.fr (ORCID)

Authors:

Julien Chiquet julien.chiquet@inrae.fr (ORCID)

Mahendra Mariadassou mahendra.mariadassou@inrae.fr (ORCID)

Stéphane Robin stephane.robin@inrae.fr

François Gindraud francois.gindraud@gmail.com

Other contributors:

Julie Aubert julie.aubert@inrae.fr [contributor]

Bastien Batardière bastien.batardiere@inrae.fr [contributor]

Giovanni Poggiato giov.poggiato@gmail.com [contributor]

Cole Trapnell coletrap@uw.edu [contributor]

Maddy Duran duran@uw.edu [contributor]