Display various outputs (goodness-of-fit criteria, robustness, diagnostic) associated with a collection of PLNnetwork fits (a PLNnetworkfamily)

Display various outputs (goodness-of-fit criteria, robustness, diagnostic) associated with a collection of PLNnetwork fits (a PLNnetworkfamily)

# S3 method for PLNnetworkfamily
plot(
  x,
  type = c("criteria", "stability", "diagnostic"),
  criteria = c("loglik", "pen_loglik", "BIC", "EBIC"),
  log.x = TRUE,
  stability = 0.9,
  ...
)

Arguments

x
an R6 object with class PLNnetworkfamily

type
a character, either "criteria", "stability" or "diagnostic" for the type of plot.

criteria
vector of characters. The criteria to plot in c("loglik", "BIC", "ICL", "R_squared", "EBIC", "pen_loglik"). Default is c("loglik", "pen_loglik", "BIC", "EBIC"). Only relevant when type = "criteria".

log.x
logical: should the x-axis be represented in log-scale? Default is TRUE.

stability
scalar: the targeted level of stability in stability plot. Default is .9.

...
additional parameters for S3 compatibility. Not used

Value

Produces a plot representing the evolution of the criteria of the different models considered, highlighting the best model in terms of BIC and EBIC (the greater, the better). These criteria have the form 'loglik - 1/2 * penalty' so that they are on the same scale as the model loglikelihood.

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
fits <- PLNnetwork(Abundance ~ 1, data = trichoptera)

#> 
#>  Initialization...
#>  Adjusting 30 PLN with sparse inverse covariance estimation
#> 	Joint optimization alternating gradient descent and graphical-lasso
#> 	sparsifying penalty = 7.510862 	sparsifying penalty = 6.937564 	sparsifying penalty = 6.408026 	sparsifying penalty = 5.918907 	sparsifying penalty = 5.467121 	sparsifying penalty = 5.049821 	sparsifying penalty = 4.664372 	sparsifying penalty = 4.308345 	sparsifying penalty = 3.979492 	sparsifying penalty = 3.675741 	sparsifying penalty = 3.395175 	sparsifying penalty = 3.136024 	sparsifying penalty = 2.896654 	sparsifying penalty = 2.675555 	sparsifying penalty = 2.471332 	sparsifying penalty = 2.282698 	sparsifying penalty = 2.108461 	sparsifying penalty = 1.947524 	sparsifying penalty = 1.798871 	sparsifying penalty = 1.661565 	sparsifying penalty = 1.534739 	sparsifying penalty = 1.417594 	sparsifying penalty = 1.30939 	sparsifying penalty = 1.209446 	sparsifying penalty = 1.11713 	sparsifying penalty = 1.03186 	sparsifying penalty = 0.953099 	sparsifying penalty = 0.8803498 	sparsifying penalty = 0.8131535 	sparsifying penalty = 0.7510862 
#>  Post-treatments
#>  DONE!

if (FALSE) {
plot(fits)
}