Prediction for a PLNmixturefit object

Predict either posterior probabilities for each group or latent positions based on new samples

Usage

# S3 method for class 'PLNmixturefit'
predict(
  object,
  newdata,
  type = c("posterior", "response", "position"),
  prior = matrix(rep(1/object$k, object$k), nrow(newdata), object$k, byrow = TRUE),
  control = PLNmixture_param(),
  ...
)

Arguments

object

an R6 object with class PLNmixturefit

newdata

A data frame in which to look for variables, offsets and counts with which to predict.

type

The type of prediction required. The default posterior are posterior probabilities for each group , response is the group with maximal posterior probability and latent is the averaged latent in the latent space, with weights equal to the posterior probabilities.

prior

User-specified prior group probabilities in the new data. The default uses a uniform prior.

control

a list-like structure for controlling the fit. See PLNmixture_param() for details.

...

additional parameters for S3 compatibility. Not used

Value

A matrix of posterior probabilities for each group (if type = "posterior"), a matrix of (average) position in the latent space (if type = "position") or a vector of predicted groups (if type = "response").

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLNmixture(Abundance ~ 1 + offset(log(Offset)),
           data = trichoptera, control = PLNmixture_param(smoothing = "none"))  %>% getBestModel()
#> 
#>  Initialization...
#> 
#>  Adjusting 5 PLN mixture models.
#> 	number of cluster = 1 	number of cluster = 2 	number of cluster = 3 	number of cluster = 4 	number of cluster = 5 
#>  Post-treatments
#>  DONE!
predict(myPLN, trichoptera, "posterior")
#>            [,1]         [,2]         [,3]         [,4]
#> 1  1.846404e-05 2.693550e-09 9.723611e-01 2.762043e-02
#> 2  4.288843e-02 1.872627e-09 9.570432e-01 6.834636e-05
#> 3  6.913514e-01 1.874787e-08 3.086482e-01 3.820535e-07
#> 4  9.998512e-01 2.220446e-16 1.488463e-04 2.220446e-16
#> 5  9.834603e-01 1.909651e-10 1.653967e-02 6.504137e-10
#> 6  3.887886e-01 4.964487e-07 6.111839e-01 2.705985e-05
#> 7  7.753928e-03 1.349880e-06 9.842668e-01 7.977963e-03
#> 8  9.325778e-01 4.895405e-07 6.741926e-02 2.459188e-06
#> 9  4.912441e-02 3.476890e-07 9.501757e-01 6.995413e-04
#> 10 4.837733e-02 6.787513e-05 9.330545e-01 1.850029e-02
#> 11 1.683353e-01 4.055854e-03 7.969957e-01 3.061311e-02
#> 12 3.493874e-01 9.882617e-02 5.118642e-01 3.992221e-02
#> 13 3.754293e-04 2.220446e-16 9.996246e-01 2.220446e-16
#> 14 3.865201e-04 2.220446e-16 9.995306e-01 8.286911e-05
#> 15 9.999643e-01 2.220446e-16 3.566450e-05 2.220446e-16
#> 16 9.960090e-01 2.220446e-16 3.991011e-03 2.220446e-16
#> 17 1.491471e-04 7.310397e-09 9.998508e-01 6.601541e-08
#> 18 8.340828e-06 2.220446e-16 9.999909e-01 8.027453e-07
#> 19 5.214391e-03 2.220446e-16 9.947856e-01 6.934109e-09
#> 20 2.220446e-16 2.220446e-16 9.996107e-01 3.892575e-04
#> 21 1.612306e-03 2.220446e-16 9.982847e-01 1.029838e-04
#> 22 2.753295e-03 1.714122e-07 6.064696e-01 3.907770e-01
#> 23 3.252643e-10 2.317034e-06 1.067660e-03 9.989300e-01
#> 24 1.318892e-01 2.973938e-03 8.210206e-01 4.411622e-02
#> 25 2.220446e-16 9.226979e-06 2.495940e-04 9.997412e-01
#> 26 2.220446e-16 2.220446e-16 5.221422e-08 9.999999e-01
#> 27 2.220446e-16 2.220446e-16 5.234078e-06 9.999948e-01
#> 28 5.668373e-08 1.080759e-03 9.510140e-04 9.979682e-01
#> 29 4.395234e-04 2.203574e-02 4.378171e-03 9.731466e-01
#> 30 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 31 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 32 5.966838e-06 2.220446e-16 9.999933e-01 7.582211e-07
#> 33 5.340459e-04 2.220446e-16 9.994660e-01 2.220446e-16
#> 34 9.943996e-01 2.220446e-16 5.600361e-03 2.220446e-16
#> 35 4.833743e-02 2.220446e-16 9.516626e-01 4.903383e-10
#> 36 9.737803e-01 2.220446e-16 2.621974e-02 5.828882e-10
#> 37 2.987385e-07 2.220446e-16 9.999996e-01 1.115953e-07
#> 38 1.346727e-04 2.220446e-16 9.971908e-01 2.674558e-03
#> 39 2.674697e-03 2.220446e-16 9.944108e-01 2.914475e-03
#> 40 6.242136e-01 4.164275e-04 3.752960e-01 7.405508e-05
#> 41 3.768345e-02 2.220446e-16 9.623165e-01 2.220446e-16
#> 42 4.486671e-05 2.220446e-16 9.998612e-01 9.396783e-05
#> 43 4.222961e-02 2.220446e-16 9.577689e-01 1.523022e-06
#> 44 2.220446e-16 2.220446e-16 9.999997e-01 3.422378e-07
#> 45 1.965158e-10 2.220446e-16 1.000000e+00 2.220446e-16
#> 46 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 47 7.319719e-03 1.590838e-05 9.896158e-01 3.048604e-03
#> 48 1.333165e-07 2.220446e-16 9.999997e-01 1.872559e-07
#> 49 2.220446e-16 2.220446e-16 5.477745e-06 9.999945e-01
predict(myPLN, trichoptera, "position")
#>          [,1]       [,2]      [,3]       [,4]        [,5]      [,6]       [,7]
#> 1   -6.693574  -7.612554 -1.998480  -5.997781 -0.56540692 -4.191958  -5.862564
#> 2   -6.785920  -7.881257 -1.936207  -6.268274 -0.54113577 -3.969593  -6.130427
#> 3  -10.874071 -11.221113 -3.585630 -10.713402 -0.19350132 -4.017961 -10.671098
#> 4  -12.805420 -12.805573 -4.090995 -12.805353 -0.09119944 -3.796877 -12.805337
#> 5  -12.703350 -12.721456 -3.336227 -12.695167 -0.14455067 -3.639043 -12.693041
#> 6   -8.965377  -9.662364 -2.700933  -8.637536 -0.29849395 -3.468920  -8.550446
#> 7   -6.603876  -7.682619 -1.983583  -6.014499 -0.72459119 -3.926236  -5.871201
#> 8  -12.383705 -12.459706 -3.729133 -12.348397 -0.17909638 -3.556201 -12.339086
#> 9   -6.827544  -7.911689 -1.631776  -6.308575 -0.47448864 -3.955917  -6.171419
#> 10  -6.915522  -7.869133 -1.773021  -6.280511 -0.71630117 -3.869776  -6.143106
#> 11  -7.764959  -8.465961 -2.392947  -7.123995 -0.44947789 -3.867114  -6.982899
#> 12  -9.527228  -9.459500 -3.419825  -8.980223 -0.24389116 -4.144993  -8.314548
#> 13  -5.964619  -7.677336 -3.210614  -6.049037 -0.22863996 -3.549911  -5.914599
#> 14  -6.543497  -7.670772 -2.260334  -6.018976 -0.52773142 -3.363479  -5.296881
#> 15 -12.806874 -12.806906 -4.860008 -12.806861 -0.18344621 -3.474789 -12.806858
#> 16 -12.781848 -12.786000 -4.227605 -12.780038 -0.19544731 -3.083267 -12.779575
#> 17  -6.570572  -7.679762 -3.062725  -6.063549 -0.29717001 -3.972673  -5.931016
#> 18  -6.561774  -7.676251 -2.760647  -5.481079 -0.51619250 -3.012276  -5.916128
#> 19  -6.599924  -7.704942 -2.323783  -6.094211 -0.63665763 -2.660121  -5.961933
#> 20  -6.642879  -7.100157 -2.856321  -6.169253 -0.58812543 -3.432119  -5.552486
#> 21  -6.553689  -7.677841 -2.119978  -6.031169 -0.58221191 -4.201981  -5.893595
#> 22  -8.772072  -6.895603 -2.225564  -5.847560 -0.97881246 -4.618357  -5.764567
#> 23 -12.239262  -5.672358 -3.380870  -5.367458 -0.91537755 -5.613587  -5.615518
#> 24  -7.605224  -8.250005 -2.293842  -6.860397 -0.55220975 -3.881867  -6.722093
#> 25 -12.243963  -5.679855 -2.895209  -5.614033 -1.05890232 -5.390170  -5.390561
#> 26 -12.245403  -5.690910 -3.353593  -5.626196 -1.40647273 -5.637078  -5.637078
#> 27 -12.245378  -5.694424 -3.379447  -5.629925 -1.16264100 -5.640759  -5.640767
#> 28 -12.240185  -5.666837 -3.308227  -5.603055 -1.24492836 -5.606755  -5.607648
#> 29 -12.226177  -5.737270 -3.376619  -5.753223 -0.99764617 -5.638217  -5.631826
#> 30 -12.504319  -8.378705 -6.665056 -12.504319 -0.30982859 -7.223315  -6.458349
#> 31 -12.505498  -8.624824 -5.751617 -12.505498 -0.10983721 -7.527903  -6.591442
#> 32  -6.602035  -7.690750 -2.554801  -6.112176 -0.97637881 -3.366641  -5.985197
#> 33  -6.651521  -7.711142 -2.450760  -6.183908 -0.78083416 -3.039717  -6.063819
#> 34 -12.772939 -12.778393 -3.196247 -12.770640 -0.39702361 -2.876079 -12.770060
#> 35  -6.879541  -7.929836 -2.470071  -6.401622 -0.82402441 -2.777540  -6.277006
#> 36 -12.643524 -12.672036 -3.411481 -12.630707 -0.40767037 -3.488497 -12.627386
#> 37  -6.575213  -7.680977 -1.373227  -5.514481 -1.08323044 -3.678174  -5.939671
#> 38  -6.534471  -7.656914 -1.734597  -5.979552 -1.35703249 -4.011562  -5.836578
#> 39  -6.550547  -7.669087 -1.473029  -5.994666 -0.98089519 -3.999099  -5.851783
#> 40 -10.448466 -10.873773 -3.299095 -10.247401 -0.14971399 -3.873246 -10.191662
#> 41  -6.817395  -7.876757 -2.817396  -6.336400 -0.37257781 -4.420540  -6.211128
#> 42  -6.564811  -7.677053 -2.786999  -5.487067 -0.49221665 -4.331915  -5.920377
#> 43  -6.870903  -7.909313 -3.217021  -6.405187 -0.30371760 -4.194327  -6.284663
#> 44  -6.552470  -7.673000 -2.116138  -5.457283 -0.45281691 -3.847605  -4.800067
#> 45  -6.574102  -7.680581 -2.547131  -5.511781 -0.53105679 -3.383594  -5.389398
#> 46  -6.124585  -7.713545 -3.369825  -6.198924 -0.37699883 -4.439011  -5.597808
#> 47  -6.583145  -7.693558 -2.312816  -6.030675 -0.43534828 -4.021756  -5.889002
#> 48  -5.925537  -7.668136 -3.287404  -6.012463 -0.62324367 -4.168802  -5.873621
#> 49 -12.245432  -5.507071 -3.450462  -5.667356 -1.16847657 -5.677820  -5.677828
#>         [,8]       [,9]     [,10]     [,11]     [,12]     [,13]      [,14]
#> 1  -5.448897  -7.060195 -3.017519 -4.330228 -4.001481 -4.220581  -5.447476
#> 2  -5.458717  -7.152645 -3.598628 -4.275074 -3.297915 -4.142186  -6.359162
#> 3  -6.558779 -10.989445 -4.616335 -5.302608 -3.285674 -4.821552 -10.741466
#> 4  -7.122966 -12.805470 -5.021923 -5.932994 -4.468484 -5.409795 -12.805365
#> 5  -7.041717 -12.709315 -5.073151 -5.752649 -4.675153 -5.157357 -12.696584
#> 6  -6.038585  -9.198362 -4.108157 -4.803759 -4.045305 -4.472644  -8.695047
#> 7  -5.383679  -6.983024 -2.490780 -4.162014 -4.274454 -4.039924  -6.093436
#> 8  -6.934660 -12.409000 -4.862709 -5.600200 -4.879976 -4.977380 -12.354559
#> 9  -5.465462  -7.192016 -3.589545 -4.273188 -4.326916 -4.135231  -6.397765
#> 10 -5.438221  -7.276125 -3.422065 -4.163497 -4.266604 -4.004802  -6.333670
#> 11 -5.651185  -8.050496 -3.604494 -4.331705 -4.367513 -4.097943  -7.104577
#> 12 -6.186328  -9.169684 -4.068152 -4.543419 -4.884687 -4.303912  -8.072542
#> 13 -4.973171  -6.931669 -3.291583 -4.493190 -4.134522 -4.405453  -6.138364
#> 14 -5.457981  -6.918905 -3.419133 -4.387689 -4.472875 -4.293375  -6.110469
#> 15 -6.523583 -12.806884 -3.720888 -5.093740 -5.346344 -5.876867 -12.806863
#> 16 -7.108253 -12.783196 -4.422502 -4.852574 -4.665356 -5.372145 -12.780348
#> 17 -5.530840  -6.332392 -2.907705 -3.461812 -4.621183 -4.459430  -6.151711
#> 18 -5.511282  -6.931145 -3.619970 -2.490845 -4.583586 -4.417701  -6.139261
#> 19 -5.533139  -6.965748 -3.370911 -2.945132 -4.612642 -4.451155  -6.182182
#> 20 -5.231338  -6.990663 -3.291733 -2.161372 -4.898336 -4.760464  -5.734795
#> 21 -5.466241  -6.927913 -2.359330 -3.946229 -4.486998 -4.309588  -6.122205
#> 22 -5.471493  -9.004461 -2.779102 -3.842700 -4.825693 -3.516171  -5.151300
#> 23 -5.593359 -12.239656 -3.210367 -3.289852 -5.627446 -2.650588  -3.482909
#> 24 -5.587196  -7.906487 -3.525309 -4.252325 -4.350604 -4.028759  -6.826057
#> 25 -5.603349 -12.244005 -3.488276 -2.993614 -5.638015 -2.773777  -3.351082
#> 26 -5.615818 -12.245403 -3.384645 -3.105888 -5.418948 -2.474332  -3.261183
#> 27 -5.619581 -12.245380 -2.773473 -2.978324 -5.653966 -2.061842  -3.834364
#> 28 -5.347369 -12.234474 -3.347215 -3.207956 -5.621983 -2.509518  -3.413668
#> 29 -5.634300 -12.103965 -3.405213 -3.012526 -5.679568 -2.579389  -3.660002
#> 30 -7.757917  -6.844346 -5.376867 -3.152877 -8.625790 -4.189472  -3.179658
#> 31 -7.861751  -6.911695 -5.920752 -4.072668 -8.491332 -5.168532  -4.018888
#> 32 -5.605543  -6.960136 -4.217111 -4.687290 -3.747407 -1.137644  -5.655442
#> 33 -5.704626  -6.997340 -3.689960 -4.515860 -2.430401 -1.936938  -6.264300
#> 34 -7.176072 -12.774683 -5.007711 -6.072555 -4.401209 -2.683996 -12.771030
#> 35 -5.624638  -7.226503 -3.248775 -4.642112 -3.909641 -3.507548  -6.484593
#> 36 -7.033783 -12.652898 -5.095824 -5.760899 -5.085274 -4.455585 -12.632923
#> 37 -5.543952  -6.940730 -3.434091 -4.164051 -3.868796 -2.978990  -6.158758
#> 38 -5.396522  -6.915785 -2.062140 -4.231982 -4.333216 -3.609345  -6.068886
#> 39 -5.397033  -6.931192 -3.097711 -4.225241 -4.325449 -3.595093  -6.083446
#> 40 -6.425626 -10.588900 -4.419157 -5.140354 -4.191833 -4.674651 -10.278815
#> 41 -5.618510  -7.167064 -3.278051 -4.649070 -4.287018 -4.547520  -6.419843
#> 42 -4.513033  -6.933476 -3.638877 -4.516084 -4.594697 -3.662941  -6.142614
#> 43 -5.682834  -7.212023 -2.349972 -4.387894 -4.799411 -3.719181  -6.485648
#> 44 -4.946079  -6.924538 -2.944975 -4.455613 -4.087862 -4.365898  -6.125655
#> 45 -4.562236  -6.939933 -2.441808 -4.157460 -4.223918 -4.481804  -6.157160
#> 46 -5.729866  -7.003898 -3.966869 -2.940842 -4.323900 -4.832315  -5.326832
#> 47 -4.818262  -6.961081 -3.147505 -3.731850 -4.346457 -4.140865  -6.118268
#> 48 -4.888252  -6.292283 -3.047988 -4.372440 -4.458654 -3.071534  -6.104465
#> 49 -5.657368 -12.245434 -3.624065 -3.124179 -5.690582 -1.045403  -3.996850
#>        [,15]      [,16]     [,17]
#> 1  -3.998076  -6.092985 -3.062285
#> 2  -5.019823  -5.959087 -3.147996
#> 3  -4.750036  -7.138793 -3.515608
#> 4  -4.653548  -7.719812 -3.920507
#> 5  -4.961023  -7.659975 -3.355636
#> 6  -5.101588  -6.584115 -3.249389
#> 7  -4.961537  -5.934876 -3.016753
#> 8  -5.242765  -7.556360 -2.643500
#> 9  -5.013718  -5.971971 -3.124717
#> 10 -4.910370  -6.065773 -2.884311
#> 11 -4.883521  -6.367349 -2.899840
#> 12 -4.646972  -6.936936 -3.150005
#> 13 -5.170733  -5.394030 -3.658248
#> 14 -4.574319  -5.927941 -2.353073
#> 15 -5.506472  -7.382706 -2.526253
#> 16 -5.551968  -7.708255 -2.791628
#> 17 -5.203005  -5.976828 -2.943533
#> 18 -5.178053  -5.963215 -2.653766
#> 19 -5.196609  -5.981699 -2.248152
#> 20 -3.782240  -6.093132 -2.432176
#> 21 -5.115558  -5.934413 -2.393076
#> 22 -3.005196  -8.373622 -2.657498
#> 23 -1.496870 -12.238579 -2.438974
#> 24 -4.827593  -6.383593 -2.853049
#> 25 -1.428719 -12.243765 -2.744798
#> 26 -1.175048 -12.245403 -2.207228
#> 27 -1.506850 -12.245375 -2.496706
#> 28 -1.320227 -12.234506 -2.474013
#> 29 -1.385115 -12.117740 -2.557536
#> 30 -1.964394  -7.757917 -5.120733
#> 31 -2.907274  -7.861751 -6.418033
#> 32 -5.297573  -6.029742 -2.610304
#> 33 -3.823193  -6.103192 -3.080550
#> 34 -4.531103  -7.747721 -2.336128
#> 35 -5.235311  -6.073184 -1.503704
#> 36 -5.393750  -7.647448 -1.596545
#> 37 -5.220047  -5.985886 -1.907366
#> 38 -5.015085  -5.904022 -2.109414
#> 39 -5.009553  -5.907743 -2.385776
#> 40 -5.146735  -7.005269 -3.282838
#> 41 -5.247001  -5.023760 -1.999486
#> 42 -5.185067  -5.967727 -1.759216
#> 43 -5.321120  -6.111016 -2.943810
#> 44 -5.147873  -5.947382 -3.597834
#> 45 -5.216682  -5.432317 -2.988440
#> 46 -4.380273  -5.202992 -2.185313
#> 47 -5.020290  -5.921535 -3.210741
#> 48 -5.098598  -5.922491 -1.938729
#> 49 -1.930137 -12.245429 -2.841893
predict(myPLN, trichoptera, "response")
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 
#>  3  3  1  1  1  3  3  1  3  3  3  3  3  3  1  1  3  3  3  3  3  3  4  3  4  4 
#> 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 
#>  4  4  4  2  2  3  3  1  3  1  3  3  3  1  3  3  3  3  3  3  3  3  4 
#> Levels: 1 2 3 4