Prepare data for use in PLN models from a phyloseq object

Wrapper around prepare_data, extracts the count table and the covariates data.frame from a "phyloseq" class object before passing them to prepare_data. See prepare_data for details.

prepare_data_from_phyloseq(physeq, offset = "TSS", ...)

Arguments

physeq
Required. A phyloseq class object from which the count table and covariates data.frame are extracted.

offset
Optional. Normalisation scheme used to compute scaling factors used as offset during PLN inference. Available schemes are "TSS" (Total Sum Scaling, default), "CSS" (Cumulative Sum Scaling, used in metagenomeSeq), "RLE" (Relative Log Expression, used in DESeq2), "GMPR" (Geometric Mean of Pairwise Ratio, introduced in Chen et al., 2018) or "none". Alternatively the user can supply its own vector or matrix of offsets (see note for specification of the user-supplied offsets).

...
Addtional arguments passed on to compute_offset

Details

This functions depends on the phyloseq package which is not a proper dependency of PLNmodels as it is not available on CRAN

See also

Examples

## Requires the phyloseq package
if (FALSE) {
library(phyloseq)
data(enterotypes)
prepare_data_from_phyloseq(enterotypes)
}