Calculate Variance-Covariance Matrix for a fitted PLN() model object

Returns the variance-covariance matrix of the main parameters of a fitted PLN() model object. The main parameters of the model correspond to $$B$$, as returned by coef.PLNfit(). The function can also be used to return the variance-covariance matrix of the residuals. The latter matrix can also be accessed via sigma.PLNfit()

Usage

# S3 method for class 'PLNfit'
vcov(object, type = c("main", "covariance"), ...)

Arguments

object

an R6 object with class PLNfit

type

type of parameter that should be extracted. Either "main" (default) for $$B$$ or "covariance" for $$\Sigma$$

...

additional parameters for S3 compatibility. Not used

Value

A matrix of variance/covariance extracted from the PLNfit model. If type="main" and \(B\) is a matrix of size d * p, the result is a block-diagonal matrix with p (number of species) blocks of size d (number of covariates). if type="main", it is a symmetric matrix of size p. .

See also

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLN(Abundance ~ 1 + offset(log(Offset)), data = trichoptera)
#> 
#>  Initialization...
#>  Adjusting a full covariance PLN model with nlopt optimizer
#>  Post-treatments...
#>  DONE!
vcov(myPLN, type = "covariance") ## Sigma
#>             Che          Hyc         Hym         Hys           Psy
#> Che  2.30755024  0.017446095  0.36678458  0.44761668 -2.265394e-02
#> Hyc  0.01744609  1.599138365  0.12426898  0.26506874 -1.351268e-01
#> Hym  0.36678458  0.124268980  1.24603742  0.86498919 -1.170216e-01
#> Hys  0.44761668  0.265068740  0.86498919  2.64658334 -1.442111e-01
#> Psy -0.02265394 -0.135126818 -0.11702161 -0.14421113  7.574884e-02
#> Aga  0.30089336 -0.098177919  0.51666810  0.43070546  2.084023e-04
#> Glo  0.09280033  0.154056996  0.20909150  0.33743179 -4.843917e-02
#> Ath  0.34270567  0.108450184  0.38780447  0.60493923 -4.678147e-02
#> Cea -0.01161195  0.006174725 -0.02623102 -0.01948513  1.508924e-05
#> Ced  0.33555793  0.215621628  0.50878071  0.67923660 -9.742270e-02
#> Set -0.05281945  0.424190738 -0.18937194  0.12967616 -9.118035e-02
#> All  0.26745636 -0.213709182  0.68383567  0.37547391 -5.431725e-03
#> Han -0.15994027  0.579972756  0.19852972  0.12290471 -2.909041e-01
#> Hfo -0.35112900  0.500226793 -0.41684825 -0.20585363 -1.476073e-01
#> Hsp -0.42513800  0.601987486 -0.36101282 -0.20716736 -1.923562e-01
#> Hve  0.29301036 -0.107157122  0.16865694  0.24811121  4.322799e-02
#> Sta  0.36006557  0.253538044  0.47777282  0.61305387 -1.215417e-01
#>               Aga          Glo          Ath           Cea          Ced
#> Che  0.3008933623  0.092800329  0.342705673 -1.161195e-02  0.335557934
#> Hyc -0.0981779194  0.154056996  0.108450184  6.174725e-03  0.215621628
#> Hym  0.5166680963  0.209091502  0.387804466 -2.623102e-02  0.508780713
#> Hys  0.4307054586  0.337431794  0.604939233 -1.948513e-02  0.679236601
#> Psy  0.0002084023 -0.048439171 -0.046781469  1.508924e-05 -0.097422695
#> Aga  1.0973832958 -0.058307096  0.201305668 -2.780459e-02  0.220999231
#> Glo -0.0583070958  1.100797383  0.221212520  2.304059e-03  0.224337574
#> Ath  0.2013056679  0.221212520  1.360798062 -8.304967e-03  0.441359890
#> Cea -0.0278045870  0.002304059 -0.008304967  5.223097e-02 -0.008965286
#> Ced  0.2209992305  0.224337574  0.441359890 -8.965286e-03  0.957367917
#> Set -0.3010745398  0.287062683  0.090954808  1.989664e-02  0.192525343
#> All  0.5903649515 -0.071141315  0.093824466 -3.099178e-02  0.082761092
#> Han -0.1756393223 -0.195997275 -0.206378496  6.875147e-03  0.052015653
#> Hfo -0.8268580779  0.295570258 -0.173899028  3.562476e-02 -0.143331047
#> Hsp -0.7899721786  0.242656717 -0.238169159  3.339798e-02 -0.128495684
#> Hve  0.1860557220  0.090515022  0.242715303 -8.974721e-03  0.207206840
#> Sta  0.4397273972  0.017926863  0.307770868 -1.643916e-02  0.434472222
#>             Set          All          Han         Hfo         Hsp          Hve
#> Che -0.05281945  0.267456363 -0.159940271 -0.35112900 -0.42513800  0.293010361
#> Hyc  0.42419074 -0.213709182  0.579972756  0.50022679  0.60198749 -0.107157122
#> Hym -0.18937194  0.683835672  0.198529719 -0.41684825 -0.36101282  0.168656937
#> Hys  0.12967616  0.375473912  0.122904705 -0.20585363 -0.20716736  0.248111213
#> Psy -0.09118035 -0.005431725 -0.290904135 -0.14760733 -0.19235622  0.043227994
#> Aga -0.30107454  0.590364952 -0.175639322 -0.82685808 -0.78997218  0.186055722
#> Glo  0.28706268 -0.071141315 -0.195997275  0.29557026  0.24265672  0.090515022
#> Ath  0.09095481  0.093824466 -0.206378496 -0.17389903 -0.23816916  0.242715303
#> Cea  0.01989664 -0.030991776  0.006875147  0.03562476  0.03339798 -0.008974721
#> Ced  0.19252534  0.082761092  0.052015653 -0.14333105 -0.12849568  0.207206840
#> Set  1.60438352 -0.684430941 -0.041170388  0.83218815  0.78435930 -0.040040611
#> All -0.68443094  1.800793128  0.023413955 -0.75607230 -0.66890927  0.147808855
#> Han -0.04117039  0.023413955  3.759440511  0.64486539  1.01483791 -0.597370730
#> Hfo  0.83218815 -0.756072299  0.644865392  2.87806607  1.70054442 -0.347305165
#> Hsp  0.78435930 -0.668909270  1.014837905  1.70054442  2.62371286 -0.485878346
#> Hve -0.04004061  0.147808855 -0.597370730 -0.34730516 -0.48587835  1.322574835
#> Sta  0.05566164  0.199563581  0.470754592 -0.26031330 -0.19873593  0.145765237
#>             Sta
#> Che  0.36006557
#> Hyc  0.25353804
#> Hym  0.47777282
#> Hys  0.61305387
#> Psy -0.12154171
#> Aga  0.43972740
#> Glo  0.01792686
#> Ath  0.30777087
#> Cea -0.01643916
#> Ced  0.43447222
#> Set  0.05566164
#> All  0.19956358
#> Han  0.47075459
#> Hfo -0.26031330
#> Hsp -0.19873593
#> Hve  0.14576524
#> Sta  1.03042934