An R6 Class to represent a PLNfit in a PCA framework

The function PLNPCA produces a collection of models which are instances of object with class PLNPCAfit.

This class comes with a set of methods, some of them being useful for the user: See the documentation for the methods inherited by PLNfit and the plot.PLNPCAfit methods for PCA vizualization

See also

Super class

PLNmodels::PLNfit -> PLNPCAfit

Public fields

var_par

a list with two matrices, M and S, which are the estimated parameters in the variational approximation

latent

a matrix: values of the latent vector (Z in the model)

optim_par

a list with parameters useful for monitoring the optimization

loglik

variational lower bound of the loglikelihood

BIC

variational lower bound of the BIC

ICL

variational lower bound of the ICL

R_squared

approximated goodness-of-fit criterion

criteria

a vector with loglik, BIC, ICL, R_squared and number of parameters

Active bindings

rank

the dimension of the current model

model_par

a list with the matrices associated with the estimated parameters of the pPCA model: Theta (covariates), Sigma (latent covariance) and B (latent loadings)

var_par

a list with two matrices, M and S, which are the estimated parameters in the variational approximation

latent

a matrix: values of the latent vector (Z in the model)

optim_par

a list with parameters useful for monitoring the optimization

loglik

variational lower bound of the loglikelihood

BIC

variational lower bound of the BIC

ICL

variational lower bound of the ICL

R_squared

approximated goodness-of-fit criterion

criteria

a vector with loglik, BIC, ICL, R_squared and number of parameters

nb_param

number of parameters in the current PLN model

percent_var

the percent of variance explained by each axis

corr_circle

a matrix of correlations to plot the correlation circles

scores

a matrix of scores to plot the individual factor maps (a.k.a. principal comonents)

rotation

a matrix of rotation of the latent space

Methods

Public methods

PLNPCAfit$new()

PLNPCAfit$update()

PLNPCAfit$setVisualization()

PLNPCAfit$optimize()

PLNPCAfit$postTreatment()

PLNPCAfit$latent_pos()

PLNPCAfit$plot_individual_map()

PLNPCAfit$plot_correlation_circle()

PLNPCAfit$plot_PCA()

PLNPCAfit$compute_fisher()

PLNPCAfit$show()

PLNPCAfit$clone()

Inherited methods

PLNmodels::PLNfit$compute_standard_error()

PLNmodels::PLNfit$predict()

PLNmodels::PLNfit$print()

PLNmodels::PLNfit$set_R2()

PLNmodels::PLNfit$VEstep()

Method new()

Usage

PLNPCAfit$new(
  rank,
  responses,
  covariates,
  offsets,
  weights,
  model,
  xlevels,
  control
)

Method update()

Usage

PLNPCAfit$update(
  Theta = NA,
  Sigma = NA,
  B = NA,
  M = NA,
  S = NA,
  Z = NA,
  A = NA,
  Ji = NA,
  R2 = NA,
  monitoring = NA
)

Method setVisualization()

Usage

PLNPCAfit$setVisualization(scale.unit = FALSE)

Method optimize()

Usage

PLNPCAfit$optimize(responses, covariates, offsets, weights, control)

Method postTreatment()

Usage

PLNPCAfit$postTreatment(responses, covariates, offsets, weights, nullModel)

Method latent_pos()

Usage

PLNPCAfit$latent_pos(covariates, offsets)

Method plot_individual_map()

Usage

PLNPCAfit$plot_individual_map(
  axes = 1:min(2, self$rank),
  main = "Individual Factor Map",
  plot = TRUE,
  cols = "default"
)

Method plot_correlation_circle()

Usage

PLNPCAfit$plot_correlation_circle(
  axes = 1:min(2, self$rank),
  main = "Variable Factor Map",
  cols = "default",
  plot = TRUE
)

Method plot_PCA()

Usage

PLNPCAfit$plot_PCA(
  nb_axes = min(3, self$rank),
  ind_cols = "ind_cols",
  var_cols = "var_cols",
  plot = TRUE
)

Method compute_fisher()

Usage

PLNPCAfit$compute_fisher(type = c("wald", "louis"), X = NULL)

Method show()

Usage

PLNPCAfit$show()

Method clone()

The objects of this class are cloneable with this method.

Usage

PLNPCAfit$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPCAs <- PLNPCA(Abundance ~ 1 + offset(log(Offset)), data = trichoptera, ranks = 1:5)

#> 
#>  Initialization...
#> 
#>  Adjusting 5 PLN models for PCA analysis.
#> 	 Rank approximation = 1 	 Rank approximation = 2 	 Rank approximation = 3 	 Rank approximation = 4 	 Rank approximation = 5 
#>  Post-treatments
#>  DONE!

myPCA <- getBestModel(myPCAs)
class(myPCA)

#> [1] "PLNPCAfit" "PLNfit"    "R6"       

print(myPCA)

#> Poisson Lognormal with rank constrained for PCA (rank = 3)
#> ==================================================================
#>  nb_param    loglik       BIC      ICL R_squared
#>        68 -1750.015 -1882.337 -1942.39     0.803
#> ==================================================================
#> * Useful fields
#>     $model_par, $latent, $var_par, $optim_par
#>     $loglik, $BIC, $ICL, $loglik_vec, $nb_param, $criteria
#> * Useful S3 methods
#>     print(), coef(), sigma(), vcov(), fitted(), predict(), standard_error()
#> * Additional fields for PCA
#>     $percent_var, $corr_circle, $scores, $rotation
#> * Additional S3 methods for PCA
#>     plot.PLNPCAfit()