An R6 Class to represent a collection of PLNfit

super class for PLNPCAfamily and PLNnetworkfamily

This class comes with a set of methods, some of them being useful for the user: See the documentation for getBestModel, getModel and plot.

Public fields

responses

the matrix of responses common to every models

covariates

the matrix of covariates common to every models

offsets

the matrix of offsets common to every models

weights

the vector of observation weights

models

a list of PLNnetworkfit object, one per penalty.

inception

a PLNfit object, obtained when no sparsifying penalty is applied.

Active bindings

criteria

a data frame with the values of some criteria (variational lower bound J, BIC, ICL and R2) for the different models.

Methods

Public methods

PLNfamily$new()

PLNfamily$postTreatment()

PLNfamily$getModel()

PLNfamily$plot()

PLNfamily$show()

PLNfamily$print()

PLNfamily$clone()

Method new()

Usage

PLNfamily$new(responses, covariates, offsets, weights, control)

Method postTreatment()

Usage

PLNfamily$postTreatment()

Method getModel()

Usage

PLNfamily$getModel(var, index = NULL)

Method plot()

Usage

PLNfamily$plot(criteria, annotate = TRUE)

Method show()

Usage

PLNfamily$show()

Method print()

Usage

PLNfamily$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

PLNfamily$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.