An R6 Class to represent a PLNfit in a mixture framework

The function PLNmixture produces a collection of models which are instances of object with class PLNmixturefit. A PLNmixturefit (say, with k components) is itself a collection of k PLNfit.

This class comes with a set of methods, some of them being useful for the user: See the documentation for ...

See also

Active bindings

n

number of samples

p

number of dimensions of the latent space

k

number of components

components

components of the mixture (PLNfits)

posteriorProb

matrix ofposterior probability for cluster belonging

memberships

vector for cluster index

mixtureParam

vector of cluster proporitions

optim_par

a list with parameters useful for monitoring the optimization

nb_param

number of parameters in the current PLN model

entropy_clustering

Entropy of the variational distribution of the cluster (multinomial)

entropy_latent

Entropy of the variational distribution of the latent vector (Gaussian)

entropy

Full entropy of the variational distribution (latent vector + clustering)

loglik

variational lower bound of the loglikelihood

loglik_vec

element-wise variational lower bound of the loglikelihood

BIC

variational lower bound of the BIC

ICL

variational lower bound of the ICL (include only entropy of the clustering distribution)

ICL_variant

variational lower bound of the ICL (include entropy of both the clustering and latent distributions)

R_squared

approximated goodness-of-fit criterion

criteria

a vector with loglik, BIC, ICL, ICL_variant and number of parameters

model_par

a list with the matrices of parameters found in the model (Theta, Sigma, plus some others depending on the variant)

var_par

a list with two matrices, M and S2, which are the estimated parameters in the variational approximation

latent

a matrix: values of the latent vector (Z in the model)

fitted

a matrix: fitted values of the observations (A in the model)

Methods

Public methods

PLNmixturefit$new()

PLNmixturefit$optimize()

PLNmixturefit$plot_clustering_data()

PLNmixturefit$plot_clustering_pca()

PLNmixturefit$postTreatment()

PLNmixturefit$show()

PLNmixturefit$print()

PLNmixturefit$clone()

Method new()

Initialize a PLNmixturefit model

Usage

PLNmixturefit$new(
  responses,
  covariates,
  offsets,
  posteriorProb,
  model,
  xlevels,
  control
)

Arguments

responses

the matrix of responses common to every models

covariates

the matrix of covariates common to every models

offsets

the matrix of offsets common to every models

posteriorProb

matrix ofposterior probability for cluster belonging

model

model used for fitting, extracted from the formula in the upper-level call

xlevels

named listed of factor levels included in the models, extracted from the formula in the upper-level call #'

control

a list for controlling the optimization. See details.

control

a list for controlling the optimization. See details.

Method optimize()

Optimize a PLNmixturefit model

Usage

PLNmixturefit$optimize(responses, covariates, offsets, control)

Arguments

responses

the matrix of responses common to every models

covariates

the matrix of covariates common to every models

offsets

the matrix of offsets common to every models

control

a list for controlling the optimization. See details.

control

a list for controlling the optimization. See details.

Method plot_clustering_data()

Plot the matrix of mean counts (without offsets, without covariate effects) reordered according the inferred clustering

Usage

PLNmixturefit$plot_clustering_data(
  main = "Expected counts reorder by clustering",
  plot = TRUE,
  log_scale = TRUE
)

Arguments

main

character. A title for the plot. An hopefully appropriate title will be used by default.

plot

logical. Should the plot be displayed or sent back as ggplot object

log_scale

logical. Should the color scale values be log-transform before plotting? Default is TRUE.

Returns

a ggplot graphic

Method plot_clustering_pca()

Plot the individual map of a PCA performed on the latent coordinate, where individuals are colored according to the memberships

Usage

PLNmixturefit$plot_clustering_pca(
  main = "Clustering labels in Individual Factor Map",
  plot = TRUE
)

Arguments

main

character. A title for the plot. An hopefully appropriate title will be used by default.

plot

logical. Should the plot be displayed or sent back as ggplot object

Returns

a ggplot graphic

Method postTreatment()

Update fields after optimization

Usage

PLNmixturefit$postTreatment(responses, covariates, offsets, weights, nullModel)

Arguments

responses

the matrix of responses common to every models

covariates

the matrix of covariates common to every models

offsets

the matrix of offsets common to every models

weights

an optional vector of observation weights to be used in the fitting process.

nullModel

null model used for approximate R2 computations. Defaults to a GLM model with same design matrix but not latent variable.

Method show()

User friendly print method

Usage

PLNmixturefit$show()

Method print()

User friendly print method

Usage

PLNmixturefit$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

PLNmixturefit$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.