PLNmodels

The Poisson lognormal model and variants can be used for a variety of multivariate problems when count data are at play (including PCA or LDA for count data, network inference). This package implements efficients variational algorithms to fit such models accompanied with a set of functions for vizualisation and diagnostic.

Unpenalized Poisson lognormal model (aka PLN)

See the main function PLN and the associated methods for manipulation.

Also try vignette("PLN_trichoptera", package="PLNmodels") for an overview.

Rank Contraint Poisson lognormal for Poisson Principal Component Analysis (ala PLNPCA)

See the main function PLNPCA and the associated methods for manipulation.

The Poisson PCA and the associated variational inference is fully explained in CHiquet el al (2018), see reference below.

Also try vignette("PLNPCA_trichoptera", package="PLNmodels") for an overview.

Sparse Poisson lognormal model for sparse covariance inference for counts (aka PLNnetwork)

See the main function PLNnetwork and the associated methods for manipulation.

Also try vignette("PLNnetwork_trichoptera", package="PLNmodels") for an overview.

Poisson lognormal discriminant analysis (aka PLNLDA)

See the main function PLNLDA and the associated methods for manipulation.

Also try vignette("PLNLDA_trichoptera", package="PLNmodels") for an overview.