Poisson lognormal model towards sparse network inference

Fit the sparse inverse covariance variant of the Poisson lognormal with a variational algorithm. Use the (g)lm syntax for model specification (covariates, offsets).

PLNnetwork(
  formula,
  data,
  subset,
  weights,
  penalties = NULL,
  control = PLNnetwork_param()
)

Arguments

formula

an object of class "formula": a symbolic description of the model to be fitted.

data

an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which lm is called.

subset

an optional vector specifying a subset of observations to be used in the fitting process.

weights

an optional vector of observation weights to be used in the fitting process.

penalties

an optional vector of positive real number controlling the level of sparsity of the underlying network. if NULL (the default), will be set internally. See PLNnetwork_param() for additional tuning of the penalty.

control

a list-like structure for controlling the optimization, with default generated by PLNnetwork_param(). See the corresponding documentation for details;

Value

an R6 object with class PLNnetworkfamily, which contains a collection of models with class PLNnetworkfit

See also

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
fits <- PLNnetwork(Abundance ~ 1, data = trichoptera)
#> 
#>  Initialization...
#>  Adjusting 30 PLN with sparse inverse covariance estimation
#> 	Joint optimization alternating gradient descent and graphical-lasso
#> 	sparsifying penalty = 7.541317 	sparsifying penalty = 6.965695 	sparsifying penalty = 6.43401 	sparsifying penalty = 5.942907 	sparsifying penalty = 5.48929 	sparsifying penalty = 5.070297 	sparsifying penalty = 4.683286 	sparsifying penalty = 4.325815 	sparsifying penalty = 3.995629 	sparsifying penalty = 3.690646 	sparsifying penalty = 3.408942 	sparsifying penalty = 3.148741 	sparsifying penalty = 2.9084 	sparsifying penalty = 2.686404 	sparsifying penalty = 2.481353 	sparsifying penalty = 2.291954 	sparsifying penalty = 2.117011 	sparsifying penalty = 1.955421 	sparsifying penalty = 1.806166 	sparsifying penalty = 1.668303 	sparsifying penalty = 1.540962 	sparsifying penalty = 1.423342 	sparsifying penalty = 1.3147 	sparsifying penalty = 1.21435 	sparsifying penalty = 1.121659 	sparsifying penalty = 1.036044 	sparsifying penalty = 0.9569638 	sparsifying penalty = 0.8839195 	sparsifying penalty = 0.8164507 	sparsifying penalty = 0.7541317 
#>  Post-treatments
#>  DONE!