Poisson lognormal model towards sparse network inference

Fit the sparse inverse covariance variant of the Poisson lognormal with a variational algorithm. Use the (g)lm syntax for model specification (covariates, offsets).

PLNnetwork(formula, data, subset, weights, penalties = NULL,
  control_init = list(), control_main = list())

Arguments

formula
an object of class "formula": a symbolic description of the model to be fitted.

data
an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which lm is called.

subset
an optional vector specifying a subset of observations to be used in the fitting process.

weights
an optional vector of weights to be used in the fitting process. Should be NULL or a numeric vector.

penalties
an optional vector of positive real number controling the level of sparisty of the underlying network. if NULL (the default), will be set internally

control_init
a list for controling the optimization of the PLN model used at initialization. See details.

control_main
a list for controling the main optimization process. See details.

Value

an R6 object with class PLNnetworkfamily, which contains a collection of models with class PLNnetworkfit

Details

The list of parameters control_main controls the optimization of the main process, with the following entries

"ftol_rel"stop when an optimization step changes the objective function by less than ftol_rel multiplied by the absolute value of the parameter.

"ftol_abs"stop when an optimization step changes the objective function by less than ftol_abs .

"xtol_rel"stop when an optimization step changes every parameters by less than xtol_rel multiplied by the absolute value of the parameter.

"xtol_abs"stop when an optimization step changes every parameters by less than xtol_abs.

"maxeval"stop when the number of iteration exceeds maxeval. Default is 10000

"algorithm"the optimization method used by NLOPT among LD type, i.e. "MMA", "LBFGS", "TNEWTON", "TNEWTON_RESTART", "TNEWTON_PRECOND", "TNEWTON_PRECOND_RESTART", "VAR1", "VAR2". See NLOPTR documentation for further details. Default is "MMA".

"cores"integer for number of cores used. Default is 1.

"trace"integer for verbosity. Useless when cores > 1

"ftol_out"outer solver stops when an optimization step changes the objective function by less than xtol multiply by the absolute value of the parameter. Default is 1e-6

"maxit_out"outer solver stops when the number of iteration exceeds out.maxit. Default is 50

"penalize_diagonal"boolean: should the diagonal terms be penalized in the graphical-Lasso? Default is FALSE.

The list of parameters control_init controls the optimization process in the initialization and in the function PLN, plus two additional parameters:

"nPenalties"an integer that specified the number of values for the penalty grid when internally generated. Ignored when penalties is non NULL

"min.ratio"the penalty grid ranges from the minimal value that produces a sparse to this value multiplied by min.ratio. Default is 0.1.

See also

Examples

data(trichoptera)
fits <- PLNnetwork(Abundance ~ 1, data = trichoptera)

#> 
#>  Initialization...
#>  Adjusting 30 PLN with sparse inverse covariance estimation
#> 	Joint optimization alternating gradient descent and graphical-lasso
#> 	sparsifying penalty = 7.288053 	sparsifying penalty = 6.731762 	sparsifying penalty = 6.217932 	sparsifying penalty = 5.743323 	sparsifying penalty = 5.30494 	sparsifying penalty = 4.900018 	sparsifying penalty = 4.526004 	sparsifying penalty = 4.180538 	sparsifying penalty = 3.861441 	sparsifying penalty = 3.566701 	sparsifying penalty = 3.294458 	sparsifying penalty = 3.042994 	sparsifying penalty = 2.810725 	sparsifying penalty = 2.596185 	sparsifying penalty = 2.39802 	sparsifying penalty = 2.214982 	sparsifying penalty = 2.045914 	sparsifying penalty = 1.889751 	sparsifying penalty = 1.745508 	sparsifying penalty = 1.612275 	sparsifying penalty = 1.489211 	sparsifying penalty = 1.375541 	sparsifying penalty = 1.270547 	sparsifying penalty = 1.173567 	sparsifying penalty = 1.08399 	sparsifying penalty = 1.00125 	sparsifying penalty = 0.9248255 	sparsifying penalty = 0.8542343 	sparsifying penalty = 0.7890314 	sparsifying penalty = 0.7288053 
#>  Post-treatments
#>  DONE!