Poisson lognormal model towards sparse network inference

Fit the sparse inverse covariance variant of the Poisson lognormal with a variational algorithm. Use the (g)lm syntax for model specification (covariates, offsets).

PLNnetwork(
  formula,
  data,
  subset,
  weights,
  penalties = NULL,
  control_init = list(),
  control_main = list()
)

Arguments

formula
an object of class "formula": a symbolic description of the model to be fitted.

data
an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which lm is called.

subset
an optional vector specifying a subset of observations to be used in the fitting process.

weights
an optional vector of weights to be used in the fitting process. Should be NULL or a numeric vector.

penalties
an optional vector of positive real number controling the level of sparisty of the underlying network. if NULL (the default), will be set internally

control_init
a list for controling the optimization of the PLN model used at initialization. See details.

control_main
a list for controling the main optimization process. See details.

Value

an R6 object with class PLNnetworkfamily, which contains a collection of models with class PLNnetworkfit

Details

The list of parameters control_main controls the optimization of the main process, with the following entries

"ftol_rel"stop when an optimization step changes the objective function by less than ftol multiplied by the absolute value of the parameter. Default is 1e-6 when n < p, 1e-8 otherwise.

"ftol_abs"stop when an optimization step changes the objective function by less than ftol multiplied by the absolute value of the parameter. Default is 0

"xtol_rel"stop when an optimization step changes every parameters by less than xtol_rel multiplied by the absolute value of the parameter. Default is 1e-4

"xtol_abs"stop when an optimization step changes every parameters by less than xtol_abs. Default is 0

"maxeval"stop when the number of iteration exceeds maxeval. Default is 10000

"algorithm"the optimization method used by NLOPT among LD type, i.e. "CCSAQ", "MMA", "LBFGS", "VAR1", "VAR2". See NLOPT documentation for further details. Default is "CCSAQ".

"cores"integer for number of cores used. Default is 1.

"trace"integer for verbosity. Useless when cores > 1

"ftol_out"outer solver stops when an optimization step changes the objective function by less than xtol multiply by the absolute value of the parameter. Default is 1e-6

"maxit_out"outer solver stops when the number of iteration exceeds out.maxit. Default is 50

"penalize_diagonal"boolean: should the diagonal terms be penalized in the graphical-Lasso? Default is FALSE.

"penalty_weights"p x p matrix of weights (defaut filled with 1) to adapt the amount of shrinkage to each pairs of node. Must be symmetric with positive values.

The list of parameters control_init controls the optimization process in the initialization and in the function PLN, plus two additional parameters:

"nPenalties"an integer that specified the number of values for the penalty grid when internally generated. Ignored when penalties is non NULL

"min.ratio"the penalty grid ranges from the minimal value that produces a sparse to this value multiplied by min.ratio. Default is 0.1.

See also

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
fits <- PLNnetwork(Abundance ~ 1, data = trichoptera)

#> 
#>  Initialization...
#>  Adjusting 30 PLN with sparse inverse covariance estimation
#> 	Joint optimization alternating gradient descent and graphical-lasso
#> 	sparsifying penalty = 7.357248 	sparsifying penalty = 6.795675 	sparsifying penalty = 6.276967 	sparsifying penalty = 5.797852 	sparsifying penalty = 5.355307 	sparsifying penalty = 4.946541 	sparsifying penalty = 4.568976 	sparsifying penalty = 4.22023 	sparsifying penalty = 3.898103 	sparsifying penalty = 3.600564 	sparsifying penalty = 3.325736 	sparsifying penalty = 3.071886 	sparsifying penalty = 2.837411 	sparsifying penalty = 2.620834 	sparsifying penalty = 2.420788 	sparsifying penalty = 2.236011 	sparsifying penalty = 2.065339 	sparsifying penalty = 1.907693 	sparsifying penalty = 1.76208 	sparsifying penalty = 1.627582 	sparsifying penalty = 1.50335 	sparsifying penalty = 1.388601 	sparsifying penalty = 1.28261 	sparsifying penalty = 1.18471 	sparsifying penalty = 1.094282 	sparsifying penalty = 1.010756 	sparsifying penalty = 0.9336061 	sparsifying penalty = 0.8623447 	sparsifying penalty = 0.7965227 	sparsifying penalty = 0.7357248 
#>  Post-treatments
#>  DONE!