An R6 Class to represent a collection of PLNnetworkfits

The function PLNnetwork() produces an instance of this class.

This class comes with a set of methods mostly used to compare network fits (in terms of goodness of fit) or extract one from the family (based on penalty parameter and/or goodness of it). See the documentation for getBestModel(), getModel() and plot() for the user-facing ones.

See also

Super classes

PLNmodels::PLNfamily -> PLNmodels::Networkfamily -> PLNnetworkfamily

Methods

Public methods

PLNnetworkfamily$new()

PLNnetworkfamily$stability_selection()

PLNnetworkfamily$clone()

Inherited methods

PLNmodels::PLNfamily$getModel()

PLNmodels::PLNfamily$postTreatment()

PLNmodels::PLNfamily$print()

PLNmodels::Networkfamily$coefficient_path()

PLNmodels::Networkfamily$getBestModel()

PLNmodels::Networkfamily$optimize()

PLNmodels::Networkfamily$plot()

PLNmodels::Networkfamily$plot_objective()

PLNmodels::Networkfamily$plot_stars()

PLNmodels::Networkfamily$show()

Method new()

Initialize all models in the collection

Usage

PLNnetworkfamily$new(penalties, data, control)

Arguments

penalties

a vector of positive real number controlling the level of sparsity of the underlying network.

data

a named list used internally to carry the data matrices

control

a list for controlling the optimization.

Returns

Update current PLNnetworkfit with smart starting values

Method stability_selection()

Compute the stability path by stability selection

Usage

PLNnetworkfamily$stability_selection(
  subsamples = NULL,
  control = PLNnetwork_param()
)

Arguments

subsamples

a list of vectors describing the subsamples. The number of vectors (or list length) determines the number of subsamples used in the stability selection. Automatically set to 20 subsamples with size 10*sqrt(n) if n >= 144 and 0.8*n otherwise following Liu et al. (2010) recommendations.

control

a list controlling the main optimization process in each call to PLNnetwork(). See PLNnetwork() and PLN_param() for details.

Method clone()

The objects of this class are cloneable with this method.

Usage

PLNnetworkfamily$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
fits <- PLNnetwork(Abundance ~ 1, data = trichoptera)
#> 
#>  Initialization...
#>  Adjusting 30 PLN with sparse inverse covariance estimation
#> 	Joint optimization alternating gradient descent and graphical-lasso
#> 	sparsifying penalty = 7.264447 	sparsifying penalty = 6.709958 	sparsifying penalty = 6.197792 	sparsifying penalty = 5.72472 	sparsifying penalty = 5.287757 	sparsifying penalty = 4.884147 	sparsifying penalty = 4.511344 	sparsifying penalty = 4.166997 	sparsifying penalty = 3.848934 	sparsifying penalty = 3.555148 	sparsifying penalty = 3.283787 	sparsifying penalty = 3.033138 	sparsifying penalty = 2.801621 	sparsifying penalty = 2.587776 	sparsifying penalty = 2.390253 	sparsifying penalty = 2.207807 	sparsifying penalty = 2.039287 	sparsifying penalty = 1.88363 	sparsifying penalty = 1.739854 	sparsifying penalty = 1.607053 	sparsifying penalty = 1.484388 	sparsifying penalty = 1.371086 	sparsifying penalty = 1.266432 	sparsifying penalty = 1.169766 	sparsifying penalty = 1.080479 	sparsifying penalty = 0.998007 	sparsifying penalty = 0.9218299 	sparsifying penalty = 0.8514675 	sparsifying penalty = 0.7864757 	sparsifying penalty = 0.7264447 
#>  Post-treatments
#>  DONE!
class(fits)
#> [1] "PLNnetworkfamily" "Networkfamily"    "PLNfamily"        "R6"