Extract the regularization path of a PLNnetwork fit

Extract the regularization path of a PLNnetwork fit

coefficient_path(Robject, precision = TRUE, corr = TRUE)

Arguments

Robject
an object with class PLNnetworkfamily, i.e. an output from PLNnetwork

precision
a logical, should the coefficients of the precision matrix Omega or the covariance matrice Sigma be sent back. Default is TRUE.

corr
a logical, should the correlation (partial in case precision = TRUE) be sent back. Default is TRUE.

Value

Send back a tibble/data.frame.

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
fits <- PLNnetwork(Abundance ~ 1, data = trichoptera)

#> 
#>  Initialization...
#>  Adjusting 30 PLN with sparse inverse covariance estimation
#> 	Joint optimization alternating gradient descent and graphical-lasso
#> 	sparsifying penalty = 7.331617 	sparsifying penalty = 6.772001 	sparsifying penalty = 6.2551 	sparsifying penalty = 5.777653 	sparsifying penalty = 5.33665 	sparsifying penalty = 4.929308 	sparsifying penalty = 4.553058 	sparsifying penalty = 4.205527 	sparsifying penalty = 3.884523 	sparsifying penalty = 3.588021 	sparsifying penalty = 3.31415 	sparsifying penalty = 3.061184 	sparsifying penalty = 2.827526 	sparsifying penalty = 2.611704 	sparsifying penalty = 2.412354 	sparsifying penalty = 2.228222 	sparsifying penalty = 2.058143 	sparsifying penalty = 1.901047 	sparsifying penalty = 1.755942 	sparsifying penalty = 1.621912 	sparsifying penalty = 1.498113 	sparsifying penalty = 1.383763 	sparsifying penalty = 1.278142 	sparsifying penalty = 1.180582 	sparsifying penalty = 1.09047 	sparsifying penalty = 1.007235 	sparsifying penalty = 0.9303536 	sparsifying penalty = 0.8593405 	sparsifying penalty = 0.7937478 	sparsifying penalty = 0.7331617 
#>  Post-treatments
#>  DONE!

head(coefficient_path(fits))

#>   Node1 Node2 Coeff  Penalty    Edge
#> 1   Aga   Che     0 7.331617 Aga|Che
#> 2   Ath   Che     0 7.331617 Ath|Che
#> 3   Cea   Che     0 7.331617 Cea|Che
#> 4   Ced   Che     0 7.331617 Ced|Che
#> 5   All   Che     0 7.331617 All|Che
#> 6   Che   Hyc     0 7.331617 Che|Hyc