Extract the regularization path of a PLNnetwork fit

Extract the regularization path of a PLNnetwork fit

coefficient_path(Robject, precision = TRUE, corr = TRUE)

Arguments

Robject
an object with class PLNnetworkfamily, i.e. an output from PLNnetwork()

precision
a logical, should the coefficients of the precision matrix Omega or the covariance matrix Sigma be sent back. Default is TRUE.

corr
a logical, should the correlation (partial in case precision = TRUE) be sent back. Default is TRUE.

Value

Sends back a tibble/data.frame.

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
fits <- PLNnetwork(Abundance ~ 1, data = trichoptera)

#> 
#>  Initialization...
#>  Adjusting 30 PLN with sparse inverse covariance estimation
#> 	Joint optimization alternating gradient descent and graphical-lasso
#> 	sparsifying penalty = 7.499547 	sparsifying penalty = 6.927113 	sparsifying penalty = 6.398373 	sparsifying penalty = 5.90999 	sparsifying penalty = 5.458886 	sparsifying penalty = 5.042214 	sparsifying penalty = 4.657346 	sparsifying penalty = 4.301855 	sparsifying penalty = 3.973498 	sparsifying penalty = 3.670204 	sparsifying penalty = 3.390061 	sparsifying penalty = 3.1313 	sparsifying penalty = 2.892291 	sparsifying penalty = 2.671525 	sparsifying penalty = 2.46761 	sparsifying penalty = 2.279259 	sparsifying penalty = 2.105285 	sparsifying penalty = 1.944591 	sparsifying penalty = 1.796162 	sparsifying penalty = 1.659062 	sparsifying penalty = 1.532427 	sparsifying penalty = 1.415458 	sparsifying penalty = 1.307418 	sparsifying penalty = 1.207624 	sparsifying penalty = 1.115447 	sparsifying penalty = 1.030306 	sparsifying penalty = 0.9516633 	sparsifying penalty = 0.8790237 	sparsifying penalty = 0.8119286 	sparsifying penalty = 0.7499547 
#>  Post-treatments
#>  DONE!

head(coefficient_path(fits))

#>   Node1 Node2 Coeff  Penalty    Edge
#> 1   Aga   Che     0 7.499547 Aga|Che
#> 2   Ath   Che     0 7.499547 Ath|Che
#> 3   Cea   Che     0 7.499547 Cea|Che
#> 4   Ced   Che     0 7.499547 Ced|Che
#> 5   All   Che     0 7.499547 All|Che
#> 6   Che   Hyc     0 7.499547 Che|Hyc