Extract the regularization path of a PLNnetwork fit

Extract the regularization path of a PLNnetwork fit

coefficient_path(Robject, precision = TRUE, corr = TRUE)

Arguments

Robject
an object with class PLNnetworkfamily, i.e. an output from PLNnetwork

precision
a logical, should the coefficients of the precision matrix Omega or the covariance matrice Sigma be sent back. Default is TRUE.

corr
a logical, should the correlation (partial in case precision = TRUE) be sent back. Default is TRUE.

Value

Send back a tibble/data.frame.

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
fits <- PLNnetwork(Abundance ~ 1, data = trichoptera)

#> 
#>  Initialization...
#>  Adjusting 30 PLN with sparse inverse covariance estimation
#> 	Joint optimization alternating gradient descent and graphical-lasso
#> 	sparsifying penalty = 7.59819 	sparsifying penalty = 7.018226 	sparsifying penalty = 6.482531 	sparsifying penalty = 5.987725 	sparsifying penalty = 5.530687 	sparsifying penalty = 5.108535 	sparsifying penalty = 4.718604 	sparsifying penalty = 4.358437 	sparsifying penalty = 4.025762 	sparsifying penalty = 3.718479 	sparsifying penalty = 3.43465 	sparsifying penalty = 3.172486 	sparsifying penalty = 2.930333 	sparsifying penalty = 2.706663 	sparsifying penalty = 2.500066 	sparsifying penalty = 2.309238 	sparsifying penalty = 2.132976 	sparsifying penalty = 1.970168 	sparsifying penalty = 1.819787 	sparsifying penalty = 1.680884 	sparsifying penalty = 1.552583 	sparsifying penalty = 1.434076 	sparsifying penalty = 1.324614 	sparsifying penalty = 1.223508 	sparsifying penalty = 1.130118 	sparsifying penalty = 1.043857 	sparsifying penalty = 0.9641806 	sparsifying penalty = 0.8905855 	sparsifying penalty = 0.8226079 	sparsifying penalty = 0.759819 
#>  Post-treatments
#>  DONE!

head(coefficient_path(fits))

#>   Node1 Node2 Coeff Penalty    Edge
#> 1   Aga   Che     0 7.59819 Aga|Che
#> 2   Ath   Che     0 7.59819 Ath|Che
#> 3   Cea   Che     0 7.59819 Cea|Che
#> 4   Ced   Che     0 7.59819 Ced|Che
#> 5   All   Che     0 7.59819 All|Che
#> 6   Che   Hyc     0 7.59819 Che|Hyc