Extract the regularization path of a PLNnetwork fit

Extract the regularization path of a PLNnetwork fit

coefficient_path(Robject, precision = TRUE, corr = TRUE)

Arguments

Robject

an object with class PLNnetworkfamily, i.e. an output from PLNnetwork()

precision

a logical, should the coefficients of the precision matrix Omega or the covariance matrix Sigma be sent back. Default is TRUE.

corr

a logical, should the correlation (partial in case precision = TRUE) be sent back. Default is TRUE.

Value

Sends back a tibble/data.frame.

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
fits <- PLNnetwork(Abundance ~ 1, data = trichoptera)
#> 
#>  Initialization...
#>  Adjusting 30 PLN with sparse inverse covariance estimation
#> 	Joint optimization alternating gradient descent and graphical-lasso
#> 	sparsifying penalty = 7.598712 	sparsifying penalty = 7.018709 	sparsifying penalty = 6.482977 	sparsifying penalty = 5.988137 	sparsifying penalty = 5.531067 	sparsifying penalty = 5.108886 	sparsifying penalty = 4.718929 	sparsifying penalty = 4.358737 	sparsifying penalty = 4.026038 	sparsifying penalty = 3.718734 	sparsifying penalty = 3.434887 	sparsifying penalty = 3.172705 	sparsifying penalty = 2.930535 	sparsifying penalty = 2.70685 	sparsifying penalty = 2.500238 	sparsifying penalty = 2.309397 	sparsifying penalty = 2.133123 	sparsifying penalty = 1.970303 	sparsifying penalty = 1.819912 	sparsifying penalty = 1.680999 	sparsifying penalty = 1.55269 	sparsifying penalty = 1.434175 	sparsifying penalty = 1.324705 	sparsifying penalty = 1.223592 	sparsifying penalty = 1.130196 	sparsifying penalty = 1.043929 	sparsifying penalty = 0.9642469 	sparsifying penalty = 0.8906468 	sparsifying penalty = 0.8226645 	sparsifying penalty = 0.7598712 
#>  Post-treatments
#>  DONE!
head(coefficient_path(fits))
#>   Node1 Node2 Coeff  Penalty    Edge
#> 1   Aga   Che     0 7.598712 Aga|Che
#> 2   Ath   Che     0 7.598712 Ath|Che
#> 3   Cea   Che     0 7.598712 Cea|Che
#> 4   Ced   Che     0 7.598712 Ced|Che
#> 5   All   Che     0 7.598712 All|Che
#> 6   Che   Hyc     0 7.598712 Che|Hyc