Compute offsets from a count data using one of several normalization schemes

Computes offsets from the count table using one of several normalization schemes (TSS, CSS, RLE, GMPR, etc) described in the literature.

compute_offset(counts, offset = c("TSS", "GMPR", "RLE", "CSS", "none"),
  ...)

Arguments

counts
Required. An abundance count table, preferably with dimensions names and species as columns.

offset
Optional. Normalisation scheme used to compute scaling factors used as offset during PLN inference. Can be "TSS" (Total Sum Scaling, default), "CSS" (Cumulative Sum Scaling, used in metagenomeSeq), "RLE" (Relative Log Expression, used in DESeq2), "GMPR" (Geometric Mean of Pairwise Ratio, introduced in Chen et al., 2018) or "none".

...
Additional parameters passed on to specific methods (for now CSS and RLE)

Value

If offset = "none", NULL else a vector of length nrow(counts) with one offset per sample.

Details

RLE has an additional pseudocounts arguments to add pseudocounts to the observed counts (defaults to 0). CSS has an additional reference argument to choose the location function used to compute the reference quantiles (defaults to median as in the Nature publication but can be set to mean to reproduce behavior of functions cumNormStat* from metagenomeSeq). Note that (i) CSS normalization fails when the median absolute deviation around quantiles does not become instable for high quantiles (limited count variations both within and across samples) and/or one sample has less than two positive counts, (ii) RLE fails when there are no common species across all samples and (iii) GMPR fails if a sample does not share any species with all other samples.

References

Chen, L., Reeve, J., Zhang, L., Huang, S., Wang, X. and Chen, J. (2018) GMPR: A robust normalization method for zero-inflated count data with application to microbiome sequencing data. PeerJ, 6, e4600 https://doi.org/10.7717/peerj.4600

Paulson, J. N., Colin Stine, O., Bravo, H. C. and Pop, M. (2013) Differential abundance analysis for microbial marker-gene surveys. Nature Methods, 10, 1200-1202 http://dx.doi.org/10.1038/nmeth.2658

Anders, S. and Huber, W. (2010) Differential expression analysis for sequence count data. Genome Biology, 11, R106 https://doi.org/10.1186/gb-2010-11-10-r106

Examples

data(trichoptera)
counts <- trichoptera$Abundance
compute_offset(counts)

#>    1    2    3    4    5    6    7    8    9   10   11   12   13   14   15   16 
#>   29   13   38  192   79   18    8   34   12    4    4    3   49   33  600  172 
#>   17   18   19   20   21   22   23   24   25   26   27   28   29   30   31   32 
#>   58   51   56  127   35   13   17    3   27   40   44    8    9 1599 2980   88 
#>   33   34   35   36   37   38   39   40   41   42   43   44   45   46   47   48 
#>  135  327   66   90   63   15   14   20   70   53   95   43   62  149   16   31 
#>   49 
#>   86