Extract edge selection frequency in bootstrap subsamples

Extracts edge selection frequency in networks reconstructed from bootstrap subsamples during the stars stability selection procedure, as either a matrix or a named vector. In the latter case, edge names follow igraph naming convention.

extract_probs(Robject, penalty = NULL, index = NULL,
  crit = c("StARS", "BIC", "EBIC"), format = c("matrix", "vector"),
  tol = 1e-05)

Arguments

Robject
an object with class PLNnetworkfamily, i.e. an output from PLNnetwork

penalty
penalty used for the boostrap subsamples

index
Integer index of the model to be returned. Only the first value is taken into account.

crit
a character for the criterion used to performed the selection. Either "BIC", "ICL", "EBIC", "StARS", "R_squared". Default is ICL for PLNPCA, and BIC for PLNnetwork. If StARS (Stability Approach to Regularization Selection) is chosen and stability selection was not yet performed, the function will call the method stability_selection with default argument.

format
output format. Either a matrix (default) or a named vector.

tol
tolerance for rounding error when comparing penalties.

Value

Either a matrix or named vector of edge-wise probabilities. In the latter case, edge names follow igraph convention.

Examples

data(mollusk)
mollusc <- prepare_data(mollusk$Abundance, mollusk$Covariate)

#> Warning: Sample(s) 134 and 137 and 145 and 146 dropped for lack of positive counts.

nets <- PLNnetwork(Abundance ~ 0 + site + offset(log(Offset)), data = mollusc)

#> 
#>  Initialization...
#>  Adjusting 30 PLN with sparse inverse covariance estimation
#> 	Joint optimization alternating gradient descent and graphical-lasso
#> 	sparsifying penalty = 5.756715 	sparsifying penalty = 5.31731 	sparsifying penalty = 4.911444 	sparsifying penalty = 4.536558 	sparsifying penalty = 4.190286 	sparsifying penalty = 3.870445 	sparsifying penalty = 3.575018 	sparsifying penalty = 3.30214 	sparsifying penalty = 3.05009 	sparsifying penalty = 2.817279 	sparsifying penalty = 2.602239 	sparsifying penalty = 2.403612 	sparsifying penalty = 2.220146 	sparsifying penalty = 2.050685 	sparsifying penalty = 1.894158 	sparsifying penalty = 1.749578 	sparsifying penalty = 1.616034 	sparsifying penalty = 1.492684 	sparsifying penalty = 1.378749 	sparsifying penalty = 1.27351 	sparsifying penalty = 1.176304 	sparsifying penalty = 1.086518 	sparsifying penalty = 1.003585 	sparsifying penalty = 0.926982 	sparsifying penalty = 0.8562262 	sparsifying penalty = 0.7908712 	sparsifying penalty = 0.7305047 	sparsifying penalty = 0.6747459 	sparsifying penalty = 0.6232431 	sparsifying penalty = 0.5756715 
#>  Post-treatments
#>  DONE!

stability_selection(nets)

#> 
#> Stability Selection for PLNnetwork: 
#> subsampling: ++++++++++++++++++++

probs <- extract_probs(nets, crit = "StARS", format = "vector")
probs

#> Bit|Vap Bit|Vac Vap|Vac Bit|Gac Vap|Gac Vac|Gac Bit|Raa Vap|Raa Vac|Raa Gac|Raa 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Bit|Rap Vap|Rap Vac|Rap Gac|Rap Raa|Rap Bit|Lit Vap|Lit Vac|Lit Gac|Lit Raa|Lit 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Rap|Lit Bit|Pha Vap|Pha Vac|Pha Gac|Pha Raa|Pha Rap|Pha Lit|Pha Bit|Phf Vap|Phf 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Vac|Phf Gac|Phf Raa|Phf Rap|Phf Lit|Phf Pha|Phf Bit|Plc Vap|Plc Vac|Plc Gac|Plc 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Raa|Plc Rap|Plc Lit|Plc Pha|Plc Phf|Plc Bit|Bac Vap|Bac Vac|Bac Gac|Bac Raa|Bac 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Rap|Bac Lit|Bac Pha|Bac Phf|Bac Plc|Bac Bit|Ans Vap|Ans Vac|Ans Gac|Ans Raa|Ans 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Rap|Ans Lit|Ans Pha|Ans Phf|Ans Plc|Ans Bac|Ans Bit|Anl Vap|Anl Vac|Anl Gac|Anl 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Raa|Anl Rap|Anl Lit|Anl Pha|Anl Phf|Anl Plc|Anl Bac|Anl Ans|Anl Bit|An- Vap|An- 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.60    0.00    0.00 
#> Vac|An- Gac|An- Raa|An- Rap|An- Lit|An- Pha|An- Phf|An- Plc|An- Bac|An- Ans|An- 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.95 
#> Anl|An- Bit|Gya Vap|Gya Vac|Gya Gac|Gya Raa|Gya Rap|Gya Lit|Gya Pha|Gya Phf|Gya 
#>    0.35    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Plc|Gya Bac|Gya Ans|Gya Anl|Gya An-|Gya Bit|Hic Vap|Hic Vac|Hic Gac|Hic Raa|Hic 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Rap|Hic Lit|Hic Pha|Hic Phf|Hic Plc|Hic Bac|Hic Ans|Hic Anl|Hic An-|Hic Gya|Hic 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Bit|Arc Vap|Arc Vac|Arc Gac|Arc Raa|Arc Rap|Arc Lit|Arc Pha|Arc Phf|Arc Plc|Arc 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Bac|Arc Ans|Arc Anl|Arc An-|Arc Gya|Arc Hic|Arc Bit|Acl Vap|Acl Vac|Acl Gac|Acl 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Raa|Acl Rap|Acl Lit|Acl Pha|Acl Phf|Acl Plc|Acl Bac|Acl Ans|Acl Anl|Acl An-|Acl 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Gya|Acl Hic|Acl Arc|Acl Bit|Few Vap|Few Vac|Few Gac|Few Raa|Few Rap|Few Lit|Few 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Pha|Few Phf|Few Plc|Few Bac|Few Ans|Few Anl|Few An-|Few Gya|Few Hic|Few Arc|Few 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Acl|Few Bit|Spc Vap|Spc Vac|Spc Gac|Spc Raa|Spc Rap|Spc Lit|Spc Pha|Spc Phf|Spc 
#>    0.00    0.00    0.10    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Plc|Spc Bac|Spc Ans|Spc Anl|Spc An-|Spc Gya|Spc Hic|Spc Arc|Spc Acl|Spc Few|Spc 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Bit|Pio Vap|Pio Vac|Pio Gac|Pio Raa|Pio Rap|Pio Lit|Pio Pha|Pio Phf|Pio Plc|Pio 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Bac|Pio Ans|Pio Anl|Pio An-|Pio Gya|Pio Hic|Pio Arc|Pio Acl|Pio Few|Pio Spc|Pio 
#>    0.00    0.05    0.10    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Bit|Pic Vap|Pic Vac|Pic Gac|Pic Raa|Pic Rap|Pic Lit|Pic Pha|Pic Phf|Pic Plc|Pic 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Bac|Pic Ans|Pic Anl|Pic An-|Pic Gya|Pic Hic|Pic Arc|Pic Acl|Pic Few|Pic Spc|Pic 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Pio|Pic Bit|Pip Vap|Pip Vac|Pip Gac|Pip Raa|Pip Rap|Pip Lit|Pip Pha|Pip Phf|Pip 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Plc|Pip Bac|Pip Ans|Pip Anl|Pip An-|Pip Gya|Pip Hic|Pip Arc|Pip Acl|Pip Few|Pip 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Spc|Pip Pio|Pip Pic|Pip Bit|Pin Vap|Pin Vac|Pin Gac|Pin Raa|Pin Rap|Pin Lit|Pin 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Pha|Pin Phf|Pin Plc|Pin Bac|Pin Ans|Pin Anl|Pin An-|Pin Gya|Pin Hic|Pin Arc|Pin 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Acl|Pin Few|Pin Spc|Pin Pio|Pin Pic|Pin Pip|Pin Bit|Pis Vap|Pis Vac|Pis Gac|Pis 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Raa|Pis Rap|Pis Lit|Pis Pha|Pis Phf|Pis Plc|Pis Bac|Pis Ans|Pis Anl|Pis An-|Pis 
#>    0.00    0.00    0.85    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Gya|Pis Hic|Pis Arc|Pis Acl|Pis Few|Pis Spc|Pis Pio|Pis Pic|Pis Pip|Pis Pin|Pis 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.25    0.00    0.05 
#> Bit|Pim Vap|Pim Vac|Pim Gac|Pim Raa|Pim Rap|Pim Lit|Pim Pha|Pim Phf|Pim Plc|Pim 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Bac|Pim Ans|Pim Anl|Pim An-|Pim Gya|Pim Hic|Pim Arc|Pim Acl|Pim Few|Pim Spc|Pim 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Pio|Pim Pic|Pim Pip|Pim Pin|Pim Pis|Pim Bit|Ppe Vap|Ppe Vac|Ppe Gac|Ppe Raa|Ppe 
#>    0.00    0.00    0.00    0.10    0.10    0.00    0.00    0.00    0.00    0.00 
#> Rap|Ppe Lit|Ppe Pha|Ppe Phf|Ppe Plc|Ppe Bac|Ppe Ans|Ppe Anl|Ppe An-|Ppe Gya|Ppe 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Hic|Ppe Arc|Ppe Acl|Ppe Few|Ppe Spc|Ppe Pio|Ppe Pic|Ppe Pip|Ppe Pin|Ppe Pis|Ppe 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Pim|Ppe Bit|Pi- Vap|Pi- Vac|Pi- Gac|Pi- Raa|Pi- Rap|Pi- Lit|Pi- Pha|Pi- Phf|Pi- 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Plc|Pi- Bac|Pi- Ans|Pi- Anl|Pi- An-|Pi- Gya|Pi- Hic|Pi- Arc|Pi- Acl|Pi- Few|Pi- 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Spc|Pi- Pio|Pi- Pic|Pi- Pip|Pi- Pin|Pi- Pis|Pi- Pim|Pi- Ppe|Pi- Bit|Cam Vap|Cam 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Vac|Cam Gac|Cam Raa|Cam Rap|Cam Lit|Cam Pha|Cam Phf|Cam Plc|Cam Bac|Cam Ans|Cam 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Anl|Cam An-|Cam Gya|Cam Hic|Cam Arc|Cam Acl|Cam Few|Cam Spc|Cam Pio|Cam Pic|Cam 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Pip|Cam Pin|Cam Pis|Cam Pim|Cam Ppe|Cam Pi-|Cam Bit|Cat Vap|Cat Vac|Cat Gac|Cat 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Raa|Cat Rap|Cat Lit|Cat Pha|Cat Phf|Cat Plc|Cat Bac|Cat Ans|Cat Anl|Cat An-|Cat 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Gya|Cat Hic|Cat Arc|Cat Acl|Cat Few|Cat Spc|Cat Pio|Cat Pic|Cat Pip|Cat Pin|Cat 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.80 
#> Pis|Cat Pim|Cat Ppe|Cat Pi-|Cat Cam|Cat Bit|Euf Vap|Euf Vac|Euf Gac|Euf Raa|Euf 
#>    0.00    0.00    0.00    0.00    0.15    0.00    0.00    0.00    0.00    0.00 
#> Rap|Euf Lit|Euf Pha|Euf Phf|Euf Plc|Euf Bac|Euf Ans|Euf Anl|Euf An-|Euf Gya|Euf 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Hic|Euf Arc|Euf Acl|Euf Few|Euf Spc|Euf Pio|Euf Pic|Euf Pip|Euf Pin|Euf Pis|Euf 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Pim|Euf Ppe|Euf Pi-|Euf Cam|Euf Cat|Euf Bit|Vea Vap|Vea Vac|Vea Gac|Vea Raa|Vea 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Rap|Vea Lit|Vea Pha|Vea Phf|Vea Plc|Vea Bac|Vea Ans|Vea Anl|Vea An-|Vea Gya|Vea 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Hic|Vea Arc|Vea Acl|Vea Few|Vea Spc|Vea Pio|Vea Pic|Vea Pip|Vea Pin|Vea Pis|Vea 
#>    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00    0.00 
#> Pim|Vea Ppe|Vea Pi-|Vea Cam|Vea Cat|Vea Euf|Vea 
#>    0.00    0.00    0.00    0.00    0.00    0.00 

# NOT RUN {
## Add edge attributes to graph using igraph
net_stars <- getBestModel(nets, "StARS")
g <- plot(net_BIC, type = "partial_cor", plot=F)
library(igraph)
E(g)$prob <- probs[as_ids(E(g))]
g
# }