Extract edge selection frequency in bootstrap subsamples

Extracts edge selection frequency in networks reconstructed from bootstrap subsamples during the stars stability selection procedure, as either a matrix or a named vector. In the latter case, edge names follow igraph naming convention.

extract_probs(
  Robject,
  penalty = NULL,
  index = NULL,
  crit = c("StARS", "BIC", "EBIC"),
  format = c("matrix", "vector"),
  tol = 1e-05
)

Arguments

Robject

an object with class PLNnetworkfamily, i.e. an output from PLNnetwork()

penalty

penalty used for the bootstrap subsamples

index

Integer index of the model to be returned. Only the first value is taken into account.

crit

a character for the criterion used to performed the selection. Either "BIC", "ICL", "EBIC", "StARS", "R_squared". Default is ICL for PLNPCA, and BIC for PLNnetwork. If StARS (Stability Approach to Regularization Selection) is chosen and stability selection was not yet performed, the function will call the method stability_selection() with default argument.

format

output format. Either a matrix (default) or a named vector.

tol

tolerance for rounding error when comparing penalties.

Value

Either a matrix or named vector of edge-wise probabilities. In the latter case, edge names follow igraph convention.

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
nets <- PLNnetwork(Abundance ~ 1 + offset(log(Offset)), data = trichoptera)
#> 
#>  Initialization...
#>  Adjusting 30 PLN with sparse inverse covariance estimation
#> 	Joint optimization alternating gradient descent and graphical-lasso
#> 	sparsifying penalty = 3.478205 	sparsifying penalty = 3.212716 	sparsifying penalty = 2.967492 	sparsifying penalty = 2.740986 	sparsifying penalty = 2.531769 	sparsifying penalty = 2.338521 	sparsifying penalty = 2.160024 	sparsifying penalty = 1.995151 	sparsifying penalty = 1.842863 	sparsifying penalty = 1.702199 	sparsifying penalty = 1.572272 	sparsifying penalty = 1.452261 	sparsifying penalty = 1.341412 	sparsifying penalty = 1.239023 	sparsifying penalty = 1.144449 	sparsifying penalty = 1.057094 	sparsifying penalty = 0.9764073 	sparsifying penalty = 0.9018789 	sparsifying penalty = 0.8330393 	sparsifying penalty = 0.7694541 	sparsifying penalty = 0.7107223 	sparsifying penalty = 0.6564735 	sparsifying penalty = 0.6063654 	sparsifying penalty = 0.5600821 	sparsifying penalty = 0.5173315 	sparsifying penalty = 0.477844 	sparsifying penalty = 0.4413706 	sparsifying penalty = 0.4076812 	sparsifying penalty = 0.3765632 	sparsifying penalty = 0.3478205 
#>  Post-treatments
#>  DONE!
if (FALSE) {
stability_selection(nets)
probs <- extract_probs(nets, crit = "StARS", format = "vector")
probs
}

if (FALSE) {
## Add edge attributes to graph using igraph
net_stars <- getBestModel(nets, "StARS")
g <- plot(net_stars, type = "partial_cor", plot=F)
library(igraph)
E(g)$prob <- probs[as_ids(E(g))]
g
}