Extract edge selection frequency in bootstrap subsamples

Extracts edge selection frequency in networks reconstructed from bootstrap subsamples during the stars stability selection procedure, as either a matrix or a named vector. In the latter case, edge names follow igraph naming convention.

extract_probs(
  Robject,
  penalty = NULL,
  index = NULL,
  crit = c("StARS", "BIC", "EBIC"),
  format = c("matrix", "vector"),
  tol = 1e-05
)

Arguments

Robject
an object with class PLNnetworkfamily, i.e. an output from PLNnetwork()

penalty
penalty used for the bootstrap subsamples

index
Integer index of the model to be returned. Only the first value is taken into account.

crit
a character for the criterion used to performed the selection. Either "BIC", "ICL", "EBIC", "StARS", "R_squared". Default is ICL for PLNPCA, and BIC for PLNnetwork. If StARS (Stability Approach to Regularization Selection) is chosen and stability selection was not yet performed, the function will call the method stability_selection() with default argument.

format
output format. Either a matrix (default) or a named vector.

tol
tolerance for rounding error when comparing penalties.

Value

Either a matrix or named vector of edge-wise probabilities. In the latter case, edge names follow igraph convention.

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
nets <- PLNnetwork(Abundance ~ 1 + offset(log(Offset)), data = trichoptera)

#> 
#>  Initialization...
#>  Adjusting 30 PLN with sparse inverse covariance estimation
#> 	Joint optimization alternating gradient descent and graphical-lasso
#> 	sparsifying penalty = 3.632478 	sparsifying penalty = 3.355214 	sparsifying penalty = 3.099114 	sparsifying penalty = 2.862561 	sparsifying penalty = 2.644064 	sparsifying penalty = 2.442245 	sparsifying penalty = 2.25583 	sparsifying penalty = 2.083645 	sparsifying penalty = 1.924602 	sparsifying penalty = 1.777699 	sparsifying penalty = 1.642009 	sparsifying penalty = 1.516675 	sparsifying penalty = 1.400909 	sparsifying penalty = 1.293979 	sparsifying penalty = 1.19521 	sparsifying penalty = 1.103981 	sparsifying penalty = 1.019715 	sparsifying penalty = 0.9418812 	sparsifying penalty = 0.8699882 	sparsifying penalty = 0.8035827 	sparsifying penalty = 0.742246 	sparsifying penalty = 0.685591 	sparsifying penalty = 0.6332604 	sparsifying penalty = 0.5849242 	sparsifying penalty = 0.5402774 	sparsifying penalty = 0.4990385 	sparsifying penalty = 0.4609473 	sparsifying penalty = 0.4257636 	sparsifying penalty = 0.3932654 	sparsifying penalty = 0.3632478 
#>  Post-treatments
#>  DONE!

if (FALSE) {
stability_selection(nets)
probs <- extract_probs(nets, crit = "StARS", format = "vector")
probs
}

if (FALSE) {
## Add edge attributes to graph using igraph
net_stars <- getBestModel(nets, "StARS")
g <- plot(net_stars, type = "partial_cor", plot=F)
library(igraph)
E(g)$prob <- probs[as_ids(E(g))]
g
}