Best model extraction from a collection of models

Best model extraction from a collection of models

# S3 method for PLNPCAfamily
getBestModel(Robject, crit = c("ICL", "BIC"), ...)

getBestModel(Robject, crit, ...)

# S3 method for PLNmixturefamily
getBestModel(Robject, crit = c("ICL", "BIC", "ICL_variant"), ...)

# S3 method for PLNnetworkfamily
getBestModel(Robject, crit = c("BIC", "EBIC", "StARS"), ...)

Arguments

Robject
an object with class PLNPCAfamilly ot PLNnetworkfamily

crit
a character for the criterion used to performed the selection. Either "BIC", "ICL", "EBIC", "StARS", "R_squared". Default is ICL for PLNPCA, and BIC for PLNnetwork. If StARS (Stability Approach to Regularization Selection) is chosen and stability selection was not yet performed, the function will call the method stability_selection() with default argument.

...
additional parameters for StARS criterion (only for PLNnetwork). stability, a scalar indicating the target stability (= 1 - 2 beta) at which the network is selected. Default is 0.9.

Value

Send back an object with class PLNPCAfit or PLNnetworkfit

Methods (by class)

PLNPCAfamily: Model extraction for PLNPCAfamily

PLNmixturefamily: Model extraction for PLNmixturefamily

PLNnetworkfamily: Model extraction for PLNnetworkfamily

Examples

if (FALSE) {
data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPCA <- PLNPCA(Abundance ~ 1 + offset(log(Offset)), data = trichoptera, ranks = 1:4)
myModel <- getBestModel(myPCA)
}