Extract and plot the network (aprtial correlation, support or inverse covariance) from a PLNnetworkfit object

Extract and plot the network (aprtial correlation, support or inverse covariance) from a PLNnetworkfit object

# S3 method for PLNnetworkfit
plot(x, type = c("partial_cor", "support"),
  output = c("igraph", "corrplot"), edge.color = c("#F8766D",
  "#00BFC4"), remove.isolated = FALSE, node.labels = NULL,
  layout = layout_in_circle, plot = TRUE, ...)

Arguments

x
an R6 object with class PLNnetworkfit

type
character. Value of the weigth of the edges in the network, either "partial_cor" (partial correlation) or "support" (binary). Default is "partial_cor".

output
the type of output used: either 'igraph' or 'corrplot'. Default is 'igraph'.

edge.color
2D numeric. Color for positive/negative edges. Default is c("#F8766D", "#00BFC4"). Only relevant for igraph output.

remove.isolated
if TRUE, isolated node are remove before plotting. Only relevant for igraph output.

node.labels
vector of character. The labels of the nodes. The Default will use the column names ot the response matrix.

layout
an optional igraph layout. Only relevant for igraph output.

plot
logical. Should the final network be displayed or only sent back to the user. Default is TRUE.

...
Not used (S3 compatibility).

Value

Send back an invisible object (igraph or Matrix, depending on the output chosen) and optionaly displays a graph (via igraph or corrplot for large ones)

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
fits <- PLNnetwork(Abundance ~ 1, data = trichoptera)

#> 
#>  Initialization...
#>  Adjusting 30 PLN with sparse inverse covariance estimation
#> 	Joint optimization alternating gradient descent and graphical-lasso
#> 	sparsifying penalty = 7.331617 	sparsifying penalty = 6.772001 	sparsifying penalty = 6.2551 	sparsifying penalty = 5.777653 	sparsifying penalty = 5.33665 	sparsifying penalty = 4.929308 	sparsifying penalty = 4.553058 	sparsifying penalty = 4.205527 	sparsifying penalty = 3.884523 	sparsifying penalty = 3.588021 	sparsifying penalty = 3.31415 	sparsifying penalty = 3.061184 	sparsifying penalty = 2.827526 	sparsifying penalty = 2.611704 	sparsifying penalty = 2.412354 	sparsifying penalty = 2.228222 	sparsifying penalty = 2.058143 	sparsifying penalty = 1.901047 	sparsifying penalty = 1.755942 	sparsifying penalty = 1.621912 	sparsifying penalty = 1.498113 	sparsifying penalty = 1.383763 	sparsifying penalty = 1.278142 	sparsifying penalty = 1.180582 	sparsifying penalty = 1.09047 	sparsifying penalty = 1.007235 	sparsifying penalty = 0.9303536 	sparsifying penalty = 0.8593405 	sparsifying penalty = 0.7937478 	sparsifying penalty = 0.7331617 
#>  Post-treatments
#>  DONE!

myNet <- getBestModel(fits)
# NOT RUN {
plot(myNet)
# }