Extract and plot the network (partial correlation, support or inverse covariance) from a PLNnetworkfit object

Extract and plot the network (partial correlation, support or inverse covariance) from a PLNnetworkfit object

# S3 method for PLNnetworkfit
plot(
  x,
  type = c("partial_cor", "support"),
  output = c("igraph", "corrplot"),
  edge.color = c("#F8766D", "#00BFC4"),
  remove.isolated = FALSE,
  node.labels = NULL,
  layout = layout_in_circle,
  plot = TRUE,
  ...
)

Arguments

x
an R6 object with class PLNnetworkfit

type
character. Value of the weight of the edges in the network, either "partial_cor" (partial correlation) or "support" (binary). Default is "partial_cor".

output
the type of output used: either 'igraph' or 'corrplot'. Default is 'igraph'.

edge.color
Length 2 color vector. Color for positive/negative edges. Default is c("#F8766D", "#00BFC4"). Only relevant for igraph output.

remove.isolated
if TRUE, isolated node are remove before plotting. Only relevant for igraph output.

node.labels
vector of character. The labels of the nodes. The default will use the column names ot the response matrix.

layout
an optional igraph layout. Only relevant for igraph output.

plot
logical. Should the final network be displayed or only sent back to the user. Default is TRUE.

...
Not used (S3 compatibility).

Value

Send back an invisible object (igraph or Matrix, depending on the output chosen) and optionally displays a graph (via igraph or corrplot for large ones)

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
fits <- PLNnetwork(Abundance ~ 1, data = trichoptera)

#> 
#>  Initialization...
#>  Adjusting 30 PLN with sparse inverse covariance estimation
#> 	Joint optimization alternating gradient descent and graphical-lasso
#> 	sparsifying penalty = 7.499547 	sparsifying penalty = 6.927113 	sparsifying penalty = 6.398373 	sparsifying penalty = 5.90999 	sparsifying penalty = 5.458886 	sparsifying penalty = 5.042214 	sparsifying penalty = 4.657346 	sparsifying penalty = 4.301855 	sparsifying penalty = 3.973498 	sparsifying penalty = 3.670204 	sparsifying penalty = 3.390061 	sparsifying penalty = 3.1313 	sparsifying penalty = 2.892291 	sparsifying penalty = 2.671525 	sparsifying penalty = 2.46761 	sparsifying penalty = 2.279259 	sparsifying penalty = 2.105285 	sparsifying penalty = 1.944591 	sparsifying penalty = 1.796162 	sparsifying penalty = 1.659062 	sparsifying penalty = 1.532427 	sparsifying penalty = 1.415458 	sparsifying penalty = 1.307418 	sparsifying penalty = 1.207624 	sparsifying penalty = 1.115447 	sparsifying penalty = 1.030306 	sparsifying penalty = 0.9516633 	sparsifying penalty = 0.8790237 	sparsifying penalty = 0.8119286 	sparsifying penalty = 0.7499547 
#>  Post-treatments
#>  DONE!

myNet <- getBestModel(fits)
if (FALSE) {
plot(myNet)
}