Extract and plot the network (partial correlation, support or inverse covariance) from a PLNnetworkfit object

Extract and plot the network (partial correlation, support or inverse covariance) from a PLNnetworkfit object

# S3 method for PLNnetworkfit
plot(
  x,
  type = c("partial_cor", "support"),
  output = c("igraph", "corrplot"),
  edge.color = c("#F8766D", "#00BFC4"),
  remove.isolated = FALSE,
  node.labels = NULL,
  layout = layout_in_circle,
  plot = TRUE,
  ...
)

Arguments

x
an R6 object with class PLNnetworkfit

type
character. Value of the weight of the edges in the network, either "partial_cor" (partial correlation) or "support" (binary). Default is "partial_cor".

output
the type of output used: either 'igraph' or 'corrplot'. Default is 'igraph'.

edge.color
Length 2 color vector. Color for positive/negative edges. Default is c("#F8766D", "#00BFC4"). Only relevant for igraph output.

remove.isolated
if TRUE, isolated node are remove before plotting. Only relevant for igraph output.

node.labels
vector of character. The labels of the nodes. The default will use the column names ot the response matrix.

layout
an optional igraph layout. Only relevant for igraph output.

plot
logical. Should the final network be displayed or only sent back to the user. Default is TRUE.

...
Not used (S3 compatibility).

Value

Send back an invisible object (igraph or Matrix, depending on the output chosen) and optionally displays a graph (via igraph or corrplot for large ones)

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
fits <- PLNnetwork(Abundance ~ 1, data = trichoptera)

#> 
#>  Initialization...
#>  Adjusting 30 PLN with sparse inverse covariance estimation
#> 	Joint optimization alternating gradient descent and graphical-lasso
#> 	sparsifying penalty = 7.510862 	sparsifying penalty = 6.937564 	sparsifying penalty = 6.408026 	sparsifying penalty = 5.918907 	sparsifying penalty = 5.467121 	sparsifying penalty = 5.049821 	sparsifying penalty = 4.664372 	sparsifying penalty = 4.308345 	sparsifying penalty = 3.979492 	sparsifying penalty = 3.675741 	sparsifying penalty = 3.395175 	sparsifying penalty = 3.136024 	sparsifying penalty = 2.896654 	sparsifying penalty = 2.675555 	sparsifying penalty = 2.471332 	sparsifying penalty = 2.282698 	sparsifying penalty = 2.108461 	sparsifying penalty = 1.947524 	sparsifying penalty = 1.798871 	sparsifying penalty = 1.661565 	sparsifying penalty = 1.534739 	sparsifying penalty = 1.417594 	sparsifying penalty = 1.30939 	sparsifying penalty = 1.209446 	sparsifying penalty = 1.11713 	sparsifying penalty = 1.03186 	sparsifying penalty = 0.953099 	sparsifying penalty = 0.8803498 	sparsifying penalty = 0.8131535 	sparsifying penalty = 0.7510862 
#>  Post-treatments
#>  DONE!

myNet <- getBestModel(fits)
if (FALSE) {
plot(myNet)
}