Extract and plot the network (aprtial correlation, support or inverse covariance) from a PLNnetworkfit object

Extract and plot the network (aprtial correlation, support or inverse covariance) from a PLNnetworkfit object

# S3 method for PLNnetworkfit
plot(x, type = c("partial_cor", "support"),
  output = c("igraph", "corrplot"), edge.color = c("#F8766D",
  "#00BFC4"), remove.isolated = FALSE, node.labels = NULL,
  layout = layout_in_circle, plot = TRUE, ...)

Arguments

x
an R6 object with class PLNnetworkfit

type
character. Value of the weigth of the edges in the network, either "partial_cor" (partial correlation) or "support" (binary). Default is "partial_cor".

output
the type of output used: either 'igraph' or 'corrplot'. Default is 'igraph'.

edge.color
2D numeric. Color for positive/negative edges. Default is c("#F8766D", "#00BFC4"). Only relevant for igraph output.

remove.isolated
if TRUE, isolated node are remove before plotting. Only relevant for igraph output.

node.labels
vector of character. The labels of the nodes. The Default will use the column names ot the response matrix.

layout
an optional igraph layout. Only relevant for igraph output.

plot
logical. Should the final network be displayed or only sent back to the user. Default is TRUE.

...
Not used (S3 compatibility).

Value

Send back an invisible object (igraph or Matrix, depending on the output chosen) and optionaly displays a graph (via igraph or corrplot for large ones)