Prediction for a PLNmixturefit object

Predict either posterior probabilities for each group or latent positions based on new samples

# S3 method for class 'PLNmixturefit'
predict(
  object,
  newdata,
  type = c("posterior", "response", "position"),
  prior = matrix(rep(1/object$k, object$k), nrow(newdata), object$k, byrow = TRUE),
  control = PLNmixture_param(),
  ...
)

Arguments

object

an R6 object with class PLNmixturefit

newdata

A data frame in which to look for variables, offsets and counts with which to predict.

type

The type of prediction required. The default posterior are posterior probabilities for each group , response is the group with maximal posterior probability and latent is the averaged latent in the latent space, with weights equal to the posterior probabilities.

prior

User-specified prior group probabilities in the new data. The default uses a uniform prior.

control

a list-like structure for controlling the fit. See PLNmixture_param() for details.

...

additional parameters for S3 compatibility. Not used

Value

A matrix of posterior probabilities for each group (if type = "posterior"), a matrix of (average) position in the latent space (if type = "position") or a vector of predicted groups (if type = "response").

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLNmixture(Abundance ~ 1 + offset(log(Offset)),
           data = trichoptera, control = PLNmixture_param(smoothing = "none"))  %>% getBestModel()
#> 
#>  Initialization...
#> 
#>  Adjusting 5 PLN mixture models.
#> 	number of cluster = 1 	number of cluster = 2 	number of cluster = 3 	number of cluster = 4 	number of cluster = 5 
#>  Post-treatments
#>  DONE!
predict(myPLN, trichoptera, "posterior")
#>            [,1]         [,2]         [,3]         [,4]
#> 1  1.836915e-05 2.712981e-09 9.723639e-01 2.761771e-02
#> 2  4.288422e-02 1.894677e-09 9.570474e-01 6.835215e-05
#> 3  6.913416e-01 1.913536e-08 3.086580e-01 3.821442e-07
#> 4  9.998512e-01 2.220446e-16 1.488425e-04 2.220446e-16
#> 5  9.834608e-01 1.924651e-10 1.653919e-02 6.505828e-10
#> 6  3.887899e-01 4.998733e-07 6.111826e-01 2.706201e-05
#> 7  7.754676e-03 1.351288e-06 9.842658e-01 7.978185e-03
#> 8  9.325808e-01 4.902379e-07 6.741627e-02 2.459389e-06
#> 9  4.911711e-02 3.483857e-07 9.501830e-01 6.995773e-04
#> 10 4.837556e-02 6.795658e-05 9.330558e-01 1.850064e-02
#> 11 1.683385e-01 4.058742e-03 7.969896e-01 3.061315e-02
#> 12 3.493942e-01 9.883364e-02 5.118507e-01 3.992147e-02
#> 13 3.744029e-04 2.220446e-16 9.996256e-01 2.220446e-16
#> 14 3.845561e-04 2.220446e-16 9.995326e-01 8.286954e-05
#> 15 9.999645e-01 2.220446e-16 3.553961e-05 2.220446e-16
#> 16 9.960118e-01 2.220446e-16 3.988228e-03 2.220446e-16
#> 17 1.484505e-04 7.318103e-09 9.998515e-01 6.591001e-08
#> 18 8.306431e-06 2.220446e-16 9.999909e-01 8.028372e-07
#> 19 5.218836e-03 2.220446e-16 9.947812e-01 6.935745e-09
#> 20 2.220446e-16 2.220446e-16 9.996111e-01 3.889287e-04
#> 21 1.612736e-03 2.220446e-16 9.982843e-01 1.029937e-04
#> 22 2.753659e-03 1.716709e-07 6.064670e-01 3.907791e-01
#> 23 3.220125e-10 2.266795e-06 1.067917e-03 9.989298e-01
#> 24 1.318915e-01 2.975999e-03 8.210162e-01 4.411632e-02
#> 25 2.220446e-16 9.243117e-06 2.496648e-04 9.997411e-01
#> 26 2.220446e-16 2.220446e-16 5.222386e-08 9.999999e-01
#> 27 2.220446e-16 2.220446e-16 5.232957e-06 9.999948e-01
#> 28 5.646162e-08 1.082016e-03 9.511947e-04 9.979667e-01
#> 29 4.397012e-04 2.203689e-02 4.378153e-03 9.731453e-01
#> 30 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 31 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 32 5.938693e-06 2.220446e-16 9.999933e-01 7.584720e-07
#> 33 5.344896e-04 2.220446e-16 9.994655e-01 2.220446e-16
#> 34 9.944052e-01 2.220446e-16 5.594800e-03 2.220446e-16
#> 35 4.835489e-02 2.220446e-16 9.516451e-01 4.904721e-10
#> 36 9.737851e-01 2.220446e-16 2.621490e-02 5.829769e-10
#> 37 2.973109e-07 2.220446e-16 9.999996e-01 1.116082e-07
#> 38 1.346859e-04 2.220446e-16 9.971906e-01 2.674735e-03
#> 39 2.674809e-03 2.220446e-16 9.944106e-01 2.914584e-03
#> 40 6.242175e-01 4.188542e-04 3.752896e-01 7.406014e-05
#> 41 3.779012e-02 2.220446e-16 9.622099e-01 2.220446e-16
#> 42 4.473706e-05 2.220446e-16 9.998613e-01 9.396939e-05
#> 43 4.225896e-02 2.220446e-16 9.577395e-01 1.523429e-06
#> 44 2.220446e-16 2.220446e-16 9.999997e-01 3.421927e-07
#> 45 1.952198e-10 2.220446e-16 1.000000e+00 2.220446e-16
#> 46 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 47 7.329901e-03 1.594761e-05 9.896056e-01 3.048600e-03
#> 48 1.320882e-07 2.220446e-16 9.999997e-01 1.866283e-07
#> 49 2.220446e-16 2.220446e-16 5.480927e-06 9.999945e-01
predict(myPLN, trichoptera, "position")
#>          [,1]       [,2]      [,3]       [,4]        [,5]      [,6]       [,7]
#> 1   -6.693310  -7.611498 -1.998481  -5.997648 -0.56540658 -4.191939  -5.862450
#> 2   -6.785825  -7.880402 -1.936211  -6.268328 -0.54113569 -3.969582  -6.130500
#> 3  -10.877416 -11.224223 -3.585615 -10.716789 -0.19350171 -4.017927 -10.674490
#> 4  -12.810473 -12.810626 -4.090980 -12.810407 -0.09119959 -3.796861 -12.810390
#> 5  -12.708319 -12.726412 -3.336244 -12.700140 -0.14455087 -3.639026 -12.698014
#> 6   -8.967169  -9.663666 -2.700949  -8.639426 -0.29849389 -3.468919  -8.552348
#> 7   -6.603637  -7.681578 -1.983590  -6.014405 -0.72458920 -3.926225  -5.871127
#> 8  -12.388419 -12.464363 -3.729131 -12.353123 -0.17909652 -3.556185 -12.343813
#> 9   -6.827464  -7.910856 -1.631782  -6.308641 -0.47448913 -3.955907  -6.171503
#> 10  -6.915503  -7.868339 -1.773035  -6.280614 -0.71629791 -3.869768  -6.143224
#> 11  -7.765739  -8.465938 -2.392969  -7.124863 -0.44947595 -3.867110  -6.983677
#> 12  -9.531152  -9.460426 -3.419788  -8.984186 -0.24389134 -4.144947  -8.316292
#> 13  -5.964355  -7.676251 -3.210601  -6.048899 -0.22864035 -3.549910  -5.914480
#> 14  -6.543195  -7.669675 -2.260333  -6.018829 -0.52773152 -3.363482  -5.296774
#> 15 -12.811925 -12.811956 -4.859972 -12.811912 -0.18344623 -3.474788 -12.811908
#> 16 -12.786898 -12.791044 -4.227582 -12.785090 -0.19544733 -3.083272 -12.784628
#> 17  -6.570283  -7.678680 -3.062716  -6.063414 -0.29717029 -3.972665  -5.930899
#> 18  -6.561486  -7.675169 -2.760642  -5.480969 -0.51619255 -3.012281  -5.916014
#> 19  -6.599694  -7.703916 -2.323785  -6.094136 -0.63665738 -2.660128  -5.961878
#> 20  -6.642607  -7.099159 -2.856319  -6.169129 -0.58812540 -3.432119  -5.552402
#> 21  -6.553409  -7.676765 -2.119980  -6.031045 -0.58221181 -4.201966  -5.893491
#> 22  -8.772453  -6.894939 -2.225570  -5.847486 -0.97881098 -4.618346  -5.764504
#> 23 -12.240622  -5.672332 -3.380869  -5.367445 -0.91537793 -5.613570  -5.615501
#> 24  -7.605796  -8.249770 -2.293862  -6.861042 -0.55220644 -3.881863  -6.722677
#> 25 -12.245325  -5.679830 -2.895210  -5.614019 -1.05890237 -5.390155  -5.390546
#> 26 -12.246766  -5.690885 -3.353592  -5.626182 -1.40647217 -5.637061  -5.637061
#> 27 -12.246742  -5.694399 -3.379446  -5.629911 -1.16264091 -5.640742  -5.640751
#> 28 -12.241568  -5.666811 -3.308229  -5.603073 -1.24492689 -5.606739  -5.607632
#> 29 -12.227990  -5.737176 -3.376607  -5.753704 -0.99764411 -5.638189  -5.631811
#> 30 -12.526281  -8.375778 -6.664445 -12.526281 -0.30982839 -7.222894  -6.458280
#> 31 -12.527435  -8.621891 -5.751466 -12.527435 -0.10983729 -7.527404  -6.591352
#> 32  -6.601756  -7.689682 -2.554800  -6.112048 -0.97637823 -3.366643  -5.985087
#> 33  -6.651257  -7.710097 -2.450760  -6.183791 -0.78083399 -3.039720  -6.063719
#> 34 -12.777995 -12.783440 -3.196256 -12.775699 -0.39702342 -2.876084 -12.775120
#> 35  -6.879624  -7.929146 -2.470079  -6.401860 -0.82402337 -2.777546  -6.277264
#> 36 -12.648469 -12.676955 -3.411496 -12.635659 -0.40766965 -3.488489 -12.632339
#> 37  -6.574928  -7.679900 -1.373230  -5.514373 -1.08322933 -3.678172  -5.939559
#> 38  -6.534180  -7.655820 -1.734603  -5.979414 -1.35702520 -4.011550  -5.836460
#> 39  -6.550270  -7.668009 -1.473037  -5.994542 -0.98089195 -3.999087  -5.851679
#> 40 -10.451559 -10.876539 -3.299109 -10.250557 -0.14971457 -3.873224 -10.194803
#> 41  -6.817978  -7.876461 -2.817444  -6.337183 -0.37257664 -4.420490  -6.211943
#> 42  -6.564524  -7.675971 -2.786994  -5.486956 -0.49221673 -4.331898  -5.920263
#> 43  -6.871033  -7.908657 -3.217028  -6.405475 -0.30371752 -4.194308  -6.284973
#> 44  -6.552181  -7.671915 -2.116140  -5.457171 -0.45281706 -3.847599  -4.799994
#> 45  -6.573818  -7.679503 -2.547129  -5.511673 -0.53105682 -3.383596  -5.389308
#> 46  -6.124352  -7.712499 -3.369817  -6.198802 -0.37699892 -4.438998  -5.597724
#> 47  -6.582957  -7.692562 -2.312826  -6.030645 -0.43534811 -4.021741  -5.888992
#> 48  -5.925271  -7.667046 -3.287388  -6.012327 -0.62324346 -4.168787  -5.873505
#> 49 -12.246795  -5.507048 -3.450461  -5.667342 -1.16847652 -5.677804  -5.677812
#>         [,8]       [,9]     [,10]     [,11]     [,12]     [,13]      [,14]
#> 1  -5.448847  -7.059787 -3.017519 -4.330213 -4.001474 -4.220567  -5.447329
#> 2  -5.458613  -7.152408 -3.598606 -4.275039 -3.297926 -4.142156  -6.359175
#> 3  -6.557981 -10.992747 -4.616237 -5.302372 -3.285719 -4.821373 -10.744841
#> 4  -7.121925 -12.810523 -5.021836 -5.932712 -4.468483 -5.409581 -12.810418
#> 5  -7.040645 -12.714282 -5.073033 -5.752352 -4.675135 -5.157130 -12.701556
#> 6  -6.038120  -9.200063 -4.108105 -4.803626 -4.045309 -4.472539  -8.696910
#> 7  -5.383621  -6.982637 -2.490793 -4.161999 -4.274443 -4.039907  -6.093297
#> 8  -6.933624 -12.413703 -4.862600 -5.599908 -4.879935 -4.977156 -12.359282
#> 9  -5.465346  -7.191797 -3.589519 -4.273147 -4.326899 -4.135198  -6.397789
#> 10 -5.438115  -7.275965 -3.422050 -4.163465 -4.266592 -4.004775  -6.333728
#> 11 -5.650961  -8.051047 -3.604479 -4.331643 -4.367493 -4.097894  -7.105294
#> 12 -6.185803  -9.171282 -4.068133 -4.543304 -4.884199 -4.303827  -8.074241
#> 13 -4.973135  -6.931229 -3.291579 -4.493169 -4.134515 -4.405432  -6.138184
#> 14 -5.457927  -6.918455 -3.419125 -4.387668 -4.472859 -4.293355  -6.110280
#> 15 -6.522868 -12.811935 -3.720888 -5.093647 -5.346303 -5.876673 -12.811914
#> 16 -7.107211 -12.788245 -4.422465 -4.852434 -4.665343 -5.371931 -12.785399
#> 17 -5.530789  -6.331987 -2.907706 -3.461813 -4.621165 -4.459409  -6.151534
#> 18 -5.511232  -6.930709 -3.619961 -2.490854 -4.583569 -4.417681  -6.139086
#> 19 -5.533090  -6.965369 -3.370911 -2.945142 -4.612626 -4.451136  -6.182065
#> 20 -5.231304  -6.990246 -3.291731 -2.161377 -4.898316 -4.760442  -5.734660
#> 21 -5.466189  -6.927484 -2.359336 -3.946222 -4.486983 -4.309569  -6.122038
#> 22 -5.471455  -9.004750 -2.779107 -3.842687 -4.825681 -3.516156  -5.151200
#> 23 -5.593346 -12.241016 -3.210367 -3.289851 -5.627427 -2.650585  -3.482910
#> 24 -5.587009  -7.906856 -3.525296 -4.252274 -4.350587 -4.028718  -6.826581
#> 25 -5.603336 -12.245368 -3.488274 -2.993614 -5.637996 -2.773774  -3.351083
#> 26 -5.615806 -12.246766 -3.384644 -3.105887 -5.418931 -2.474331  -3.261184
#> 27 -5.619569 -12.246744 -2.773475 -2.978324 -5.653947 -2.061842  -3.834362
#> 28 -5.347359 -12.235826 -3.347217 -3.207956 -5.621962 -2.509518  -3.413668
#> 29 -5.634260 -12.105280 -3.405214 -3.012529 -5.679440 -2.579388  -3.660003
#> 30 -7.756767  -6.844141 -5.376905 -3.152891 -8.620946 -4.189506  -3.179671
#> 31 -7.860654  -6.911503 -5.920692 -4.072648 -8.486850 -5.168476  -4.018871
#> 32 -5.605494  -6.959710 -4.217092 -4.687269 -3.747407 -1.137647  -5.655300
#> 33 -5.704578  -6.996932 -3.689954 -4.515847 -2.430406 -1.936941  -6.264144
#> 34 -7.175087 -12.779737 -5.007646 -6.072292 -4.401212 -2.683996 -12.776089
#> 35 -5.624565  -7.226439 -3.248786 -4.642098 -3.909647 -3.507552  -6.484790
#> 36 -7.032733 -12.657838 -5.095711 -5.760612 -5.085221 -4.455456 -12.637873
#> 37 -5.543902  -6.940298 -3.434085 -4.164041 -3.868795 -2.978995  -6.158585
#> 38 -5.396470  -6.915344 -2.062153 -4.231966 -4.333203 -3.609342  -6.068704
#> 39 -5.396979  -6.930766 -3.097710 -4.225224 -4.325437 -3.595089  -6.083278
#> 40 -6.424920 -10.591912 -4.419088 -5.140152 -4.191848 -4.674495 -10.281924
#> 41 -5.618581  -7.167466 -3.278127 -4.649162 -4.287049 -4.547557  -6.420577
#> 42 -4.513013  -6.933041 -3.638868 -4.516065 -4.594680 -3.662937  -6.142439
#> 43 -5.682784  -7.212003 -2.349986 -4.387902 -4.799400 -3.719188  -6.485895
#> 44 -4.946045  -6.924100 -2.944976 -4.455594 -4.087856 -4.365878  -6.125478
#> 45 -4.562215  -6.939501 -2.441812 -4.157449 -4.223910 -4.481784  -6.156988
#> 46 -5.729818  -7.003487 -3.966859 -2.940846 -4.323893 -4.832292  -5.326722
#> 47 -4.818240  -6.960740 -3.147515 -3.731860 -4.346448 -4.140854  -6.118192
#> 48 -4.888219  -6.291869 -3.047988 -4.372422 -4.458639 -3.071540  -6.104286
#> 49 -5.657356 -12.246797 -3.624063 -3.124178 -5.690563 -1.045404  -3.996848
#>        [,15]      [,16]     [,17]
#> 1  -3.998077  -6.092870 -3.062277
#> 2  -5.019787  -5.958873 -3.147980
#> 3  -4.749979  -7.137604 -3.515562
#> 4  -4.653514  -7.718241 -3.920460
#> 5  -4.960946  -7.658385 -3.355604
#> 6  -5.101523  -6.583393 -3.249361
#> 7  -4.961504  -5.934736 -3.016740
#> 8  -5.242652  -7.554831 -2.643500
#> 9  -5.013681  -5.971743 -3.124701
#> 10 -4.910333  -6.065582 -2.884300
#> 11 -4.883469  -6.367004 -2.899834
#> 12 -4.646897  -6.936230 -3.149992
#> 13 -5.170700  -5.393915 -3.658228
#> 14 -4.574301  -5.927799 -2.353073
#> 15 -5.506401  -7.381386 -2.526254
#> 16 -5.551847  -7.706687 -2.791626
#> 17 -5.202972  -5.976693 -2.943528
#> 18 -5.178020  -5.963080 -2.653764
#> 19 -5.196577  -5.981565 -2.248154
#> 20 -3.782241  -6.093005 -2.432176
#> 21 -5.115525  -5.934274 -2.393077
#> 22 -3.005191  -8.374078 -2.657494
#> 23 -1.496870 -12.239939 -2.438974
#> 24 -4.827547  -6.383322 -2.853044
#> 25 -1.428719 -12.245127 -2.744797
#> 26 -1.175049 -12.246766 -2.207228
#> 27 -1.506849 -12.246739 -2.496706
#> 28 -1.320229 -12.235858 -2.474016
#> 29 -1.385119 -12.119029 -2.557537
#> 30 -1.964399  -7.756767 -5.120755
#> 31 -2.907268  -7.860654 -6.417764
#> 32 -5.297540  -6.029612 -2.610303
#> 33 -3.823193  -6.103067 -3.080547
#> 34 -4.531085  -7.746212 -2.336130
#> 35 -5.235275  -6.073008 -1.503708
#> 36 -5.393632  -7.645884 -1.596556
#> 37 -5.220014  -5.985753 -1.907368
#> 38 -5.015052  -5.903885 -2.109419
#> 39 -5.009520  -5.907602 -2.385778
#> 40 -5.146644  -7.004203 -3.282807
#> 41 -5.246976  -5.023804 -1.999493
#> 42 -5.185034  -5.967592 -1.759219
#> 43 -5.321086  -6.110873 -2.943807
#> 44 -5.147840  -5.947245 -3.597815
#> 45 -5.216650  -5.432207 -2.988434
#> 46 -4.380266  -5.202912 -2.185314
#> 47 -5.020259  -5.921405 -3.210726
#> 48 -5.098566  -5.922352 -1.938733
#> 49 -1.930137 -12.246792 -2.841892
predict(myPLN, trichoptera, "response")
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 
#>  3  3  1  1  1  3  3  1  3  3  3  3  3  3  1  1  3  3  3  3  3  3  4  3  4  4 
#> 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 
#>  4  4  4  2  2  3  3  1  3  1  3  3  3  1  3  3  3  3  3  3  3  3  4 
#> Levels: 1 2 3 4