Prediction for a PLNmixturefit object

Predict either posterior probabilities for each group or latent positions based on new samples

# S3 method for PLNmixturefit
predict(
  object,
  newdata,
  type = c("posterior", "response", "position"),
  prior = matrix(rep(1/object$k, object$k), nrow(newdata), object$k, byrow = TRUE),
  control = list(),
  ...
)

Arguments

object
an R6 object with class PLNmixturefit

newdata
A data frame in which to look for variables, offsets and counts with which to predict.

type
The type of prediction required. The default posterior are posterior probabilities for each group , response is the group with maximal posterior probability and latent is the averaged latent in the latent space, with weights equal to the posterior probabilities.

prior
User-specified prior group probabilities in the new data. The default uses a uniform prior.

control
a list for controlling the optimization. See PLN() for details.

...
additional parameters for S3 compatibility. Not used

Value

A matrix of posterior probabilities for each group (if type = "posterior"), a matrix of (average) position in the latent space (if type = "position") or a vector of predicted groups (if type = "response").

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
fits <- PLNmixture(Abundance ~ 1, clusters = 1:4, data = trichoptera)

#> 
#>  Initialization...
#> 
#>  Adjusting 4 PLN mixture models.
#> 	number of cluster = 1 	number of cluster = 2 	number of cluster = 3 	number of cluster = 4 
#> 
#>  Smoothing PLN mixture models.
#>    Going backward +++                                                                                                       Going forward +++                                                                                                       Going backward +++                                                                                                       Going forward +++                                                                                                    
#>  Post-treatments
#>  DONE!

myMixture <- getBestModel(fits)
predict(myMixture, trichoptera, "posterior")

#>            [,1]         [,2]         [,3]         [,4]
#> 1  2.220446e-16 1.000000e+00 4.382172e-12 6.382089e-16
#> 2  2.220446e-16 2.171992e-12 1.000000e+00 2.046603e-09
#> 3  2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 4  2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 5  2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 6  2.220446e-16 6.912108e-08 9.999999e-01 4.120025e-12
#> 7  2.220446e-16 5.907042e-14 1.183318e-01 8.816682e-01
#> 8  2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 9  2.220446e-16 1.163075e-12 1.000000e+00 2.589062e-09
#> 10 2.220446e-16 2.220446e-16 4.919124e-01 5.080876e-01
#> 11 2.220446e-16 2.220446e-16 3.069129e-01 6.930871e-01
#> 12 2.220446e-16 2.220446e-16 5.586897e-01 4.413103e-01
#> 13 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 14 2.220446e-16 1.000000e+00 2.697108e-16 2.220446e-16
#> 15 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 17 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 18 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 19 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 20 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 21 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 22 2.220446e-16 9.466207e-12 1.260723e-01 8.739277e-01
#> 23 2.220446e-16 6.742852e-12 1.000000e+00 2.860743e-10
#> 24 2.220446e-16 2.220446e-16 5.172536e-01 4.827464e-01
#> 25 2.220446e-16 1.776205e-09 1.986986e-11 1.000000e+00
#> 26 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 27 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 28 2.220446e-16 2.220446e-16 6.729536e-02 9.327046e-01
#> 29 2.220446e-16 5.568445e-14 6.507983e-02 9.349202e-01
#> 30 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 31 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 32 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 33 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 34 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 35 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 36 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 37 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 38 2.220446e-16 2.421665e-11 9.820803e-08 9.999999e-01
#> 39 2.220446e-16 3.545812e-10 8.942552e-03 9.910574e-01
#> 40 2.220446e-16 1.134428e-06 9.999989e-01 1.976538e-14
#> 41 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 42 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 43 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 44 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 45 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 46 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 47 2.220446e-16 4.792841e-10 1.000000e+00 1.880326e-10
#> 48 2.220446e-16 1.000000e+00 3.388226e-14 2.220446e-16
#> 49 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16

predict(myMixture, trichoptera, "position")

#>          [,1]         [,2]        [,3]       [,4]      [,5]         [,6]
#> 1  -2.3024550 -2.425401957  1.18560214 -2.0478779 2.4535743 -0.007759000
#> 2  -1.6155381 -1.616239857  0.42569087 -1.6156089 1.8252772 -1.611587518
#> 3  -2.2968928 -2.420489836  0.52782588 -2.0406785 3.1968000  0.054239406
#> 4  -2.2897513 -2.414185534  1.08014512 -2.0314263 5.1109270  1.093431257
#> 5  -2.2924412 -2.416559720  1.22450285 -2.0349124 4.0975072  0.596864820
#> 6  -1.6155290 -1.616230709  0.39695567 -1.6155997 1.9945392 -0.621184944
#> 7  -1.6694550 -1.669925379  0.08239675 -1.6273950 0.7332526 -1.622815448
#> 8  -2.2984246 -2.421842446  0.50318540 -2.0426620 3.0599173  0.279719887
#> 9  -1.6155760 -1.616277684  0.45419154 -1.6156468 1.8240923 -1.611625521
#> 10 -1.5372995 -1.537946642  0.09548593 -1.5098615 0.3548069 -1.504903314
#> 11 -1.5881795 -1.588750304 -0.39654925 -1.5531738 0.6503833 -1.548322402
#> 12 -1.5271735 -1.527842289 -0.87175534 -1.5026816 0.7719170 -1.497729658
#> 13 -1.9879246 -2.419253062  0.55087494 -2.0388644 3.4634002  0.553971455
#> 14 -2.3007127 -2.423863112  1.10235226 -2.0456235 2.6591534  0.481502041
#> 15 -2.2890462 -2.413563331  1.09606902 -2.0305118 6.2094344  2.826216541
#> 16 -2.2900521 -2.414450966  0.86303646 -2.0318161 4.8879723  1.999710263
#> 17 -2.2947137 -2.418566020  0.78026489 -2.0378563 3.5868181  0.326779287
#> 18 -2.2964424 -2.420092208  0.95196980 -1.7344597 3.1994553  0.980333849
#> 19 -2.2961802 -2.419860785  1.49215055 -2.0397556 3.1869881  1.376012246
#> 20 -2.2915002 -2.106875376  1.76416475 -2.0336924 4.1771202  1.276723534
#> 21 -2.3002301 -2.423436916  1.23715649 -2.0449988 2.6822202  0.016889566
#> 22 -1.5729186 -1.573432047  0.47151754 -1.5250493 1.1204981 -1.519930215
#> 23 -1.6154911 -1.616192837 -1.16419807 -0.6250132 1.8267502 -1.611540279
#> 24 -1.5316198 -1.532272144 -0.34977855 -1.5046651 0.3187298 -1.499694454
#> 25 -1.7195625 -1.719963671  0.47863074 -1.6725648 1.7841272 -0.671224384
#> 26 -2.3020092 -2.425008460  0.44907456 -2.0473001 1.9622344 -0.002350105
#> 27 -2.3007139 -2.423864271  0.46874100 -2.0456244 2.3805356  0.011751729
#> 28 -1.6934065 -1.693847079 -1.12069567 -1.6491336 0.1764743 -1.644626334
#> 29 -1.7050983 -1.705532977 -1.14897948 -1.6614911 0.6628334 -1.657052395
#> 30 -0.4186995 -0.003070577  1.37557198 -0.4186995 7.0646364  0.619959981
#> 31 -1.6935555 -1.693967149  2.38937430 -1.6453007 7.8861870 -0.602842597
#> 32 -2.2944670 -2.418348357  1.69759616 -2.0375359 3.3437535  1.019095531
#> 33 -2.2915225 -2.415749223  2.36037362 -2.0337206 4.0215191  1.825951607
#> 34 -2.2891950 -2.413694676  2.62516190 -2.0307046 5.3560382  2.846760621
#> 35 -2.2954968 -2.419257513  1.51161322 -2.0388699 3.1743769  1.392365420
#> 36 -2.2921967 -2.416344160  1.23275644 -2.0345948 3.9757749  0.833535303
#> 37 -2.2960197 -2.419719603  2.60623878 -1.7337441 2.8624236  0.549602613
#> 38 -1.7381017 -1.738495465  0.58357452 -1.6919874 0.5034843 -1.687675366
#> 39 -1.7290336 -1.729435003  0.77379775 -1.6830383 0.8868653 -1.678686283
#> 40 -1.6156022 -1.616304039 -0.18696496 -1.6156727 2.1037917 -1.611648583
#> 41 -2.2936676 -2.417642565  1.20121697 -2.0365009 3.7289257  0.091178443
#> 42 -2.2958214 -2.419543943  0.96467676 -1.7335091 3.2794998  0.066506749
#> 43 -2.2920248 -2.416192256  1.03752106 -2.0343727 4.1493801  0.361047863
#> 44 -2.2981119 -2.421566311  1.44515009 -1.7369718 3.0505906  0.283976700
#> 45 -2.2953961 -2.419168467  1.34414028 -1.7328708 3.4018335  0.783184153
#> 46 -1.9823162 -2.414976599  1.26209428 -2.0325877 4.5368795  0.378562657
#> 47 -1.6155013 -1.616203026  0.39716294 -1.6155721 1.9110830 -1.611550514
#> 48 -1.9955059 -2.424976036  0.23051535 -2.0472539 2.4734696 -0.002426986
#> 49 -2.2949954 -2.110403487  0.77910803 -2.0382201 3.1369860  0.076364183
#>          [,7]       [,8]       [,9]       [,10]         [,11]       [,12]
#> 1  -1.8512220 -1.5610980 -2.4351330  0.44269098 -0.2784404267 -0.15824730
#> 2  -1.6114559 -1.5628443 -1.6157098 -1.43490212 -1.4838696194 -0.26161303
#> 3  -1.8424288 -1.5492703 -2.4302689  0.03559807 -0.2326588824  0.64955590
#> 4  -1.8311178 -1.5340292 -2.4240265  0.37086018 -0.1725274342  0.47805578
#> 5  -1.8353812 -1.5397779 -2.4263773  0.08539241 -0.1953765296 -0.07777379
#> 6  -1.6114466 -1.5628346 -1.6157006 -1.43489095 -1.4838590010 -0.57292524
#> 7  -1.6645276 -1.5715607 -1.6695674 -0.24779395 -1.6388606186 -1.56943092
#> 8  -1.8448523 -1.5525326 -2.4316083  0.01863929 -0.2453886302 -0.39481325
#> 9  -1.6114939 -1.5628842 -1.6157476 -1.43494747 -1.4839128007 -1.56221641
#> 10 -1.5321071 -1.4472643 -1.5374561 -1.30593796 -1.4520202518 -1.44552331
#> 11 -1.5830366 -1.4928543 -1.5883170 -1.36177403 -1.5280907771 -1.49089398
#> 12 -1.5220087 -1.4398499 -1.5273356 -1.29636210 -1.4337360271 -1.43820407
#> 13 -1.8402119 -1.2454025 -2.4290443  0.53779127 -0.2209679633 -0.10241238
#> 14 -1.5457047 -1.5573976 -2.4336092  0.23504309 -0.2642019179 -0.41087360
#> 15 -1.8299988 -0.9285406 -2.4234104  2.73733811  1.1088114783  0.22155638
#> 16 -1.8315947 -1.5346724 -2.4242893  0.84901651  0.6123564244  0.21193693
#> 17 -1.8389797 -1.5446254 -2.1199317  0.99669673  0.5574016524 -0.36845332
#> 18 -1.8417158 -1.5483102 -2.4298752  0.28749775  1.4072363120 -0.38075438
#> 19 -1.8413007 -1.5477510 -2.4296460  0.52905235  1.0030180877 -0.37887743
#> 20 -1.5287215 -1.2354949 -2.4255551  1.26424552  2.6887110186 -0.34536751
#> 21 -1.8477063 -1.5563712 -2.4331871  1.10163335 -0.0008184138 -0.40747691
#> 22 -1.5674477 -1.4625229 -1.5730412 -0.20363694 -1.5422889800 -1.46008211
#> 23 -1.6114086 -1.5627947 -1.6156627 -0.44877053 -1.4838159432 -1.56212685
#> 24 -1.5264233 -1.4418517 -1.5317777 -1.29985650 -1.4448693980 -1.44012087
#> 25 -0.7094538 -1.6197921 -1.7196577 -1.51644237 -0.2440782810 -1.61754173
#> 26 -1.8505154 -1.5601451 -2.4347433  0.22085444  0.2333873542 -0.15376791
#> 27 -1.8484699 -1.5573966 -2.4336103  1.09826681  0.4883828765 -0.41080128
#> 28 -1.6885377 -0.7931323 -1.6935115 -1.48431049 -1.6718172831 -1.59233510
#> 29 -1.7002975 -1.6077448 -1.7052019 -1.49950583 -0.8083017393 -1.60559433
#> 30  1.2922454  0.3521596  0.9986512  2.10622880  4.1872022836 -0.01863987
#> 31  0.4900445 -0.5614373  0.1719343  1.45265241  3.8420227163 -0.55951761
#> 32 -1.8385870 -1.5440941 -2.4281484  0.06306301 -0.2121987598  0.43212078
#> 33 -1.8339222 -1.5378066 -2.4255748  0.82995754  0.0816537518  2.54471734
#> 34 -1.8302343 -1.5328357 -2.4235404  0.60791426 -0.1676643324  0.98617711
#> 35 -1.8402178 -1.5462910 -2.4290487  0.76822447 -0.2208605372  0.16311390
#> 36 -1.8349917 -1.5392501 -2.4261639  0.08853670 -0.1931285421 -0.35035636
#> 37 -1.8410419 -1.5473961 -2.4295063  0.53345794  0.0394814532  0.15904425
#> 38 -1.7334079 -1.6402548 -1.7381951  0.37592333 -1.7283544265 -1.63804997
#> 39 -1.7243019 -1.6307488 -1.7291289 -0.59869039 -1.7176116538 -1.62853938
#> 40 -1.6115197 -1.5629109 -1.6157740 -1.43497569 -1.4839400064 -0.57313205
#> 41 -1.8373226 -1.5423923 -2.4274496  0.79507474 -0.2056007725 -0.08747431
#> 42 -1.8407328 -0.9477619 -2.4293323  0.29536918 -0.2236796146 -0.37632771
#> 43 -1.8347210 -1.5388873 -2.4260135  2.20274549  0.0762592311 -0.34921306
#> 44 -1.2393539 -1.2518427 -2.4313348  0.72697373 -0.2427567179 -0.12336592
#> 45 -1.5358802 -0.9465492 -2.4289605  1.55819253  0.0444613889 -0.10147159
#> 46 -1.5272265 -1.5359438 -2.4248098  0.59823561  2.0378787973  0.20659901
#> 47 -1.6114189 -0.5734986 -1.6156729 -0.44880328 -0.4964789103 -1.56213761
#> 48 -1.8504601 -1.2614484 -2.1273661  0.44946133 -0.2744981777 -0.41967138
#> 49 -1.8394231 -1.5452193 -2.4286105  0.30684740  0.7968458820 -0.37031321
#>           [,13]      [,14]       [,15]      [,16]       [,17]
#> 1  -0.115082844 -1.2945682  0.21954197 -1.8363122  0.56281663
#> 2  -1.564571191 -1.6066891 -1.36602287 -1.6160654 -1.42393802
#> 3  -0.060167604 -1.5820428  0.02226258 -1.8273844  0.45740024
#> 4   0.012308616 -1.5673026  0.60837435 -1.8158994  1.00764905
#> 5  -0.015282738 -1.5728618  0.06400209 -1.8202285  0.95258412
#> 6  -1.564561364 -1.6066798 -1.36601094 -1.6160562 -1.42392667
#> 7  -1.618730241 -1.5402629 -1.34093872 -1.6698115 -1.45442618
#> 8  -0.075469070 -1.5851983 -0.25350779 -1.8298452  1.01449392
#> 9  -1.564611024 -1.6067273 -1.36607145 -1.6161032 -1.42398387
#> 10 -1.479220472 -1.4478162 -1.18980030 -1.5377877 -1.29871475
#> 11 -1.532763154 -1.4778458 -1.24321186 -1.5886109 -1.35678426
#> 12 -1.468667317 -1.4459687 -1.18251611 -1.5276778 -1.28843142
#> 13 -0.046099089 -1.5791556 -0.22930009 -1.5211548  0.48200943
#> 14 -0.098026840 -1.5899044 -0.01294344 -1.8335175  1.13532740
#> 15  0.019727161 -1.5658418  0.36089480 -1.5091695  3.85054626
#> 16  0.009206847 -1.5679246 -0.18383105 -1.8163837  2.30418125
#> 17 -0.038223499 -1.5775502 -0.22282293 -1.8238823  1.09593784
#> 18 -0.055614080 -1.5811142 -0.23714583 -1.8266606  1.23280292
#> 19 -0.052946482 -1.5805733 -0.23495632 -1.8262391  1.69244723
#> 20 -0.005433631 -1.2682758  1.06762266 -1.8187128  2.36369774
#> 21 -0.093241742 -1.5889115 -0.26821071 -1.8327428  1.14766880
#> 22 -1.516699661 -1.4246684 -0.09162600 -1.5733078 -0.73847979
#> 23 -1.564521677 -0.6162954  0.08265706 -1.6160184 -0.06415358
#> 24 -1.473361861 -1.4432924 -1.18347547 -1.5321119 -1.29254784
#> 25 -1.672070142 -0.1278109  1.23676683 -1.7198673 -0.54880560
#> 26  0.603392477 -0.7032370  2.19822934 -1.8355947  1.47376166
#> 27  1.413986810 -1.5899037  1.81611949 -1.8335180  1.14298084
#> 28 -1.644029940 -0.7639481  0.05271202 -1.6937407 -0.69951370
#> 29 -1.656429505 -1.5720568  0.08530062 -1.7054281 -0.65773419
#> 30  3.126519250  4.1726861  5.39635421  0.3521596  1.99548647
#> 31  2.519547052  3.9061614  5.06294563 -0.6473789  0.04627995
#> 32  3.303234289 -1.2752456 -0.22061025 -1.8234836  1.75070879
#> 33  2.917753061 -1.5709560  1.06898391 -1.8187469  1.68127592
#> 34  3.129262141 -1.5661486  1.09908741 -1.8150022  3.44953286
#> 35  0.691605033 -1.5791614 -0.22919551 -1.8251394  2.57294296
#> 36  0.496302539 -1.5723517 -0.20170751 -1.8198330  2.86140211
#> 37  1.124408294 -1.5802308 -0.23323913 -1.8259760  2.15445245
#> 38 -0.712879898 -1.5984343 -1.42811288 -1.7384009  0.11896795
#> 39 -0.721231368 -1.5895482 -1.41596991 -1.7293385 -0.15114402
#> 40 -1.564635590 -1.6067529 -1.36610241 -1.6161289 -1.42401137
#> 41 -0.027564877 -1.5753905 -0.21406874 -1.2149843  2.20551504
#> 42  0.450354919 -1.5798336 -0.23198857 -1.8256624  2.14058281
#> 43  0.738910985 -1.5720006 -0.20039980 -1.8195582  1.50342623
#> 44 -0.072289932 -1.5845532 -0.25089926 -1.8293424  0.43660087
#> 45 -0.045081775 -1.5789571 -0.22847417 -1.5209422  1.08181461
#> 46  0.003168601 -0.9655588  0.59809729 -1.2085853  2.75060328
#> 47 -1.564532405 -1.6066519 -1.36597556 -1.6160286 -1.42389338
#> 48  0.599713576 -1.5924935 -0.28236773 -1.8355387  1.47268188
#> 49  3.419785298 -1.5781250  2.09235635 -1.8243324  1.27579178

predict(myMixture, trichoptera, "response")

#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 
#>  2  3  2  2  2  3  4  2  3  4  4  3  2  2  2  2  2  2  2  2  2  4  3  3  4  2 
#> 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 
#>  2  4  4  1  4  2  2  2  2  2  2  4  4  3  2  2  2  2  2  2  3  2  2 
#> Levels: 1 2 3 4