Prediction for a PLNmixturefit object

Predict either posterior probabilities for each group or latent positions based on new samples

# S3 method for PLNmixturefit
predict(
  object,
  newdata,
  type = c("posterior", "response", "position"),
  prior = matrix(rep(1/object$k, object$k), nrow(newdata), object$k, byrow = TRUE),
  control = list(),
  ...
)

Arguments

object

an R6 object with class PLNmixturefit

newdata

A data frame in which to look for variables, offsets and counts with which to predict.

type

The type of prediction required. The default posterior are posterior probabilities for each group , response is the group with maximal posterior probability and latent is the averaged latent in the latent space, with weights equal to the posterior probabilities.

prior

User-specified prior group probabilities in the new data. The default uses a uniform prior.

control

a list for controlling the optimization. See PLN() for details.

...

additional parameters for S3 compatibility. Not used

Value

A matrix of posterior probabilities for each group (if type = "posterior"), a matrix of (average) position in the latent space (if type = "position") or a vector of predicted groups (if type = "response").

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLNmixture(Abundance ~ 1 + offset(log(Offset)),
           data = trichoptera, control_main = list(iterates = 0))  %>% getBestModel()
#> 
#>  Initialization...
#> 
#>  Adjusting 5 PLN mixture models.
#> 	number of cluster = 1 	number of cluster = 2 	number of cluster = 3 	number of cluster = 4 	number of cluster = 5 
#> 
#>  Smoothing PLN mixture models.
#> 
#>  Post-treatments
#>  DONE!
predict(myPLN, trichoptera, "posterior")
#>            [,1]         [,2]         [,3]         [,4]
#> 1  1.185519e-04 6.278938e-11 2.220446e-16 9.998814e-01
#> 2  9.990542e-01 1.274502e-12 2.220446e-16 9.458238e-04
#> 3  9.990197e-01 8.926873e-15 2.220446e-16 9.802644e-04
#> 4  2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 5  5.129610e-12 2.220446e-16 2.220446e-16 1.000000e+00
#> 6  9.593809e-01 1.548675e-11 2.220446e-16 4.061905e-02
#> 7  9.904230e-01 6.998576e-09 2.220446e-16 9.576953e-03
#> 8  4.366554e-06 2.377237e-12 2.220446e-16 9.999956e-01
#> 9  9.964871e-01 1.553247e-10 2.220446e-16 3.512920e-03
#> 10 9.957490e-01 2.624593e-08 2.220446e-16 4.250980e-03
#> 11 9.150074e-01 6.755218e-02 2.220446e-16 1.744042e-02
#> 12 1.502618e-02 8.020125e-01 2.220446e-16 1.829613e-01
#> 13 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 14 2.466391e-08 1.753362e-14 2.220446e-16 1.000000e+00
#> 15 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 17 7.348354e-15 2.564106e-15 2.220446e-16 1.000000e+00
#> 18 7.055559e-13 2.220446e-16 2.220446e-16 1.000000e+00
#> 19 1.118807e-09 2.220446e-16 2.220446e-16 1.000000e+00
#> 20 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 21 2.166557e-05 1.620105e-15 2.220446e-16 9.999783e-01
#> 22 6.250837e-01 3.698874e-01 2.220446e-16 5.028937e-03
#> 23 6.784667e-08 9.999955e-01 2.220446e-16 4.414131e-06
#> 24 9.268220e-01 6.200592e-02 2.220446e-16 1.117213e-02
#> 25 2.220446e-16 1.000000e+00 2.220446e-16 9.112383e-10
#> 26 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 27 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 28 2.430174e-08 9.999993e-01 2.220446e-16 6.940381e-07
#> 29 1.402441e-07 9.999755e-01 2.220446e-16 2.435833e-05
#> 30 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 31 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 32 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 33 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 34 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 35 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 36 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 37 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 38 1.000000e+00 2.334079e-14 2.220446e-16 5.053465e-11
#> 39 1.000000e+00 3.059432e-13 2.220446e-16 3.657822e-10
#> 40 4.587616e-05 3.199274e-10 2.220446e-16 9.999541e-01
#> 41 6.974185e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 42 1.000000e+00 2.220446e-16 2.220446e-16 9.580397e-15
#> 43 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 44 1.144075e-14 2.220446e-16 2.220446e-16 1.000000e+00
#> 45 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 46 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 47 3.497724e-05 2.934837e-09 2.220446e-16 9.999650e-01
#> 48 3.730217e-10 1.000000e+00 2.220446e-16 2.220446e-16
#> 49 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
predict(myPLN, trichoptera, "position")
#>          [,1]       [,2]      [,3]      [,4]        [,5]      [,6]       [,7]
#> 1   -7.636128  -7.630930 -2.580009 -6.882541 -0.55732607 -3.692200  -6.164740
#> 2  -16.599781 -16.599776 -2.171127 -6.290297 -0.61042396 -3.449767 -16.598381
#> 3  -16.599472 -16.599467 -2.734661 -6.298060 -0.18412962 -3.574690 -16.598021
#> 4   -7.649200  -7.644074 -3.705449 -6.912374 -0.09032065 -3.765399  -6.223658
#> 5   -7.639422  -7.634246 -3.073975 -6.891837 -0.14293739 -3.619001  -6.182345
#> 6  -16.243687 -16.243476 -2.404317 -6.314835 -0.35555894 -3.343052 -16.183576
#> 7  -16.522310 -16.522260 -2.242684 -6.293804 -0.76046851 -3.423550 -16.508135
#> 8   -7.635540  -7.630344 -3.185636 -6.883536 -0.17951916 -3.565126  -6.165449
#> 9  -16.576739 -16.576721 -2.023324 -6.291485 -0.55805101 -3.444546 -16.571540
#> 10 -16.570114 -16.570092 -2.175023 -6.289441 -0.76397375 -3.401118 -16.563822
#> 11 -15.898556 -15.852300 -2.454887 -6.440285 -0.63329268 -3.644726 -15.735870
#> 12  -8.398237  -7.848919 -4.591548 -8.094139 -0.34027155 -6.314054  -6.500639
#> 13  -7.476817  -7.631620 -3.283550 -6.886313 -0.21671074 -3.488698  -6.171083
#> 14  -7.635414  -7.630217 -2.753530 -6.883354 -0.52292075 -3.411769  -6.005638
#> 15  -7.683725  -7.678773 -4.621161 -6.983195 -0.18368383 -3.480765  -6.360114
#> 16  -7.647476  -7.642342 -3.744011 -6.908770 -0.19877460 -3.154195  -6.216470
#> 17  -7.637594  -7.632408 -3.203259 -6.887974 -0.28755090 -3.671612  -6.174475
#> 18  -7.636984  -7.631795 -3.023899 -6.727067 -0.51258738 -3.214966  -6.171838
#> 19  -7.637420  -7.632233 -2.670566 -6.887605 -0.63758936 -2.967544  -6.173722
#> 20  -7.643587  -7.478938 -3.000289 -6.900613 -0.58769425 -3.424644  -6.043208
#> 21  -7.635783  -7.630587 -2.632333 -6.883711 -0.58055403 -3.718914  -6.166009
#> 22 -13.535294 -13.283562 -2.497673 -7.075816 -0.91975534 -4.754656 -12.785347
#> 23  -8.423718  -7.744454 -5.093976 -7.475203 -0.80703221 -7.017326  -6.424042
#> 24 -16.000215 -15.957762 -2.424886 -6.424553 -0.69877487 -3.619080 -15.857910
#> 25  -8.426991  -7.750810 -3.291227 -8.413625 -1.01436274 -6.135741  -5.583710
#> 26  -8.431196  -7.758904 -4.545523 -8.417886 -1.46228861 -7.046033  -6.473012
#> 27  -8.432480  -7.761358 -4.577516 -8.419186 -1.15081833 -7.050800  -6.480923
#> 28  -8.420750  -7.738632 -5.022575 -8.407300 -1.29153551 -7.005453  -6.402958
#> 29  -8.421064  -7.739282 -5.031128 -8.407600 -0.88409556 -7.006712  -6.405358
#> 30  -8.756392  -7.682219 -6.842534 -8.746114 -0.31067561 -6.902678  -6.396534
#> 31  -8.928140  -8.489951 -5.503338 -8.918924 -0.10992945 -7.652457  -6.641076
#> 32 -16.608271 -16.608271 -2.533577 -6.312840 -0.99441439 -3.289083 -16.608271
#> 33 -16.608272 -16.608272 -2.453975 -6.327044 -0.78591823 -3.052780 -16.608272
#> 34 -16.608274 -16.608274 -3.079533 -6.383055 -0.40781745 -2.834292 -16.608274
#> 35 -16.608271 -16.608271 -2.440903 -6.306121 -0.83877552 -2.935304 -16.608271
#> 36 -16.608271 -16.608271 -2.843922 -6.313448 -0.41641345 -3.419555 -16.608271
#> 37 -16.608271 -16.608271 -1.373459 -6.156624 -1.11718950 -3.424101 -16.608271
#> 38 -16.608270 -16.608270 -2.065842 -6.290368 -1.21269534 -3.460340 -16.608270
#> 39 -16.608270 -16.608270 -1.916872 -6.290056 -0.95967861 -3.455052 -16.608270
#> 40  -7.634690  -7.629487 -3.084934 -6.880916 -0.14366421 -3.650862  -6.160553
#> 41  -7.638639  -7.633459 -3.018555 -6.890183 -0.36789809 -3.861982  -6.178980
#> 42 -16.608271 -16.608271 -2.643605 -6.153345 -0.48796651 -3.643172 -16.608271
#> 43 -16.608271 -16.608271 -2.982029 -6.314968 -0.30112322 -3.687433 -16.608271
#> 44  -7.636286  -7.631094 -2.566553 -6.725482 -0.44640894 -3.606372  -5.850716
#> 45  -7.637942  -7.632759 -2.839709 -6.729243 -0.52827231 -3.407731  -6.017381
#> 46  -7.486129  -7.640346 -3.427509 -6.904609 -0.38632639 -3.997990  -6.051799
#> 47  -7.634242  -7.629037 -2.903322 -6.880179 -0.41693866 -3.631918  -6.158911
#> 48  -7.499830  -7.753320 -5.185300 -8.414942 -0.63839960 -7.035075  -6.454599
#> 49  -8.445684  -6.912346 -4.347852 -8.432559 -1.16172966 -7.097542  -6.555780
#>         [,8]       [,9]     [,10]     [,11]     [,12]       [,13]     [,14]
#> 1  -6.047100  -7.613592 -3.298235 -3.771219 -4.870532 -15.7163227 -6.032097
#> 2  -6.301636 -16.599759 -3.623992 -5.900267 -3.671838  -2.7457782 -7.038581
#> 3  -6.309276 -16.599450 -3.729906 -5.911574 -3.478112  -2.9875665 -7.042218
#> 4  -6.114409  -7.626981 -4.072135 -4.292503 -4.769416 -15.7178846 -6.250022
#> 5  -6.068216  -7.616982 -3.818793 -3.960373 -4.932052 -15.7178815 -6.209806
#> 6  -6.292256 -16.242769 -3.634417 -5.812569 -3.884578  -3.3029904 -7.005105
#> 7  -6.297784 -16.522094 -3.310515 -5.878459 -3.957937  -2.8015432 -7.030431
#> 8  -6.049208  -7.613011 -3.621362 -3.791593 -5.034317 -15.7178229 -6.193331
#> 9  -6.300634 -16.576660 -3.618920 -5.893995 -3.964535  -2.7678487 -7.036226
#> 10 -6.297989 -16.570018 -3.582408 -5.888674 -3.939573  -2.6857036 -7.034422
#> 11 -6.317840 -15.778758 -3.635813 -5.703065 -4.211538  -2.9122447 -6.807776
#> 12 -6.527231  -6.975155 -4.215303 -3.609684 -7.202091  -5.6743893 -4.303066
#> 13 -5.896800  -7.614311 -3.404355 -3.850995 -4.895655 -15.7178807 -6.198850
#> 14 -6.048780  -7.612883 -3.467876 -3.787498 -5.033145 -15.7178799 -6.192958
#> 15 -5.976204  -7.662268 -3.587268 -4.504958 -5.289932 -15.7178956 -6.383148
#> 16 -6.106387  -7.625219 -3.771648 -3.864403 -4.890709 -15.7178840 -6.243021
#> 17 -6.059390  -7.455186 -3.173005 -3.455374 -5.062166 -15.7178809 -6.202150
#> 18 -6.056431  -7.614489 -3.556697 -2.866825 -5.054125 -15.7178808 -6.199585
#> 19 -6.058546  -7.614935 -3.439053 -3.167058 -5.059875 -15.7178809 -6.201418
#> 20 -5.931606  -7.621242 -3.370457 -2.155024 -5.138164 -15.7178828 -6.070101
#> 21 -6.049638  -7.613256 -2.898435 -3.645752 -5.035473 -15.7175957 -6.193714
#> 22 -6.433355 -12.886560 -3.302305 -5.118529 -5.370534  -3.1360813 -5.924956
#> 23 -6.670120  -6.675958 -3.868222 -3.890301 -7.776086  -3.7976241 -3.455945
#> 24 -6.317198 -15.890263 -3.627173 -5.727561 -4.180890  -2.8131799 -6.829049
#> 25 -6.687757  -6.693492 -4.684656 -3.033016 -7.782262  -3.9543771 -3.162315
#> 26 -6.709490  -6.715114 -4.168774 -3.264762 -6.881867  -2.8874351 -3.026002
#> 27 -6.715932  -6.721524 -2.714946 -2.994420 -7.792498  -2.1109038 -4.402684
#> 28 -5.732586  -6.659401 -4.479384 -3.705723 -7.770454  -3.6006821 -3.209064
#> 29 -6.655351  -6.661302 -4.492933 -3.042191 -7.771019  -3.6271229 -3.964618
#> 30 -7.142148  -6.773891 -5.137146 -3.126065 -8.285000  -4.0690417 -3.159013
#> 31 -7.512049  -6.922214 -5.895165 -4.103393 -7.997167  -5.1940359 -4.051242
#> 32 -6.324711 -16.608271 -3.914152 -5.936245 -3.779114  -0.9857076 -6.901352
#> 33 -6.338749 -16.608272 -3.697478 -5.810993 -2.577371  -1.7987753 -7.057195
#> 34 -6.394126 -16.608274 -4.296100 -6.037004 -4.113649  -2.5117146 -7.084577
#> 35 -6.318072 -16.608271 -3.440476 -5.926439 -3.846139  -2.8425727 -7.047167
#> 36 -6.325313 -16.608271 -3.920861 -5.937132 -4.183816  -3.1236146 -7.050667
#> 37 -6.317163 -16.608271 -3.559302 -5.777130 -3.836964  -2.5947455 -7.046729
#> 38 -6.302508 -16.608270 -3.054421 -5.903327 -3.970587  -2.6162715 -7.039690
#> 39 -6.302200 -16.608270 -3.483060 -5.902869 -3.967415  -2.6049127 -7.039542
#> 40 -6.043229  -7.612134 -3.551260 -3.732899 -4.859124 -15.7172772 -6.188183
#> 41 -6.064443  -7.616181 -3.369342 -3.928763 -4.921271 -15.7178813 -6.206532
#> 42 -6.016639 -16.608271 -3.652891 -5.920599 -4.084227  -2.8543793 -7.045266
#> 43 -6.326815 -16.608271 -2.626612 -5.792323 -4.196557  -3.0474669 -7.051394
#> 44 -5.894072  -7.613775 -3.230556 -3.827645 -4.888213 -15.7178805 -6.196645
#> 45 -5.744496  -7.615469 -2.799759 -3.755996 -4.911590 -15.7178811 -6.203613
#> 46 -6.097082  -7.623188 -3.815359 -2.991602 -4.864171 -15.7178834 -5.922416
#> 47 -5.881719  -7.611678 -3.376200 -3.560285 -5.012877 -15.7174203 -6.186688
#> 48 -5.825068  -5.830415 -3.533905 -4.088990 -7.784697  -2.7142838 -4.278530
#> 49 -6.777882  -6.783191 -4.515735 -3.228111 -7.816601  -0.9559795 -4.683854
#>        [,15]      [,16]     [,17]
#> 1  -4.284050  -5.800907 -2.837816
#> 2  -4.978985 -16.598032 -2.328746
#> 3  -4.860674 -16.597660 -2.672957
#> 4  -4.439843  -5.885082 -3.697461
#> 5  -4.532971  -5.826627 -3.114981
#> 6  -4.966234 -16.168603 -2.405253
#> 7  -4.969402 -16.504603 -2.245948
#> 8  -4.605493  -5.802408 -2.613595
#> 9  -4.976677 -16.570244 -2.312668
#> 10 -4.967130 -16.562255 -2.171068
#> 11 -4.783551 -15.845008 -2.262002
#> 12 -2.713834  -7.806966 -3.295386
#> 13 -4.632228  -5.652154 -3.148795
#> 14 -4.447747  -5.801815 -2.469919
#> 15 -5.075361  -5.929672 -2.514504
#> 16 -4.828317  -5.874981 -2.777121
#> 17 -4.647934  -5.815376 -2.834413
#> 18 -4.635739  -5.811598 -2.649452
#> 19 -4.644465  -5.814298 -2.340679
#> 20 -4.044892  -5.851875 -2.439493
#> 21 -4.607303  -5.803140 -2.491459
#> 22 -3.726928 -13.409758 -2.513613
#> 23 -1.612652  -8.109925 -2.727067
#> 24 -4.794618 -15.959997 -2.233215
#> 25 -1.472284  -8.114385 -3.425606
#> 26 -1.139197  -8.120087 -2.203005
#> 27 -1.558306  -8.121823 -2.694409
#> 28 -1.324024  -8.105878 -2.928756
#> 29 -1.404432  -8.106277 -2.968624
#> 30 -1.953842  -7.895687 -4.735560
#> 31 -2.914293  -8.167693 -6.415940
#> 32 -5.062471 -16.608271 -2.452730
#> 33 -4.417160 -16.608272 -2.885142
#> 34 -4.649593 -16.608274 -2.307114
#> 35 -5.038814 -16.608271 -1.441397
#> 36 -5.064594 -16.608271 -1.491039
#> 37 -5.035545 -16.608271 -1.847708
#> 38 -4.981720 -16.608270 -1.963772
#> 39 -4.980567 -16.608270 -2.077080
#> 40 -4.579892  -5.795185 -2.892258
#> 41 -4.668501  -5.506983 -2.063354
#> 42 -5.024568 -16.608271 -1.715643
#> 43 -5.069882 -16.608271 -2.692209
#> 44 -4.621619  -5.807262 -3.100662
#> 45 -4.654836  -5.659666 -2.867089
#> 46 -4.366462  -5.554373 -2.165904
#> 47 -4.572447  -5.792864 -2.851308
#> 48 -3.238059  -8.116149 -1.997595
#> 49 -2.074687  -8.139554 -3.202492
predict(myPLN, trichoptera, "response")
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 
#>  4  1  1  4  4  1  1  4  1  1  1  2  4  4  4  4  4  4  4  4  4  1  2  1  2  2 
#> 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 
#>  2  2  2  2  2  1  1  1  1  1  1  1  1  4  4  1  1  4  4  4  4  2  2 
#> Levels: 1 2 4