Prediction for a PLNmixturefit object

Predict either posterior probabilities for each group or latent positions based on new samples

# S3 method for PLNmixturefit
predict(
  object,
  newdata,
  type = c("posterior", "response", "position"),
  prior = matrix(rep(1/object$k, object$k), nrow(newdata), object$k, byrow = TRUE),
  control = list(),
  ...
)

Arguments

object
an R6 object with class PLNmixturefit

newdata
A data frame in which to look for variables, offsets and counts with which to predict.

type
The type of prediction required. The default posterior are posterior probabilities for each group , response is the group with maximal posterior probability and latent is the averaged latent in the latent space, with weights equal to the posterior probabilities.

prior
User-specified prior group probabilities in the new data. The default uses a uniform prior.

control
a list for controlling the optimization. See PLN() for details.

...
additional parameters for S3 compatibility. Not used

Value

A matrix of posterior probabilities for each group (if type = "posterior"), a matrix of (average) position in the latent space (if type = "position") or a vector of predicted groups (if type = "response").

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
fits <- PLNmixture(Abundance ~ 1, clusters = 1:4, data = trichoptera)

#> 
#>  Initialization...
#> 
#>  Adjusting 4 PLN mixture models.
#> 	number of cluster = 1 	number of cluster = 2 	number of cluster = 3 	number of cluster = 4 
#> 
#>  Smoothing PLN mixture models.
#>    Going backward +++                                                                                                       Going forward +++                                                                                                       Going backward +++                                                                                                       Going forward +++                                                                                                    
#>  Post-treatments
#>  DONE!

myMixture <- getBestModel(fits)
predict(myMixture, trichoptera, "posterior")

#>            [,1]         [,2]         [,3]         [,4]
#> 1  2.220446e-16 2.220446e-16 1.000000e+00 3.177591e-15
#> 2  2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 3  2.220446e-16 2.220446e-16 9.999998e-01 1.877628e-07
#> 4  1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 5  3.824298e-09 2.220446e-16 1.000000e+00 2.220446e-16
#> 6  2.220446e-16 2.220446e-16 1.000000e+00 1.231162e-13
#> 7  2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 8  2.220446e-16 2.220446e-16 9.999530e-01 4.703871e-05
#> 9  2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 10 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 11 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 12 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 13 2.773541e-14 2.220446e-16 9.999995e-01 4.702169e-07
#> 14 2.220446e-16 2.220446e-16 9.999959e-01 4.050208e-06
#> 15 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 16 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 17 2.722600e-01 2.220446e-16 7.277400e-01 2.148188e-12
#> 18 6.932869e-01 2.220446e-16 3.067131e-01 4.689487e-08
#> 19 1.000000e+00 2.220446e-16 1.030388e-08 1.521363e-13
#> 20 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 21 2.220446e-16 2.220446e-16 9.996327e-01 3.672560e-04
#> 22 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 23 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 24 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 25 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 26 1.000000e+00 2.220446e-16 2.725291e-14 2.220446e-16
#> 27 1.000000e+00 2.220446e-16 7.141450e-16 2.220446e-16
#> 28 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 29 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 30 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 31 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 32 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 33 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 34 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 35 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 36 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 37 1.000000e+00 2.220446e-16 2.220446e-16 3.987530e-13
#> 38 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 39 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 40 2.220446e-16 2.220446e-16 1.000000e+00 8.565707e-13
#> 41 9.638287e-06 2.220446e-16 2.220446e-16 9.999904e-01
#> 42 1.273711e-01 2.220446e-16 8.726289e-01 7.907362e-13
#> 43 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 44 2.220446e-16 2.220446e-16 9.998967e-01 1.032682e-04
#> 45 9.347583e-01 2.220446e-16 6.524161e-02 1.175668e-07
#> 46 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 47 2.220446e-16 2.220446e-16 1.000000e+00 3.597739e-15
#> 48 2.220446e-16 2.220446e-16 1.000000e+00 2.970773e-11
#> 49 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16

predict(myMixture, trichoptera, "position")

#>          [,1]       [,2]      [,3]      [,4]     [,5]       [,6]       [,7]
#> 1   -2.597035 -14.639094 0.9500722 -2.597097 2.634334 -1.2598618 -1.9311007
#> 2   -2.597073 -14.639094 0.7417714 -2.597135 2.081433 -1.2600064 -1.9311746
#> 3   -2.596968 -14.639093 0.5327241 -2.597027 3.198492 -1.2595956 -1.9309653
#> 4   -2.312640  -1.659396 1.2458744 -9.677088 5.120868  1.3586922 -1.6578922
#> 5   -2.591027 -14.639094 0.7302305 -2.591088 3.966456 -1.1458000 -1.9193740
#> 6   -2.597052 -14.639094 0.6370483 -2.597114 2.469285 -1.1426315 -1.9311336
#> 7   -2.597086 -14.639094 0.6362408 -2.597148 1.780363 -1.2600555 -1.9311997
#> 8   -2.597537 -14.638933 0.5324258 -2.596980 3.114363 -1.1422410 -1.9310258
#> 9   -2.597073 -14.639094 0.8459478 -2.597134 2.081482 -1.2600042 -1.9311735
#> 10  -2.597092 -14.639094 0.6360903 -2.597154 1.617995 -1.2600797 -1.9312120
#> 11  -2.597091 -14.639094 0.5299438 -2.597152 1.700028 -1.2600730 -1.9312086
#> 12  -2.597092 -14.639094 0.4228134 -2.597153 1.699970 -1.2600762 -1.9312102
#> 13  -2.477927 -14.639092 0.5337083 -2.596982 3.381967 -1.0249469 -1.9308795
#> 14  -2.597070 -14.639080 0.8473355 -2.597078 2.779759 -1.0253881 -1.8126551
#> 15  -2.311972  -1.658121 1.2595357 -9.677087 6.191858  2.7624331 -1.6566161
#> 16  -2.312752  -1.659610 1.0642362 -9.677088 4.898394  2.0068655 -1.6581067
#> 17  -2.502881 -11.086542 0.8857275 -4.517576 2.727627 -0.5042967 -1.8240800
#> 18  -2.372395  -5.592825 1.2559117 -7.502546 3.482369  0.6057338 -1.6881384
#> 19  -2.277549  -1.591137 1.5278508 -9.677066 3.977413  1.4176361 -1.5895279
#> 20  -2.313163  -1.341103 1.8045270 -9.677088 4.204288  1.5128880 -1.3396240
#> 21  -2.601359 -14.637837 0.9506165 -2.596588 2.780443 -1.2589471 -1.9313069
#> 22  -2.597081 -14.639094 0.7415751 -2.597142 1.858810 -1.2600359 -1.9311896
#> 23  -2.597073 -14.639094 0.4231942 -2.478146 2.081447 -1.2600059 -1.9311743
#> 24  -2.597093 -14.639094 0.5298841 -2.597155 1.617937 -1.2600833 -1.9312138
#> 25  -2.597063 -14.639094 0.7420462 -2.597124 2.219805 -1.1426742 -1.8127364
#> 26  -2.314659  -1.663252 0.8354469 -9.677089 2.415459  0.5730244 -1.6617547
#> 27  -2.314407  -1.662768 0.8406278 -9.677089 2.732119  0.5773281 -1.6612694
#> 28  -2.597092 -14.639094 0.4227988 -2.597154 1.618009 -1.2600790 -1.9312116
#> 29  -2.597087 -14.639094 0.4229099 -2.597148 1.780310 -1.2600584 -1.9312011
#> 30  -2.311733  -1.337701 0.6730424 -9.677087 7.056329  1.0364560 -0.7089493
#> 31  -7.693206  -0.696493 1.7892731 -7.693206 7.846205  0.4021840  1.2503098
#> 32  -2.313595  -1.661218 1.7942466 -9.677088 3.459362  1.3458701 -1.6597173
#> 33  -2.313144  -1.660357 2.2848921 -9.677088 4.071562  1.8914799 -1.6588546
#> 34  -2.312295  -1.658738 2.5291660 -9.677087 5.352703  2.8057526 -1.6572334
#> 35  -2.303482  -1.641969 1.6577663 -9.677082 3.336218  1.6224223 -1.6404397
#> 36  -2.279041  -1.594099 1.5123318 -9.677067 4.034538  1.3936472 -1.5924954
#> 37  -2.304275  -1.643496 2.6640849 -9.431219 3.056059  1.0400430 -1.6419698
#> 38  -2.597084 -14.639094 0.8456315 -2.597146 1.700451 -1.2600489 -1.9311963
#> 39  -2.597080 -14.639094 0.9487686 -2.597141 1.858916 -1.2600305 -1.9311869
#> 40  -2.597036 -14.639094 0.5311544 -2.597098 2.684265 -1.2598649 -1.9311023
#> 41 -14.420896 -11.215654 1.5043852 -1.263490 3.858187  1.0334675 -2.5941367
#> 42  -2.547786 -12.977108 0.7583372 -3.490348 2.467727 -0.8862997 -1.8708616
#> 43  -2.284082  -1.604080 1.4228311 -9.677070 4.199982  1.2272895 -1.6024917
#> 44  -2.598202 -14.638740 1.0536115 -2.477939 3.084272 -1.1421319 -1.6943616
#> 45  -2.297558  -2.442014 1.4682297 -9.214518 3.915427  1.2423723 -1.6101714
#> 46  -1.986245  -1.659976 1.3990264 -9.677088 4.572416  0.8144484 -1.3391009
#> 47  -2.597063 -14.639094 0.6367779 -2.597125 2.285135 -1.2599693 -1.9311556
#> 48  -2.478046 -14.639094 0.4239297 -2.597099 2.634079 -1.2598703 -1.9311051
#> 49  -2.313620  -1.342185 1.0480862 -9.677088 3.282881  0.5889354 -1.6597661
#>            [,8]        [,9]       [,10]      [,11]      [,12]       [,13]
#> 1  -1.697032198  -3.2845144 -0.06080508 -1.8376320 -0.8336886 -1.70384390
#> 2  -1.697125567  -3.2845335 -0.28773762 -1.8377131 -0.7177134 -1.70393664
#> 3  -1.696860702  -3.2844793 -0.28664789 -1.8374826 -0.4854485 -1.70367333
#> 4  -1.303290747 -15.8989116  0.65380653  0.4999837  0.9391429  0.91542951
#> 5  -1.682190838  -3.2814973 -0.22524062 -1.8247492 -0.8731924 -1.68910399
#> 6  -1.697073719  -3.2845229 -0.28752458 -1.8376680 -0.8337837 -1.70388514
#> 7  -1.697157261  -3.2845399 -0.06140445 -1.8377406 -0.9504191 -1.70396812
#> 8  -1.696877711  -3.2844571 -0.28673931 -1.8374099 -0.9498357 -1.70362834
#> 9  -1.697124164  -3.2845332 -0.28773186 -1.8377118 -0.9503492 -1.70393524
#> 10 -1.697172891  -3.2845431 -0.28793196 -1.8377542 -0.9504521 -1.70398364
#> 11 -1.697168516  -3.2845422 -0.28791399 -1.8377504 -0.9504429 -1.70397929
#> 12 -1.697170638  -3.2845427 -0.28792271 -1.8377522 -0.9504474 -1.70398140
#> 13 -1.578616951  -3.2844569 -0.05946772 -1.8373877 -0.8330493 -1.70356494
#> 14 -1.696998387  -3.2845053 -0.17368920 -1.8375950 -0.9500840 -1.70380497
#> 15 -0.679617626 -15.8989116  2.60471842  1.4766364  0.7356916  0.92724805
#> 16 -1.303594362 -15.8989116  1.05309858  1.1232460  0.7251945  0.91351465
#> 17 -1.545218029  -6.7153372  0.15955306 -0.9905064 -0.4378800 -0.65355193
#> 18 -1.347897949 -12.0284063  0.69116279  0.2869885  0.2741804  0.93342700
#> 19 -1.204771681 -15.8989114  1.07630604  1.2204405  0.7883581  2.07297205
#> 20 -0.991862860 -15.8989116  1.40382562  2.6476994  0.2485723  0.90623308
#> 21 -1.696826492  -3.2842712  0.27525554 -1.7185512 -0.9497790 -1.70315470
#> 22 -1.697144609  -3.2845373 -0.06134375 -1.8377296 -0.9503924 -1.70395555
#> 23 -1.697125240  -3.2845334 -0.17424895 -1.8377128 -0.9503514 -1.70393631
#> 24 -1.697175170  -3.2845436 -0.28794131 -1.8377562 -0.9504569 -1.70398590
#> 25 -1.697099245  -3.2845281 -0.28762950 -1.6011955 -0.9502965 -1.70391049
#> 26 -1.308764622 -15.8989116  0.62344997  0.8975504  0.4713320  1.39511448
#> 27 -1.308072943 -15.8989116  1.37707724  1.0921896  0.2344345  1.89791573
#> 28 -1.579071452  -3.2845430 -0.28792989 -1.8377537 -0.9504511 -1.70398314
#> 29 -1.697159127  -3.2845403 -0.28787545 -1.7194445 -0.9504230 -1.70396997
#> 30 -0.987064400 -15.2294024  1.95892929  4.1644867  0.2649900  3.20765555
#> 31 -0.002789687   0.9138627  2.07546952  4.0769189 -0.6956053  3.03293692
#> 32 -1.305874505 -15.8989116  0.41073611  0.4869867  0.9233199  3.36615240
#> 33 -1.304652752 -15.8989116  1.04894338  0.7169246  2.3892928  2.99645663
#> 34 -1.302357531 -15.8989116  0.87455750  0.5047234  1.3196875  3.18877875
#> 35 -1.278648498 -15.8989116  1.04908196  0.6057932  0.7228642  1.44912992
#> 36 -1.209129277 -15.8989116  1.03054752  1.1627303  0.7567286  1.99206064
#> 37 -1.280832009 -15.8989116  0.87767893  0.7745847  0.7208830  1.69643458
#> 38 -1.697152962  -3.2845391  0.16294948 -1.8377369 -0.9504100 -1.58585110
#> 39 -1.697141126  -3.2845366 -0.17431915 -1.8377266 -0.9503850 -1.58583841
#> 40 -1.697034227  -3.2845148 -0.28736221 -1.8376337 -0.8336933 -1.70384592
#> 41 -1.265125107  -2.6487912  1.04017066  0.4053678 -0.1973880  0.04592196
#> 42 -1.613115272  -4.8869452 -0.02341266 -1.4301501 -0.6829233 -1.19968172
#> 43 -1.223768609 -15.8989116  1.52250180  1.0711721  0.6557979  1.80890888
#> 44 -1.578761014  -3.2844216  0.05243491 -1.8372850 -0.8333367 -1.70353129
#> 45 -1.234728217 -15.0756065  0.99607320  1.0196656  0.6825750  1.84363452
#> 46 -1.304114744 -15.8989116  0.86381551  2.0587624  0.7222798  0.91008471
#> 47 -1.578994817  -3.2845286 -0.17414429 -1.7193904 -0.9503015 -1.70391285
#> 48 -1.578925468  -3.1652251 -0.06083146 -1.8376367 -0.9501665 -1.34986745
#> 49 -1.305944070 -15.8989116  0.63942321  1.2816857  0.2434070  3.45995641
#>           [,14]       [,15]        [,16]       [,17]
#> 1   -1.58176452 -0.07449097 -3.284501363 -0.06091344
#> 2   -1.69997454 -0.30159211 -3.284520432 -0.17412816
#> 3   -1.69971229 -0.18686185 -3.284466046 -0.17290626
#> 4   -0.59674597  1.27422838 -1.304457685  1.37820463
#> 5   -1.68508259 -0.19496660 -3.281484170 -0.01409262
#> 6   -1.69992283 -0.30138201 -3.284509842 -0.17388935
#> 7   -1.70000614 -0.30172048 -3.284526906 -0.17427405
#> 8   -1.70019526 -0.30071802 -3.284392464  0.16383061
#> 9   -1.69997314 -0.30158643 -3.284520145 -0.17412170
#> 10  -1.70002173 -0.30178377 -3.284530099 -0.17434598
#> 11  -1.70001736 -0.30176605 -3.284529205 -0.17432585
#> 12  -1.70001948 -0.30177465 -3.284529639 -0.17433561
#> 13  -1.69960666 -0.30007624 -3.165148424 -0.17240293
#> 14  -1.69988800 -0.18747670 -3.284487798  0.27421287
#> 15  -0.59316810  1.11036191 -0.989015724  3.82095345
#> 16  -0.59734479  0.69473385 -1.304760955  2.34844817
#> 17  -1.32515990  0.29102458 -2.714887811  0.60344591
#> 18  -0.78302606  1.04972906 -1.841611436  1.62539844
#> 19  -0.38948591  1.59223800 -1.206057689  2.36161672
#> 20  -0.30795281  1.55599881 -1.305873686  2.41467259
#> 21  -1.70332918 -0.30140454 -3.283855399  0.27474490
#> 22  -1.69999352 -0.07502228 -3.284524322 -0.06147206
#> 23  -1.58186513  0.14962438 -3.284520365  0.05083067
#> 24  -1.70002400 -0.30179299 -3.284530565 -0.17435646
#> 25  -1.46385105  0.48185514 -3.284515056 -0.06124671
#> 26   0.23002795  2.40506746 -1.309925307  1.84074815
#> 27  -0.60618815  2.09670919 -1.309234434  1.61577087
#> 28  -1.58191604  0.03739196 -3.284529996 -0.06160987
#> 29  -1.70000800  0.03745911 -3.284527287 -0.06154423
#> 30   4.08125286  5.38354414 -0.988210198  2.53207954
#> 31   4.08852460  5.22600001 -0.002789687  2.00328193
#> 32  -0.31071712  0.68170464 -1.307038503  1.96706932
#> 33  -0.59942971  1.55629462 -1.305818142  1.86969166
#> 34  -0.59490111  1.56960636 -1.303525534  3.42057368
#> 35  -0.54875836  0.80528720 -1.279843210  2.71214384
#> 36  -0.39931173  1.52264847 -1.210409678  2.51011379
#> 37  -0.55317886  0.79205366 -1.282024168  2.32691385
#> 38  -1.70000185 -0.30170307 -3.284526028  0.16228848
#> 39  -1.69999005 -0.30165514 -3.284523610  0.05074609
#> 40  -1.69988346 -0.30122189 -3.284501778 -0.17370733
#> 41 -11.21564400 -1.33925956 -1.551710272  2.00291575
#> 42  -1.51172568  0.02992880 -3.015410459  0.25177324
#> 43  -0.43197502  1.31499499 -1.225030387  2.07023342
#> 44  -1.70073165 -0.30078205 -3.284295302 -0.17286084
#> 45  -0.47209477  1.48488365 -1.340760289  2.21154846
#> 46  -0.02438311  1.26799788 -0.684749748  2.79277470
#> 47  -1.69995066 -0.30149509 -3.284515541 -0.17401788
#> 48  -1.69988692 -0.30123602 -3.284502487  0.49387942
#> 49  -0.60198339  2.20370917 -1.307107987  1.62855691

predict(myMixture, trichoptera, "response")

#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 
#>  3  3  3  1  3  3  3  3  3  3  3  3  3  3  1  1  3  1  1  1  3  3  3  3  3  1 
#> 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 
#>  1  3  3  1  2  1  1  1  1  1  1  3  3  3  4  3  1  3  1  1  3  3  1 
#> Levels: 1 2 3 4