Prediction for a PLNmixturefit object

Predict either posterior probabilities for each group or latent positions based on new samples

# S3 method for PLNmixturefit
predict(
  object,
  newdata,
  type = c("posterior", "response", "position"),
  prior = matrix(rep(1/object$k, object$k), nrow(newdata), object$k, byrow = TRUE),
  control = PLNmixture_param(),
  ...
)

Arguments

object

an R6 object with class PLNmixturefit

newdata

A data frame in which to look for variables, offsets and counts with which to predict.

type

The type of prediction required. The default posterior are posterior probabilities for each group , response is the group with maximal posterior probability and latent is the averaged latent in the latent space, with weights equal to the posterior probabilities.

prior

User-specified prior group probabilities in the new data. The default uses a uniform prior.

control

a list-like structure for controlling the fit. See PLNmixture_param() for details.

...

additional parameters for S3 compatibility. Not used

Value

A matrix of posterior probabilities for each group (if type = "posterior"), a matrix of (average) position in the latent space (if type = "position") or a vector of predicted groups (if type = "response").

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLNmixture(Abundance ~ 1 + offset(log(Offset)),
           data = trichoptera, control = PLNmixture_param(smoothing = "none"))  %>% getBestModel()
#> 
#>  Initialization...
#> 
#>  Adjusting 5 PLN mixture models.
#> 	number of cluster = 1 	number of cluster = 2 	number of cluster = 3 	number of cluster = 4 	number of cluster = 5 
#>  Post-treatments
#>  DONE!
predict(myPLN, trichoptera, "posterior")
#>            [,1]         [,2]         [,3]         [,4]         [,5]
#> 1  2.220446e-16 1.930004e-03 6.388136e-02 9.341886e-01 2.220446e-16
#> 2  2.220446e-16 1.848068e-05 9.030065e-01 7.480969e-02 2.216534e-02
#> 3  2.220446e-16 2.220446e-16 2.082227e-01 5.786095e-04 7.911987e-01
#> 4  2.220446e-16 2.220446e-16 3.338341e-05 2.220446e-16 9.999666e-01
#> 5  2.220446e-16 2.220446e-16 8.003573e-03 3.762691e-04 9.916202e-01
#> 6  2.220446e-16 1.063942e-05 4.765951e-01 1.690796e-01 3.543147e-01
#> 7  2.220446e-16 2.704537e-03 1.215082e-01 8.718804e-01 3.906840e-03
#> 8  2.220446e-16 1.489087e-06 2.454425e-02 1.908656e-02 9.563677e-01
#> 9  2.220446e-16 1.663717e-04 4.240073e-01 5.637123e-01 1.211394e-02
#> 10 1.414095e-14 8.285968e-03 3.810689e-01 5.881683e-01 2.247678e-02
#> 11 3.309018e-03 1.922542e-02 2.975348e-01 5.633069e-01 1.166239e-01
#> 12 8.610894e-02 4.330951e-02 1.946605e-01 3.451421e-01 3.307789e-01
#> 13 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 14 2.220446e-16 3.688121e-06 2.220446e-16 9.999963e-01 2.220446e-16
#> 15 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 16 2.220446e-16 2.220446e-16 2.104663e-05 2.220446e-16 9.999790e-01
#> 17 2.220446e-16 4.037879e-06 2.220446e-16 9.999960e-01 2.220446e-16
#> 18 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 19 2.220446e-16 2.220446e-16 1.232618e-04 9.998138e-01 6.295580e-05
#> 20 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 21 2.220446e-16 4.354306e-06 5.531362e-03 9.943815e-01 8.275676e-05
#> 22 2.220446e-16 6.858450e-02 5.956009e-02 8.714913e-01 3.641434e-04
#> 23 2.220446e-16 9.981433e-01 4.424426e-13 1.856715e-03 2.220446e-16
#> 24 2.518061e-03 2.773055e-02 3.141326e-01 5.679867e-01 8.763207e-02
#> 25 2.220446e-16 9.990219e-01 2.220446e-16 9.780917e-04 2.220446e-16
#> 26 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 27 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 28 5.961238e-04 9.985165e-01 2.220446e-16 8.873662e-04 2.220446e-16
#> 29 2.019238e-02 9.704544e-01 1.699138e-04 9.034309e-03 1.490181e-04
#> 30 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 31 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 32 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 33 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 34 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 35 2.220446e-16 2.220446e-16 9.999344e-01 6.555270e-05 2.220446e-16
#> 36 2.220446e-16 2.220446e-16 8.760472e-01 4.484134e-04 1.235043e-01
#> 37 2.220446e-16 2.220446e-16 9.999609e-01 3.910285e-05 2.220446e-16
#> 38 2.220446e-16 2.346605e-03 6.552176e-01 3.424358e-01 2.220446e-16
#> 39 2.220446e-16 5.382949e-04 8.454057e-01 1.540353e-01 2.066900e-05
#> 40 3.188151e-04 4.127334e-05 1.546951e-01 4.838753e-02 7.965573e-01
#> 41 2.220446e-16 2.220446e-16 3.528479e-01 5.265090e-01 1.206431e-01
#> 42 2.220446e-16 4.365236e-03 2.220446e-16 9.956348e-01 2.220446e-16
#> 43 2.220446e-16 3.486338e-06 8.830486e-01 1.146472e-01 2.300669e-03
#> 44 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 45 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 46 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 47 2.220446e-16 6.827417e-04 6.118319e-15 9.961082e-01 3.209085e-03
#> 48 2.220446e-16 9.787654e-01 2.220446e-16 2.123457e-02 2.220446e-16
#> 49 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
predict(myPLN, trichoptera, "position")
#>          [,1]       [,2]      [,3]       [,4]        [,5]      [,6]       [,7]
#> 1   -6.985772  -7.630229 -2.129649  -5.502742 -0.55742838 -3.691834  -5.969395
#> 2  -18.816719 -18.868541 -2.011512  -7.127479 -0.53849080 -3.550453 -18.734473
#> 3  -20.080111 -20.080509 -3.726329 -17.398975 -0.19283204 -3.810043 -20.079485
#> 4  -20.154199 -20.154199 -4.092181 -20.153772 -0.09232115 -3.767971 -20.154199
#> 5  -20.146396 -20.146652 -3.591584 -20.043241 -0.14549181 -3.653498 -20.145999
#> 6  -17.627717 -17.744685 -2.797384 -11.367328 -0.31024972 -3.398246 -17.442445
#> 7   -7.824570  -8.429451 -2.180076  -5.635347 -0.63891349 -3.594120  -6.862796
#> 8  -19.878384 -19.891533 -3.878272 -19.548893 -0.17916821 -3.603534 -19.857685
#> 9  -12.096827 -12.487429 -1.892837  -6.230927 -0.48787441 -3.582935 -11.476666
#> 10 -11.647349 -12.056098 -2.060039  -6.313986 -0.63153491 -3.542647 -10.996755
#> 11 -11.853582 -12.217973 -2.450766  -7.614649 -0.46620217 -3.595665 -11.196412
#> 12 -14.326271 -13.851271 -3.330254 -11.569147 -0.30371233 -3.979072 -13.058270
#> 13  -5.965030  -6.803129 -2.722031  -5.417306 -0.26339783 -3.500865  -5.042514
#> 14  -6.111126  -6.800174 -2.281343  -5.405758 -0.52990589 -3.433389  -4.876850
#> 15 -20.154210 -20.154210 -4.673928 -20.154210 -0.18343109 -3.498414 -20.154210
#> 16 -20.154203 -20.154203 -4.184244 -20.153933 -0.19515250 -3.203008 -20.154203
#> 17  -6.120189  -6.804777 -2.684798  -5.423682 -0.32215308 -3.658427  -5.051601
#> 18  -6.117670  -6.803497 -2.540125  -5.269890 -0.51988204 -3.266803  -5.044548
#> 19  -6.122046  -6.806862 -2.315134  -5.423390 -0.62420301 -3.064859  -5.052368
#> 20  -6.144337  -6.666162 -2.671917  -5.469946 -0.58570440 -3.448111  -4.979618
#> 21  -6.188940  -6.873760 -2.204330  -5.417036 -0.57422798 -3.709276  -5.111203
#> 22  -6.896787  -7.503203 -2.203610  -5.507283 -0.78812070 -3.759464  -5.938444
#> 23  -5.669420  -5.710220 -3.794654  -5.456118 -0.89979105 -5.661607  -5.664146
#> 24 -11.661675 -12.036343 -2.388914  -7.204976 -0.52956951 -3.598363 -11.007007
#> 25  -5.676626  -5.716563 -3.316716  -5.664460 -1.03863724 -5.472191  -5.473490
#> 26  -5.685891  -5.724830 -3.722219  -5.674514 -1.36927314 -5.682701  -5.682701
#> 27  -5.688832  -5.727663 -3.739351  -5.677487 -1.14315779 -5.685651  -5.685651
#> 28  -5.668704  -5.704156 -3.754058  -5.656436 -1.18607277 -5.657599  -5.658332
#> 29  -5.896135  -5.773045 -3.796470  -5.876193 -0.96686182 -5.674627  -5.675320
#> 30 -16.812722  -8.370839 -6.655526 -16.812722 -0.30985256 -7.223395  -6.458305
#> 31 -16.812737  -8.613470 -5.752187 -16.812737 -0.10985394 -7.524125  -6.590181
#> 32 -19.837365 -19.837365 -2.523024  -6.984970 -0.96480380 -3.246781 -19.837365
#> 33 -19.837365 -19.837365 -2.437125  -7.004629 -0.77737849 -2.998858 -19.837365
#> 34 -19.837366 -19.837366 -3.004850  -7.077046 -0.40048526 -2.880044 -19.837366
#> 35 -19.836466 -19.836511 -2.425858  -6.975352 -0.81661530 -2.759478 -19.836397
#> 36 -19.870351 -19.870656 -2.981855  -8.611487 -0.41714457 -3.447203 -19.869881
#> 37 -19.836829 -19.836856 -1.419575  -6.588205 -1.06154078 -3.450900 -19.836787
#> 38 -15.101486 -15.338662 -1.903375  -6.415257 -1.15273640 -3.591953 -14.725504
#> 39 -17.714339 -17.821034 -1.646127  -6.711297 -0.90534947 -3.565048 -17.545103
#> 40 -19.423791 -19.454596 -3.605644 -17.396564 -0.15330966 -3.710199 -19.367561
#> 41 -12.655632 -13.014973 -2.750583  -7.753380 -0.38003350 -3.886567 -12.096992
#> 42  -6.116545  -6.799194 -2.559224  -5.272477 -0.49911351 -3.777357  -5.049391
#> 43 -18.266911 -18.344667 -3.033373  -6.841808 -0.30869467 -3.845759 -18.146987
#> 44  -6.114774  -6.802025 -2.196390  -5.263287 -0.46593410 -3.605221  -4.742460
#> 45  -6.121624  -6.805514 -2.445352  -5.278822 -0.53271726 -3.431473  -4.910900
#> 46  -6.005225  -6.821166 -2.964707  -5.483823 -0.38493146 -3.902738  -5.000907
#> 47  -6.149651  -6.839142 -2.310095  -5.440718 -0.46314595 -3.638692  -5.056703
#> 48  -5.491047  -5.741356 -3.832642  -5.662160 -0.72479478 -5.633924  -5.662146
#> 49  -5.718647  -5.565651 -3.747462  -5.707619 -1.15693741 -5.715554  -5.715554
#>          [,8]       [,9]     [,10]     [,11]     [,12]     [,13]      [,14]
#> 1   -5.690847  -7.618026 -2.875823 -3.465071 -4.790892 -5.247929  -5.371444
#> 2  -18.411711 -18.867566 -3.832592 -5.626710 -3.167028 -3.276450  -7.141722
#> 3   -9.373631 -20.080502 -4.487512 -5.396955 -3.535970 -5.810308 -17.401544
#> 4   -6.674328 -20.154199 -4.707808 -5.439645 -4.339823 -6.519715 -20.153772
#> 5   -6.742233 -20.146647 -4.685406 -5.327960 -4.418592 -6.454275 -20.043352
#> 6  -12.594893 -17.742485 -3.993635 -5.192305 -3.947730 -4.675339 -11.380202
#> 7   -6.542461  -8.417981 -2.717673 -3.510442 -4.835397 -5.073589  -5.668001
#> 8   -6.910142 -19.891285 -4.548958 -5.246447 -4.505000 -6.361387 -19.549974
#> 9  -11.141743 -12.480082 -3.374839 -4.345232 -4.432882 -4.384910  -6.259343
#> 10 -10.506363 -12.048096 -3.287637 -4.218101 -4.449294 -4.406019  -6.326991
#> 11  -9.434976 -12.204515 -3.378208 -4.192643 -4.533209 -4.682681  -7.560308
#> 12  -8.551790 -13.705992 -3.862111 -4.399787 -4.862344 -5.036131 -10.364151
#> 13  -4.608432  -6.790196 -2.970973 -3.395381 -4.900334 -5.404914  -5.457916
#> 14  -4.730591  -6.787202 -2.988488 -3.321734 -5.030438 -5.393216  -5.446808
#> 15  -6.396800 -20.154210 -3.786611 -5.027467 -4.976612 -6.647968 -20.154210
#> 16  -6.667892 -20.154203 -4.354234 -4.858238 -4.452307 -6.512505 -20.153933
#> 17  -4.764342  -6.638788 -2.815464 -3.091249 -5.056001 -5.411355  -5.464042
#> 18  -4.755125  -6.790568 -3.093708 -2.635311 -5.048977 -5.406355  -5.459286
#> 19  -4.763701  -6.793947 -3.010860 -2.868982 -5.053840 -5.409769  -5.463805
#> 20  -4.715233  -6.804502 -3.057357 -2.320331 -5.120312 -5.458152  -5.366730
#> 21  -4.817081  -6.860866 -2.545658 -3.217112 -5.026020 -5.383245  -5.457866
#> 22  -5.621024  -7.489133 -2.760214 -3.374076 -4.974971 -5.037763  -5.411495
#> 23  -4.956506  -5.670681 -3.286863 -3.313959 -5.674309 -2.515198  -3.555100
#> 24  -9.629365 -12.023759 -3.339639 -4.172138 -4.524095 -4.574867  -7.144522
#> 25  -4.971828  -5.677289 -3.525683 -3.050965 -5.682436 -2.633539  -3.434575
#> 26  -4.991009  -5.685891 -3.431353 -3.146443 -5.495790 -2.387107  -3.346059
#> 27  -4.996686  -5.688832 -2.873099 -3.029867 -5.695627 -2.021957  -3.853267
#> 28  -4.743833  -5.663386 -3.410075 -3.246404 -5.669766 -2.375715  -3.499004
#> 29  -5.002850  -5.701696 -3.457294 -3.075325 -5.720907 -2.463254  -3.714506
#> 30  -7.754678  -6.843467 -5.378335 -3.153473 -8.608524 -4.190815  -3.180275
#> 31  -7.857307  -6.910221 -5.918703 -4.072087 -8.477077 -5.166626  -4.018456
#> 32 -19.837365 -19.837365 -4.300901 -5.929648 -3.494129 -1.126505  -6.615598
#> 33 -19.837365 -19.837365 -3.833781 -5.643259 -2.431954 -1.887426  -7.016383
#> 34 -19.837366 -19.837366 -4.511201 -6.143677 -4.016376 -2.530102  -7.088070
#> 35 -19.836378 -19.836510 -3.470684 -5.904058 -3.559444 -2.981977  -6.987434
#> 36 -18.200813 -19.870650 -4.357857 -5.857278 -4.217163 -3.664056  -8.621983
#> 37 -19.836776 -19.836855 -3.675906 -5.536928 -3.538907 -2.645664  -6.986175
#> 38 -14.620485 -15.334112 -2.583861 -4.937645 -4.152109 -3.606714  -6.433066
#> 39 -17.497892 -17.819008 -3.441542 -5.432775 -3.892594 -3.126709  -6.727114
#> 40  -8.570213 -19.453452 -4.386985 -5.245309 -4.154916 -5.887198 -17.396220
#> 41 -10.316529 -13.008189 -3.298344 -4.557690 -4.469720 -4.901127  -7.778724
#> 42  -4.473547  -6.786158 -3.105982 -3.412788 -5.053836 -5.100794  -5.453781
#> 43 -18.084157 -18.343197 -2.524915 -5.341998 -4.283675 -3.412956  -6.856865
#> 44  -4.599189  -6.789077 -2.823125 -3.368941 -4.893295 -5.400565  -5.453783
#> 45  -4.488015  -6.792611 -2.560619 -3.332836 -4.915242 -5.414205  -5.466749
#> 46  -4.870876  -6.808460 -3.379956 -2.895272 -4.873099 -5.472176  -5.243792
#> 47  -4.561256  -6.826153 -2.870800 -3.096998 -5.010922 -5.381950  -5.480799
#> 48  -4.780314  -5.505391 -3.209087 -3.402302 -5.672097 -2.336963  -3.835600
#> 49  -5.052585  -5.718647 -3.638810 -3.153531 -5.725252 -1.055798  -3.994288
#>        [,15]      [,16]     [,17]
#> 1  -4.037697  -5.517075 -2.712102
#> 2  -4.872952  -6.209587 -2.845498
#> 3  -4.944033  -7.063691 -3.248279
#> 4  -4.866212  -7.300839 -3.652679
#> 5  -5.083499  -7.272493 -3.213535
#> 6  -4.924952  -6.474141 -2.961720
#> 7  -4.342422  -5.566994 -2.648507
#> 8  -5.218071  -7.213189 -2.727871
#> 9  -4.557072  -5.798468 -2.746974
#> 10 -4.497503  -5.888945 -2.615655
#> 11 -4.501075  -6.154613 -2.653301
#> 12 -4.429034  -6.998442 -2.968016
#> 13 -4.354777  -5.302274 -2.955374
#> 14 -4.179365  -5.440400 -2.404747
#> 15 -5.405208  -7.153436 -2.544644
#> 16 -5.397476  -7.297392 -2.811853
#> 17 -4.371761  -5.457746 -2.704009
#> 18 -4.358603  -5.452897 -2.559224
#> 19 -4.368181  -5.456542 -2.331785
#> 20 -3.897375  -5.502541 -2.440995
#> 21 -4.330825  -5.446542 -2.423647
#> 22 -3.837321  -6.430209 -2.542841
#> 23 -1.574981 -19.318227 -2.323464
#> 24 -4.460806  -6.230465 -2.622192
#> 25 -1.493232 -19.330461 -2.599786
#> 26 -1.228467 -19.344064 -2.145859
#> 27 -1.551734 -19.344064 -2.404952
#> 28 -1.433454 -19.324833 -2.335774
#> 29 -1.500600 -18.981308 -2.418026
#> 30 -1.964651  -7.754678 -5.123511
#> 31 -2.907161  -7.857307 -6.411268
#> 32 -5.115198  -6.312684 -2.563314
#> 33 -3.979325  -6.348785 -2.983611
#> 34 -4.475730  -6.472973 -2.321873
#> 35 -5.064283  -6.294685 -1.539520
#> 36 -5.142927  -6.433541 -1.612440
#> 37 -5.056943  -6.292203 -1.928086
#> 38 -4.696000  -5.999111 -2.184693
#> 39 -4.817809  -6.129544 -2.342374
#> 40 -5.132944  -7.024814 -3.077268
#> 41 -4.741503  -5.509299 -2.104635
#> 42 -4.354400  -5.514904 -1.992907
#> 43 -5.051199  -6.224669 -2.835630
#> 44 -4.343119  -5.447361 -2.918564
#> 45 -4.379138  -5.312514 -2.729828
#> 46 -4.164752  -5.237933 -2.246297
#> 47 -4.288393  -5.443592 -2.714696
#> 48 -2.340145 -19.048794 -2.027290
#> 49 -1.949100 -19.344064 -2.737674
predict(myPLN, trichoptera, "response")
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 
#>  4  3  5  5  5  3  4  5  4  4  4  4  4  4  5  5  4  4  4  4  4  4  2  4  2  2 
#> 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 
#>  2  2  2  1  1  3  3  3  3  3  3  3  3  5  4  4  3  4  4  4  4  2  2 
#> Levels: 1 2 3 4 5