Prediction for a PLNmixturefit object

Predict either posterior probabilities for each group or latent positions based on new samples

# S3 method for PLNmixturefit
predict(
  object,
  newdata,
  type = c("posterior", "response", "position"),
  prior = matrix(rep(1/object$k, object$k), nrow(newdata), object$k, byrow = TRUE),
  control = PLNmixture_param(),
  ...
)

Arguments

object

an R6 object with class PLNmixturefit

newdata

A data frame in which to look for variables, offsets and counts with which to predict.

type

The type of prediction required. The default posterior are posterior probabilities for each group , response is the group with maximal posterior probability and latent is the averaged latent in the latent space, with weights equal to the posterior probabilities.

prior

User-specified prior group probabilities in the new data. The default uses a uniform prior.

control

a list-like structure for controlling the fit. See PLNmixture_param() for details.

...

additional parameters for S3 compatibility. Not used

Value

A matrix of posterior probabilities for each group (if type = "posterior"), a matrix of (average) position in the latent space (if type = "position") or a vector of predicted groups (if type = "response").

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLNmixture(Abundance ~ 1 + offset(log(Offset)),
           data = trichoptera, control = PLNmixture_param(smoothing = "none"))  %>% getBestModel()
#> 
#>  Initialization...
#> 
#>  Adjusting 5 PLN mixture models.
#> 	number of cluster = 1 	number of cluster = 2 	number of cluster = 3 	number of cluster = 4 	number of cluster = 5 
#>  Post-treatments
#>  DONE!
predict(myPLN, trichoptera, "posterior")
#>            [,1]         [,2]         [,3]         [,4]         [,5]
#> 1  2.075454e-12 2.220446e-16 1.000000e+00 2.220446e-16 6.424339e-12
#> 2  1.182492e-13 2.220446e-16 2.084102e-08 1.503029e-10 1.000000e+00
#> 3  2.220446e-16 2.220446e-16 9.659413e-12 1.000000e+00 6.230272e-13
#> 4  2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 5  2.220446e-16 2.220446e-16 2.560782e-13 1.000000e+00 4.137427e-16
#> 6  3.211941e-14 2.343805e-14 1.058280e-07 1.731135e-08 9.999999e-01
#> 7  4.930929e-11 1.436619e-14 1.000000e+00 2.962746e-11 1.174848e-11
#> 8  2.220446e-16 3.272443e-15 1.450345e-09 1.000000e+00 6.363972e-13
#> 9  1.663344e-12 7.561223e-15 3.088539e-07 1.417948e-10 9.999997e-01
#> 10 2.893677e-10 2.009802e-12 1.000000e+00 3.834231e-10 5.920676e-11
#> 11 1.716587e-09 3.377947e-10 5.475832e-07 7.362231e-09 9.999994e-01
#> 12 2.429586e-09 6.190759e-09 2.044824e-07 9.999998e-01 5.073781e-11
#> 13 2.220446e-16 2.220446e-16 1.000000e+00 1.211417e-15 2.220446e-16
#> 14 1.631856e-15 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 15 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 17 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 18 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 19 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 20 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 21 4.283113e-16 2.220446e-16 1.000000e+00 1.195147e-14 1.703986e-13
#> 22 1.526172e-09 4.043409e-16 1.000000e+00 5.461648e-13 3.156503e-12
#> 23 1.000000e+00 2.220446e-16 1.946094e-10 2.220446e-16 2.220446e-16
#> 24 1.219347e-09 1.096617e-10 1.000000e+00 2.316785e-09 6.283741e-11
#> 25 1.000000e+00 8.722912e-15 7.513817e-12 2.220446e-16 2.220446e-16
#> 26 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 27 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 28 1.000000e+00 3.962547e-12 9.039736e-11 2.220446e-16 2.220446e-16
#> 29 1.000000e+00 2.816704e-10 1.470165e-09 6.219415e-13 4.800307e-15
#> 30 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 31 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 32 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 33 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 34 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 35 2.220446e-16 2.220446e-16 3.378036e-14 2.220446e-16 1.000000e+00
#> 36 2.220446e-16 2.220446e-16 5.972804e-13 1.000000e+00 4.504597e-13
#> 37 2.220446e-16 2.220446e-16 6.886163e-16 2.220446e-16 1.000000e+00
#> 38 1.463553e-11 2.220446e-16 1.000000e+00 3.831061e-16 6.396380e-12
#> 39 1.539561e-11 2.220446e-16 1.000000e+00 9.747562e-14 2.592122e-10
#> 40 8.422060e-15 1.094083e-11 6.226495e-09 1.000000e+00 7.531773e-12
#> 41 2.220446e-16 2.220446e-16 1.124455e-07 2.135303e-10 9.999999e-01
#> 42 1.117565e-14 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 43 2.220446e-16 2.220446e-16 8.322292e-10 2.603341e-14 1.000000e+00
#> 44 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 45 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 46 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 47 3.272878e-12 7.772385e-14 1.000000e+00 4.164156e-11 2.220446e-16
#> 48 1.000000e+00 2.220446e-16 1.955147e-09 2.220446e-16 2.220446e-16
#> 49 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
predict(myPLN, trichoptera, "position")
#>          [,1]       [,2]      [,3]       [,4]        [,5]      [,6]       [,7]
#> 1   -6.110182  -6.799941 -2.134668  -5.403469 -0.55688210 -3.688139  -5.022274
#> 2  -18.529808 -18.529808 -1.955237  -6.949181 -0.54742536 -3.533558 -18.529808
#> 3  -24.358992 -24.358992 -3.955424 -24.358992 -0.18481066 -3.810601 -24.358992
#> 4  -24.358992 -24.358992 -4.092619 -24.358992 -0.09206127 -3.762713 -24.358992
#> 5  -24.358992 -24.358992 -3.617314 -24.358992 -0.14439183 -3.651533 -24.358992
#> 6  -18.529807 -18.529807 -2.324391  -6.951523 -0.33038008 -3.269938 -18.529807
#> 7   -6.102318  -6.795996 -2.189533  -5.387560 -0.63172469 -3.603598  -4.999308
#> 8  -24.358992 -24.358992 -3.939424 -24.358992 -0.17475406 -3.607124 -24.358992
#> 9  -18.529805 -18.529805 -1.700861  -6.948710 -0.48717876 -3.521847 -18.529804
#> 10  -6.100806  -6.795239 -2.120386  -5.384492 -0.60863656 -3.585673  -4.994765
#> 11 -18.529802 -18.529802 -2.153728  -6.944931 -0.47996959 -3.415940 -18.529801
#> 12 -24.358989 -24.358989 -4.013541 -24.358988 -0.13420844 -3.624724 -24.358988
#> 13  -5.964736  -6.803678 -2.723391  -5.417897 -0.26300789 -3.500699  -5.043207
#> 14  -6.111592  -6.800688 -2.280998  -5.406289 -0.52978149 -3.432884  -4.876673
#> 15 -24.358992 -24.358992 -4.659984 -24.358992 -0.18314176 -3.499379 -24.358992
#> 16 -24.358992 -24.358992 -4.181649 -24.358992 -0.19467198 -3.211196 -24.358992
#> 17  -6.120702  -6.805290 -2.685922  -5.424297 -0.32182502 -3.658945  -5.052326
#> 18  -6.118169  -6.804050 -2.540682  -5.269709 -0.51969406 -3.265683  -5.045251
#> 19  -6.119980  -6.804921 -2.314947  -5.422881 -0.62411787 -3.063113  -5.050312
#> 20  -6.144964  -6.665361 -2.672686  -5.470784 -0.58554898 -3.447837  -4.979959
#> 21  -6.112328  -6.801061 -2.204073  -5.407756 -0.57413278 -3.709864  -5.028650
#> 22  -6.104198  -6.796939 -2.138407  -5.391367 -0.76358723 -3.625102  -5.004842
#> 23  -5.668569  -5.708153 -3.796903  -5.456468 -0.90015464 -5.665348  -5.665348
#> 24  -6.100427  -6.795050 -2.249059  -5.383722 -0.53918255 -3.581082  -4.993627
#> 25  -5.676162  -5.715458 -3.317827  -5.664670 -1.03874196 -5.474062  -5.474062
#> 26  -5.685850  -5.724785 -3.722198  -5.674467 -1.36921006 -5.682683  -5.682683
#> 27  -5.688789  -5.727622 -3.739324  -5.677439 -1.14308950 -5.685632  -5.685632
#> 28  -5.661627  -5.701478 -3.753709  -5.649968 -1.18669416 -5.658383  -5.658383
#> 29  -5.662403  -5.702224 -3.758707  -5.650754 -0.97865587 -5.659162  -5.659162
#> 30 -16.848772  -8.370860 -6.654430 -16.848772 -0.30984634 -7.223777  -6.458345
#> 31 -16.848642  -8.612756 -5.752849 -16.848642 -0.10981096 -7.523556  -6.590010
#> 32 -18.529809 -18.529809 -2.518679  -6.983021 -0.96713075 -3.242054 -18.529809
#> 33 -18.529809 -18.529809 -2.434398  -7.002944 -0.77876832 -2.995067 -18.529809
#> 34 -18.529810 -18.529810 -3.003542  -7.076246 -0.40097135 -2.878169 -18.529810
#> 35 -18.529809 -18.529809 -2.420933  -6.973369 -0.81907259 -2.753273 -18.529809
#> 36 -24.358992 -24.358992 -3.661824 -24.358992 -0.39701586 -3.552160 -24.358992
#> 37 -18.529809 -18.529809 -1.415859  -6.582228 -1.06487072 -3.445892 -18.529809
#> 38  -6.104948  -6.797316 -2.048360  -5.392881 -0.99426464 -3.633440  -5.007057
#> 39  -6.104574  -6.797128 -1.913930  -5.392125 -0.80059692 -3.629289  -5.005951
#> 40 -24.358992 -24.358992 -3.878994 -24.358992 -0.13812577 -3.723802 -24.358992
#> 41 -18.529808 -18.529808 -2.717527  -6.975140 -0.37776097 -3.961394 -18.529808
#> 42  -6.118894  -6.804422 -2.555857  -5.271354 -0.49861973 -3.770004  -5.047276
#> 43 -18.529809 -18.529809 -3.057054  -6.986047 -0.30838174 -3.852748 -18.529809
#> 44  -6.115258  -6.802561 -2.195877  -5.263068 -0.46569135 -3.605531  -4.741611
#> 45  -6.122140  -6.806025 -2.445598  -5.278696 -0.53254126 -3.431040  -4.910917
#> 46  -6.004316  -6.821753 -2.966199  -5.484723 -0.38470713 -3.904111  -5.001367
#> 47  -6.105323  -6.797504 -2.304706  -5.393635 -0.46314327 -3.637556  -5.008160
#> 48  -5.480910  -5.718347 -3.857267  -5.667706 -0.72733793 -5.675976  -5.675976
#> 49  -5.718598  -5.565624 -3.747439  -5.707568 -1.15685808 -5.715531  -5.715531
#>          [,8]       [,9]     [,10]     [,11]     [,12]     [,13]      [,14]
#> 1   -4.725503  -6.786786 -2.839136 -3.303525 -4.876151 -5.390690  -5.292232
#> 2  -18.529808 -18.529808 -3.881874 -5.833595 -3.008714 -3.016967  -6.962054
#> 3   -6.615300 -24.358992 -4.583266 -5.264453 -3.746516 -6.460525 -24.358992
#> 4   -6.663363 -24.358992 -4.689943 -5.423000 -4.345786 -6.516044 -24.358992
#> 5   -6.628592 -24.358992 -4.669009 -5.311914 -4.427155 -6.475967 -24.358992
#> 6  -18.529807 -18.529807 -3.923274 -5.840582 -3.394980 -3.096962  -6.964367
#> 7   -4.694663  -6.782784 -2.652784 -3.181902 -5.003848 -5.374573  -5.429480
#> 8   -6.613982 -24.358992 -4.574036 -5.259557 -4.511855 -6.458991 -24.358992
#> 9  -18.529804 -18.529805 -3.873140 -5.832183 -3.661773 -2.999401  -6.961590
#> 10  -4.688579  -6.782017 -2.906680 -3.154894 -4.999354 -5.371465  -5.426551
#> 11 -18.529801 -18.529802 -3.796532 -5.820719 -3.568570 -2.831080  -6.957859
#> 12  -6.603637 -24.358989 -4.495689 -5.219742 -4.428573 -6.446916 -24.358988
#> 13  -4.608374  -6.790574 -2.971536 -3.397582 -4.900224 -5.405291  -5.458617
#> 14  -4.731157  -6.787544 -2.989160 -3.323723 -5.031002 -5.393540  -5.447460
#> 15  -6.399770 -24.358992 -3.793019 -5.028595 -4.963365 -6.638194 -24.358992
#> 16  -6.657394 -24.358992 -4.350485 -4.867584 -4.454489 -6.509196 -24.358992
#> 17  -4.765057  -6.638321 -2.815416 -3.092073 -5.056682 -5.411769  -5.464774
#> 18  -4.755805  -6.790952 -3.094756 -2.634449 -5.049635 -5.406740  -5.459993
#> 19  -4.762429  -6.791893 -3.011400 -2.868770 -5.054677 -5.410342  -5.463417
#> 20  -4.715702  -6.804895 -3.057982 -2.319137 -5.121282 -5.458788  -5.366963
#> 21  -4.733959  -6.787922 -2.543352 -3.205678 -5.033109 -5.395025  -5.448869
#> 22  -4.702189  -6.783741 -2.703599 -3.213675 -5.009407 -5.378429  -5.433132
#> 23  -4.956965  -5.668569 -3.287558 -3.314002 -5.675626 -2.509868  -3.551891
#> 24  -4.687033  -6.781825 -2.897770 -3.147895 -4.998223 -5.370685  -5.425810
#> 25  -4.972068  -5.676162 -3.526042 -3.050912 -5.683166 -2.630817  -3.432894
#> 26  -4.991010  -5.685850 -3.431296 -3.146374 -5.495897 -2.387094  -3.346097
#> 27  -4.996687  -5.688789 -2.873052 -3.029803 -5.695709 -2.021940  -3.853254
#> 28  -4.742303  -5.661627 -3.409062 -3.246116 -5.668732 -2.371724  -3.497437
#> 29  -4.944519  -5.662403 -3.415967 -3.066627 -5.669503 -2.389011  -3.696688
#> 30  -7.754605  -6.843484 -5.378608 -3.153921 -8.608318 -4.191009  -3.180111
#> 31  -7.856946  -6.910185 -5.918497 -4.071927 -8.477000 -5.165918  -4.018249
#> 32 -18.529809 -18.529809 -4.300928 -5.928268 -3.490288 -1.124052  -6.610318
#> 33 -18.529809 -18.529809 -3.831564 -5.639573 -2.428545 -1.885008  -7.015177
#> 34 -18.529810 -18.529810 -4.512129 -6.144420 -4.015608 -2.528696  -7.087711
#> 35 -18.529809 -18.529809 -3.466020 -5.902509 -3.555169 -2.977630  -6.985944
#> 36  -6.632093 -24.358992 -4.689729 -5.323887 -4.633544 -6.052042 -24.358992
#> 37 -18.529809 -18.529809 -3.672009 -5.531841 -3.534513 -2.640784  -6.984626
#> 38  -4.705168  -6.784123 -2.459894 -3.225839 -5.011610 -5.226195  -5.434585
#> 39  -4.703681  -6.783932 -2.849394 -3.219801 -5.010509 -5.225304  -5.433859
#> 40  -6.609339 -24.358992 -4.540267 -5.242002 -4.266069 -6.453577 -24.358992
#> 41 -18.529808 -18.529808 -3.493774 -5.907301 -3.812520 -3.587598  -6.987693
#> 42  -4.472109  -6.791329 -3.105344 -3.414557 -5.051654 -5.110512  -5.461365
#> 43 -18.529809 -18.529809 -2.526976 -5.582703 -4.177306 -3.196501  -6.998472
#> 44  -4.599081  -6.789433 -2.823086 -3.371090 -4.893153 -5.400919  -5.454464
#> 45  -4.487383  -6.793021 -2.559798 -3.334586 -4.915214 -5.414622  -5.467487
#> 46  -4.872049  -6.808889 -3.381596 -2.895119 -4.872906 -5.472878  -5.243388
#> 47  -4.554923  -6.784314 -2.866013 -3.091812 -5.012709 -5.380726  -5.435310
#> 48  -4.784674  -5.480910 -3.216726 -3.404184 -5.686149 -2.280476  -3.800664
#> 49  -5.052563  -5.718598 -3.638586 -3.153479 -5.725339 -1.055784  -3.994269
#>        [,15]      [,16]     [,17]
#> 1  -4.025878  -5.438084 -2.703454
#> 2  -4.912559  -6.245884 -2.845294
#> 3  -5.033563  -7.267200 -3.243823
#> 4  -4.879566  -7.293043 -3.631568
#> 5  -5.091884  -7.274225 -3.202168
#> 6  -4.929283  -6.250563 -2.935152
#> 7  -4.269546  -5.422838 -2.634343
#> 8  -5.240027  -7.266509 -2.741581
#> 9  -4.909135  -6.244942 -2.825384
#> 10 -4.260234  -5.419877 -2.588748
#> 11 -4.880728  -6.237329 -2.630285
#> 12 -5.199434  -7.261116 -2.924840
#> 13 -4.355793  -5.302134 -2.957230
#> 14 -4.179449  -5.440936 -2.404471
#> 15 -5.397695  -7.153158 -2.546188
#> 16 -5.387874  -7.289768 -2.812457
#> 17 -4.372855  -5.458359 -2.704757
#> 18 -4.359635  -5.453549 -2.559458
#> 19 -4.369113  -5.456993 -2.331308
#> 20 -3.896441  -5.503414 -2.440876
#> 21 -4.328030  -5.442354 -2.423825
#> 22 -3.979795  -5.426513 -2.551350
#> 23 -1.571077 -18.789748 -2.323225
#> 24 -4.257911  -5.419134 -2.576645
#> 25 -1.491463 -18.789748 -2.599716
#> 26 -1.228548 -18.789749 -2.145872
#> 27 -1.551811 -18.789749 -2.404968
#> 28 -1.431042 -18.789748 -2.333747
#> 29 -1.466246 -18.789748 -2.351582
#> 30 -1.964561  -7.754605 -5.122116
#> 31 -2.907274  -7.856946 -6.408932
#> 32 -5.116131  -6.311247 -2.559464
#> 33 -3.973990  -6.347784 -2.980966
#> 34 -4.473480  -6.473240 -2.320653
#> 35 -5.064803  -6.293063 -1.535715
#> 36 -5.305486  -7.276091 -1.755610
#> 37 -5.057374  -6.290525 -1.923579
#> 38 -4.285406  -5.427975 -2.316421
#> 39 -4.283173  -5.427245 -2.431373
#> 40 -5.222138  -7.264081 -3.104159
#> 41 -5.074537  -5.556805 -2.006938
#> 42 -4.363448  -5.454930 -1.992261
#> 43 -5.131441  -6.316882 -2.843485
#> 44 -4.344086  -5.447985 -2.920336
#> 45 -4.380286  -5.312431 -2.730652
#> 46 -4.164865  -5.237481 -2.245894
#> 47 -4.287632  -5.428704 -2.714650
#> 48 -2.297349 -18.789748 -2.024014
#> 49 -1.949156 -18.789749 -2.737653
predict(myPLN, trichoptera, "response")
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 
#>  3  5  4  4  4  5  3  4  5  3  5  4  3  3  4  4  3  3  3  3  3  3  1  3  1  1 
#> 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 
#>  1  1  1  2  2  5  5  5  5  4  5  3  3  4  5  3  5  3  3  3  3  1  1 
#> Levels: 1 2 3 4 5