Prediction for a PLNmixturefit object

Predict either posterior probabilities for each group or latent positions based on new samples

# S3 method for PLNmixturefit
predict(
  object,
  newdata,
  type = c("posterior", "response", "position"),
  prior = matrix(rep(1/object$k, object$k), nrow(newdata), object$k, byrow = TRUE),
  control = list(),
  ...
)

Arguments

object

an R6 object with class PLNmixturefit

newdata

A data frame in which to look for variables, offsets and counts with which to predict.

type

The type of prediction required. The default posterior are posterior probabilities for each group , response is the group with maximal posterior probability and latent is the averaged latent in the latent space, with weights equal to the posterior probabilities.

prior

User-specified prior group probabilities in the new data. The default uses a uniform prior.

control

a list for controlling the optimization. See PLN() for details.

...

additional parameters for S3 compatibility. Not used

Value

A matrix of posterior probabilities for each group (if type = "posterior"), a matrix of (average) position in the latent space (if type = "position") or a vector of predicted groups (if type = "response").

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLNmixture(Abundance ~ 1 + offset(log(Offset)),
           data = trichoptera, control_main = list(iterates = 0))  %>% getBestModel()
#> 
#>  Initialization...
#> 
#>  Adjusting 5 PLN mixture models.
#> 	number of cluster = 1 	number of cluster = 2 	number of cluster = 3 	number of cluster = 4 	number of cluster = 5 
#> 
#>  Smoothing PLN mixture models.
#> 
#>  Post-treatments
#>  DONE!
predict(myPLN, trichoptera, "posterior")
#>            [,1]         [,2]         [,3]         [,4]
#> 1  2.220446e-16 9.998800e-01 1.200127e-04 6.232338e-11
#> 2  2.220446e-16 7.861321e-04 9.992139e-01 1.045780e-12
#> 3  2.220446e-16 5.283390e-04 9.994717e-01 4.762182e-15
#> 4  2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 5  2.220446e-16 1.000000e+00 5.201698e-12 2.220446e-16
#> 6  2.220446e-16 2.270700e-02 9.772930e-01 1.310331e-11
#> 7  2.220446e-16 8.245412e-03 9.917546e-01 6.052766e-09
#> 8  2.220446e-16 9.999955e-01 4.461543e-06 2.343023e-12
#> 9  2.220446e-16 2.903486e-03 9.970965e-01 1.270952e-10
#> 10 2.220446e-16 3.974405e-03 9.960256e-01 2.414778e-08
#> 11 2.220446e-16 1.713396e-02 9.828657e-01 3.626211e-07
#> 12 2.220446e-16 1.543929e-01 1.066017e-02 8.349469e-01
#> 13 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 14 2.220446e-16 1.000000e+00 2.513378e-08 1.783128e-14
#> 15 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 17 2.220446e-16 1.000000e+00 7.488721e-15 2.192629e-15
#> 18 2.220446e-16 1.000000e+00 7.094821e-13 2.220446e-16
#> 19 2.220446e-16 1.000000e+00 1.126091e-09 2.220446e-16
#> 20 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 21 2.220446e-16 9.999781e-01 2.189109e-05 1.549477e-15
#> 22 2.220446e-16 3.320148e-03 7.774509e-01 2.192290e-01
#> 23 2.220446e-16 4.321724e-06 6.711160e-08 9.999956e-01
#> 24 2.220446e-16 1.113895e-02 9.888607e-01 3.344488e-07
#> 25 2.220446e-16 9.334111e-10 2.220446e-16 1.000000e+00
#> 26 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 27 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 28 2.220446e-16 6.873120e-07 2.431445e-08 9.999993e-01
#> 29 2.220446e-16 2.489039e-05 1.447840e-07 9.999750e-01
#> 30 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 31 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 32 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 33 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 34 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 35 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 36 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 37 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 38 2.220446e-16 3.484760e-11 1.000000e+00 1.626435e-14
#> 39 2.220446e-16 3.185852e-10 1.000000e+00 2.664207e-13
#> 40 2.220446e-16 9.999532e-01 4.684253e-05 3.139114e-10
#> 41 2.220446e-16 1.000000e+00 7.062356e-16 2.220446e-16
#> 42 2.220446e-16 7.309555e-15 1.000000e+00 2.220446e-16
#> 43 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 44 2.220446e-16 1.000000e+00 1.149932e-14 2.220446e-16
#> 45 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 46 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 47 2.220446e-16 9.999643e-01 3.569949e-05 2.768652e-09
#> 48 2.220446e-16 2.220446e-16 3.569780e-10 1.000000e+00
#> 49 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
predict(myPLN, trichoptera, "position")
#>          [,1]       [,2]      [,3]      [,4]        [,5]      [,6]       [,7]
#> 1   -7.636177  -7.630979 -2.574225 -6.882617 -0.55759139 -3.694063  -6.164912
#> 2  -16.601218 -16.601214 -2.106121 -6.291779 -0.58923673 -3.476052 -16.600054
#> 3  -16.603532 -16.603529 -2.824838 -6.302360 -0.19746778 -3.639981 -16.602750
#> 4   -7.649434  -7.644310 -3.710832 -6.912865 -0.09038705 -3.767534  -6.224637
#> 5   -7.639519  -7.634344 -3.074054 -6.892043 -0.14190390 -3.619925  -6.182764
#> 6  -16.404463 -16.404344 -2.388631 -6.306564 -0.32369608 -3.326760 -16.370856
#> 7  -16.534265 -16.534222 -2.191223 -6.293979 -0.76314990 -3.439966 -16.522061
#> 8   -7.635583  -7.630387 -3.187030 -6.883626 -0.17962461 -3.565389  -6.165633
#> 9  -16.582213 -16.582198 -1.913463 -6.292577 -0.52815075 -3.468855 -16.577916
#> 10 -16.572600 -16.572580 -2.099284 -6.289762 -0.76604902 -3.409680 -16.566718
#> 11 -16.454481 -16.454391 -2.303628 -6.297502 -0.60747946 -3.411183 -16.429120
#> 12  -8.385213  -7.813536 -4.661976 -8.147047 -0.33475129 -6.428618  -6.463890
#> 13  -7.474743  -7.631681 -3.286060 -6.886441 -0.21643106 -3.488060  -6.171346
#> 14  -7.635455  -7.630258 -2.749809 -6.883441 -0.52304835 -3.410122  -6.003696
#> 15  -7.684434  -7.679486 -4.629192 -6.984628 -0.18365846 -3.480636  -6.362802
#> 16  -7.647687  -7.642553 -3.750103 -6.909211 -0.20540604 -3.150975  -6.217352
#> 17  -7.637666  -7.632480 -3.204793 -6.888125 -0.28720426 -3.673185  -6.174785
#> 18  -7.637047  -7.631858 -3.023408 -6.725075 -0.51269707 -3.211117  -6.172111
#> 19  -7.637489  -7.632303 -2.666367 -6.887751 -0.63766834 -2.961022  -6.174022
#> 20  -7.643743  -7.476972 -2.999693 -6.900941 -0.58768079 -3.423503  -6.041813
#> 21  -7.635828  -7.630632 -2.627288 -6.883803 -0.58071984 -3.721108  -6.166200
#> 22 -14.783550 -14.633950 -2.393825 -6.755926 -0.93463633 -4.230326 -14.336224
#> 23  -8.424936  -7.746554 -5.119730 -7.333848 -0.80743975 -7.021415  -6.431669
#> 24 -16.508291 -16.508232 -2.277475 -6.293552 -0.68499733 -3.402932 -16.491803
#> 25  -8.428843  -7.754131 -3.237577 -8.415179 -1.01478427 -6.011609  -5.473851
#> 26  -8.433858  -7.763764 -4.515598 -8.420262 -1.46280289 -7.055454  -6.489283
#> 27  -8.435389  -7.766681 -4.551668 -8.421813 -1.15103775 -7.061083  -6.498512
#> 28  -8.421392  -7.739605 -5.037170 -8.407626 -1.28800102 -7.007263  -6.406626
#> 29  -8.421770  -7.740382 -5.047188 -8.407990 -0.88406427 -7.008779  -6.409484
#> 30  -8.801780  -7.682727 -6.840411 -8.791457 -0.31068001 -6.902032  -6.399878
#> 31  -8.978591  -8.533009 -5.503143 -8.969151 -0.10995760 -7.653891  -6.639319
#> 32 -16.608275 -16.608275 -2.553264 -6.323185 -0.99230054 -3.272768 -16.608275
#> 33 -16.608276 -16.608276 -2.456539 -6.342606 -0.78596891 -3.000045 -16.608276
#> 34 -16.608279 -16.608279 -3.088498 -6.417999 -0.39878646 -2.831170 -16.608279
#> 35 -16.608275 -16.608275 -2.449964 -6.313952 -0.83809864 -2.833161 -16.608275
#> 36 -16.608275 -16.608275 -2.921887 -6.324019 -0.41632187 -3.435061 -16.608275
#> 37 -16.608275 -16.608275 -1.329739 -6.112434 -1.11426931 -3.442207 -16.608275
#> 38 -16.608274 -16.608274 -1.974210 -6.292187 -1.30222266 -3.490294 -16.608274
#> 39 -16.608274 -16.608274 -1.783871 -6.291755 -0.99705938 -3.483208 -16.608274
#> 40  -7.634723  -7.629520 -3.085077 -6.880968 -0.14298985 -3.652199  -6.160671
#> 41  -7.638725  -7.633545 -3.018049 -6.890366 -0.36780288 -3.865818  -6.179352
#> 42 -16.608275 -16.608275 -2.700397 -6.107790 -0.49328495 -3.726681 -16.608275
#> 43 -16.608275 -16.608275 -3.087588 -6.326103 -0.30114896 -3.773199 -16.608275
#> 44  -7.636340  -7.631147 -2.561004 -6.723466 -0.44651703 -3.607157  -5.846726
#> 45  -7.638019  -7.632836 -2.837303 -6.727283 -0.52842436 -3.406187  -6.015612
#> 46  -7.484193  -7.640529 -3.430688 -6.904993 -0.38557580 -4.002879  -6.050528
#> 47  -7.634268  -7.629064 -2.901158 -6.880221 -0.41698739 -3.633007  -6.159005
#> 48  -7.361538  -7.757120 -5.222638 -8.416750 -0.63864487 -7.042489  -6.467716
#> 49  -8.451103  -6.797501 -4.333777 -8.437732 -1.16174007 -7.115851  -6.584502
#>         [,8]       [,9]     [,10]     [,11]     [,12]       [,13]     [,14]
#> 1  -6.047276  -7.613641 -3.296547 -3.772944 -4.868940 -15.7163035 -6.030128
#> 2  -6.303235 -16.601200 -3.646204 -5.902947 -3.587671  -2.7961218 -7.039464
#> 3  -6.313908 -16.603520 -3.786420 -5.919281 -3.342536  -3.0951367 -7.044777
#> 4  -6.115501  -7.627221 -4.078487 -4.299472 -4.766840 -15.7178846 -6.250975
#> 5  -6.068685  -7.617082 -3.823331 -3.964299 -4.931328 -15.7178815 -6.210213
#> 6  -6.299263 -16.403949 -3.668580 -5.856249 -3.835311  -3.1414432 -7.021545
#> 7  -6.299086 -16.534079 -3.224820 -5.882879 -3.966420  -2.8185181 -7.032027
#> 8  -6.049414  -7.613055 -3.623714 -3.793574 -5.034885 -15.7178216 -6.193509
#> 9  -6.302229 -16.582147 -3.639562 -5.897507 -3.978589  -2.8088650 -7.037435
#> 10 -6.298540 -16.572511 -3.589626 -5.890011 -3.944340  -2.7004886 -7.034888
#> 11 -6.295032 -16.454093 -3.587355 -5.860287 -3.957976  -2.8716495 -7.023591
#> 12 -6.546336  -6.905250 -4.255037 -3.655088 -7.298390  -5.3478886 -4.250843
#> 13 -5.894991  -7.614372 -3.404045 -3.853694 -4.894422 -15.7178807 -6.199106
#> 14 -6.048980  -7.612925 -3.468309 -3.789427 -5.033697 -15.7178799 -6.193131
#> 15 -5.975631  -7.662992 -3.587518 -4.510280 -5.293208 -15.7178957 -6.385774
#> 16 -6.107373  -7.625434 -3.774969 -3.866796 -4.889565 -15.7178840 -6.243880
#> 17 -6.059738  -7.453126 -3.170065 -3.453288 -5.063110 -15.7178809 -6.202451
#> 18 -6.056737  -7.614553 -3.558218 -2.858035 -5.054961 -15.7178807 -6.199850
#> 19 -6.058881  -7.615005 -3.439167 -3.161647 -5.060788 -15.7178808 -6.201709
#> 20 -5.930301  -7.621401 -3.369897 -2.145868 -5.140088 -15.7178828 -6.068687
#> 21 -6.049850  -7.613302 -2.892034 -3.645893 -5.036056 -15.7175927 -6.193898
#> 22 -6.377932 -14.397327 -3.320844 -5.431425 -4.802578  -2.9793461 -6.367361
#> 23 -6.675417  -6.682164 -3.826335 -3.977417 -7.778159  -3.8469285 -3.451307
#> 24 -6.296210 -16.508038 -3.582078 -5.873200 -3.947319  -2.7774535 -7.028536
#> 25 -6.696322  -6.702933 -4.729824 -3.049314 -7.785519  -4.0011903 -3.157878
#> 26 -6.721943  -6.728405 -4.161015 -3.283912 -6.751992  -2.8879800 -3.028627
#> 27 -6.729508  -6.735927 -2.734319 -3.011626 -7.797700  -2.1111148 -4.461488
#> 28 -5.596495  -6.662473 -4.502290 -3.787108 -7.771435  -3.6418932 -3.173451
#> 29 -6.657840  -6.664734 -4.517703 -3.045933 -7.772121  -3.6703075 -4.016224
#> 30 -7.146055  -6.773597 -5.137296 -3.126916 -8.335346  -4.0689051 -3.154632
#> 31 -7.508997  -6.918904 -5.894669 -4.104018 -8.003684  -5.1940792 -4.051342
#> 32 -6.334935 -16.608275 -4.017791 -5.951249 -3.744319  -1.0273295 -6.854210
#> 33 -6.354133 -16.608276 -3.733408 -5.784230 -2.403031  -1.8058468 -7.064755
#> 34 -6.428699 -16.608279 -4.432459 -6.085711 -4.150276  -2.5135701 -7.102213
#> 35 -6.325811 -16.608275 -3.411530 -5.937845 -3.828176  -2.9031261 -7.050915
#> 36 -6.335761 -16.608275 -4.025990 -5.952458 -4.266472  -3.2376287 -7.055753
#> 37 -6.324560 -16.608275 -3.563098 -5.736958 -3.816003  -2.6011945 -7.050309
#> 38 -6.304305 -16.608274 -2.889991 -5.906004 -3.988786  -2.6316014 -7.040550
#> 39 -6.303878 -16.608274 -3.457924 -5.905369 -3.984475  -2.6166376 -7.040346
#> 40 -6.043351  -7.612168 -3.552738 -3.734132 -4.857365 -15.7172645 -6.188289
#> 41 -6.064860  -7.616270 -3.368676 -3.932346 -4.920398 -15.7178812 -6.206894
#> 42 -5.919303 -16.608275 -3.684225 -5.929832 -4.139500  -2.9246775 -7.048284
#> 43 -6.337820 -16.608275 -2.463023 -5.758266 -4.282424  -3.1392675 -7.056758
#> 44 -5.892222  -7.613830 -3.228151 -3.830066 -4.886875 -15.7178805 -6.196869
#> 45 -5.740686  -7.615547 -2.792929 -3.757502 -4.910582 -15.7178810 -6.203934
#> 46 -6.097943  -7.623375 -3.819319 -2.986097 -4.862676 -15.7178834 -5.919091
#> 47 -5.879686  -7.611705 -3.375419 -3.559275 -5.013149 -15.7174107 -6.186773
#> 48 -5.711416  -5.717452 -3.502906 -4.170780 -7.788420  -2.7127848 -4.341226
#> 49 -6.801256  -6.807305 -4.521828 -3.241207 -7.826261  -0.9579311 -4.728352
#>        [,15]      [,16]     [,17]
#> 1  -4.280729  -5.801148 -2.839962
#> 2  -4.984873 -16.599764 -2.405388
#> 3  -4.827048 -16.602555 -2.821025
#> 4  -4.438821  -5.886456 -3.705296
#> 5  -4.532950  -5.827225 -3.119541
#> 6  -4.982309 -16.362484 -2.500227
#> 7  -4.973541 -16.519020 -2.297473
#> 8  -4.606366  -5.802672 -2.613134
#> 9  -4.982303 -16.576845 -2.384543
#> 10 -4.969053 -16.565252 -2.199806
#> 11 -4.963331 -16.422802 -2.204722
#> 12 -2.685742  -7.836796 -3.344942
#> 13 -4.633455  -5.650438 -3.154329
#> 14 -4.446564  -5.802071 -2.467857
#> 15 -5.080580  -5.931555 -2.513857
#> 16 -4.831898  -5.876224 -2.777637
#> 17 -4.649365  -5.815820 -2.836284
#> 18 -4.637012  -5.811989 -2.649410
#> 19 -4.645851  -5.814727 -2.337703
#> 20 -4.039168  -5.852814 -2.438223
#> 21 -4.608200  -5.803414 -2.489655
#> 22 -4.086873 -14.708479 -2.396546
#> 23 -1.612224  -8.111989 -2.712810
#> 24 -4.964352 -16.487696 -2.176960
#> 25 -1.472709  -8.117305 -3.430623
#> 26 -1.139396  -8.124097 -2.201837
#> 27 -1.557792  -8.126163 -2.694348
#> 28 -1.322644  -8.107158 -2.899077
#> 29 -1.403868  -8.107643 -2.943943
#> 30 -1.953737  -7.886204 -4.735198
#> 31 -2.914245  -8.176097 -6.409542
#> 32 -5.097684 -16.608275 -2.495829
#> 33 -4.258722 -16.608276 -2.967544
#> 34 -4.576156 -16.608279 -2.305132
#> 35 -5.066077 -16.608275 -1.417362
#> 36 -5.100506 -16.608275 -1.491151
#> 37 -5.061683 -16.608275 -1.824516
#> 38 -4.988420 -16.608274 -1.934253
#> 39 -4.986830 -16.608274 -2.078710
#> 40 -4.580419  -5.795351 -2.895133
#> 41 -4.670196  -5.503380 -2.058755
#> 42 -5.046888 -16.608275 -1.674444
#> 43 -5.107523 -16.608275 -2.773893
#> 44 -4.622706  -5.807593 -3.105744
#> 45 -4.656356  -5.658059 -2.869267
#> 46 -4.364523  -5.551477 -2.164309
#> 47 -4.572872  -5.792997 -2.853699
#> 48 -3.239720  -8.119408 -1.994539
#> 49 -2.074239  -8.147219 -3.203324
predict(myPLN, trichoptera, "response")
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 
#>  2  3  3  2  2  3  3  2  3  3  3  4  2  2  2  2  2  2  2  2  2  3  4  3  4  4 
#> 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 
#>  4  4  4  4  4  3  3  3  3  3  3  3  3  2  2  3  3  2  2  2  2  4  4 
#> Levels: 2 3 4