Prediction for a PLNmixturefit object

Predict either posterior probabilities for each group or latent positions based on new samples

# S3 method for PLNmixturefit
predict(
  object,
  newdata,
  type = c("posterior", "response", "position"),
  prior = matrix(rep(1/object$k, object$k), nrow(newdata), object$k, byrow = TRUE),
  control = list(),
  ...
)

Arguments

object
an R6 object with class PLNmixturefit

newdata
A data frame in which to look for variables, offsets and counts with which to predict.

type
The type of prediction required. The default posterior are posterior probabilities for each group , response is the group with maximal posterior probability and latent is the averaged latent in the latent space, with weights equal to the posterior probabilities.

prior
User-specified prior group probabilities in the new data. The default uses a uniform prior.

control
a list for controlling the optimization. See PLN() for details.

...
additional parameters for S3 compatibility. Not used

Value

A matrix of posterior probabilities for each group (if type = "posterior"), a matrix of (average) position in the latent space (if type = "position") or a vector of predicted groups (if type = "response").

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLNmixture(Abundance ~ 1 + offset(log(Offset)),
           data = trichoptera, control_main = list(iterates = 0))  %>% getBestModel()

#> 
#>  Initialization...
#> 
#>  Adjusting 5 PLN mixture models.
#> 	number of cluster = 1 	number of cluster = 2 	number of cluster = 3 	number of cluster = 4 	number of cluster = 5 
#> 
#>  Smoothing PLN mixture models.
#> 
#>  Post-treatments
#>  DONE!

predict(myPLN, trichoptera, "posterior")

#>            [,1]         [,2]         [,3]         [,4]         [,5]
#> 1  9.988543e-01 2.220446e-16 3.697914e-05 1.108717e-03 1.084402e-12
#> 2  8.374114e-02 2.220446e-16 9.162589e-01 2.220446e-16 2.578708e-10
#> 3  8.967514e-09 2.220446e-16 9.993886e-01 6.113714e-04 5.003958e-10
#> 4  2.220446e-16 2.220446e-16 1.000000e+00 3.163645e-16 2.220446e-16
#> 5  2.220446e-16 2.220446e-16 9.999055e-01 9.453561e-05 3.046240e-12
#> 6  1.303740e-04 2.220446e-16 9.998696e-01 2.220446e-16 2.369846e-08
#> 7  9.921336e-01 2.220446e-16 7.866358e-03 2.220446e-16 4.528460e-08
#> 8  4.361831e-08 2.220446e-16 9.999999e-01 2.220446e-16 1.910108e-08
#> 9  8.611430e-01 2.220446e-16 1.388570e-01 2.220446e-16 4.975096e-08
#> 10 8.779196e-01 2.220446e-16 1.220769e-01 2.220446e-16 3.537713e-06
#> 11 5.845550e-01 2.220446e-16 4.149298e-01 2.220446e-16 5.151574e-04
#> 12 3.248892e-01 2.220446e-16 6.751108e-01 2.220446e-16 2.220446e-16
#> 13 3.672516e-09 2.220446e-16 1.000000e+00 2.220446e-16 3.157563e-08
#> 14 9.978320e-01 2.220446e-16 2.167977e-03 2.220446e-16 1.852920e-12
#> 15 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 16 2.220446e-16 2.220446e-16 9.999998e-01 2.055988e-07 2.220446e-16
#> 17 8.345598e-01 2.220446e-16 1.652735e-01 2.220446e-16 1.666745e-04
#> 18 9.764061e-01 2.220446e-16 2.359394e-02 2.220446e-16 7.461648e-14
#> 19 5.334120e-06 2.220446e-16 9.999947e-01 2.220446e-16 2.220446e-16
#> 20 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 21 9.998461e-01 2.220446e-16 1.538801e-04 2.220446e-16 2.220446e-16
#> 22 9.958137e-01 2.220446e-16 1.591434e-05 4.170350e-03 4.363044e-11
#> 23 9.999909e-01 2.220446e-16 1.042371e-11 9.139021e-06 1.771298e-10
#> 24 6.882972e-01 2.220446e-16 3.114705e-01 2.220446e-16 2.322166e-04
#> 25 1.000000e+00 2.220446e-16 2.220446e-16 5.142556e-09 2.833511e-09
#> 26 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 27 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 28 9.999844e-01 2.220446e-16 7.304544e-07 7.430118e-06 7.478380e-06
#> 29 9.993023e-01 2.220446e-16 5.039733e-05 4.303221e-07 6.468392e-04
#> 30 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.861815e-10
#> 31 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 32 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 33 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 34 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 35 2.980975e-13 2.220446e-16 9.999943e-01 5.741007e-06 2.220446e-16
#> 36 2.220446e-16 2.220446e-16 9.999979e-01 2.131514e-06 2.220446e-16
#> 37 9.999988e-01 2.220446e-16 6.103144e-15 1.154165e-06 2.220446e-16
#> 38 9.983424e-01 2.220446e-16 1.722278e-06 1.655847e-03 2.220446e-16
#> 39 9.996465e-01 2.220446e-16 3.535033e-04 2.351085e-10 2.115239e-15
#> 40 6.087198e-05 2.220446e-16 9.999033e-01 2.220446e-16 3.583985e-05
#> 41 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 42 9.999976e-01 2.220446e-16 2.440043e-06 2.220446e-16 2.220446e-16
#> 43 2.738017e-11 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 44 1.000000e+00 2.220446e-16 2.622022e-14 2.220446e-16 2.220446e-16
#> 45 9.986512e-01 2.220446e-16 1.348767e-03 2.220446e-16 2.220446e-16
#> 46 6.916794e-07 2.220446e-16 9.999993e-01 2.220446e-16 6.914087e-15
#> 47 9.711778e-01 2.220446e-16 2.881988e-02 2.220446e-16 2.283831e-06
#> 48 9.826023e-01 2.220446e-16 5.954495e-12 1.739768e-02 2.220446e-16
#> 49 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16

predict(myPLN, trichoptera, "position")

#>          [,1]       [,2]      [,3]       [,4]       [,5]      [,6]       [,7]
#> 1   -5.915434  -5.940632 -2.148854  -5.907552 -0.6964969 -5.909709  -5.912297
#> 2  -16.825423 -16.827495 -3.086349 -16.824840 -0.5601018 -3.489272 -16.825260
#> 3  -17.816792 -17.817216 -3.513056 -17.816228 -0.2698057 -3.417743 -17.816119
#> 4  -17.823751 -17.823751 -3.787812 -17.823751 -0.1295969 -3.559434 -17.823751
#> 5  -17.822675 -17.822741 -3.304132 -17.822588 -0.2038142 -3.359727 -17.822571
#> 6  -17.822196 -17.822200 -3.287201 -17.822196 -0.3599129 -3.169389 -17.822196
#> 7   -5.998501  -6.022997 -2.216574  -5.991621 -0.8818447 -5.881515  -5.996577
#> 8  -17.823750 -17.823750 -3.495947 -17.823750 -0.2625239 -3.261600 -17.823750
#> 9   -7.561134  -7.582362 -2.206780  -7.555171 -0.6651270 -5.530143  -7.559466
#> 10  -7.358144  -7.379832 -2.298441  -7.352045 -0.8603459 -5.571258  -7.356403
#> 11 -10.854522 -10.864803 -2.763414 -10.850459 -0.6516343 -4.784750 -10.848202
#> 12 -13.950407 -13.958453 -3.134647 -13.948147 -0.5236668 -4.082708 -13.949776
#> 13 -17.823737 -17.823751 -3.478932 -17.823751 -0.4540978 -3.196190 -17.823750
#> 14  -5.927130  -5.951850 -2.376465  -5.920186 -0.9363246 -5.887412  -5.922156
#> 15 -17.823753 -17.823753 -4.442494 -17.823753 -0.1993903 -3.413326 -17.823753
#> 16 -17.823749 -17.823749 -3.839499 -17.823749 -0.2322928 -3.051329 -17.823749
#> 17  -7.873506  -7.892829 -2.564098  -7.867698 -0.8468368 -5.451459  -7.870205
#> 18  -6.182298  -6.206495 -2.402407  -6.175498 -0.9452782 -5.835263  -6.180398
#> 19 -17.823687 -17.823687 -3.473858 -17.823687 -0.4630561 -3.191864 -17.823687
#> 20  -5.946407  -5.844151 -2.691366  -5.939754 -0.6451783 -5.319705  -5.818689
#> 21  -5.915012  -5.939494 -2.233505  -5.908137 -0.7011789 -5.910839  -5.913090
#> 22  -5.909525  -5.936946 -2.153842  -5.898791 -0.9870251 -5.895059  -5.903011
#> 23  -5.908962  -5.933556 -2.597766  -5.742527 -0.8827516 -5.907004  -5.907021
#> 24  -9.617943  -9.633093 -2.636523  -9.613158 -0.7360300 -5.060860  -9.614268
#> 25  -5.913530  -5.938008 -2.339842  -5.906655 -1.0203926 -5.752782  -5.752782
#> 26  -5.919367  -5.943707 -2.686051  -5.912533 -1.3266013 -5.917457  -5.917457
#> 27  -5.921134  -5.945431 -2.716633  -5.914311 -1.1295240 -5.919226  -5.919226
#> 28  -5.904869  -5.929503 -2.492286  -5.897928 -1.0999469 -5.902829  -5.902847
#> 29  -5.912724  -5.932134 -2.507338  -5.905797 -0.9288751 -5.904085  -5.904278
#> 30  -6.356877  -6.262924 -4.715532  -6.352095 -0.3182501 -6.138741  -6.036340
#> 31  -6.578427  -6.593301 -4.727702  -6.574265 -0.1162567 -6.481662  -6.217472
#> 32 -17.823751 -17.823751 -3.002576 -17.823751 -0.9556081 -3.173149 -17.823751
#> 33 -17.823751 -17.823751 -2.751003 -17.823751 -0.7920828 -2.988505 -17.823751
#> 34 -17.823752 -17.823752 -3.182404 -17.823752 -0.4179202 -2.884179 -17.823752
#> 35 -17.823686 -17.823690 -3.007654 -17.823680 -0.8208259 -2.832473 -17.823679
#> 36 -17.823727 -17.823728 -3.350289 -17.823725 -0.4532422 -3.293468 -17.823725
#> 37  -5.929429  -5.953532 -1.608710  -5.766256 -1.1069618 -5.615351  -5.927535
#> 38  -5.908942  -5.934660 -2.059678  -5.900515 -1.2320481 -5.902101  -5.905191
#> 39  -5.911782  -5.936395 -1.920248  -5.904869 -1.0231476 -5.904678  -5.909849
#> 40 -17.822979 -17.822690 -3.432337 -17.822979 -0.2357152 -3.319262 -17.822619
#> 41 -17.823751 -17.823751 -3.445745 -17.823751 -0.4264005 -3.254363 -17.823751
#> 42  -5.921258  -5.945551 -2.572670  -5.782174 -0.5963093 -5.919315  -5.919351
#> 43 -17.823751 -17.823751 -3.487458 -17.823751 -0.3467133 -3.541936 -17.823751
#> 44  -5.901232  -5.926008 -2.372932  -5.892340 -0.9369556 -5.897354  -5.895421
#> 45  -5.917073  -5.941822 -2.377874  -5.910122 -0.9564995 -5.895401  -5.915130
#> 46 -17.823743 -17.823743 -3.478923 -17.823743 -0.4542012 -3.196191 -17.823743
#> 47  -6.250654  -6.274556 -2.362637  -6.243935 -0.6340194 -5.827257  -6.248752
#> 48  -5.768674  -5.960544 -2.726645  -5.901693 -0.7388374 -5.871251  -5.903448
#> 49  -5.939192  -5.807945 -2.875672  -5.932489 -1.1579662 -5.937318  -5.937318
#>         [,8]       [,9]     [,10]     [,11]     [,12]     [,13]     [,14]
#> 1  -5.224661  -5.916200 -2.828836 -3.530542 -4.670011 -2.770853 -4.037321
#> 2  -6.645299 -16.825423 -4.255976 -5.121589 -3.820875 -3.103012 -7.103847
#> 3  -6.781251 -17.817214 -4.432728 -5.299439 -3.507921 -3.324345 -7.378619
#> 4  -6.808757 -17.823751 -4.550694 -5.411926 -4.048000 -3.901401 -7.394307
#> 5  -6.789268 -17.822740 -4.505710 -5.331642 -4.036569 -3.522849 -7.383425
#> 6  -6.778266 -17.822196 -4.393424 -5.283781 -3.871162 -3.198992 -7.377123
#> 7  -5.217767  -5.998501 -2.648193 -3.442980 -4.789972 -2.563724 -4.156450
#> 8  -6.781447 -17.823750 -4.425361 -5.297205 -4.056970 -3.299173 -7.379185
#> 9  -5.426172  -7.561133 -3.154917 -3.700052 -4.685144 -2.669399 -4.589406
#> 10 -5.393728  -7.358110 -3.073420 -3.633734 -4.689492 -2.577850 -4.517271
#> 11 -5.855737 -10.849513 -3.502350 -4.178537 -4.449570 -2.747616 -5.469782
#> 12 -6.263487 -13.950407 -3.877925 -4.660830 -4.231722 -2.890837 -6.315889
#> 13 -6.775257 -17.823750 -4.358301 -5.269685 -3.958669 -3.077728 -7.375792
#> 14 -5.201873  -5.927130 -2.863216 -3.391344 -4.784285 -2.462527 -4.117597
#> 15 -6.589939 -17.823753 -3.727930 -5.053274 -4.714991 -4.542245 -7.431316
#> 16 -6.805278 -17.823749 -4.267453 -4.980023 -4.121399 -3.844325 -7.392370
#> 17 -5.459447  -7.870548 -3.104877 -3.694178 -4.649882 -2.562875 -4.649985
#> 18 -5.235088  -6.182298 -2.892366 -3.428161 -4.765696 -2.466888 -4.185866
#> 19 -6.775344 -17.823687 -4.356805 -5.263986 -3.960036 -3.080976 -7.375819
#> 20 -5.164714  -5.946407 -3.051518 -2.397584 -4.915973 -3.251884 -4.236089
#> 21 -5.222539  -5.915012 -2.567563 -3.412680 -4.821689 -2.766155 -4.179774
#> 22 -5.210072  -5.912406 -2.688501 -3.454652 -4.801261 -2.626181 -4.151276
#> 23 -5.213842  -5.908968 -2.863526 -3.475130 -4.809148 -2.660046 -4.001142
#> 24 -5.691640  -9.615685 -3.344337 -3.982397 -4.532464 -2.677476 -5.131599
#> 25 -5.222911  -5.913530 -3.074208 -3.241537 -4.822605 -2.752479 -3.881059
#> 26 -5.234382  -5.919367 -3.029149 -3.310252 -4.686799 -2.520090 -3.778785
#> 27 -5.237830  -5.921134 -2.603646 -3.203031 -4.844508 -2.174969 -4.219958
#> 28 -5.045928  -5.904802 -2.931434 -3.428861 -4.796688 -2.560340 -3.973790
#> 29 -5.208137  -5.906435 -2.941765 -3.281343 -4.800376 -2.573689 -4.133629
#> 30 -5.815357  -6.139985 -4.448980 -3.140631 -5.678796 -3.816249 -3.229272
#> 31 -6.111906  -6.302920 -4.999052 -4.032425 -5.923325 -4.652371 -4.057372
#> 32 -6.791063 -17.823751 -4.521653 -5.338844 -3.799082 -1.341539 -7.234836
#> 33 -6.799166 -17.823751 -4.203515 -5.230701 -2.677065 -2.016079 -7.388974
#> 34 -6.830935 -17.823752 -4.610689 -5.497573 -4.088473 -2.604430 -7.406734
#> 35 -6.787230 -17.823690 -4.066093 -5.322446 -3.869584 -3.104894 -7.382366
#> 36 -6.791351 -17.823728 -4.524434 -5.340080 -4.196467 -3.337425 -7.384645
#> 37 -5.253882  -5.929430 -3.045633 -3.535981 -4.572008 -2.512908 -4.260943
#> 38 -5.211970  -5.910086 -2.436144 -3.465063 -4.805476 -2.507354 -4.152453
#> 39 -5.211630  -5.911782 -2.838155 -3.460807 -4.804730 -2.492216 -4.148282
#> 40 -6.778860 -17.822631 -4.398425 -5.285851 -3.874912 -3.214832 -7.377462
#> 41 -6.776929 -17.823751 -4.318282 -5.277138 -3.966685 -3.143010 -7.376702
#> 42 -4.977760  -5.921258 -3.086821 -3.601680 -4.845284 -2.713503 -4.222610
#> 43 -6.791952 -17.823751 -3.129043 -5.200452 -4.204552 -3.247261 -7.384978
#> 44 -5.196421  -5.901232 -2.856167 -3.386644 -4.783994 -2.459627 -4.110218
#> 45 -5.199997  -5.917073 -2.858961 -3.386223 -4.784080 -2.452958 -4.113188
#> 46 -6.775269 -17.823743 -4.358319 -5.269680 -3.958670 -3.077694 -7.375790
#> 47 -5.127681  -6.250631 -2.898463 -3.386877 -4.779646 -2.638972 -4.238366
#> 48 -5.072021  -5.780691 -2.849147 -3.536023 -4.817477 -2.471548 -4.216901
#> 49 -5.272497  -5.939192 -3.258246 -3.281742 -4.894381 -1.163128 -4.306285
#>        [,15]      [,16]     [,17]
#> 1  -2.282110 -17.883784 -2.686079
#> 2  -4.478176  -8.313341 -2.730230
#> 3  -4.584267  -7.438579 -2.983554
#> 4  -4.526648  -7.453699 -3.509747
#> 5  -4.642875  -7.443395 -3.027989
#> 6  -4.700256  -7.439279 -2.806820
#> 7  -2.257642 -17.815791 -2.608337
#> 8  -4.724064  -7.439456 -2.579589
#> 9  -2.625129 -16.445416 -2.647532
#> 10 -2.474104 -16.620913 -2.567284
#> 11 -3.204239 -13.552009 -2.592527
#> 12 -3.847063 -10.835313 -2.605443
#> 13 -4.674748  -7.436247 -2.621932
#> 14 -2.105284 -17.875407 -2.502590
#> 15 -5.050959  -7.342511 -2.528162
#> 16 -4.895344  -7.451874 -2.740208
#> 17 -2.527046 -16.167331 -2.528184
#> 18 -2.150717 -17.651252 -2.508578
#> 19 -4.675393  -7.436359 -2.618596
#> 20 -2.707902 -17.898088 -2.433288
#> 21 -2.508790 -17.896478 -2.413215
#> 22 -2.065934 -17.845570 -2.524700
#> 23 -1.887265 -17.897973 -2.439020
#> 24 -2.934162 -14.637213 -2.572919
#> 25 -1.758546 -17.898088 -2.668973
#> 26 -1.443174 -17.898088 -2.254239
#> 27 -1.763558 -17.898088 -2.486195
#> 28 -1.827004 -17.897911 -2.462900
#> 29 -1.849793 -17.890982 -2.477986
#> 30 -1.960655 -17.737329 -4.223217
#> 31 -2.890761 -17.737333 -5.111914
#> 32 -4.796561  -7.444448 -2.598030
#> 33 -4.184854  -7.448674 -2.948087
#> 34 -4.457099  -7.465415 -2.356514
#> 35 -4.768323  -7.442448 -1.656646
#> 36 -4.798684  -7.444596 -1.692650
#> 37 -2.768490 -17.898074 -2.073607
#> 38 -2.346258 -17.877284 -2.284313
#> 39 -2.329671 -17.894390 -2.401947
#> 40 -4.703931  -7.438795 -2.829902
#> 41 -4.688228  -7.397849 -2.515832
#> 42 -2.681708 -17.898063 -1.977214
#> 43 -4.803092  -7.444910 -2.823047
#> 44 -2.100485 -17.898088 -2.512960
#> 45 -2.093191 -17.883978 -2.505995
#> 46 -4.674740  -7.436268 -2.621879
#> 47 -2.395598 -17.596557 -2.670164
#> 48 -2.558562 -17.679689 -2.156764
#> 49 -2.096693 -17.898088 -2.770621

predict(myPLN, trichoptera, "response")

#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 
#>  1  3  3  3  3  3  1  3  1  1  1  3  3  1  3  3  1  1  3  1  1  1  1  1  1  1 
#> 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 
#>  1  1  1  1  1  3  3  3  3  3  1  1  1  3  3  1  3  1  1  3  1  1  1 
#> Levels: 1 3