Prediction for a PLNmixturefit object

Predict either posterior probabilities for each group or latent positions based on new samples

# S3 method for PLNmixturefit
predict(
  object,
  newdata,
  type = c("posterior", "response", "position"),
  prior = matrix(rep(1/object$k, object$k), nrow(newdata), object$k, byrow = TRUE),
  control = PLNmixture_param(),
  ...
)

Arguments

object

an R6 object with class PLNmixturefit

newdata

A data frame in which to look for variables, offsets and counts with which to predict.

type

The type of prediction required. The default posterior are posterior probabilities for each group , response is the group with maximal posterior probability and latent is the averaged latent in the latent space, with weights equal to the posterior probabilities.

prior

User-specified prior group probabilities in the new data. The default uses a uniform prior.

control

a list-like structure for controlling the fit. See PLNmixture_param() for details.

...

additional parameters for S3 compatibility. Not used

Value

A matrix of posterior probabilities for each group (if type = "posterior"), a matrix of (average) position in the latent space (if type = "position") or a vector of predicted groups (if type = "response").

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLNmixture(Abundance ~ 1 + offset(log(Offset)),
           data = trichoptera, control = PLNmixture_param(smoothing = "none"))  %>% getBestModel()
#> 
#>  Initialization...
#> 
#>  Adjusting 5 PLN mixture models.
#> 	number of cluster = 1 	number of cluster = 2 	number of cluster = 3 	number of cluster = 4 	number of cluster = 5 
#>  Post-treatments
#>  DONE!
predict(myPLN, trichoptera, "posterior")
#>            [,1]         [,2]         [,3]         [,4]         [,5]
#> 1  6.359439e-02 2.220446e-16 2.220446e-16 1.927636e-03 9.344780e-01
#> 2  9.028578e-01 2.220446e-16 2.204658e-02 1.852247e-05 7.507710e-02
#> 3  2.061844e-01 2.220446e-16 7.932390e-01 2.220446e-16 5.766090e-04
#> 4  3.224804e-05 2.220446e-16 9.999678e-01 2.220446e-16 2.220446e-16
#> 5  7.909108e-03 2.220446e-16 9.917165e-01 2.220446e-16 3.744194e-04
#> 6  4.768653e-01 1.007879e-14 3.540325e-01 1.062451e-05 1.690916e-01
#> 7  1.210390e-01 1.339442e-14 3.851953e-03 2.705034e-03 8.724040e-01
#> 8  2.462063e-02 4.026924e-15 9.562175e-01 1.500152e-06 1.916040e-02
#> 9  4.230344e-01 3.009655e-15 1.206735e-02 1.664718e-04 5.647318e-01
#> 10 3.802777e-01 1.265844e-12 2.226239e-02 8.295868e-03 5.891640e-01
#> 11 2.975222e-01 3.294424e-03 1.162495e-01 1.923400e-02 5.636999e-01
#> 12 1.946669e-01 8.593737e-02 3.313020e-01 4.325663e-02 3.448371e-01
#> 13 2.220446e-16 2.220446e-16 4.884496e-14 2.220446e-16 1.000000e+00
#> 14 2.220446e-16 2.220446e-16 2.220446e-16 3.678773e-06 9.999963e-01
#> 15 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 16 1.420280e-05 2.220446e-16 9.999858e-01 2.220446e-16 2.220446e-16
#> 17 2.220446e-16 2.220446e-16 2.335369e-16 4.026580e-06 9.999960e-01
#> 18 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 19 1.220750e-04 2.220446e-16 6.898360e-05 2.220446e-16 9.998089e-01
#> 20 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 21 5.492451e-03 2.220446e-16 8.359961e-05 4.344516e-06 9.944196e-01
#> 22 5.919120e-02 3.768192e-16 3.550393e-04 6.855122e-02 8.719025e-01
#> 23 2.220446e-16 2.220446e-16 2.220446e-16 9.981405e-01 1.859512e-03
#> 24 3.139904e-01 2.504303e-03 8.720939e-02 2.774972e-02 5.685462e-01
#> 25 2.220446e-16 9.067287e-15 2.220446e-16 9.990195e-01 9.805480e-04
#> 26 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 27 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 28 2.220446e-16 4.117463e-12 2.220446e-16 9.991109e-01 8.891248e-04
#> 29 1.683601e-04 2.034883e-02 1.459413e-04 9.702962e-01 9.040714e-03
#> 30 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 31 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 32 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 33 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 34 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 35 9.999344e-01 2.220446e-16 2.220446e-16 2.220446e-16 6.557330e-05
#> 36 8.785680e-01 2.220446e-16 1.209788e-01 2.220446e-16 4.532225e-04
#> 37 9.999635e-01 2.220446e-16 2.220446e-16 2.220446e-16 3.653734e-05
#> 38 6.523186e-01 2.220446e-16 2.220446e-16 2.358996e-03 3.453225e-01
#> 39 8.445108e-01 2.220446e-16 4.345542e-05 5.401967e-04 1.549056e-01
#> 40 1.539193e-01 1.084517e-11 7.978395e-01 4.104209e-05 4.820014e-02
#> 41 3.518373e-01 2.220446e-16 1.188681e-01 2.220446e-16 5.292946e-01
#> 42 2.220446e-16 2.220446e-16 2.220446e-16 4.343917e-03 9.956561e-01
#> 43 8.820619e-01 2.220446e-16 2.319971e-03 3.493606e-06 1.156147e-01
#> 44 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 45 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 46 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 47 2.220446e-16 8.257046e-14 3.190495e-03 6.827257e-04 9.961268e-01
#> 48 2.220446e-16 2.220446e-16 2.220446e-16 9.787104e-01 2.128965e-02
#> 49 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
predict(myPLN, trichoptera, "position")
#>          [,1]       [,2]      [,3]       [,4]       [,5]      [,6]       [,7]
#> 1   -6.952071  -7.596685 -2.129745  -5.502248 -0.5574785 -3.691462  -5.935307
#> 2  -18.385813 -18.437817 -2.012161  -7.122251 -0.5377094 -3.547866 -18.303266
#> 3  -19.889187 -19.889584 -3.728202 -17.331109 -0.1914377 -3.809501 -19.888564
#> 4  -20.034883 -20.034883 -4.092156 -20.034485 -0.0919893 -3.767888 -20.034883
#> 5  -20.024415 -20.024670 -3.591105 -19.926189 -0.1446813 -3.653259 -20.024020
#> 6  -17.361000 -17.477969 -2.796954 -11.321019 -0.3089472 -3.396950 -17.175701
#> 7   -7.760104  -8.365311 -2.180245  -5.633252 -0.6388868 -3.593599  -6.797677
#> 8  -19.751666 -19.764864 -3.877825 -19.432697 -0.1774550 -3.603112 -19.730885
#> 9  -11.882519 -12.273803 -1.893514  -6.226989 -0.4875883 -3.581674 -11.261189
#> 10 -11.452064 -11.861480 -2.060518  -6.306665 -0.6307709 -3.541530 -10.800317
#> 11 -11.694405 -12.059047 -2.450449  -7.594976 -0.4650402 -3.594557 -11.036782
#> 12 -14.203964 -13.727043 -3.330324 -11.538336 -0.3001596 -3.977522 -12.934938
#> 13  -5.965096  -6.803240 -2.722298  -5.417313 -0.2634149 -3.500598  -5.042602
#> 14  -6.111280  -6.800283 -2.281432  -5.405758 -0.5299761 -3.433071  -4.876841
#> 15 -20.034904 -20.034904 -4.674314 -20.034904 -0.1833140 -3.498295 -20.034904
#> 16 -20.034895 -20.034895 -4.184310 -20.034719 -0.1947491 -3.202597 -20.034895
#> 17  -6.120347  -6.804889 -2.685032  -5.423693 -0.3221733 -3.658238  -5.051694
#> 18  -6.117827  -6.803608 -2.540304  -5.269811 -0.5199293 -3.266443  -5.044636
#> 19  -6.122207  -6.806972 -2.315217  -5.423479 -0.6242623 -3.064436  -5.052467
#> 20  -6.144507  -6.666190 -2.672075  -5.469984 -0.5857304 -3.447892  -4.979668
#> 21  -6.185978  -6.870780 -2.204385  -5.416975 -0.5743011 -3.709080  -5.108126
#> 22  -6.863851  -7.470515 -2.203727  -5.506466 -0.7881852 -3.759035  -5.904984
#> 23  -5.669279  -5.710099 -3.794550  -5.456167 -0.8998317 -5.661475  -5.664019
#> 24 -11.495734 -11.870793 -2.388648  -7.187824 -0.5284434 -3.597261 -10.840449
#> 25  -5.676480  -5.716437 -3.316927  -5.664351 -1.0386345 -5.472207  -5.473509
#> 26  -5.685737  -5.724695 -3.722172  -5.674398 -1.3691875 -5.682563  -5.682563
#> 27  -5.688676  -5.727526 -3.739294  -5.677369 -1.1431355 -5.685511  -5.685511
#> 28  -5.661920  -5.702376 -3.752523  -5.649683 -1.1862924 -5.656451  -5.657693
#> 29  -5.898319  -5.773204 -3.796755  -5.878510 -0.9667296 -5.674767  -5.675170
#> 30 -16.848714  -8.370759 -6.654046 -16.848714 -0.3098205 -7.223517  -6.458335
#> 31 -16.848729  -8.612931 -5.752108 -16.848729 -0.1098406 -7.523765  -6.590055
#> 32 -19.367107 -19.367107 -2.523073  -6.984219 -0.9645127 -3.245682 -19.367107
#> 33 -19.367107 -19.367107 -2.437199  -7.003838 -0.7772346 -2.998319 -19.367107
#> 34 -19.367107 -19.367107 -3.004718  -7.076121 -0.4004542 -2.879844 -19.367107
#> 35 -19.366238 -19.366283 -2.425984  -6.974621 -0.8163133 -2.758952 -19.366168
#> 36 -19.441898 -19.442205 -2.979669  -8.563127 -0.4169893 -3.445622 -19.441422
#> 37 -19.366623 -19.366648 -1.419995  -6.588414 -1.0610829 -3.449181 -19.366583
#> 38 -14.754962 -14.994122 -1.904666  -6.410291 -1.1509825 -3.590243 -14.375781
#> 39 -17.305247 -17.412538 -1.647394  -6.709658 -0.9041945 -3.562684 -17.135042
#> 40 -19.260176 -19.293503 -3.605871 -17.315345 -0.1510372 -3.708507 -19.207425
#> 41 -12.437320 -12.798536 -2.748698  -7.711256 -0.3799476 -3.885146 -11.875692
#> 42  -6.116710  -6.799327 -2.559387  -5.272395 -0.4991557 -3.777175  -5.049466
#> 43 -17.838604 -17.917011 -3.032804  -6.839632 -0.3086293 -3.843663 -17.717662
#> 44  -6.114929  -6.802135 -2.196438  -5.263204 -0.4659807 -3.604995  -4.742374
#> 45  -6.121783  -6.805626 -2.445480  -5.278750 -0.5327596 -3.431192  -4.910911
#> 46  -6.005313  -6.821286 -2.964957  -5.483868 -0.3849427 -3.902682  -5.000967
#> 47  -6.149159  -6.838620 -2.310185  -5.440055 -0.4632306 -3.638454  -5.056108
#> 48  -5.491075  -5.741290 -3.832448  -5.662036 -0.7248462 -5.633680  -5.661981
#> 49  -5.718471  -5.565635 -3.747405  -5.707480 -1.1569233 -5.715394  -5.715394
#>          [,8]       [,9]     [,10]     [,11]     [,12]     [,13]      [,14]
#> 1   -5.656842  -7.584282 -2.875737 -3.464605 -4.791220 -5.248290  -5.371153
#> 2  -17.984691 -18.436823 -3.831776 -5.625706 -3.167877 -3.274700  -7.136768
#> 3   -9.250324 -19.889576 -4.488556 -5.395153 -3.536873 -5.816221 -17.333711
#> 4   -6.675084 -20.034883 -4.708068 -5.439350 -4.339586 -6.519839 -20.034485
#> 5   -6.738014 -20.024665 -4.685713 -5.327399 -4.418421 -6.454579 -19.926301
#> 6  -12.374492 -17.475734 -3.993196 -5.192044 -3.947243 -4.673350 -11.334054
#> 7   -6.478545  -8.353650 -2.717467 -3.509280 -4.836027 -5.074297  -5.666074
#> 8   -6.900140 -19.764611 -4.548927 -5.245743 -4.504850 -6.361026 -19.433791
#> 9  -10.928137 -12.266323 -3.373732 -4.342558 -4.433927 -4.386174  -6.255631
#> 10 -10.315295 -11.853340 -3.286506 -4.215539 -4.450360 -4.406889  -6.319873
#> 11  -9.294326 -12.045489 -3.377504 -4.191759 -4.533223 -4.681589  -7.540858
#> 12  -8.461099 -13.581989 -3.862437 -4.400226 -4.861206 -5.036424 -10.332444
#> 13  -4.608613  -6.790097 -2.971084 -3.395724 -4.900309 -5.404800  -5.458056
#> 14  -4.730848  -6.787101 -2.988618 -3.322059 -5.030488 -5.393093  -5.446942
#> 15  -6.397121 -20.034904 -3.786341 -5.026949 -4.976943 -6.648263 -20.034904
#> 16  -6.668576 -20.034895 -4.354103 -4.857264 -4.452178 -6.512629 -20.034719
#> 17  -4.764612  -6.638599 -2.815504 -3.091425 -5.056066 -5.411246  -5.464185
#> 18  -4.755391  -6.790470 -3.093875 -2.635294 -5.049038 -5.406243  -5.459426
#> 19  -4.763906  -6.793849 -3.010989 -2.869057 -5.053901 -5.409665  -5.464025
#> 20  -4.715456  -6.804414 -3.057465 -2.320256 -5.120410 -5.458075  -5.366813
#> 21  -4.814169  -6.857676 -2.545577 -3.217281 -5.026111 -5.383196  -5.457940
#> 22  -5.587790  -7.456255 -2.760044 -3.373302 -4.975434 -5.038445  -5.410894
#> 23  -4.956431  -5.670541 -3.286777 -3.313715 -5.674303 -2.514769  -3.555166
#> 24  -9.478900 -11.858102 -3.338744 -4.170821 -4.524341 -4.573990  -7.127601
#> 25  -4.971742  -5.677143 -3.525440 -3.050904 -5.682425 -2.633076  -3.434711
#> 26  -4.990909  -5.685737 -3.431186 -3.146331 -5.495920 -2.386838  -3.346229
#> 27  -4.996583  -5.688676 -2.873241 -3.029823 -5.695606 -2.021870  -3.853125
#> 28  -4.742219  -5.662525 -3.408559 -3.245950 -5.668053 -2.373943  -3.499142
#> 29  -5.003123  -5.701592 -3.457464 -3.075237 -5.721360 -2.463214  -3.714339
#> 30  -7.754543  -6.843394 -5.378603 -3.153264 -8.608358 -4.190844  -3.180123
#> 31  -7.857033  -6.910069 -5.918532 -4.071652 -8.477155 -5.166205  -4.018122
#> 32 -19.367107 -19.367107 -4.299759 -5.929019 -3.493917 -1.126668  -6.616028
#> 33 -19.367107 -19.367107 -3.833523 -5.643253 -2.432348 -1.887437  -7.015861
#> 34 -19.367107 -19.367107 -4.510286 -6.142702 -4.015863 -2.529970  -7.087399
#> 35 -19.366150 -19.366282 -3.470866 -5.903481 -3.559082 -2.980839  -6.986979
#> 36 -17.821006 -19.442199 -4.355863 -5.858155 -4.214806 -3.655816  -8.573894
#> 37 -19.366573 -19.366647 -3.675804 -5.537188 -3.538590 -2.645003  -6.985727
#> 38 -14.269945 -14.989461 -2.584360 -4.929888 -4.155436 -3.612744  -6.428391
#> 39 -17.087291 -17.410468 -3.441099 -5.430165 -3.893135 -3.127641  -6.725755
#> 40  -8.488147 -19.292864 -4.387367 -5.245119 -4.153687 -5.890189 -17.319211
#> 41 -10.132799 -12.791606 -3.295897 -4.551898 -4.471471 -4.900236  -7.736865
#> 42  -4.473647  -6.786083 -3.106143 -3.413129 -5.053884 -5.100565  -5.453954
#> 43 -17.654602 -17.915504 -2.525431 -5.340075 -4.283406 -3.413607  -6.854971
#> 44  -4.599366  -6.788977 -2.823170 -3.369278 -4.893266 -5.400448  -5.453921
#> 45  -4.488126  -6.792514 -2.560565 -3.333129 -4.915227 -5.414097  -5.466893
#> 46  -4.871179  -6.808374 -3.380181 -2.895330 -4.873066 -5.472109  -5.243807
#> 47  -4.561380  -6.825421 -2.870847 -3.097177 -5.010971 -5.381798  -5.480272
#> 48  -4.780358  -5.505451 -3.209033 -3.402023 -5.672049 -2.336870  -3.835585
#> 49  -5.052446  -5.718471 -3.638554 -3.153439 -5.725209 -1.055879  -3.994076
#>        [,15]      [,16]     [,17]
#> 1  -4.037579  -5.514833 -2.711954
#> 2  -4.871993  -6.208749 -2.844309
#> 3  -4.944650  -7.065609 -3.250574
#> 4  -4.865924  -7.300879 -3.654926
#> 5  -5.083479  -7.272587 -3.215892
#> 6  -4.924507  -6.473584 -2.962181
#> 7  -4.342009  -5.563677 -2.648224
#> 8  -5.218136  -7.212952 -2.729975
#> 9  -4.556084  -5.797241 -2.746258
#> 10 -4.496509  -5.879112 -2.615143
#> 11 -4.500485  -6.133518 -2.653135
#> 12 -4.429221  -6.952344 -2.968163
#> 13 -4.354861  -5.302308 -2.955446
#> 14 -4.179351  -5.440512 -2.404590
#> 15 -5.405506  -7.153268 -2.544899
#> 16 -5.397863  -7.297431 -2.812881
#> 17 -4.371853  -5.457866 -2.703979
#> 18 -4.358689  -5.453020 -2.559139
#> 19 -4.368276  -5.456676 -2.331637
#> 20 -3.897247  -5.502687 -2.440918
#> 21 -4.330864  -5.446620 -2.423501
#> 22 -3.837253  -6.356854 -2.542687
#> 23 -1.575031 -18.260610 -2.323546
#> 24 -4.460048  -6.200294 -2.621902
#> 25 -1.493299 -18.271917 -2.599709
#> 26 -1.228569 -18.284515 -2.145981
#> 27 -1.551766 -18.284515 -2.404973
#> 28 -1.433222 -18.273079 -2.333934
#> 29 -1.500699 -17.951391 -2.418503
#> 30 -1.964506  -7.754543 -5.122210
#> 31 -2.906906  -7.857033 -6.408023
#> 32 -5.114034  -6.312002 -2.563039
#> 33 -3.979990  -6.348033 -2.983056
#> 34 -4.475780  -6.472015 -2.321847
#> 35 -5.063188  -6.294041 -1.539756
#> 36 -5.141446  -6.430490 -1.612266
#> 37 -5.055859  -6.291567 -1.928209
#> 38 -4.693587  -5.994083 -2.185220
#> 39 -4.816547  -6.127865 -2.342203
#> 40 -5.134123  -7.025637 -3.079013
#> 41 -4.738923  -5.506231 -2.104721
#> 42 -4.354525  -5.510128 -1.992686
#> 43 -5.049502  -6.223279 -2.835059
#> 44 -4.343197  -5.447481 -2.918622
#> 45 -4.379235  -5.312554 -2.729808
#> 46 -4.164739  -5.237931 -2.246202
#> 47 -4.288423  -5.442943 -2.714658
#> 48 -2.339970 -18.011040 -2.027470
#> 49 -1.949039 -18.284515 -2.737571
predict(myPLN, trichoptera, "response")
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 
#>  5  1  3  3  3  1  5  3  5  5  5  5  5  5  3  3  5  5  5  5  5  5  4  5  4  4 
#> 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 
#>  4  4  4  2  2  1  1  1  1  1  1  1  1  3  5  5  1  5  5  5  5  4  4 
#> Levels: 1 2 3 4 5