Prediction for a PLNmixturefit object

Predict either posterior probabilities for each group or latent positions based on new samples

# S3 method for PLNmixturefit
predict(
  object,
  newdata,
  type = c("posterior", "response", "position"),
  prior = matrix(rep(1/object$k, object$k), nrow(newdata), object$k, byrow = TRUE),
  control = PLNmixture_param(),
  ...
)

Arguments

object

an R6 object with class PLNmixturefit

newdata

A data frame in which to look for variables, offsets and counts with which to predict.

type

The type of prediction required. The default posterior are posterior probabilities for each group , response is the group with maximal posterior probability and latent is the averaged latent in the latent space, with weights equal to the posterior probabilities.

prior

User-specified prior group probabilities in the new data. The default uses a uniform prior.

control

a list-like structure for controlling the fit. See PLNmixture_param() for details.

...

additional parameters for S3 compatibility. Not used

Value

A matrix of posterior probabilities for each group (if type = "posterior"), a matrix of (average) position in the latent space (if type = "position") or a vector of predicted groups (if type = "response").

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLNmixture(Abundance ~ 1 + offset(log(Offset)),
           data = trichoptera, control = PLNmixture_param(smoothing = "none"))  %>% getBestModel()
#> 
#>  Initialization...
#> 
#>  Adjusting 5 PLN mixture models.
#> 	number of cluster = 1 	number of cluster = 2 	number of cluster = 3 	number of cluster = 4 	number of cluster = 5 
#>  Post-treatments
#>  DONE!
predict(myPLN, trichoptera, "posterior")
#>            [,1]         [,2]         [,3]         [,4]         [,5]
#> 1  6.362185e-02 2.220446e-16 2.220446e-16 1.927801e-03 9.344503e-01
#> 2  9.028956e-01 2.204022e-02 2.220446e-16 1.851494e-05 7.504562e-02
#> 3  2.062385e-01 7.931850e-01 2.220446e-16 2.220446e-16 5.764947e-04
#> 4  3.220276e-05 9.999678e-01 2.220446e-16 2.220446e-16 2.220446e-16
#> 5  7.911705e-03 9.917139e-01 2.220446e-16 2.220446e-16 3.743670e-04
#> 6  4.769583e-01 3.539845e-01 1.008222e-14 1.061825e-05 1.690466e-01
#> 7  1.210880e-01 3.852238e-03 1.340177e-14 2.705193e-03 8.723546e-01
#> 8  2.462863e-02 9.562121e-01 4.028814e-15 1.426900e-06 1.915781e-02
#> 9  4.231463e-01 1.206655e-02 3.010886e-15 1.664569e-04 5.646207e-01
#> 10 3.803850e-01 2.226130e-02 1.266386e-12 8.295335e-03 5.890584e-01
#> 11 2.976133e-01 1.162467e-01 3.294148e-03 1.923323e-02 5.636126e-01
#> 12 1.947197e-01 3.312823e-01 8.597301e-02 4.325339e-02 3.447716e-01
#> 13 2.220446e-16 4.885134e-14 2.220446e-16 2.220446e-16 1.000000e+00
#> 14 2.220446e-16 2.220446e-16 2.220446e-16 3.678101e-06 9.999963e-01
#> 15 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 16 2.023444e-05 9.999798e-01 2.220446e-16 2.220446e-16 2.220446e-16
#> 17 2.220446e-16 2.335674e-16 2.220446e-16 4.025133e-06 9.999960e-01
#> 18 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 19 1.220987e-04 6.900477e-05 2.220446e-16 2.220446e-16 9.998089e-01
#> 20 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 21 5.494778e-03 8.358075e-05 2.220446e-16 4.332121e-06 9.944173e-01
#> 22 5.921643e-02 3.550151e-04 3.770342e-16 6.855669e-02 8.718719e-01
#> 23 2.220446e-16 2.220446e-16 2.220446e-16 9.981407e-01 1.859312e-03
#> 24 3.140838e-01 8.720649e-02 2.508436e-03 2.774836e-02 5.684530e-01
#> 25 2.220446e-16 2.220446e-16 9.071736e-15 9.990195e-01 9.804912e-04
#> 26 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 27 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 28 2.220446e-16 2.220446e-16 5.975068e-04 9.985139e-01 8.885884e-04
#> 29 1.688744e-04 1.457433e-04 2.035822e-02 9.702876e-01 9.039600e-03
#> 30 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 31 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 32 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 33 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 34 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 35 9.999344e-01 2.220446e-16 2.220446e-16 2.220446e-16 6.563027e-05
#> 36 8.786033e-01 1.209436e-01 2.220446e-16 2.220446e-16 4.530257e-04
#> 37 9.999608e-01 2.220446e-16 2.220446e-16 2.220446e-16 3.924729e-05
#> 38 6.524230e-01 2.220446e-16 2.220446e-16 2.358557e-03 3.452185e-01
#> 39 8.445721e-01 4.249359e-05 2.220446e-16 5.400489e-04 1.548454e-01
#> 40 1.539171e-01 7.975642e-01 3.004464e-04 4.102459e-05 4.817721e-02
#> 41 3.519370e-01 1.188625e-01 2.220446e-16 2.220446e-16 5.292005e-01
#> 42 2.220446e-16 2.220446e-16 2.220446e-16 4.344415e-03 9.956556e-01
#> 43 8.821096e-01 2.319318e-03 2.220446e-16 3.489002e-06 1.155676e-01
#> 44 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 45 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 46 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 47 2.220446e-16 3.190910e-03 8.262044e-14 6.828039e-04 9.961263e-01
#> 48 2.220446e-16 2.220446e-16 2.220446e-16 9.787127e-01 2.128725e-02
#> 49 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
predict(myPLN, trichoptera, "position")
#>          [,1]       [,2]      [,3]       [,4]        [,5]      [,6]       [,7]
#> 1   -6.952435  -7.597030 -2.129742  -5.502291 -0.55747858 -3.691462  -5.935701
#> 2  -18.386226 -18.438208 -2.012147  -7.122217 -0.53771032 -3.547862 -18.303714
#> 3  -19.889153 -19.889549 -3.728144 -17.330404 -0.19144853 -3.809499 -19.888529
#> 4  -20.034883 -20.034883 -4.092156 -20.034485 -0.09198931 -3.767888 -20.034883
#> 5  -20.024414 -20.024669 -3.591103 -19.926156 -0.14468138 -3.653258 -20.024019
#> 6  -17.361565 -17.478503 -2.796892 -11.320461 -0.30894859 -3.396929 -17.176315
#> 7   -7.760758  -8.365931 -2.180238  -5.633332 -0.63888963 -3.593593  -6.798386
#> 8  -19.751698 -19.764894 -3.877809 -19.432631 -0.17745510 -3.603111 -19.730919
#> 9  -11.883992 -12.275199 -1.893481  -6.227152 -0.48758735 -3.581663 -11.262784
#> 10 -11.453472 -11.862815 -2.060487  -6.306817 -0.63077961 -3.541512 -10.801842
#> 11 -11.695572 -12.060157 -2.450405  -7.595074 -0.46504512 -3.594540 -11.038049
#> 12 -14.204775 -13.727510 -3.330433 -11.538539 -0.30016491 -3.977641 -12.935454
#> 13  -5.965096  -6.803240 -2.722298  -5.417313 -0.26341489 -3.500598  -5.042602
#> 14  -6.111280  -6.800283 -2.281432  -5.405758 -0.52997607 -3.433071  -4.876841
#> 15 -20.034904 -20.034904 -4.674314 -20.034904 -0.18331417 -3.498295 -20.034904
#> 16 -20.034891 -20.034891 -4.184308 -20.034640 -0.19474917 -3.202596 -20.034891
#> 17  -6.120347  -6.804889 -2.685032  -5.423693 -0.32217330 -3.658238  -5.051694
#> 18  -6.117827  -6.803608 -2.540304  -5.269811 -0.51993008 -3.266443  -5.044636
#> 19  -6.122207  -6.806973 -2.315217  -5.423479 -0.62426238 -3.064435  -5.052468
#> 20  -6.144507  -6.666190 -2.672075  -5.469984 -0.58573044 -3.447892  -4.979668
#> 21  -6.186009  -6.870809 -2.204385  -5.416978 -0.57430117 -3.709080  -5.108159
#> 22  -6.864183  -7.470826 -2.203726  -5.506507 -0.78818945 -3.759044  -5.905349
#> 23  -5.669279  -5.710099 -3.794556  -5.456168 -0.89983170 -5.661476  -5.664019
#> 24 -11.496978 -11.871937 -2.388592  -7.187974 -0.52845041 -3.597258 -10.841753
#> 25  -5.676480  -5.716437 -3.316927  -5.664351 -1.03863455 -5.472207  -5.473509
#> 26  -5.685737  -5.724695 -3.722172  -5.674398 -1.36918752 -5.682563  -5.682563
#> 27  -5.688676  -5.727526 -3.739294  -5.677369 -1.14313568 -5.685511  -5.685511
#> 28  -5.668604  -5.704043 -3.753989  -5.656374 -1.18598993 -5.657478  -5.658212
#> 29  -5.898427  -5.773234 -3.796777  -5.878613 -0.96674506 -5.674786  -5.675184
#> 30 -16.848714  -8.370739 -6.654046 -16.848714 -0.30982049 -7.223520  -6.458342
#> 31 -16.848729  -8.612932 -5.752107 -16.848729 -0.10984059 -7.523765  -6.590055
#> 32 -19.367107 -19.367107 -2.523073  -6.984219 -0.96451269 -3.245682 -19.367107
#> 33 -19.367107 -19.367107 -2.437199  -7.003838 -0.77723464 -2.998319 -19.367107
#> 34 -19.367107 -19.367107 -3.004718  -7.076121 -0.40045420 -2.879843 -19.367107
#> 35 -19.366238 -19.366282 -2.425984  -6.974621 -0.81631330 -2.758952 -19.366168
#> 36 -19.441877 -19.442184 -2.979642  -8.562669 -0.41699007 -3.445618 -19.441402
#> 37 -19.366587 -19.366614 -1.419997  -6.588410 -1.06108310 -3.449182 -19.366545
#> 38 -14.756347 -14.995435 -1.904642  -6.410454 -1.15100752 -3.590235 -14.377281
#> 39 -17.306047 -17.413297 -1.647374  -6.709739 -0.90420212 -3.562679 -17.135908
#> 40 -19.259540 -19.290340 -3.606569 -17.314753 -0.15103646 -3.709604 -19.203714
#> 41 -12.438563 -12.799715 -2.748706  -7.711328 -0.37994658 -3.885159 -11.877034
#> 42  -6.116710  -6.799326 -2.559388  -5.272395 -0.49915572 -3.777176  -5.049466
#> 43 -17.839227 -17.917602 -3.032815  -6.839696 -0.30862877 -3.843672 -17.718334
#> 44  -6.114929  -6.802135 -2.196438  -5.263204 -0.46598068 -3.604995  -4.742374
#> 45  -6.121783  -6.805626 -2.445480  -5.278750 -0.53275958 -3.431192  -4.910911
#> 46  -6.005313  -6.821286 -2.964957  -5.483868 -0.38494270 -3.902682  -5.000967
#> 47  -6.149165  -6.838626 -2.310185  -5.440061 -0.46323052 -3.638454  -5.056115
#> 48  -5.491074  -5.741288 -3.832451  -5.662037 -0.72484676 -5.633685  -5.661983
#> 49  -5.718471  -5.565635 -3.747405  -5.707480 -1.15692325 -5.715394  -5.715394
#>          [,8]       [,9]     [,10]     [,11]     [,12]     [,13]      [,14]
#> 1   -5.657244  -7.584628 -2.875753 -3.464675 -4.791183 -5.248231  -5.371188
#> 2  -17.985234 -18.437215 -3.831800 -5.625793 -3.167818 -3.274604  -7.136733
#> 3   -9.251014 -19.889542 -4.488527 -5.395186 -3.536828 -5.816052 -17.333006
#> 4   -6.675084 -20.034883 -4.708068 -5.439349 -4.339586 -6.519839 -20.034485
#> 5   -6.738047 -20.024664 -4.685711 -5.327400 -4.418417 -6.454570 -19.926268
#> 6  -12.375764 -17.476268 -3.993207 -5.192190 -3.947137 -4.673086 -11.333496
#> 7   -6.479265  -8.354270 -2.717491 -3.509410 -4.835960 -5.074177  -5.666153
#> 8   -6.900246 -19.764641 -4.548925 -5.245753 -4.504846 -6.361004 -19.433725
#> 9  -10.929777 -12.267721 -3.373832 -4.342851 -4.433778 -4.385907  -6.255791
#> 10 -10.316867 -11.854676 -3.286600 -4.215823 -4.450208 -4.406618  -6.320022
#> 11  -9.295658 -12.046600 -3.377581 -4.192000 -4.533093 -4.681359  -7.540962
#> 12  -8.461947 -13.582400 -3.862549 -4.400343 -4.861266 -5.036254 -10.332179
#> 13  -4.608613  -6.790097 -2.971084 -3.395724 -4.900309 -5.404800  -5.458056
#> 14  -4.730848  -6.787101 -2.988618 -3.322059 -5.030488 -5.393093  -5.446942
#> 15  -6.397121 -20.034904 -3.786341 -5.026949 -4.976944 -6.648263 -20.034904
#> 16  -6.668653 -20.034891 -4.354101 -4.857264 -4.452177 -6.512615 -20.034641
#> 17  -4.764612  -6.638599 -2.815504 -3.091425 -5.056066 -5.411246  -5.464185
#> 18  -4.755391  -6.790470 -3.093875 -2.635294 -5.049038 -5.406243  -5.459426
#> 19  -4.763906  -6.793849 -3.010988 -2.869056 -5.053903 -5.409668  -5.464026
#> 20  -4.715456  -6.804414 -3.057465 -2.320256 -5.120410 -5.458075  -5.366813
#> 21  -4.814203  -6.857705 -2.545578 -3.217286 -5.026109 -5.383189  -5.457944
#> 22  -5.588161  -7.456566 -2.760057 -3.373368 -4.975405 -5.038369  -5.410923
#> 23  -4.956431  -5.670541 -3.286777 -3.313715 -5.674304 -2.514768  -3.555166
#> 24  -9.480278 -11.859240 -3.338834 -4.171072 -4.524223 -4.573751  -7.127702
#> 25  -4.971742  -5.677143 -3.525440 -3.050904 -5.682425 -2.633076  -3.434711
#> 26  -4.990909  -5.685737 -3.431186 -3.146331 -5.495920 -2.386838  -3.346229
#> 27  -4.996583  -5.688676 -2.873241 -3.029823 -5.695606 -2.021870  -3.853125
#> 28  -4.743931  -5.663252 -3.409901 -3.246195 -5.669771 -2.375330  -3.499114
#> 29  -5.003157  -5.701607 -3.457487 -3.075241 -5.721387 -2.463234  -3.714337
#> 30  -7.754544  -6.843394 -5.378603 -3.153264 -8.608358 -4.190855  -3.180123
#> 31  -7.857000  -6.910069 -5.918532 -4.071642 -8.477157 -5.166205  -4.018120
#> 32 -19.367107 -19.367107 -4.299759 -5.929019 -3.493917 -1.126668  -6.616028
#> 33 -19.367107 -19.367107 -3.833523 -5.643253 -2.432348 -1.887437  -7.015861
#> 34 -19.367107 -19.367107 -4.510286 -6.142702 -4.015863 -2.529970  -7.087399
#> 35 -19.366149 -19.366282 -3.470866 -5.903481 -3.559082 -2.980837  -6.986979
#> 36 -17.821456 -19.442178 -4.355849 -5.858177 -4.214789 -3.655721  -8.573437
#> 37 -19.366534 -19.366613 -3.675803 -5.537183 -3.538592 -2.645005  -6.985723
#> 38 -14.271476 -14.990776 -2.584380 -4.930161 -4.155298 -3.612488  -6.428552
#> 39 -17.088188 -17.411227 -3.441140 -5.430323 -3.893053 -3.127488  -6.725835
#> 40  -8.488519 -19.289217 -4.387735 -5.244693 -4.154923 -5.889696 -17.314660
#> 41 -10.134244 -12.792786 -3.295942 -4.552131 -4.471363 -4.900048  -7.736935
#> 42  -4.473647  -6.786082 -3.106143 -3.413129 -5.053884 -5.100564  -5.453954
#> 43 -17.655295 -17.916095 -2.525433 -5.340176 -4.283362 -3.413520  -6.855033
#> 44  -4.599366  -6.788977 -2.823170 -3.369278 -4.893266 -5.400448  -5.453921
#> 45  -4.488126  -6.792514 -2.560565 -3.333129 -4.915227 -5.414097  -5.466893
#> 46  -4.871179  -6.808374 -3.380181 -2.895330 -4.873066 -5.472109  -5.243807
#> 47  -4.561381  -6.825426 -2.870847 -3.097178 -5.010970 -5.381798  -5.480278
#> 48  -4.780358  -5.505448 -3.209034 -3.402023 -5.672051 -2.336863  -3.835581
#> 49  -5.052446  -5.718471 -3.638554 -3.153439 -5.725209 -1.055879  -3.994076
#>        [,15]      [,16]     [,17]
#> 1  -4.037585  -5.514858 -2.711958
#> 2  -4.872011  -6.208768 -2.844314
#> 3  -4.944630  -7.065555 -3.250565
#> 4  -4.865924  -7.300878 -3.654926
#> 5  -5.083479  -7.272585 -3.215892
#> 6  -4.924521  -6.473572 -2.962179
#> 7  -4.342040  -5.563720 -2.648230
#> 8  -5.218136  -7.212948 -2.729974
#> 9  -4.556154  -5.797331 -2.746276
#> 10 -4.496576  -5.879191 -2.615149
#> 11 -4.500550  -6.133577 -2.653138
#> 12 -4.429179  -6.952394 -2.968272
#> 13 -4.354861  -5.302308 -2.955446
#> 14 -4.179351  -5.440512 -2.404590
#> 15 -5.405506  -7.153268 -2.544899
#> 16 -5.397863  -7.297425 -2.812881
#> 17 -4.371853  -5.457866 -2.703979
#> 18 -4.358689  -5.453020 -2.559139
#> 19 -4.368276  -5.456676 -2.331637
#> 20 -3.897247  -5.502687 -2.440918
#> 21 -4.330866  -5.446621 -2.423501
#> 22 -3.837243  -6.356946 -2.542686
#> 23 -1.575031 -18.260613 -2.323541
#> 24 -4.460113  -6.200358 -2.621918
#> 25 -1.493299 -18.271917 -2.599709
#> 26 -1.228569 -18.284515 -2.145981
#> 27 -1.551766 -18.284515 -2.404973
#> 28 -1.433588 -18.266825 -2.335867
#> 29 -1.500697 -17.951303 -2.418536
#> 30 -1.964506  -7.754544 -5.122210
#> 31 -2.906906  -7.857000 -6.408023
#> 32 -5.114034  -6.312002 -2.563039
#> 33 -3.979990  -6.348033 -2.983056
#> 34 -4.475780  -6.472015 -2.321847
#> 35 -5.063188  -6.294041 -1.539756
#> 36 -5.141440  -6.430456 -1.612261
#> 37 -5.055857  -6.291565 -1.928209
#> 38 -4.693654  -5.994163 -2.185199
#> 39 -4.816585  -6.127911 -2.342196
#> 40 -5.133374  -7.025842 -3.079819
#> 41 -4.738985  -5.506261 -2.104707
#> 42 -4.354523  -5.510134 -1.992686
#> 43 -5.049535  -6.223318 -2.835064
#> 44 -4.343197  -5.447481 -2.918622
#> 45 -4.379235  -5.312554 -2.729808
#> 46 -4.164739  -5.237931 -2.246202
#> 47 -4.288423  -5.442945 -2.714658
#> 48 -2.339965 -18.011070 -2.027469
#> 49 -1.949039 -18.284515 -2.737571
predict(myPLN, trichoptera, "response")
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 
#>  5  1  2  2  2  1  5  2  5  5  5  5  5  5  2  2  5  5  5  5  5  5  4  5  4  4 
#> 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 
#>  4  4  4  3  3  1  1  1  1  1  1  1  1  2  5  5  1  5  5  5  5  4  4 
#> Levels: 1 2 3 4 5