Prediction for a PLNmixturefit object

Predict either posterior probabilities for each group or latent positions based on new samples

# S3 method for PLNmixturefit
predict(
  object,
  newdata,
  type = c("posterior", "response", "position"),
  prior = matrix(rep(1/object$k, object$k), nrow(newdata), object$k, byrow = TRUE),
  control = list(),
  ...
)

Arguments

object
an R6 object with class PLNmixturefit

newdata
A data frame in which to look for variables, offsets and counts with which to predict.

type
The type of prediction required. The default posterior are posterior probabilities for each group , response is the group with maximal posterior probability and latent is the averaged latent in the latent space, with weights equal to the posterior probabilities.

prior
User-specified prior group probabilities in the new data. The default uses a uniform prior.

control
a list for controlling the optimization. See PLN() for details.

...
additional parameters for S3 compatibility. Not used

Value

A matrix of posterior probabilities for each group (if type = "posterior"), a matrix of (average) position in the latent space (if type = "position") or a vector of predicted groups (if type = "response").

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
fits <- PLNmixture(Abundance ~ 1, clusters = 1:4, data = trichoptera)

#> 
#>  Initialization...
#> 
#>  Adjusting 4 PLN mixture models.
#> 	number of cluster = 1 	number of cluster = 2 	number of cluster = 3 	number of cluster = 4 
#> 
#>  Smoothing PLN mixture models.
#>    Going backward +++                                                                                                       Going forward +++                                                                                                       Going backward +++                                                                                                       Going forward +++                                                                                                    
#>  Post-treatments
#>  DONE!

myMixture <- getBestModel(fits)
predict(myMixture, trichoptera, "posterior")

#>            [,1]         [,2]         [,3]         [,4]
#> 1  2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 2  2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 3  3.085464e-07 2.220446e-16 9.999997e-01 2.220446e-16
#> 4  2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 5  7.819218e-01 2.220446e-16 2.180782e-01 3.694070e-10
#> 6  5.160415e-14 2.220446e-16 1.000000e+00 2.220446e-16
#> 7  2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 8  3.075177e-04 2.220446e-16 9.996925e-01 2.220446e-16
#> 9  2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 10 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 11 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 12 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 13 2.553053e-07 2.220446e-16 9.999997e-01 1.048441e-13
#> 14 3.048951e-06 2.220446e-16 9.999970e-01 2.220446e-16
#> 15 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 17 1.000000e+00 2.220446e-16 2.220446e-16 4.711336e-08
#> 18 9.963515e-01 2.220446e-16 1.273870e-14 3.648473e-03
#> 19 9.923033e-01 2.220446e-16 2.220446e-16 7.696697e-03
#> 20 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 21 1.106646e-04 2.220446e-16 9.998893e-01 2.220446e-16
#> 22 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 23 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 24 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 25 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 26 2.220446e-16 2.220446e-16 2.359743e-14 1.000000e+00
#> 27 2.220446e-16 2.220446e-16 8.896682e-16 1.000000e+00
#> 28 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 29 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 30 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 31 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 32 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 33 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 34 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 35 2.275844e-01 2.220446e-16 2.220446e-16 7.724156e-01
#> 36 9.738889e-01 2.220446e-16 2.220446e-16 2.611113e-02
#> 37 1.203751e-05 2.220446e-16 2.220446e-16 9.999880e-01
#> 38 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 39 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 40 4.049587e-13 2.220446e-16 1.000000e+00 2.220446e-16
#> 41 9.999890e-01 2.220446e-16 2.220446e-16 1.104276e-05
#> 42 9.998113e-01 2.220446e-16 5.919839e-13 1.887055e-04
#> 43 9.123119e-01 2.220446e-16 2.220446e-16 8.768814e-02
#> 44 1.079045e-02 2.220446e-16 9.892096e-01 2.220446e-16
#> 45 9.999229e-01 2.220446e-16 2.889597e-16 7.710317e-05
#> 46 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 47 1.279258e-15 2.220446e-16 1.000000e+00 2.220446e-16
#> 48 7.775731e-12 2.220446e-16 1.000000e+00 2.220446e-16
#> 49 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00

predict(myMixture, trichoptera, "position")

#>          [,1]        [,2]      [,3]      [,4]     [,5]       [,6]       [,7]
#> 1   -2.601899 -14.8265269 0.8473251 -2.601988 2.617397 -1.2479105 -1.9267890
#> 2   -2.601906 -14.8265269 0.7163164 -2.601996 2.110684 -1.2479398 -1.9268038
#> 3   -2.601891 -14.8265270 0.5857443 -2.601976 3.400703 -1.2478617 -1.9267650
#> 4   -2.331985  -1.6861989 1.2566729 -4.128497 5.120757  1.3664662 -1.6849345
#> 5  -12.823470 -15.0420267 1.3237806 -1.567396 3.936630  0.5840682 -2.4440680
#> 6   -2.601902 -14.8265269 0.6509240 -2.601992 2.450901 -1.1803654 -1.9267952
#> 7   -2.601909 -14.8265269 0.6507378 -2.601999 1.879032 -1.2479512 -1.9268095
#> 8   -2.605909 -14.8266116 0.5859453 -2.601572 3.251743 -1.1796304 -1.9269735
#> 9   -2.601906 -14.8265269 0.7817561 -2.601996 2.110695 -1.2479395 -1.9268036
#> 10  -2.601911 -14.8265269 0.6506990 -2.602000 1.761922 -1.2479571 -1.9268125
#> 11  -2.601910 -14.8265269 0.5851469 -2.602000 1.820554 -1.2479551 -1.9268114
#> 12  -2.601910 -14.8265269 0.5195162 -2.602000 1.820537 -1.2479557 -1.9268117
#> 13  -2.534049 -14.8265270 0.5859403 -2.601968 3.657922 -1.1127233 -1.9267496
#> 14  -2.601936 -14.8265277 0.7820193 -2.601982 2.780960 -1.1127894 -1.8590456
#> 15  -2.331066  -1.6844855 1.2707699 -4.128342 6.191634  2.7655099 -1.6832191
#> 16  -2.332184  -1.6865695 1.0701167 -4.128531 4.900581  2.0006415 -1.6853055
#> 17 -15.674282 -15.1021291 1.4745035 -1.278863 3.888881  1.0572667 -2.5883662
#> 18 -15.625612 -15.0531971 1.4735902 -1.287681 3.887054  1.0582789 -2.5850852
#> 19 -15.571607 -14.9989006 1.4758862 -1.300798 3.884825  1.0616401 -2.5814421
#> 20  -2.332793  -1.2855308 1.8028363 -4.128634 4.189588  1.5040359 -1.2843087
#> 21  -2.603343 -14.8265574 0.8474685 -2.601839 2.781126 -1.2476448 -1.9268569
#> 22  -2.601908 -14.8265269 0.7162679 -2.601998 1.937312 -1.2479471 -1.9268074
#> 23  -2.601906 -14.8265269 0.5196162 -2.534163 2.110682 -1.2479399 -1.9268038
#> 24  -2.601911 -14.8265269 0.5851297 -2.602001 1.761906 -1.2479577 -1.9268128
#> 25  -2.601904 -14.8265269 0.7163698 -2.601994 2.225191 -1.1803746 -1.8590591
#> 26  -2.334644  -1.6911504 0.8334300 -4.128948 2.415657  0.5257879 -1.6898920
#> 27  -2.334467  -1.6908180 0.8359173 -4.128918 2.721294  0.5279042 -1.6895592
#> 28  -2.601911 -14.8265269 0.5195069 -2.602000 1.761922 -1.2479571 -1.9268125
#> 29  -2.601909 -14.8265269 0.5195405 -2.601999 1.879015 -1.2479518 -1.9268098
#> 30  -2.330731  -1.2802224 0.6513170 -4.128285 7.056472  1.0259707 -0.5138182
#> 31  -7.693205  -0.6964936 1.7892731 -7.693205 7.846182  0.4021840  1.2503098
#> 32  -2.333249  -1.6885524 1.8022989 -4.128712 3.464705  1.3487006 -1.6872909
#> 33  -2.332619  -1.6873782 2.2942106 -4.128605 4.072760  1.8896820 -1.6861152
#> 34  -2.331492  -1.6852799 2.5349868 -4.128414 5.356350  2.8097703 -1.6840145
#> 35  -5.369851  -4.7420004 1.6272511 -3.480197 3.462635  1.4995463 -1.8930892
#> 36 -15.325929 -14.7518789 1.4726363 -1.353275 3.892562  1.0605312 -2.5648310
#> 37  -2.334043  -1.6898910 2.6687702 -3.695813 3.054089  0.9828937 -1.6884802
#> 38  -2.601909 -14.8265269 0.7816766 -2.601999 1.820662 -1.2479511 -1.9268095
#> 39  -2.601908 -14.8265269 0.8470770 -2.601998 1.937335 -1.2479462 -1.9268070
#> 40  -2.601899 -14.8265269 0.5854301 -2.601989 2.672114 -1.2479111 -1.9267893
#> 41 -15.674135 -15.1019816 1.4745026 -1.278894 3.888880  1.0572609 -2.5883563
#> 42 -15.671765 -15.0995987 1.4744566 -1.279319 3.888797  1.0571677 -2.5881964
#> 43 -14.504416 -13.9258881 1.4539509 -1.528758 3.915339  1.0330725 -2.5093226
#> 44  -2.742946 -14.8295008 0.9192945 -2.520421 3.202884 -1.1567080 -1.7993988
#> 45 -15.673254 -15.1010956 1.4745078 -1.279049 3.888855  1.0572754 -2.5882659
#> 46  -1.914429  -1.6870128 1.4138683 -4.128572 4.576509  0.7893895 -1.2833407
#> 47  -2.601904 -14.8265269 0.6508663 -2.601994 2.281953 -1.2479316 -1.9267996
#> 48  -2.534067 -14.8265269 0.5197903 -2.601989 2.617333 -1.2479124 -1.9267899
#> 49  -2.333291  -1.2867887 1.0551217 -4.128719 3.283768  0.5407832 -1.6873684
#>            [,8]       [,9]         [,10]      [,11]      [,12]       [,13]
#> 1  -1.695689245 -3.2914999 -0.0776887755 -1.7899178 -0.8697486 -1.70244121
#> 2  -1.695707880 -3.2915037 -0.2114712650 -1.7899348 -0.8023837 -1.70245971
#> 3  -1.695658514 -3.2914936 -0.2112466244 -1.7898893 -0.6675022 -1.70241029
#> 4  -1.336087015 -8.8998900  0.6309513188  0.4489903  0.9445380  0.89679370
#> 5  -1.339558023 -2.7694665  0.7611514055 -0.0672809 -0.3318015 -0.35749627
#> 6  -1.695697101 -3.2915015 -0.2114221365 -1.7899250 -0.8697659 -1.70244901
#> 7  -1.695715136 -3.2915052 -0.0778113212 -1.7899414 -0.9372254 -1.70246692
#> 8  -1.695525498 -3.2912899 -0.2108980120 -1.7892205 -0.9368864 -1.70188973
#> 9  -1.695707651 -3.2915036 -0.2114702194 -1.7899346 -0.9372093 -1.70245949
#> 10 -1.695718893 -3.2915059 -0.2115214672 -1.7899448 -0.9372335 -1.70247065
#> 11 -1.695717582 -3.2915056 -0.2115154900 -1.7899436 -0.9372307 -1.70246935
#> 12 -1.695717962 -3.2915057 -0.2115172205 -1.7899440 -0.9372315 -1.70246973
#> 13 -1.627946177 -3.2914897 -0.0774518699 -1.7898719 -0.8696385 -1.70239117
#> 14 -1.695681879 -3.2914967 -0.1444915788 -1.7899057 -0.9371544 -1.70243002
#> 15 -0.568771442 -8.8998887  2.6074251733  1.4755080  0.7342141  0.90898736
#> 16 -1.336601601 -8.8998903  1.0625911285  1.1210857  0.7206437  0.89396535
#> 17 -1.240269631 -2.6238789  1.0322252108  0.4131296 -0.1818486  0.01757275
#> 18 -1.240640056 -2.6467765  1.0306707464  0.4176298 -0.1805358  0.02067806
#> 19 -1.241047689 -2.6721832  1.0307206299  0.4209415 -0.1790687  0.02413866
#> 20 -0.949060056 -8.8998912  1.3999108790  2.6492180  0.1899592  0.88510204
#> 21 -1.695632080 -3.2914247  0.1227581731 -1.7219619 -0.9370714 -1.70224414
#> 22 -1.695712495 -3.2915046 -0.0777988221 -1.7899390 -0.9372197 -1.70246430
#> 23 -1.695707919 -3.2915037 -0.1446095782 -1.7899348 -0.9372098 -1.70245975
#> 24 -1.695719279 -3.2915060 -0.2115232243 -1.7899452 -0.9372344 -1.70247103
#> 25 -1.695702806 -3.2915027 -0.2114481392 -1.6545101 -0.9371988 -1.70245467
#> 26 -1.342962903 -8.8998938  0.6005076621  0.8890045  0.4465750  1.39970751
#> 27 -1.342499216 -8.8998936  1.3772659612  1.0881801  0.1753033  1.90634891
#> 28 -1.628027441 -3.2915059 -0.2115214711 -1.7899448 -0.9372335 -1.70247065
#> 29 -1.695715524 -3.2915052 -0.2115061088 -1.7222289 -0.9372263 -1.70246731
#> 30 -0.941792413 -8.0275928  1.9576541051  4.1647920  0.2089228  3.19926931
#> 31 -0.002789697  0.9138627  2.0754694405  4.0770062 -0.6956058  3.03293657
#> 32 -1.339354496 -8.8998918  0.3593366275  0.4358804  0.9278125  3.34624505
#> 33 -1.337723426 -8.8998909  1.0575019101  0.6925487  2.3989217  2.99711317
#> 34 -1.334809426 -8.8998893  0.8689355212  0.4542011  1.3241628  3.18200799
#> 35 -1.317812257 -7.4715695  1.0413480264  0.4267062  0.5009796  1.09622194
#> 36 -1.242846999 -2.7877524  1.0146921255  0.4137620 -0.1722060  0.05009950
#> 37 -1.340986017 -8.8998172  0.8384892641  0.6776717  0.7006702  1.68754567
#> 38 -1.695715083 -3.2915051  0.0557588014 -1.7899413 -0.9372253 -1.63477374
#> 39 -1.695711970 -3.2915045 -0.1446283880 -1.7899385 -0.9372186 -1.63477059
#> 40 -1.695689643 -3.2915000 -0.2113881530 -1.7899182 -0.8697495 -1.70244160
#> 41 -1.240270730 -2.6239480  1.0322253618  0.4131298 -0.1818415  0.01758220
#> 42 -1.240145820 -2.6250630  1.0321450591  0.4131323 -0.1817804  0.01780513
#> 43 -1.248916512 -3.1742105  1.1266947256  0.4371572 -0.1494324  0.14091182
#> 44 -1.623611086 -3.2843111  0.0004827538 -1.7665424 -0.8622452 -1.68413910
#> 45 -1.240218871 -2.6243626  1.0322740618  0.4131499 -0.1817996  0.01763858
#> 46 -1.337216807 -8.8998906  0.8569579110  2.0576446  0.7179861  0.89116030
#> 47 -1.628010925 -3.2915027 -0.1445852833 -1.7222170 -0.9371986 -1.70245456
#> 48 -1.627998420 -3.2236203 -0.0776943516 -1.7899189 -0.9371721 -1.49938177
#> 49 -1.339461872 -8.8998919  0.6160799410  1.2812785  0.1858025  3.43397558
#>           [,14]         [,15]        [,16]       [,17]
#> 1   -1.63064675 -0.1329822612 -3.291449497 -0.03547493
#> 2   -1.69835798 -0.2668798252 -3.291453248 -0.10230636
#> 3   -1.69831303 -0.1997432293 -3.291442794 -0.10205433
#> 4   -0.64569711  1.2653295677 -1.337253937  1.38026171
#> 5  -12.17902719 -1.0327800015 -1.911626918  1.57461520
#> 6   -1.69834723 -0.2668334541 -3.291451079 -0.10225128
#> 7   -1.69836522 -0.2669110432 -3.291454709 -0.10234344
#> 8   -1.70243727 -0.2670035082 -3.290902283  0.09860338
#> 9   -1.69835776 -0.2668788382 -3.291453202 -0.10230519
#> 10  -1.69836897 -0.2669272090 -3.291455465 -0.10236264
#> 11  -1.69836766 -0.2669215675 -3.291455201 -0.10235594
#> 12  -1.69836804 -0.2669232009 -3.291455278 -0.10235788
#> 13  -1.69829302 -0.2665852108 -3.223558933 -0.10195564
#> 14  -1.69837430 -0.1998541686 -3.291442914  0.16456929
#> 15  -0.64127500  1.0961675499 -0.944094345  3.81383648
#> 16  -0.64664853  0.6463334627 -1.337767968  2.36014950
#> 17 -15.10212900 -1.2652374679 -1.526801499  2.00487152
#> 18 -15.04942407 -1.2583457112 -1.526130768  2.00350797
#> 19 -14.99093766 -1.2506854342 -1.525383159  2.00465180
#> 20  -0.29956476  1.5607061477 -1.339348101  2.41216127
#> 21  -1.69981597 -0.2668810534 -3.291252851  0.16477202
#> 22  -1.69836259 -0.1330861371 -3.291454177 -0.03559598
#> 23  -1.63066573  0.0006320628 -3.291453256  0.03112940
#> 24  -1.69836935 -0.2669288676 -3.291455542 -0.10236461
#> 25  -1.56297481  0.2009614427 -3.291452227 -0.03554553
#> 26   0.30442725  2.4061403674 -1.344122404  1.86056265
#> 27  -0.65753032  2.0929223538 -1.343659226  1.62731127
#> 28  -1.63067689 -0.0662515490 -3.291455465 -0.03562928
#> 29  -1.69836561 -0.0662362239 -3.291454787 -0.03561174
#> 30   4.08015841  5.3833438439 -0.942914327  2.53383378
#> 31   4.08852542  5.2260042946 -0.002789697  2.00328183
#> 32  -0.30225385  0.6342026157 -1.340517894  1.99027040
#> 33  -0.64871221  1.5635201419 -1.338888587  1.89630964
#> 34  -0.64332819  1.5812243880 -1.335977733  3.40398603
#> 35  -3.94226960  0.1974920592 -1.383919994  2.55160677
#> 36 -14.72479535 -1.2155986231 -1.521927597  2.02888901
#> 37  -0.65490756  0.6270096543 -1.342151094  2.32700379
#> 38  -1.69836517 -0.2669108216 -3.291454698  0.09775796
#> 39  -1.69836206 -0.2668974257 -3.291454071  0.03110792
#> 40  -1.69833980 -0.2668013775 -3.291449577 -0.10221319
#> 41 -15.10197013 -1.2652166218 -1.526791069  2.00487517
#> 42 -15.09940343 -1.2648808381 -1.526766679  2.00493169
#> 43 -13.83492351 -1.0984963861 -1.510425007  1.98462145
#> 44  -1.84294718 -0.2775434654 -3.272463514 -0.08138042
#> 45 -15.10101576 -1.2650916956 -1.526757347  2.00483261
#> 46   0.02840756  1.2587849702 -0.576172066  2.79340465
#> 47  -1.69835281 -0.2668574839 -3.291452203 -0.10227982
#> 48  -1.69834057 -0.2668047334 -3.291449734  0.29767707
#> 49  -0.65192608  2.2030583778 -1.340625155  1.64586295

predict(myMixture, trichoptera, "response")

#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 
#>  3  3  3  4  1  3  3  3  3  3  3  3  3  3  4  4  1  1  1  4  3  3  3  3  3  4 
#> 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 
#>  4  3  3  4  2  4  4  4  4  1  4  3  3  3  1  1  1  3  1  4  3  3  4 
#> Levels: 1 2 3 4