Prediction for a PLNmixturefit object

Predict either posterior probabilities for each group or latent positions based on new samples

# S3 method for PLNmixturefit
predict(
  object,
  newdata,
  type = c("posterior", "response", "position"),
  prior = matrix(rep(1/object$k, object$k), nrow(newdata), object$k, byrow = TRUE),
  control = list(),
  ...
)

Arguments

object
an R6 object with class PLNmixturefit

newdata
A data frame in which to look for variables, offsets and counts with which to predict.

type
The type of prediction required. The default posterior are posterior probabilities for each group , response is the group with maximal posterior probability and latent is the averaged latent in the latent space, with weights equal to the posterior probabilities.

prior
User-specified prior group probabilities in the new data. The default uses a uniform prior.

control
a list for controlling the optimization. See PLN() for details.

...
additional parameters for S3 compatibility. Not used

Value

A matrix of posterior probabilities for each group (if type = "posterior"), a matrix of (average) position in the latent space (if type = "position") or a vector of predicted groups (if type = "response").

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
fits <- PLNmixture(Abundance ~ 1, clusters = 1:4, data = trichoptera)

#> 
#>  Initialization...
#> 
#>  Adjusting 4 PLN mixture models.
#> 	number of cluster = 1 	number of cluster = 2 	number of cluster = 3 	number of cluster = 4 
#> 
#>  Smoothing PLN mixture models.
#>    Going backward +++                                                                                                       Going forward +++                                                                                                       Going backward +++                                                                                                       Going forward +++                                                                                                    
#>  Post-treatments
#>  DONE!

myMixture <- getBestModel(fits)
predict(myMixture, trichoptera, "posterior")

#>            [,1]         [,2]         [,3]         [,4]
#> 1  2.220446e-16 1.000000e+00 4.382172e-12 6.382088e-16
#> 2  2.220446e-16 2.336003e-12 1.000000e+00 2.201069e-09
#> 3  2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 4  2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 5  2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 6  2.220446e-16 7.232890e-08 9.999999e-01 4.311223e-12
#> 7  2.220446e-16 6.509455e-14 1.237196e-01 8.762804e-01
#> 8  2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 9  2.220446e-16 1.275822e-12 1.000000e+00 2.839926e-09
#> 10 2.220446e-16 2.220446e-16 5.169920e-01 4.830080e-01
#> 11 2.220446e-16 2.220446e-16 3.210931e-01 6.789069e-01
#> 12 2.220446e-16 2.220446e-16 5.623728e-01 4.376272e-01
#> 13 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 14 2.220446e-16 1.000000e+00 2.697107e-16 2.220446e-16
#> 15 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 17 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 18 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 19 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 20 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 21 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 22 2.220446e-16 5.859071e-11 1.172195e-02 9.882780e-01
#> 23 2.220446e-16 7.055623e-12 1.000000e+00 2.993440e-10
#> 24 2.220446e-16 2.220446e-16 5.611620e-01 4.388380e-01
#> 25 2.220446e-16 1.792636e-09 2.005366e-11 1.000000e+00
#> 26 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 27 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 28 2.220446e-16 2.220446e-16 4.308753e-02 9.569125e-01
#> 29 2.220446e-16 6.099839e-14 6.796683e-02 9.320332e-01
#> 30 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 31 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 32 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 33 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 34 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 35 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 36 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 37 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 38 2.220446e-16 2.108411e-11 8.550468e-08 9.999999e-01
#> 39 2.220446e-16 3.334055e-10 6.981899e-03 9.930181e-01
#> 40 2.220446e-16 1.254740e-06 9.999987e-01 2.186009e-14
#> 41 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 42 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 43 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 44 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 45 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 46 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 47 2.220446e-16 4.928642e-10 1.000000e+00 1.933603e-10
#> 48 2.220446e-16 1.000000e+00 3.388226e-14 2.220446e-16
#> 49 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16

predict(myMixture, trichoptera, "position")

#>          [,1]         [,2]        [,3]       [,4]       [,5]         [,6]
#> 1  -2.3024550 -2.425401955  1.18560214 -2.0478779 2.45357428 -0.007758963
#> 2  -1.6151091 -1.615811028  0.42693302 -1.6151799 1.84074542 -1.611157155
#> 3  -2.2968928 -2.420489834  0.52782591 -2.0406785 3.19679997  0.054239439
#> 4  -2.2897513 -2.414185533  1.08014513 -2.0314263 5.11092697  1.093431232
#> 5  -2.2924412 -2.416559718  1.22450284 -2.0349124 4.09750724  0.596864822
#> 6  -1.6151014 -1.615803370  0.39985084 -1.6151722 2.00063121 -0.620337479
#> 7  -1.7157274 -1.716184222  0.02684663 -1.6759226 0.69580827 -1.671525376
#> 8  -2.2984246 -2.421842443  0.50318544 -2.0426620 3.05991726  0.279719907
#> 9  -1.6151409 -1.615842845  0.45381814 -1.6152117 1.83973267 -1.611189120
#> 10 -1.5539032 -1.554549525  0.09681534 -1.5286525 0.29303761 -1.523801209
#> 11 -1.6283423 -1.628904853 -0.43156482 -1.5963731 0.61461049 -1.591712434
#> 12 -1.5441472 -1.544807958 -0.90556573 -1.5206270 0.75307232 -1.515767782
#> 13 -1.9879247 -2.419253060  0.55087497 -2.0388644 3.46340018  0.553971459
#> 14 -2.3007127 -2.423863110  1.10235226 -2.0456235 2.65915344  0.481502051
#> 15 -2.2890462 -2.413563330  1.09606902 -2.0305118 6.20943437  2.826216518
#> 16 -2.2900521 -2.414450964  0.86303647 -2.0318161 4.88797233  1.999710207
#> 17 -2.2947137 -2.418566019  0.78026490 -2.0378563 3.58681807  0.326779304
#> 18 -2.2964424 -2.420092206  0.95196981 -1.7344598 3.19945530  0.980333832
#> 19 -2.2961802 -2.419860783  1.49215054 -2.0397556 3.18698815  1.376012212
#> 20 -2.2915002 -2.106875393  1.76416472 -2.0336924 4.17712016  1.276723501
#> 21 -2.3002301 -2.423436913  1.23715649 -2.0449988 2.68222024  0.016889602
#> 22 -1.7681260 -1.768517034  0.34399104 -1.7236597 0.87135918 -1.719435000
#> 23 -1.6150693 -1.615771261 -1.16369667 -0.6241873 1.84201123 -1.611117203
#> 24 -1.5471865 -1.547846267 -0.36156435 -1.5238459 0.27419363 -1.519001935
#> 25 -1.7470879 -1.747481723  0.47914805 -1.7009326 1.78903398 -0.676630044
#> 26 -2.3020092 -2.425008458  0.44907460 -2.0473001 1.96223444 -0.002350068
#> 27 -2.3007139 -2.423864269  0.46874103 -2.0456244 2.38053563  0.011751764
#> 28 -1.7668570 -1.767261892 -1.23225417 -1.7242273 0.05224387 -1.720004587
#> 29 -1.7549844 -1.755404036 -1.21634540 -1.7134972 0.63037874 -1.709241893
#> 30 -0.4186995 -0.003070577  1.37557198 -0.4186995 7.06463640  0.619959981
#> 31 -1.7025111 -1.702922353  2.38324293 -1.6542720 7.88547634 -0.571651513
#> 32 -2.2944670 -2.418348355  1.69759613 -2.0375359 3.34375350  1.019095510
#> 33 -2.2915225 -2.415749222  2.36037358 -2.0337206 4.02151909  1.825951553
#> 34 -2.2891950 -2.413694675  2.62516188 -2.0307045 5.35603823  2.846760602
#> 35 -2.2954968 -2.419257511  1.51161320 -2.0388699 3.17437694  1.392365384
#> 36 -2.2921967 -2.416344158  1.23275643 -2.0345948 3.97577488  0.833535291
#> 37 -2.2960197 -2.419719601  2.60623877 -1.7337442 2.86242362  0.549602617
#> 38 -1.7798402 -1.780221647  0.57290651 -1.7351671 0.48015849 -1.730989404
#> 39 -1.7749597 -1.775346026  0.76843315 -1.7305179 0.87310323 -1.726322434
#> 40 -1.6151631 -1.615865214 -0.18532661 -1.6152336 2.10421499 -1.611207815
#> 41 -2.2936676 -2.417642563  1.20121697 -2.0365009 3.72892575  0.091178473
#> 42 -2.2958214 -2.419543941  0.96467676 -1.7335091 3.27949982  0.066506781
#> 43 -2.2920248 -2.416192255  1.03752106 -2.0343727 4.14938008  0.361047878
#> 44 -2.2981119 -2.421566309  1.44515008 -1.7369718 3.05059064  0.283976720
#> 45 -2.2953961 -2.419168466  1.34414027 -1.7328708 3.40183352  0.783184145
#> 46 -1.9823162 -2.414976598  1.26209428 -2.0325877 4.53687955  0.378562671
#> 47 -1.6150779 -1.615779920  0.40002679 -1.6151487 1.92168820 -1.611125901
#> 48 -1.9955059 -2.424976033  0.23051540 -2.0472539 2.47346961 -0.002426950
#> 49 -2.2949954 -2.110403505  0.77910805 -2.0382201 3.13698599  0.076364213
#>          [,7]       [,8]       [,9]       [,10]         [,11]       [,12]
#> 1  -1.8512220 -1.5610980 -2.4351330  0.44269099 -0.2784403998 -0.15824729
#> 2  -1.6110255 -1.5623979 -1.6152808 -1.43441553 -1.4833978732 -0.26140573
#> 3  -1.8424288 -1.5492703 -2.4302689  0.03559810 -0.2326588591  0.64955586
#> 4  -1.8311178 -1.5340292 -2.4240265  0.37086019 -0.1725274148  0.47805575
#> 5  -1.8353812 -1.5397779 -2.4263773  0.08539243 -0.1953765089 -0.07777379
#> 6  -1.6110177 -1.5623898 -1.6152731 -1.43440613 -1.4833889474 -0.57200193
#> 7  -1.7109955 -1.6222621 -1.7158367 -0.27840755 -1.6845416073 -1.62023851
#> 8  -1.8448523 -1.5525326 -2.4316083  0.01863932 -0.2453886060 -0.39481323
#> 9  -1.6110574 -1.5624315 -1.6153126 -1.43445367 -1.4834341956 -1.56176346
#> 10 -1.5488286 -1.4672613 -1.5540598 -1.32792916 -1.4657920833 -1.46561306
#> 11 -1.6234025 -1.5382738 -1.6284779 -1.41108474 -1.5654516234 -1.53645085
#> 12 -1.5390785 -1.4589472 -1.5443073 -1.31802335 -1.4509594024 -1.45735205
#> 13 -1.8402119 -1.2454025 -2.4290443  0.53779127 -0.2209679409 -0.10241238
#> 14 -1.5457047 -1.5573976 -2.4336092  0.23504311 -0.2642018923 -0.41087358
#> 15 -1.8299988 -0.9285406 -2.4234104  2.73733808  1.1088114259  0.22155636
#> 16 -1.8315947 -1.5346724 -2.4242893  0.84901650  0.6123563988  0.21193691
#> 17 -1.8389796 -1.5446254 -2.1199317  0.99669671  0.5574016304 -0.36845330
#> 18 -1.8417158 -1.5483102 -2.4298752  0.28749776  1.4072362521 -0.38075436
#> 19 -1.8413007 -1.5477510 -2.4296460  0.52905235  1.0030180436 -0.37887741
#> 20 -1.5287215 -1.2354949 -2.4255551  1.26424549  2.6887109967 -0.34536750
#> 21 -1.8477063 -1.5563712 -2.4331871  1.10163333 -0.0008184035 -0.40747689
#> 22 -1.7635345 -1.6728673 -1.7682189 -0.16928829 -1.7565052394 -1.67072717
#> 23 -1.6109855 -1.5623560 -1.6152410 -0.44765808 -1.4833524750 -1.56168790
#> 24 -1.5421330 -1.4623512 -1.5473465 -1.32179159 -1.4539234646 -1.46076546
#> 25 -0.7138266 -1.6490630 -1.7471813 -1.54724725 -0.1953536335 -1.64684981
#> 26 -1.8505154 -1.5601451 -2.4347433  0.22085447  0.2333873515 -0.15376790
#> 27 -1.8484699 -1.5573966 -2.4336103  1.09826678  0.4883828596 -0.41080126
#> 28 -1.7622798 -0.7934813 -1.7669533 -1.57139511 -1.7498735681 -1.67115949
#> 29 -1.7503809 -1.6619442 -1.7550845 -1.55789969 -0.8286237737 -1.65989447
#> 30  1.2922454  0.3521596  0.9986512  2.10622880  4.1872022836 -0.01863987
#> 31  0.7689779 -0.5320757  0.3832906  1.69627179  3.8447179814 -0.53023965
#> 32 -1.8385870 -1.5440941 -2.4281484  0.06306304 -0.2121987383  0.43212075
#> 33 -1.8339222 -1.5378066 -2.4255748  0.82995753  0.0816537558  2.54471728
#> 34 -1.8302343 -1.5328357 -2.4235404  0.60791426 -0.1676643133  0.98617705
#> 35 -1.8402178 -1.5462910 -2.4290487  0.76822447 -0.2208605151  0.16311388
#> 36 -1.8349917 -1.5392501 -2.4261639  0.08853673 -0.1931285217 -0.35035634
#> 37 -1.8410419 -1.5473961 -2.4295063  0.53345794  0.0394814600  0.15904424
#> 38 -1.7752932 -1.6850354 -1.7799307  0.42806812 -1.7703979167 -1.68289815
#> 39 -1.7703966 -1.6801189 -1.7750514 -0.61710381 -1.7642325195 -1.67798559
#> 40 -1.6110792 -1.5624539 -1.6153350 -1.43447675 -1.4834565457 -0.57217607
#> 41 -1.8373226 -1.5423923 -2.4274496  0.79507473 -0.2056007513 -0.08747431
#> 42 -1.8407328 -0.9477620 -2.4293323  0.29536920 -0.2236795920 -0.37632769
#> 43 -1.8347210 -1.5388873 -2.4260135  2.20274544  0.0762592360 -0.34921304
#> 44 -1.2393539 -1.2518427 -2.4313348  0.72697373 -0.2427566939 -0.12336592
#> 45 -1.5358802 -0.9465492 -2.4289605  1.55819249  0.0444613958 -0.10147159
#> 46 -1.5272265 -1.5359438 -2.4248098  0.59823561  2.0378787219  0.20659899
#> 47 -1.6109942 -0.5725884 -1.6152497 -0.44768593 -0.4954427689 -1.56169704
#> 48 -1.8504601 -1.2614484 -2.1273661  0.44946134 -0.2744981512 -0.41967136
#> 49 -1.8394231 -1.5452193 -2.4286105  0.30684741  0.7968458468 -0.37031319
#>           [,13]      [,14]       [,15]      [,16]       [,17]
#> 1  -0.115082811 -1.2945682  0.21954197 -1.8363122  0.56281668
#> 2  -1.564125367 -1.6062572 -1.36551761 -1.6156365 -1.42344825
#> 3  -0.060167576 -1.5820428  0.02226259 -1.8273844  0.45740030
#> 4   0.012308640 -1.5673026  0.60837432 -1.8158994  1.00764907
#> 5  -0.015282712 -1.5728618  0.06400210 -1.8202285  0.95258415
#> 6  -1.564117100 -1.6062494 -1.36550759 -1.6156288 -1.42343868
#> 7  -1.666880806 -1.5934419 -1.40089938 -1.7160735 -1.50917804
#> 8  -0.075469041 -1.5851983 -0.25350776 -1.8298452  1.01449394
#> 9  -1.564158872 -1.6062893 -1.36555847 -1.6156683 -1.42348681
#> 10 -1.496759122 -1.4710111 -1.21621913 -1.5543907 -1.32045254
#> 11 -1.574713588 -1.5273531 -1.29931957 -1.6287673 -1.40579805
#> 12 -1.486624625 -1.4660432 -1.20739181 -1.5446454 -1.31012008
#> 13 -0.046099061 -1.5791556 -0.22930007 -1.5211548  0.48200948
#> 14 -0.098026809 -1.5899044 -0.01294343 -1.8335175  1.13532742
#> 15  0.019727185 -1.5658418  0.36089479 -1.5091695  3.85054625
#> 16  0.009206871 -1.5679246 -0.18383103 -1.8163837  2.30418123
#> 17 -0.038223472 -1.5775502 -0.22282290 -1.8238823  1.09593786
#> 18 -0.055614052 -1.5811141 -0.23714581 -1.8266605  1.23280293
#> 19 -0.052946454 -1.5805733 -0.23495630 -1.8262391  1.69244723
#> 20 -0.005433606 -1.2682758  1.06762261 -1.8187128  2.36369772
#> 21 -0.093241711 -1.5889115 -0.26821068 -1.8327428  1.14766882
#> 22 -1.722351440 -1.6331862 -0.07868866 -1.7684230 -0.62375957
#> 23 -1.564083490 -0.6154561  0.08041047 -1.6155968 -0.06336904
#> 24 -1.489810753 -1.4696738 -1.21173380 -1.5476839 -1.31387420
#> 25 -1.700446685 -0.1009662  1.32125609 -1.7473871 -0.55750384
#> 26  0.603392470 -0.7032370  2.19822929 -1.8355947  1.47376167
#> 27  1.413986767 -1.5899037  1.81611942 -1.8335180  1.14298086
#> 28 -1.720760719 -0.7649390  0.04768708 -1.7671643 -0.70955230
#> 29 -1.708244827 -1.6284089  0.06239172 -1.7553027 -0.68539584
#> 30  3.126519250  4.1726861  5.39635421  0.3521596  1.99548647
#> 31  2.596150869  3.9091116  5.06293443 -0.6141853  0.13673917
#> 32  3.303234306 -1.2752456 -0.22061023 -1.8234836  1.75070878
#> 33  2.917753045 -1.5709560  1.06898386 -1.8187469  1.68127592
#> 34  3.129262135 -1.5661486  1.09908736 -1.8150022  3.44953288
#> 35  0.691605019 -1.5791613 -0.22919548 -1.8251394  2.57294295
#> 36  0.496302535 -1.5723517 -0.20170749 -1.8198330  2.86140210
#> 37  1.124408259 -1.5802308 -0.23323911 -1.8259760  2.15445243
#> 38 -0.730261083 -1.6444809 -1.47857864 -1.7801300  0.14044773
#> 39 -0.733624754 -1.6401998 -1.47251218 -1.7752532 -0.16413018
#> 40 -1.564178902 -1.6063107 -1.36558401 -1.6156899 -1.42350901
#> 41 -0.027564851 -1.5753905 -0.21406872 -1.2149843  2.20551502
#> 42  0.450354918 -1.5798336 -0.23198855 -1.8256624  2.14058280
#> 43  0.738910968 -1.5720006 -0.20039978 -1.8195582  1.50342623
#> 44 -0.072289903 -1.5845532 -0.25089923 -1.8293424  0.43660092
#> 45 -0.045081747 -1.5789571 -0.22847415 -1.5209422  1.08181463
#> 46  0.003168625 -0.9655589  0.59809726 -1.2085853  2.75060328
#> 47 -1.564092608 -1.6062257 -1.36547765 -1.6156054 -1.42341054
#> 48  0.599713570 -1.5924935 -0.28236771 -1.8355387  1.47268188
#> 49  3.419785321 -1.5781250  2.09235627 -1.8243324  1.27579179

predict(myMixture, trichoptera, "response")

#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 
#>  2  3  2  2  2  3  4  2  3  3  4  3  2  2  2  2  2  2  2  2  2  4  3  3  4  2 
#> 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 
#>  2  4  4  1  4  2  2  2  2  2  2  4  4  3  2  2  2  2  2  2  3  2  2 
#> Levels: 1 2 3 4