Prediction for a PLNmixturefit object

Predict either posterior probabilities for each group or latent positions based on new samples

# S3 method for PLNmixturefit
predict(
  object,
  newdata,
  type = c("posterior", "response", "position"),
  prior = matrix(rep(1/object$k, object$k), nrow(newdata), object$k, byrow = TRUE),
  control = PLNmixture_param(),
  ...
)

Arguments

object

an R6 object with class PLNmixturefit

newdata

A data frame in which to look for variables, offsets and counts with which to predict.

type

The type of prediction required. The default posterior are posterior probabilities for each group , response is the group with maximal posterior probability and latent is the averaged latent in the latent space, with weights equal to the posterior probabilities.

prior

User-specified prior group probabilities in the new data. The default uses a uniform prior.

control

a list-like structure for controlling the fit. See PLNmixture_param() for details.

...

additional parameters for S3 compatibility. Not used

Value

A matrix of posterior probabilities for each group (if type = "posterior"), a matrix of (average) position in the latent space (if type = "position") or a vector of predicted groups (if type = "response").

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLNmixture(Abundance ~ 1 + offset(log(Offset)),
           data = trichoptera, control = PLNmixture_param(smoothing = "none"))  %>% getBestModel()
#> 
#>  Initialization...
#> 
#>  Adjusting 5 PLN mixture models.
#> 	number of cluster = 1 	number of cluster = 2 	number of cluster = 3 	number of cluster = 4 	number of cluster = 5 
#>  Post-treatments
#>  DONE!
predict(myPLN, trichoptera, "posterior")
#>            [,1]         [,2]         [,3]         [,4]         [,5]
#> 1  9.352633e-01 2.220446e-16 1.933323e-03 2.220446e-16 6.280338e-02
#> 2  7.569566e-02 2.039434e-02 1.867472e-05 2.220446e-16 9.038913e-01
#> 3  6.234650e-04 7.952624e-01 2.220446e-16 2.220446e-16 2.041141e-01
#> 4  2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16 2.584869e-12
#> 5  3.744386e-04 9.926138e-01 2.220446e-16 2.220446e-16 7.011720e-03
#> 6  1.739052e-01 3.600205e-01 1.094257e-05 2.220446e-16 4.660634e-01
#> 7  8.713481e-01 3.724113e-03 2.701949e-03 2.220446e-16 1.222258e-01
#> 8  1.833906e-02 9.599966e-01 1.356174e-06 2.220446e-16 2.166298e-02
#> 9  5.679572e-01 1.104030e-02 1.672886e-04 2.220446e-16 4.208352e-01
#> 10 5.862113e-01 2.137397e-02 8.250839e-03 1.276263e-14 3.841639e-01
#> 11 5.661860e-01 1.169564e-01 1.932985e-02 3.313136e-03 2.942146e-01
#> 12 3.450773e-01 3.373416e-01 4.331742e-02 8.610417e-02 1.881595e-01
#> 13 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 14 9.999963e-01 2.220446e-16 3.689499e-06 2.220446e-16 2.220446e-16
#> 15 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 17 9.999960e-01 2.220446e-16 4.042336e-06 2.220446e-16 2.220446e-16
#> 18 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 19 9.998781e-01 2.047304e-10 2.220446e-16 2.220446e-16 1.218613e-04
#> 20 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 21 9.943898e-01 7.229938e-05 4.355882e-06 2.220446e-16 5.533593e-03
#> 22 8.701808e-01 3.211876e-04 6.842080e-02 2.220446e-16 6.107717e-02
#> 23 1.859854e-03 2.220446e-16 9.981401e-01 2.220446e-16 5.590137e-13
#> 24 5.695314e-01 8.709024e-02 2.780822e-02 2.513950e-03 3.130562e-01
#> 25 9.874893e-04 2.220446e-16 9.990125e-01 2.220446e-16 2.220446e-16
#> 26 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 27 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 28 8.892801e-04 2.220446e-16 9.985131e-01 5.976355e-04 8.583541e-16
#> 29 9.030073e-03 1.377168e-04 9.703030e-01 2.036004e-02 1.691682e-04
#> 30 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 31 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 32 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 33 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 34 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 35 6.328863e-05 2.220446e-16 2.220446e-16 2.220446e-16 9.999367e-01
#> 36 4.817453e-04 1.107117e-01 2.220446e-16 2.220446e-16 8.888065e-01
#> 37 3.640423e-05 2.220446e-16 2.220446e-16 2.220446e-16 9.999636e-01
#> 38 3.226702e-01 2.220446e-16 2.195488e-03 2.220446e-16 6.751343e-01
#> 39 1.480518e-01 2.220446e-16 5.153604e-04 2.220446e-16 8.514328e-01
#> 40 4.829168e-02 8.089271e-01 4.111882e-05 3.146406e-04 1.424255e-01
#> 41 5.404847e-01 1.160632e-01 2.220446e-16 2.220446e-16 3.434521e-01
#> 42 9.956586e-01 2.220446e-16 4.341406e-03 2.220446e-16 2.220446e-16
#> 43 1.227867e-01 2.064282e-03 3.685593e-06 2.220446e-16 8.751453e-01
#> 44 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 45 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 46 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 47 9.961184e-01 3.197173e-03 6.844489e-04 2.220446e-16 2.616655e-14
#> 48 2.130700e-02 2.220446e-16 9.786930e-01 2.220446e-16 2.220446e-16
#> 49 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
predict(myPLN, trichoptera, "position")
#>          [,1]       [,2]      [,3]       [,4]        [,5]      [,6]       [,7]
#> 1   -6.889277  -7.534498 -2.128455  -5.501396 -0.55770310 -3.692142  -5.871845
#> 2  -17.707884 -17.760323 -1.998975  -7.186297 -0.54363514 -3.543614 -17.624669
#> 3  -23.157798 -23.158227 -3.735845 -20.795953 -0.19095486 -3.800739 -23.157124
#> 4  -24.358992 -24.358992 -4.092600 -24.358992 -0.09194478 -3.762667 -24.358992
#> 5  -24.311293 -24.311548 -3.611271 -24.230045 -0.14436836 -3.650788 -24.310898
#> 6  -18.467735 -18.588048 -2.808845 -12.947890 -0.30936837 -3.398109 -18.277190
#> 7   -7.688062  -8.292612 -2.177235  -5.649501 -0.64044827 -3.593239  -6.726918
#> 8  -23.898056 -23.910690 -3.891021 -23.634468 -0.17640211 -3.603002 -23.878170
#> 9  -11.534570 -11.928136 -1.886361  -6.255771 -0.49032072 -3.580180 -10.909776
#> 10 -11.262175 -11.669593 -2.050651  -6.391681 -0.63687247 -3.538510 -10.613767
#> 11 -11.920066 -12.286333 -2.448878  -8.105838 -0.46770290 -3.593607 -11.259569
#> 12 -15.504798 -15.026683 -3.339771 -13.077081 -0.30000768 -3.978282 -14.233826
#> 13  -5.964750  -6.803626 -2.723131  -5.417886 -0.26322155 -3.500718  -5.043187
#> 14  -6.111587  -6.800656 -2.280949  -5.406283 -0.52995331 -3.432926  -4.876704
#> 15 -24.358992 -24.358992 -4.659902 -24.358992 -0.18325223 -3.499413 -24.358992
#> 16 -24.358992 -24.358992 -4.181501 -24.358992 -0.19451090 -3.211233 -24.358992
#> 17  -6.120694  -6.805282 -2.685705  -5.424293 -0.32203243 -3.658935  -5.052316
#> 18  -6.118163  -6.803995 -2.540572  -5.269738 -0.51990352 -3.265752  -5.045230
#> 19  -6.121486  -6.806346 -2.314850  -5.423064 -0.62432260 -3.063161  -5.051943
#> 20  -6.144959  -6.665966 -2.672511  -5.470763 -0.58574079 -3.447853  -4.979975
#> 21  -6.182354  -6.867199 -2.203662  -5.417705 -0.57434507 -3.710010  -5.104749
#> 22  -6.838944  -7.444480 -2.201445  -5.510564 -0.78938283 -3.758895  -5.882102
#> 23  -5.669381  -5.710178 -3.794607  -5.456113 -0.89983149 -5.661582  -5.664126
#> 24 -11.596371 -11.972086 -2.384225  -7.560997 -0.53244366 -3.595629 -10.939963
#> 25  -5.676587  -5.716526 -3.316756  -5.664408 -1.03865855 -5.472197  -5.473508
#> 26  -5.685846  -5.724781 -3.722192  -5.674462 -1.36927013 -5.682678  -5.682678
#> 27  -5.688786  -5.727613 -3.739321  -5.677435 -1.14316895 -5.685627  -5.685627
#> 28  -5.668703  -5.704118 -3.754023  -5.656427 -1.18609095 -5.657580  -5.658315
#> 29  -5.898883  -5.773646 -3.796826  -5.879170 -0.96675910 -5.674928  -5.675638
#> 30 -16.848714  -8.370759 -6.654046 -16.848714 -0.30982049 -7.223517  -6.458335
#> 31 -16.848729  -8.612931 -5.752107 -16.848729 -0.10984059 -7.523765  -6.590055
#> 32 -18.529809 -18.529809 -2.519028  -6.983015 -0.96669897 -3.242058 -18.529809
#> 33 -18.529809 -18.529809 -2.434267  -7.002941 -0.77835728 -2.995073 -18.529809
#> 34 -18.529810 -18.529810 -3.003455  -7.076242 -0.40072460 -2.878185 -18.529810
#> 35 -18.529024 -18.529067 -2.420989  -6.973267 -0.81864951 -2.753370 -18.528956
#> 36 -19.169195 -19.169523 -2.970674  -8.906770 -0.41793445 -3.442004 -19.168690
#> 37 -18.529357 -18.529382 -1.415747  -6.582194 -1.06446231 -3.445897 -18.529318
#> 38 -14.492427 -14.715925 -1.890637  -6.444800 -1.16540248 -3.585714 -14.138171
#> 39 -16.683597 -16.786153 -1.634058  -6.718387 -0.91334533 -3.558553 -16.520945
#> 40 -22.644207 -22.674971 -3.630473 -20.961003 -0.15000981 -3.705215 -22.588118
#> 41 -12.501756 -12.870629 -2.746003  -8.159099 -0.38034723 -3.882597 -11.928417
#> 42  -6.117050  -6.799717 -2.559690  -5.272329 -0.49911404 -3.778303  -5.050064
#> 43 -17.019737 -17.103010 -3.028319  -6.833264 -0.30940139 -3.842821 -16.891336
#> 44  -6.115253  -6.802516 -2.195799  -5.263099 -0.46590501 -3.605513  -4.741669
#> 45  -6.122137  -6.806022 -2.445486  -5.278718 -0.53274468 -3.431072  -4.910932
#> 46  -6.005164  -6.821751 -2.966032  -5.484699 -0.38489516 -3.904031  -5.001357
#> 47  -6.163353  -6.852888 -2.309758  -5.454447 -0.46306877 -3.639109  -5.070472
#> 48  -5.491075  -5.741398 -3.832528  -5.662102 -0.72482184 -5.633774  -5.662090
#> 49  -5.718595  -5.565637 -3.747434  -5.707561 -1.15694379 -5.715524  -5.715524
#>          [,8]       [,9]     [,10]     [,11]     [,12]     [,13]      [,14]
#> 1   -5.592930  -7.522113 -2.874844 -3.463969 -4.791797 -5.250202  -5.369551
#> 2  -17.239706 -17.759323 -3.826651 -5.622975 -3.157882 -3.265685  -7.201059
#> 3   -9.046047 -23.158219 -4.475725 -5.386328 -3.567471 -5.821972 -20.798576
#> 4   -6.663353 -24.358992 -4.689783 -5.422973 -4.345787 -6.516033 -24.358992
#> 5   -6.711347 -24.311544 -4.665650 -5.315473 -4.423352 -6.455706 -24.230147
#> 6  -11.834399 -18.585754 -3.985925 -5.172480 -3.961509 -4.702517 -12.961092
#> 7   -6.393457  -8.280991 -2.715720 -3.512924 -4.834107 -5.070733  -5.682314
#> 8   -6.837606 -23.910448 -4.535685 -5.237191 -4.507090 -6.370915 -23.635496
#> 9  -10.541553 -11.920631 -3.368710 -4.334402 -4.435609 -4.388130  -6.284605
#> 10 -10.048849 -11.661511 -3.286452 -4.223645 -4.443053 -4.393147  -6.404981
#> 11  -8.999995 -12.272725 -3.372897 -4.183573 -4.536598 -4.688145  -8.051406
#> 12  -8.217524 -14.881366 -3.861574 -4.393219 -4.868669 -5.057839 -11.868824
#> 13  -4.608393  -6.790513 -2.971516 -3.397563 -4.900250 -5.405280  -5.458607
#> 14  -4.731143  -6.787504 -2.989149 -3.323719 -5.030994 -5.393523  -5.447447
#> 15  -6.399840 -24.358992 -3.793184 -5.028795 -4.963052 -6.638114 -24.358992
#> 16  -6.657385 -24.358992 -4.350506 -4.867683 -4.454474 -6.509185 -24.358992
#> 17  -4.765049  -6.638333 -2.815444 -3.092097 -5.056677 -5.411753  -5.464762
#> 18  -4.755790  -6.790887 -3.094723 -2.634593 -5.049620 -5.406728  -5.459983
#> 19  -4.764098  -6.793252 -3.011459 -2.868940 -5.054538 -5.410095  -5.463590
#> 20  -4.715684  -6.804893 -3.057987 -2.319394 -5.121261 -5.458767  -5.366970
#> 21  -4.810422  -6.854126 -2.544598 -3.218495 -5.026567 -5.383518  -5.458651
#> 22  -5.564357  -7.430277 -2.759777 -3.379210 -4.972924 -5.034367  -5.415192
#> 23  -4.956499  -5.670643 -3.286794 -3.313843 -5.674393 -2.515234  -3.555110
#> 24  -9.202239 -11.959383 -3.335114 -4.167382 -4.524385 -4.573450  -7.500627
#> 25  -4.971816  -5.677256 -3.525567 -3.050906 -5.682514 -2.633569  -3.434617
#> 26  -4.990994  -5.685846 -3.431266 -3.146372 -5.495913 -2.387142  -3.346098
#> 27  -4.996670  -5.688786 -2.873109 -3.029818 -5.695706 -2.022023  -3.853213
#> 28  -4.743879  -5.663350 -3.409980 -3.246293 -5.669857 -2.375782  -3.499024
#> 29  -5.003073  -5.702039 -3.457567 -3.075239 -5.721495 -2.463617  -3.714800
#> 30  -7.754543  -6.843394 -5.378603 -3.153264 -8.608358 -4.190844  -3.180123
#> 31  -7.857002  -6.910069 -5.918531 -4.071642 -8.477156 -5.166205  -4.018122
#> 32 -18.529809 -18.529809 -4.300917 -5.928222 -3.490288 -1.124096  -6.610331
#> 33 -18.529809 -18.529809 -3.831565 -5.639554 -2.428563 -1.885053  -7.015173
#> 34 -18.529810 -18.529810 -4.512119 -6.144357 -4.015592 -2.528785  -7.087706
#> 35 -18.528938 -18.529066 -3.465993 -5.902322 -3.555281 -2.977796  -6.985844
#> 36 -17.205980 -19.169516 -4.350414 -5.862186 -4.209788 -3.629678  -8.917855
#> 37 -18.529308 -18.529381 -3.671988 -5.531770 -3.534596 -2.640873  -6.984567
#> 38 -14.039201 -14.711580 -2.578908 -4.988471 -4.122410 -3.552051  -6.462707
#> 39 -16.475827 -16.784178 -3.439858 -5.446497 -3.880041 -3.104057  -6.734526
#> 40  -8.215825 -22.673826 -4.382013 -5.231128 -4.175657 -5.927252 -20.960624
#> 41  -9.716902 -12.863569 -3.288010 -4.525829 -4.482213 -4.912939  -8.185117
#> 42  -4.472875  -6.786504 -3.107015 -3.414998 -5.054471 -5.099783  -5.454519
#> 43 -16.820751 -17.101413 -2.520135 -5.321145 -4.290617 -3.424366  -6.848986
#> 44  -4.599102  -6.789389 -2.823124 -3.371055 -4.893175 -5.400909  -5.454455
#> 45  -4.487428  -6.792941 -2.559857 -3.334642 -4.915236 -5.414616  -5.467481
#> 46  -4.872005  -6.808870 -3.381557 -2.895363 -4.872933 -5.472853  -5.243469
#> 47  -4.560929  -6.839719 -2.870934 -3.098023 -5.011406 -5.382173  -5.494642
#> 48  -4.780324  -5.505464 -3.209015 -3.402212 -5.672141 -2.337164  -3.835688
#> 49  -5.052558  -5.718595 -3.638701 -3.153475 -5.725322 -1.055850  -3.994214
#>        [,15]      [,16]     [,17]
#> 1  -4.036488  -5.515555 -2.712357
#> 2  -4.870812  -6.204833 -2.837150
#> 3  -4.952426  -7.062269 -3.233998
#> 4  -4.879643  -7.293038 -3.631548
#> 5  -5.089401  -7.266743 -3.201752
#> 6  -4.922919  -6.472787 -2.954295
#> 7  -4.343064  -5.565823 -2.647866
#> 8  -5.217597  -7.211365 -2.732512
#> 9  -4.553903  -5.793027 -2.742762
#> 10 -4.498194  -5.883583 -2.610945
#> 11 -4.498926  -6.142092 -2.649445
#> 12 -4.430590  -6.978984 -2.966291
#> 13 -4.355767  -5.302165 -2.957137
#> 14 -4.179530  -5.440979 -2.404509
#> 15 -5.397523  -7.153265 -2.546213
#> 16 -5.387849  -7.289763 -2.812516
#> 17 -4.372817  -5.458407 -2.704750
#> 18 -4.359608  -5.453538 -2.559502
#> 19 -4.369164  -5.457082 -2.331396
#> 20 -3.896511  -5.503403 -2.440936
#> 21 -4.331684  -5.447095 -2.423445
#> 22 -3.836719  -6.391470 -2.542534
#> 23 -1.575049 -18.764900 -2.323508
#> 24 -4.459260  -6.213206 -2.617864
#> 25 -1.493305 -18.776562 -2.599785
#> 26 -1.228524 -18.789749 -2.145909
#> 27 -1.551775 -18.789749 -2.404973
#> 28 -1.433559 -18.771298 -2.335827
#> 29 -1.500628 -18.441723 -2.418512
#> 30 -1.964506  -7.754543 -5.122210
#> 31 -2.906906  -7.857002 -6.408023
#> 32 -5.116108  -6.311221 -2.559459
#> 33 -3.974089  -6.347763 -2.980957
#> 34 -4.473552  -6.473233 -2.320627
#> 35 -5.064740  -6.292991 -1.535719
#> 36 -5.140667  -6.419079 -1.609440
#> 37 -5.057330  -6.290476 -1.923566
#> 38 -4.708464  -6.010774 -2.173192
#> 39 -4.820775  -6.131945 -2.332644
#> 40 -5.135733  -7.032145 -3.068382
#> 41 -4.731298  -5.497016 -2.105348
#> 42 -4.355459  -5.512811 -1.991764
#> 43 -5.046559  -6.216499 -2.831787
#> 44 -4.344061  -5.447976 -2.920233
#> 45 -4.380224  -5.312460 -2.730639
#> 46 -4.164864  -5.237561 -2.245957
#> 47 -4.289069  -5.443714 -2.715682
#> 48 -2.340278 -18.505295 -2.027347
#> 49 -1.949108 -18.789749 -2.737656
predict(myPLN, trichoptera, "response")
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 
#>  1  5  2  2  2  5  1  2  1  1  1  1  1  1  2  2  1  1  1  1  1  1  3  1  3  3 
#> 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 
#>  3  3  3  4  4  5  5  5  5  5  5  5  5  2  1  1  5  1  1  1  1  3  3 
#> Levels: 1 2 3 4 5