Prediction for a PLNmixturefit object

Predict either posterior probabilities for each group or latent positions based on new samples

# S3 method for PLNmixturefit
predict(
  object,
  newdata,
  type = c("posterior", "response", "position"),
  prior = matrix(rep(1/object$k, object$k), nrow(newdata), object$k, byrow = TRUE),
  control = PLNmixture_param(),
  ...
)

Arguments

object

an R6 object with class PLNmixturefit

newdata

A data frame in which to look for variables, offsets and counts with which to predict.

type

The type of prediction required. The default posterior are posterior probabilities for each group , response is the group with maximal posterior probability and latent is the averaged latent in the latent space, with weights equal to the posterior probabilities.

prior

User-specified prior group probabilities in the new data. The default uses a uniform prior.

control

a list-like structure for controlling the fit. See PLNmixture_param() for details.

...

additional parameters for S3 compatibility. Not used

Value

A matrix of posterior probabilities for each group (if type = "posterior"), a matrix of (average) position in the latent space (if type = "position") or a vector of predicted groups (if type = "response").

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLNmixture(Abundance ~ 1 + offset(log(Offset)),
           data = trichoptera, control = PLNmixture_param(smoothing = "none"))  %>% getBestModel()
#> 
#>  Initialization...
#> 
#>  Adjusting 5 PLN mixture models.
#> 	number of cluster = 1 	number of cluster = 2 	number of cluster = 3 	number of cluster = 4 	number of cluster = 5 
#>  Post-treatments
#>  DONE!
predict(myPLN, trichoptera, "posterior")
#>            [,1]         [,2]         [,3]         [,4]         [,5]
#> 1  2.220446e-16 1.933071e-03 6.276718e-02 2.220446e-16 9.352998e-01
#> 2  2.040277e-02 1.868239e-05 9.038405e-01 2.220446e-16 7.573801e-02
#> 3  7.953394e-01 2.220446e-16 2.040381e-01 2.220446e-16 6.224975e-04
#> 4  9.999717e-01 2.220446e-16 2.834313e-05 2.220446e-16 2.220446e-16
#> 5  9.926173e-01 2.220446e-16 7.008459e-03 2.220446e-16 3.742409e-04
#> 6  3.600901e-01 1.094621e-05 4.659347e-01 1.037322e-14 1.739642e-01
#> 7  3.723852e-03 2.701695e-03 1.221599e-01 1.337302e-14 8.714146e-01
#> 8  9.600037e-01 1.438896e-06 2.165297e-02 3.854061e-15 1.834188e-02
#> 9  1.104154e-02 1.673031e-04 4.206874e-01 3.021474e-15 5.681038e-01
#> 10 2.137600e-02 8.251422e-03 3.840200e-01 1.258854e-12 5.863525e-01
#> 11 1.169632e-01 1.933051e-02 2.940937e-01 3.310875e-03 5.663017e-01
#> 12 3.373705e-01 4.332007e-02 1.880873e-01 8.606499e-02 3.451572e-01
#> 13 1.054828e-15 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 14 2.220446e-16 3.695524e-06 2.220446e-16 2.220446e-16 9.999963e-01
#> 15 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 16 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 17 2.220446e-16 4.050703e-06 2.220446e-16 2.220446e-16 9.999959e-01
#> 18 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 19 5.503914e-05 2.220446e-16 1.282965e-04 2.220446e-16 9.998167e-01
#> 20 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 21 7.312607e-05 4.360952e-06 5.530477e-03 2.220446e-16 9.943920e-01
#> 22 3.210848e-04 6.841253e-02 6.104259e-02 3.758731e-16 8.702238e-01
#> 23 2.220446e-16 9.981398e-01 2.220446e-16 2.220446e-16 1.860170e-03
#> 24 8.709624e-02 2.780946e-02 3.129308e-01 2.509729e-03 5.696538e-01
#> 25 2.220446e-16 9.990124e-01 2.220446e-16 9.061713e-15 9.876442e-04
#> 26 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 27 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 28 2.220446e-16 9.991101e-01 2.220446e-16 4.116861e-12 8.899024e-04
#> 29 1.377235e-04 9.703107e-01 1.702215e-04 2.034970e-02 9.031682e-03
#> 30 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 31 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 32 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 33 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 34 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 35 2.220446e-16 2.220446e-16 9.999367e-01 2.220446e-16 6.333543e-05
#> 36 1.107580e-01 2.220446e-16 8.887602e-01 2.220446e-16 4.817811e-04
#> 37 2.220446e-16 2.220446e-16 9.999640e-01 2.220446e-16 3.598293e-05
#> 38 2.220446e-16 2.196027e-03 6.749995e-01 2.220446e-16 3.228044e-01
#> 39 1.257261e-14 5.155429e-04 8.513550e-01 2.220446e-16 1.481294e-01
#> 40 8.092301e-01 4.115995e-05 1.424119e-01 1.086167e-11 4.831682e-02
#> 41 1.160727e-01 2.220446e-16 3.433191e-01 2.220446e-16 5.406081e-01
#> 42 2.220446e-16 4.340670e-03 2.220446e-16 2.220446e-16 9.956593e-01
#> 43 2.065093e-03 3.706148e-06 8.750783e-01 2.220446e-16 1.228529e-01
#> 44 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 45 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 46 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 47 3.196910e-03 6.843326e-04 2.220446e-16 8.270437e-14 9.961188e-01
#> 48 2.220446e-16 9.786900e-01 2.220446e-16 2.220446e-16 2.130997e-02
#> 49 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
predict(myPLN, trichoptera, "position")
#>          [,1]       [,2]      [,3]       [,4]        [,5]      [,6]       [,7]
#> 1   -6.888828  -7.534074 -2.128459  -5.501340 -0.55770238 -3.692141  -5.871356
#> 2  -17.707407 -17.759875 -1.998996  -7.186378 -0.54363359 -3.543620 -17.624145
#> 3  -23.158259 -23.158687 -3.735931 -20.797295 -0.19094959 -3.800742 -23.157586
#> 4  -24.358827 -24.358827 -4.092584 -24.358500 -0.09194496 -3.762666 -24.358827
#> 5  -24.311316 -24.311571 -3.611271 -24.230105 -0.14436767 -3.650789 -24.310921
#> 6  -18.467408 -18.587761 -2.808939 -12.949011 -0.30935174 -3.398138 -18.276798
#> 7   -7.687238  -8.291834 -2.177246  -5.649394 -0.64044373 -3.593247  -6.726021
#> 8  -23.898062 -23.910697 -3.891029 -23.634588 -0.17640214 -3.603004 -23.878172
#> 9  -11.532756 -11.926424 -1.886851  -6.255564 -0.49032194 -3.580192 -10.907800
#> 10 -11.260424 -11.667939 -2.050696  -6.391495 -0.63686146 -3.538536 -10.611859
#> 11 -11.918663 -12.285029 -2.448922  -8.105753 -0.46770344 -3.593621 -11.258061
#> 12 -15.504005 -15.026273 -3.339667 -13.077067 -0.30001710 -3.978153 -14.233349
#> 13  -5.964750  -6.803626 -2.723131  -5.417886 -0.26322155 -3.500718  -5.043187
#> 14  -6.111587  -6.800656 -2.280949  -5.406283 -0.52995329 -3.432926  -4.876704
#> 15 -24.358992 -24.358992 -4.659902 -24.358992 -0.18325212 -3.499413 -24.358992
#> 16 -24.358992 -24.358992 -4.181501 -24.358992 -0.19451092 -3.211231 -24.358992
#> 17  -6.120694  -6.805282 -2.685705  -5.424293 -0.32203243 -3.658935  -5.052316
#> 18  -6.118163  -6.803995 -2.540572  -5.269738 -0.51990352 -3.265752  -5.045230
#> 19  -6.122570  -6.807388 -2.314897  -5.424116 -0.62432008 -3.063155  -5.053092
#> 20  -6.144959  -6.665966 -2.672511  -5.470763 -0.58574079 -3.447853  -4.979975
#> 21  -6.182331  -6.867176 -2.203663  -5.417716 -0.57434501 -3.710012  -5.104723
#> 22  -6.838516  -7.444082 -2.201443  -5.510506 -0.78937601 -3.758881  -5.881627
#> 23  -5.669381  -5.710178 -3.794607  -5.456113 -0.89983143 -5.661581  -5.664125
#> 24 -11.594875 -11.970712 -2.384290  -7.560867 -0.53243459 -3.595638 -10.938375
#> 25  -5.676587  -5.716526 -3.316756  -5.664408 -1.03865850 -5.472196  -5.473507
#> 26  -5.685846  -5.724781 -3.722192  -5.674462 -1.36927013 -5.682678  -5.682678
#> 27  -5.688786  -5.727613 -3.739321  -5.677435 -1.14316895 -5.685627  -5.685627
#> 28  -5.662018  -5.702451 -3.752556  -5.649734 -1.18639509 -5.656553  -5.657796
#> 29  -5.898782  -5.773633 -3.796797  -5.879055 -0.96676228 -5.674905  -5.675641
#> 30 -16.848714  -8.370759 -6.654046 -16.848714 -0.30982049 -7.223517  -6.458335
#> 31 -16.848729  -8.612931 -5.752108 -16.848729 -0.10984059 -7.523765  -6.590055
#> 32 -18.529809 -18.529809 -2.519028  -6.983015 -0.96669897 -3.242058 -18.529809
#> 33 -18.529809 -18.529809 -2.434267  -7.002941 -0.77835728 -2.995073 -18.529809
#> 34 -18.529810 -18.529810 -3.003455  -7.076242 -0.40072460 -2.878185 -18.529810
#> 35 -18.529023 -18.529066 -2.420989  -6.973267 -0.81864879 -2.753370 -18.528956
#> 36 -19.169465 -19.169792 -2.970710  -8.907573 -0.41793335 -3.442010 -19.168959
#> 37 -18.529362 -18.529387 -1.415747  -6.582194 -1.06446238 -3.445896 -18.529324
#> 38 -14.490752 -14.714343 -1.890670  -6.444590 -1.16536864 -3.585726 -14.136348
#> 39 -16.682630 -16.785240 -1.634084  -6.718266 -0.91333482 -3.558560 -16.519893
#> 40 -22.646189 -22.679600 -3.629784 -20.963124 -0.15000952 -3.704065 -22.593319
#> 41 -12.500281 -12.869238 -2.745995  -8.159074 -0.38034928 -3.882579 -11.926811
#> 42  -6.117050  -6.799718 -2.559689  -5.272329 -0.49911399 -3.778302  -5.050064
#> 43 -17.018921 -17.102239 -3.028304  -6.833177 -0.30940215 -3.842815 -16.890450
#> 44  -6.115253  -6.802516 -2.195799  -5.263099 -0.46590501 -3.605513  -4.741669
#> 45  -6.122137  -6.806022 -2.445486  -5.278718 -0.53274468 -3.431072  -4.910932
#> 46  -6.005164  -6.821751 -2.966032  -5.484699 -0.38489515 -3.904031  -5.001356
#> 47  -6.163349  -6.852884 -2.309757  -5.454442 -0.46306860 -3.639125  -5.070467
#> 48  -5.491076  -5.741401 -3.832524  -5.662101 -0.72482148 -5.633768  -5.662088
#> 49  -5.718595  -5.565637 -3.747434  -5.707561 -1.15694379 -5.715524  -5.715524
#>          [,8]       [,9]     [,10]     [,11]     [,12]     [,13]      [,14]
#> 1   -5.592430  -7.521688 -2.874824 -3.463877 -4.791846 -5.250280  -5.369499
#> 2  -17.239020 -17.758875 -3.826619 -5.622859 -3.157963 -3.265814  -7.201141
#> 3   -9.045143 -23.158679 -4.475766 -5.386283 -3.567539 -5.822210 -20.799917
#> 4   -6.663689 -24.358827 -4.689774 -5.422989 -4.345775 -6.515964 -24.358500
#> 5   -6.711309 -24.311566 -4.665651 -5.315473 -4.423353 -6.455716 -24.230208
#> 6  -11.832753 -18.585466 -3.985915 -5.172284 -3.961656 -4.702886 -12.962213
#> 7   -6.392545  -8.280213 -2.715689 -3.512750 -4.834198 -5.070895  -5.682208
#> 8   -6.837481 -23.910456 -4.535686 -5.237179 -4.507099 -6.370943 -23.635615
#> 9  -10.539511 -11.918916 -3.368578 -4.334016 -4.435808 -4.388483  -6.284402
#> 10 -10.046861 -11.659856 -3.286327 -4.223266 -4.443258 -4.393513  -6.404798
#> 11  -8.998327 -12.271423 -3.372794 -4.183261 -4.536759 -4.688458  -8.051350
#> 12  -8.216457 -14.881018 -3.861450 -4.393067 -4.868620 -5.058075 -11.869326
#> 13  -4.608393  -6.790513 -2.971516 -3.397563 -4.900250 -5.405280  -5.458607
#> 14  -4.731143  -6.787504 -2.989149 -3.323719 -5.030994 -5.393523  -5.447447
#> 15  -6.399840 -24.358992 -3.793184 -5.028795 -4.963052 -6.638114 -24.358992
#> 16  -6.657385 -24.358992 -4.350506 -4.867683 -4.454474 -6.509185 -24.358992
#> 17  -4.765049  -6.638333 -2.815444 -3.092097 -5.056677 -5.411753  -5.464762
#> 18  -4.755790  -6.790887 -3.094723 -2.634593 -5.049620 -5.406728  -5.459983
#> 19  -4.764289  -6.794295 -3.011523 -2.869016 -5.054505 -5.410141  -5.464641
#> 20  -4.715684  -6.804893 -3.057987 -2.319394 -5.121261 -5.458767  -5.366970
#> 21  -4.810380  -6.854103 -2.544606 -3.218488 -5.026571 -5.383526  -5.458662
#> 22  -5.563877  -7.429878 -2.759750 -3.379119 -4.972965 -5.034473  -5.415152
#> 23  -4.956499  -5.670643 -3.286794 -3.313843 -5.674393 -2.515235  -3.555110
#> 24  -9.200502 -11.958013 -3.335001 -4.167053 -4.524549 -4.573775  -7.500549
#> 25  -4.971816  -5.677256 -3.525567 -3.050906 -5.682514 -2.633570  -3.434617
#> 26  -4.990994  -5.685846 -3.431266 -3.146372 -5.495913 -2.387142  -3.346098
#> 27  -4.996670  -5.688786 -2.873109 -3.029818 -5.695706 -2.022023  -3.853213
#> 28  -4.742168  -5.662624 -3.408638 -3.246048 -5.668139 -2.374396  -3.499050
#> 29  -5.003057  -5.702042 -3.457544 -3.075234 -5.721462 -2.463598  -3.714807
#> 30  -7.754543  -6.843394 -5.378603 -3.153264 -8.608358 -4.190844  -3.180123
#> 31  -7.857036  -6.910069 -5.918532 -4.071652 -8.477155 -5.166205  -4.018122
#> 32 -18.529809 -18.529809 -4.300917 -5.928222 -3.490288 -1.124097  -6.610331
#> 33 -18.529809 -18.529809 -3.831565 -5.639554 -2.428563 -1.885053  -7.015173
#> 34 -18.529810 -18.529810 -4.512119 -6.144357 -4.015592 -2.528785  -7.087706
#> 35 -18.528938 -18.529065 -3.465993 -5.902322 -3.555281 -2.977796  -6.985844
#> 36 -17.205429 -19.169785 -4.350431 -5.862158 -4.209810 -3.629804  -8.918658
#> 37 -18.529313 -18.529386 -3.671988 -5.531770 -3.534596 -2.640873  -6.984568
#> 38 -14.037338 -14.709996 -2.578887 -4.988119 -4.122589 -3.552382  -6.462500
#> 39 -16.474751 -16.783264 -3.439804 -5.446293 -3.880145 -3.104250  -6.734407
#> 40  -8.215239 -22.678961 -4.381629 -5.231549 -4.174401 -5.927816 -20.966880
#> 41  -9.715091 -12.862176 -3.287952 -4.525523 -4.482356 -4.913193  -8.185096
#> 42  -4.472875  -6.786504 -3.107015 -3.414998 -5.054470 -5.099785  -5.454520
#> 43 -16.819833 -17.100641 -2.520130 -5.321004 -4.290678 -3.424488  -6.848900
#> 44  -4.599102  -6.789389 -2.823124 -3.371055 -4.893175 -5.400909  -5.454455
#> 45  -4.487427  -6.792941 -2.559857 -3.334642 -4.915236 -5.414616  -5.467481
#> 46  -4.872005  -6.808870 -3.381557 -2.895363 -4.872933 -5.472853  -5.243469
#> 47  -4.560928  -6.839715 -2.870933 -3.098023 -5.011407 -5.382173  -5.494637
#> 48  -4.780328  -5.505530 -3.209014 -3.402212 -5.672139 -2.337171  -3.835693
#> 49  -5.052558  -5.718595 -3.638701 -3.153475 -5.725322 -1.055850  -3.994214
#>        [,15]      [,16]     [,17]
#> 1  -4.036479  -5.515523 -2.712352
#> 2  -4.870788  -6.204807 -2.837145
#> 3  -4.952458  -7.062347 -3.234008
#> 4  -4.879634  -7.293011 -3.631550
#> 5  -5.089402  -7.266747 -3.201752
#> 6  -4.922902  -6.472809 -2.954294
#> 7  -4.343024  -5.565765 -2.647859
#> 8  -5.217596  -7.211371 -2.732513
#> 9  -4.553811  -5.792908 -2.742740
#> 10 -4.498105  -5.883477 -2.610940
#> 11 -4.498852  -6.142009 -2.649434
#> 12 -4.430637  -6.978920 -2.966174
#> 13 -4.355767  -5.302165 -2.957137
#> 14 -4.179530  -5.440979 -2.404509
#> 15 -5.397524  -7.153265 -2.546213
#> 16 -5.387849  -7.289763 -2.812516
#> 17 -4.372817  -5.458407 -2.704750
#> 18 -4.359608  -5.453538 -2.559502
#> 19 -4.369219  -5.457188 -2.331397
#> 20 -3.896511  -5.503403 -2.440936
#> 21 -4.331683  -5.447094 -2.423440
#> 22 -3.836725  -6.391331 -2.542535
#> 23 -1.575050 -18.764896 -2.323508
#> 24 -4.459179  -6.213121 -2.617849
#> 25 -1.493305 -18.776560 -2.599785
#> 26 -1.228524 -18.789749 -2.145909
#> 27 -1.551775 -18.789749 -2.404973
#> 28 -1.433190 -18.777853 -2.333896
#> 29 -1.500624 -18.441802 -2.418475
#> 30 -1.964506  -7.754543 -5.122210
#> 31 -2.906906  -7.857036 -6.408023
#> 32 -5.116108  -6.311221 -2.559459
#> 33 -3.974089  -6.347763 -2.980957
#> 34 -4.473552  -6.473233 -2.320627
#> 35 -5.064740  -6.292991 -1.535719
#> 36 -5.140675  -6.419123 -1.609447
#> 37 -5.057330  -6.290476 -1.923566
#> 38 -4.708377  -6.010671 -2.173221
#> 39 -4.820725  -6.131883 -2.332653
#> 40 -5.136500  -7.031943 -3.067534
#> 41 -4.731216  -5.496980 -2.105367
#> 42 -4.355460  -5.512801 -1.991764
#> 43 -5.046513  -6.216444 -2.831781
#> 44 -4.344061  -5.447976 -2.920233
#> 45 -4.380224  -5.312460 -2.730639
#> 46 -4.164864  -5.237561 -2.245957
#> 47 -4.289069  -5.443712 -2.715687
#> 48 -2.340284 -18.505255 -2.027348
#> 49 -1.949108 -18.789749 -2.737656
predict(myPLN, trichoptera, "response")
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 
#>  5  3  1  1  1  3  5  1  5  5  5  5  5  5  1  1  5  5  5  5  5  5  2  5  2  2 
#> 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 
#>  2  2  2  4  4  3  3  3  3  3  3  3  3  1  5  5  3  5  5  5  5  2  2 
#> Levels: 1 2 3 4 5