Prediction for a PLNmixturefit object

Predict either posterior probabilities for each group or latent positions based on new samples

# S3 method for PLNmixturefit
predict(
  object,
  newdata,
  type = c("posterior", "response", "position"),
  prior = matrix(rep(1/object$k, object$k), nrow(newdata), object$k, byrow = TRUE),
  control = PLNmixture_param(),
  ...
)

Arguments

object

an R6 object with class PLNmixturefit

newdata

A data frame in which to look for variables, offsets and counts with which to predict.

type

The type of prediction required. The default posterior are posterior probabilities for each group , response is the group with maximal posterior probability and latent is the averaged latent in the latent space, with weights equal to the posterior probabilities.

prior

User-specified prior group probabilities in the new data. The default uses a uniform prior.

control

a list-like structure for controlling the fit. See PLNmixture_param() for details.

...

additional parameters for S3 compatibility. Not used

Value

A matrix of posterior probabilities for each group (if type = "posterior"), a matrix of (average) position in the latent space (if type = "position") or a vector of predicted groups (if type = "response").

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLNmixture(Abundance ~ 1 + offset(log(Offset)),
           data = trichoptera, control = PLNmixture_param(smoothing = "none"))  %>% getBestModel()
#> 
#>  Initialization...
#> 
#>  Adjusting 5 PLN mixture models.
#> 	number of cluster = 1 	number of cluster = 2 	number of cluster = 3 	number of cluster = 4 	number of cluster = 5 
#>  Post-treatments
#>  DONE!
predict(myPLN, trichoptera, "posterior")
#>            [,1]         [,2]         [,3]         [,4]         [,5]
#> 1  1.927800e-03 6.362164e-02 9.344506e-01 2.220446e-16 2.220446e-16
#> 2  1.851664e-05 9.028952e-01 7.504585e-02 2.204047e-02 2.220446e-16
#> 3  2.220446e-16 2.062366e-01 5.764913e-04 7.931869e-01 2.220446e-16
#> 4  2.220446e-16 3.050115e-05 2.220446e-16 9.999695e-01 2.220446e-16
#> 5  2.220446e-16 7.911615e-03 3.743665e-04 9.917140e-01 2.220446e-16
#> 6  1.062111e-05 4.769561e-01 1.690464e-01 3.539869e-01 2.220446e-16
#> 7  2.705191e-03 1.210877e-01 8.723550e-01 3.852149e-03 2.220446e-16
#> 8  1.485557e-06 2.462836e-02 1.915767e-02 9.562125e-01 2.220446e-16
#> 9  1.664583e-04 4.231454e-01 5.646215e-01 1.206667e-02 2.220446e-16
#> 10 8.295336e-03 3.803841e-01 5.890591e-01 2.226151e-02 1.283327e-14
#> 11 1.923325e-02 2.976128e-01 5.636136e-01 1.162478e-01 3.292487e-03
#> 12 4.325325e-02 1.947186e-01 3.447709e-01 3.312843e-01 8.597300e-02
#> 13 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 14 3.677104e-06 2.220446e-16 9.999963e-01 2.220446e-16 2.220446e-16
#> 15 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16
#> 17 4.024899e-06 2.220446e-16 9.999960e-01 2.220446e-16 2.220446e-16
#> 18 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 19 2.220446e-16 1.219905e-04 9.998089e-01 6.911866e-05 2.220446e-16
#> 20 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 21 4.341443e-06 5.494927e-03 9.944170e-01 8.370332e-05 2.220446e-16
#> 22 6.855663e-02 5.921624e-02 8.718721e-01 3.550780e-04 2.220446e-16
#> 23 9.981406e-01 4.172603e-13 1.859353e-03 2.220446e-16 2.220446e-16
#> 24 2.774868e-02 3.140866e-01 5.684601e-01 8.720829e-02 2.496287e-03
#> 25 9.990195e-01 2.220446e-16 9.804923e-04 2.220446e-16 2.220446e-16
#> 26 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 27 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 28 9.985182e-01 2.836782e-16 8.885933e-04 2.220446e-16 5.931790e-04
#> 29 9.702868e-01 1.690943e-04 9.039604e-03 1.458130e-04 2.035866e-02
#> 30 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 31 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 32 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 33 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 34 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 35 2.220446e-16 9.999343e-01 6.568600e-05 2.220446e-16 2.220446e-16
#> 36 2.220446e-16 8.786021e-01 4.530327e-04 1.209449e-01 2.220446e-16
#> 37 2.220446e-16 9.999604e-01 3.958189e-05 2.220446e-16 2.220446e-16
#> 38 2.358559e-03 6.524222e-01 3.452192e-01 2.220446e-16 2.220446e-16
#> 39 5.400491e-04 8.445704e-01 1.548456e-01 4.394254e-05 2.220446e-16
#> 40 4.101922e-05 1.539143e-01 4.817650e-02 7.975588e-01 3.093803e-04
#> 41 2.220446e-16 3.519358e-01 5.292006e-01 1.188635e-01 2.220446e-16
#> 42 4.344410e-03 2.220446e-16 9.956556e-01 2.220446e-16 2.220446e-16
#> 43 3.490976e-06 8.821092e-01 1.155680e-01 2.319305e-03 2.220446e-16
#> 44 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 45 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 46 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 47 6.828032e-04 9.296933e-15 9.961265e-01 3.190737e-03 2.220446e-16
#> 48 9.787127e-01 2.220446e-16 2.128728e-02 2.220446e-16 2.220446e-16
#> 49 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
predict(myPLN, trichoptera, "position")
#>          [,1]       [,2]      [,3]       [,4]        [,5]      [,6]       [,7]
#> 1   -6.952432  -7.597028 -2.129742  -5.502291 -0.55747858 -3.691462  -5.935698
#> 2  -18.386223 -18.438206 -2.012377  -7.122220 -0.53771029 -3.547862 -18.303711
#> 3  -19.889154 -19.889551 -3.728146 -17.330428 -0.19143922 -3.809499 -19.888530
#> 4  -20.034884 -20.034884 -4.092148 -20.034507 -0.09198909 -3.767888 -20.034884
#> 5  -20.024414 -20.024669 -3.591103 -19.926157 -0.14468110 -3.653258 -20.024019
#> 6  -17.361570 -17.478507 -2.796895 -11.320494 -0.30894838 -3.396929 -17.176320
#> 7   -7.760752  -8.365925 -2.180237  -5.633331 -0.63888964 -3.593593  -6.798379
#> 8  -19.751699 -19.764895 -3.877809 -19.432636 -0.17745509 -3.603111 -19.730921
#> 9  -11.883982 -12.275190 -1.893481  -6.227152 -0.48758679 -3.581663 -11.262773
#> 10 -11.453463 -11.862806 -2.060487  -6.306819 -0.63077778 -3.541512 -10.801832
#> 11 -11.695564 -12.060163 -2.450401  -7.595071 -0.46504005 -3.594534 -11.038056
#> 12 -14.204788 -13.727523 -3.330431 -11.538566 -0.30014777 -3.977641 -12.935468
#> 13  -5.965096  -6.803240 -2.722298  -5.417313 -0.26341489 -3.500598  -5.042602
#> 14  -6.111280  -6.800283 -2.281432  -5.405758 -0.52997607 -3.433071  -4.876841
#> 15 -20.034904 -20.034904 -4.674314 -20.034904 -0.18331417 -3.498295 -20.034904
#> 16 -20.034904 -20.034904 -4.184320 -20.034904 -0.19474903 -3.202599 -20.034904
#> 17  -6.120347  -6.804889 -2.685032  -5.423693 -0.32217332 -3.658238  -5.051694
#> 18  -6.117827  -6.803608 -2.540304  -5.269811 -0.51993008 -3.266443  -5.044636
#> 19  -6.122207  -6.806973 -2.315218  -5.423481 -0.62426229 -3.064437  -5.052468
#> 20  -6.144507  -6.666190 -2.672075  -5.469984 -0.58573044 -3.447892  -4.979668
#> 21  -6.186013  -6.870812 -2.204390  -5.416980 -0.57430113 -3.709080  -5.108163
#> 22  -6.864181  -7.470824 -2.203728  -5.506507 -0.78818941 -3.759044  -5.905348
#> 23  -5.669279  -5.710099 -3.794551  -5.456168 -0.89983165 -5.661475  -5.664019
#> 24 -11.496910 -11.871975 -2.388558  -7.187866 -0.52845258 -3.597212 -10.841802
#> 25  -5.676480  -5.716437 -3.316927  -5.664351 -1.03863454 -5.472207  -5.473509
#> 26  -5.685737  -5.724695 -3.722172  -5.674398 -1.36918752 -5.682563  -5.682563
#> 27  -5.688676  -5.727526 -3.739294  -5.677369 -1.14313552 -5.685511  -5.685511
#> 28  -5.668556  -5.704031 -3.753979  -5.656326 -1.18599216 -5.657471  -5.658208
#> 29  -5.898436  -5.773239 -3.796777  -5.878619 -0.96672489 -5.674786  -5.675188
#> 30 -16.848714  -8.370313 -6.654046 -16.848714 -0.30982049 -7.223519  -6.458334
#> 31 -16.848729  -8.612932 -5.752107 -16.848729 -0.10984059 -7.523765  -6.590047
#> 32 -19.367107 -19.367107 -2.523073  -6.984219 -0.96451269 -3.245682 -19.367107
#> 33 -19.367107 -19.367107 -2.437199  -7.003838 -0.77723464 -2.998319 -19.367107
#> 34 -19.367107 -19.367107 -3.004718  -7.076122 -0.40045420 -2.879844 -19.367107
#> 35 -19.366237 -19.366282 -2.425984  -6.974621 -0.81631331 -2.758952 -19.366167
#> 36 -19.441877 -19.442185 -2.979643  -8.562685 -0.41699002 -3.445621 -19.441402
#> 37 -19.366583 -19.366610 -1.419996  -6.588409 -1.06108323 -3.449182 -19.366540
#> 38 -14.756336 -14.995425 -1.904643  -6.410452 -1.15100733 -3.590235 -14.377269
#> 39 -17.306045 -17.413294 -1.647375  -6.709758 -0.90420133 -3.562679 -17.135906
#> 40 -19.259524 -19.290248 -3.606579 -17.314772 -0.15103824 -3.709638 -19.203606
#> 41 -12.438562 -12.799714 -2.748707  -7.711341 -0.37994668 -3.885159 -11.877033
#> 42  -6.116710  -6.799326 -2.559388  -5.272395 -0.49915572 -3.777176  -5.049466
#> 43 -17.839222 -17.917597 -3.032816  -6.839695 -0.30862869 -3.843673 -17.718328
#> 44  -6.114929  -6.802135 -2.196438  -5.263204 -0.46598068 -3.604995  -4.742374
#> 45  -6.121783  -6.805626 -2.445480  -5.278750 -0.53275958 -3.431192  -4.910911
#> 46  -6.005313  -6.821286 -2.964957  -5.483868 -0.38494270 -3.902682  -5.000967
#> 47  -6.149162  -6.838623 -2.310185  -5.440058 -0.46323020 -3.638454  -5.056112
#> 48  -5.491074  -5.741288 -3.832451  -5.662037 -0.72484676 -5.633685  -5.661983
#> 49  -5.718471  -5.565635 -3.747405  -5.707480 -1.15692325 -5.715394  -5.715394
#>          [,8]       [,9]     [,10]     [,11]     [,12]     [,13]      [,14]
#> 1   -5.657241  -7.584625 -2.875753 -3.464675 -4.791183 -5.248231  -5.371188
#> 2  -17.985227 -18.437212 -3.831800 -5.625792 -3.167818 -3.274606  -7.136736
#> 3   -9.250990 -19.889543 -4.488528 -5.395184 -3.536829 -5.816058 -17.333031
#> 4   -6.675062 -20.034884 -4.708067 -5.439347 -4.339587 -6.519842 -20.034507
#> 5   -6.738046 -20.024664 -4.685712 -5.327400 -4.418417 -6.454570 -19.926269
#> 6  -12.375735 -17.476272 -3.993209 -5.192189 -3.947139 -4.673094 -11.333528
#> 7   -6.479259  -8.354264 -2.717490 -3.509409 -4.835960 -5.074178  -5.666152
#> 8   -6.900243 -19.764643 -4.548925 -5.245753 -4.504843 -6.361005 -19.433730
#> 9  -10.929764 -12.267711 -3.373832 -4.342849 -4.433779 -4.385910  -6.255791
#> 10 -10.316854 -11.854667 -3.286600 -4.215822 -4.450209 -4.406620  -6.320024
#> 11  -9.295647 -12.046608 -3.377578 -4.192000 -4.533087 -4.681363  -7.540981
#> 12  -8.461935 -13.582413 -3.862551 -4.400344 -4.861266 -5.036259 -10.332207
#> 13  -4.608613  -6.790097 -2.971084 -3.395724 -4.900309 -5.404800  -5.458056
#> 14  -4.730848  -6.787101 -2.988618 -3.322059 -5.030488 -5.393093  -5.446942
#> 15  -6.397121 -20.034904 -3.786353 -5.026949 -4.976943 -6.648263 -20.034904
#> 16  -6.668396 -20.034904 -4.354109 -4.857257 -4.452186 -6.512666 -20.034904
#> 17  -4.764612  -6.638599 -2.815504 -3.091425 -5.056066 -5.411246  -5.464185
#> 18  -4.755391  -6.790470 -3.093875 -2.635294 -5.049038 -5.406243  -5.459426
#> 19  -4.763905  -6.793849 -3.010990 -2.869059 -5.053903 -5.409667  -5.464027
#> 20  -4.715456  -6.804414 -3.057465 -2.320256 -5.120410 -5.458075  -5.366813
#> 21  -4.814205  -6.857708 -2.545580 -3.217287 -5.026108 -5.383189  -5.457946
#> 22  -5.588158  -7.456565 -2.760059 -3.373368 -4.975404 -5.038370  -5.410923
#> 23  -4.956431  -5.670541 -3.286777 -3.313715 -5.674304 -2.514768  -3.555166
#> 24  -9.480285 -11.859299 -3.338807 -4.171077 -4.524175 -4.573753  -7.127754
#> 25  -4.971742  -5.677143 -3.525440 -3.050904 -5.682425 -2.633076  -3.434711
#> 26  -4.990909  -5.685737 -3.431186 -3.146331 -5.495920 -2.386838  -3.346229
#> 27  -4.996583  -5.688676 -2.873241 -3.029823 -5.695606 -2.021870  -3.853125
#> 28  -4.743910  -5.663247 -3.409892 -3.246195 -5.669759 -2.375320  -3.499108
#> 29  -5.003162  -5.701611 -3.457488 -3.075241 -5.721388 -2.463234  -3.714336
#> 30  -7.754542  -6.843394 -5.378603 -3.153264 -8.608358 -4.190844  -3.180123
#> 31  -7.857003  -6.910069 -5.918531 -4.071642 -8.477156 -5.166205  -4.018122
#> 32 -19.367107 -19.367107 -4.299759 -5.929019 -3.493916 -1.126668  -6.616028
#> 33 -19.367107 -19.367107 -3.833523 -5.643253 -2.432348 -1.887437  -7.015861
#> 34 -19.367107 -19.367107 -4.510286 -6.142702 -4.015863 -2.529970  -7.087399
#> 35 -19.366148 -19.366281 -3.470866 -5.903481 -3.559082 -2.980838  -6.986979
#> 36 -17.821440 -19.442179 -4.355849 -5.858176 -4.214790 -3.655724  -8.573453
#> 37 -19.366529 -19.366609 -3.675803 -5.537182 -3.538592 -2.645004  -6.985723
#> 38 -14.271464 -14.990766 -2.584380 -4.930159 -4.155299 -3.612490  -6.428550
#> 39 -17.088166 -17.411225 -3.441141 -5.430322 -3.893056 -3.127493  -6.725854
#> 40  -8.488495 -19.289112 -4.387750 -5.244688 -4.154948 -5.889694 -17.314570
#> 41 -10.134229 -12.792785 -3.295943 -4.552130 -4.471364 -4.900051  -7.736948
#> 42  -4.473647  -6.786082 -3.106143 -3.413129 -5.053884 -5.100564  -5.453954
#> 43 -17.655290 -17.916091 -2.525434 -5.340175 -4.283363 -3.413520  -6.855032
#> 44  -4.599366  -6.788977 -2.823170 -3.369278 -4.893266 -5.400448  -5.453921
#> 45  -4.488126  -6.792514 -2.560565 -3.333129 -4.915227 -5.414097  -5.466893
#> 46  -4.871179  -6.808374 -3.380181 -2.895330 -4.873066 -5.472109  -5.243807
#> 47  -4.561380  -6.825424 -2.870847 -3.097177 -5.010971 -5.381798  -5.480276
#> 48  -4.780358  -5.505448 -3.209034 -3.402023 -5.672051 -2.336863  -3.835581
#> 49  -5.052446  -5.718471 -3.638554 -3.153439 -5.725209 -1.055879  -3.994076
#>        [,15]      [,16]     [,17]
#> 1  -4.037585  -5.514857 -2.711958
#> 2  -4.872011  -6.208768 -2.844314
#> 3  -4.944631  -7.065557 -3.250565
#> 4  -4.865926  -7.300880 -3.654925
#> 5  -5.083479  -7.272585 -3.215892
#> 6  -4.924522  -6.473574 -2.962180
#> 7  -4.342039  -5.563720 -2.648230
#> 8  -5.218136  -7.212949 -2.729974
#> 9  -4.556154  -5.797331 -2.746276
#> 10 -4.496575  -5.879191 -2.615148
#> 11 -4.500555  -6.133575 -2.653133
#> 12 -4.429181  -6.952395 -2.968272
#> 13 -4.354861  -5.302308 -2.955446
#> 14 -4.179351  -5.440512 -2.404590
#> 15 -5.405506  -7.153268 -2.544899
#> 16 -5.397867  -7.297445 -2.812882
#> 17 -4.371853  -5.457866 -2.703979
#> 18 -4.358689  -5.453020 -2.559139
#> 19 -4.368276  -5.456676 -2.331637
#> 20 -3.897247  -5.502687 -2.440918
#> 21 -4.330866  -5.446622 -2.423501
#> 22 -3.837245  -6.356945 -2.542686
#> 23 -1.575031 -18.260612 -2.323541
#> 24 -4.460135  -6.200339 -2.621880
#> 25 -1.493299 -18.271917 -2.599709
#> 26 -1.228569 -18.284515 -2.145981
#> 27 -1.551766 -18.284515 -2.404973
#> 28 -1.433575 -18.266871 -2.335845
#> 29 -1.500698 -17.951295 -2.418536
#> 30 -1.964506  -7.754542 -5.122210
#> 31 -2.906906  -7.857003 -6.408023
#> 32 -5.114034  -6.312002 -2.563039
#> 33 -3.979990  -6.348033 -2.983056
#> 34 -4.475780  -6.472015 -2.321847
#> 35 -5.063188  -6.294041 -1.539756
#> 36 -5.141440  -6.430457 -1.612261
#> 37 -5.055857  -6.291564 -1.928210
#> 38 -4.693653  -5.994163 -2.185200
#> 39 -4.816584  -6.127913 -2.342197
#> 40 -5.133372  -7.025851 -3.079845
#> 41 -4.738985  -5.506262 -2.104708
#> 42 -4.354524  -5.510134 -1.992686
#> 43 -5.049534  -6.223318 -2.835064
#> 44 -4.343197  -5.447481 -2.918622
#> 45 -4.379235  -5.312554 -2.729808
#> 46 -4.164739  -5.237931 -2.246202
#> 47 -4.288423  -5.442945 -2.714659
#> 48 -2.339965 -18.011070 -2.027469
#> 49 -1.949039 -18.284515 -2.737571
predict(myPLN, trichoptera, "response")
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 
#>  3  2  4  4  4  2  3  4  3  3  3  3  3  3  4  4  3  3  3  3  3  3  1  3  1  1 
#> 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 
#>  1  1  1  5  5  2  2  2  2  2  2  2  2  4  3  3  2  3  3  3  3  1  1 
#> Levels: 1 2 3 4 5