Prediction for a PLNmixturefit object

Predict either posterior probabilities for each group or latent positions based on new samples

# S3 method for PLNmixturefit
predict(
  object,
  newdata,
  type = c("posterior", "response", "position"),
  prior = matrix(rep(1/object$k, object$k), nrow(newdata), object$k, byrow = TRUE),
  control = list(),
  ...
)

Arguments

object
an R6 object with class PLNmixturefit

newdata
A data frame in which to look for variables, offsets and counts with which to predict.

type
The type of prediction required. The default posterior are posterior probabilities for each group , response is the group with maximal posterior probability and latent is the averaged latent in the latent space, with weights equal to the posterior probabilities.

prior
User-specified prior group probabilities in the new data. The default uses a uniform prior.

control
a list for controlling the optimization. See PLN() for details.

...
additional parameters for S3 compatibility. Not used

Value

A matrix of posterior probabilities for each group (if type = "posterior"), a matrix of (average) position in the latent space (if type = "position") or a vector of predicted groups (if type = "response").

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLNmixture(Abundance ~ 1 + offset(log(Offset)),
           data = trichoptera, control_main = list(iterates = 0))  %>% getBestModel()

#> 
#>  Initialization...
#> 
#>  Adjusting 5 PLN mixture models.
#> 	number of cluster = 1 	number of cluster = 2 	number of cluster = 3 	number of cluster = 4 	number of cluster = 5 
#> 
#>  Smoothing PLN mixture models.
#> 
#>  Post-treatments
#>  DONE!

predict(myPLN, trichoptera, "posterior")

#>            [,1]         [,2]         [,3]         [,4]         [,5]
#> 1  5.931913e-01 9.730701e-04 4.058356e-01 2.220446e-16 8.138748e-13
#> 2  1.594856e-02 1.831007e-01 8.009507e-01 2.220446e-16 3.074607e-11
#> 3  5.050181e-05 9.016211e-01 9.832836e-02 2.220446e-16 2.855608e-10
#> 4  2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 5  2.220446e-16 7.566208e-01 2.433792e-01 2.220446e-16 1.499620e-12
#> 6  3.152599e-04 2.817906e-01 7.178942e-01 2.220446e-16 4.293726e-09
#> 7  2.271487e-01 1.478303e-02 7.580683e-01 2.220446e-16 1.310265e-08
#> 8  1.464234e-05 4.004455e-01 5.995398e-01 2.220446e-16 4.974595e-09
#> 9  5.078219e-02 4.210227e-02 6.475945e-01 2.595210e-01 3.703990e-09
#> 10 2.195136e-01 7.922614e-02 5.783246e-01 1.229345e-01 1.118498e-06
#> 11 1.186562e-01 1.676891e-01 6.582323e-01 5.119762e-02 4.224790e-03
#> 12 7.592168e-02 2.940339e-01 5.164446e-01 1.629711e-02 9.730279e-02
#> 13 2.220446e-16 2.179596e-08 1.000000e+00 2.220446e-16 4.497162e-16
#> 14 6.494475e-05 2.140910e-08 9.999350e-01 2.220446e-16 2.220446e-16
#> 15 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 16 2.220446e-16 7.147541e-01 2.852459e-01 2.220446e-16 2.220446e-16
#> 17 1.081077e-05 1.982711e-10 9.999892e-01 2.220446e-16 1.310827e-13
#> 18 6.493687e-12 3.026834e-13 1.000000e+00 2.220446e-16 2.220446e-16
#> 19 2.543694e-13 7.353071e-08 9.999999e-01 2.220446e-16 2.220446e-16
#> 20 6.100841e-06 2.220446e-16 9.999939e-01 2.220446e-16 2.220446e-16
#> 21 4.693813e-03 3.941449e-05 9.952668e-01 2.220446e-16 2.220446e-16
#> 22 9.596422e-01 6.092760e-04 3.974856e-02 2.220446e-16 5.313753e-11
#> 23 9.991401e-01 2.007690e-11 8.598579e-04 2.220446e-16 2.236650e-10
#> 24 1.684728e-01 1.534302e-01 6.181219e-01 5.683099e-02 3.144133e-03
#> 25 1.000000e+00 2.220446e-16 6.559848e-09 2.220446e-16 3.581153e-09
#> 26 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 27 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16
#> 28 3.830497e-01 1.101409e-06 1.801151e-04 6.167657e-01 3.361468e-06
#> 29 9.584852e-01 1.063159e-03 7.087673e-03 2.220446e-16 3.336400e-02
#> 30 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 31 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 32 5.211955e-06 9.999948e-01 2.220446e-16 2.220446e-16 2.220446e-16
#> 33 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 34 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 35 1.395546e-13 9.535529e-01 4.644708e-02 2.220446e-16 2.220446e-16
#> 36 2.220446e-16 9.848967e-01 1.510327e-02 2.220446e-16 2.220446e-16
#> 37 6.314438e-01 7.428752e-15 3.685562e-01 2.220446e-16 2.220446e-16
#> 38 9.358326e-01 9.954357e-04 6.317192e-02 2.220446e-16 2.220446e-16
#> 39 6.958575e-01 2.157264e-02 2.825698e-01 2.220446e-16 1.860848e-15
#> 40 1.127295e-03 4.342501e-01 5.634665e-01 2.220446e-16 1.156155e-03
#> 41 2.220446e-16 3.968078e-04 9.996032e-01 2.220446e-16 2.220446e-16
#> 42 2.439407e-03 1.134159e-08 9.975606e-01 2.220446e-16 2.220446e-16
#> 43 9.747641e-05 1.975213e-02 9.801504e-01 2.220446e-16 2.220446e-16
#> 44 1.969512e-05 2.220446e-16 9.999803e-01 2.220446e-16 2.220446e-16
#> 45 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 46 2.220446e-16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 47 1.785201e-02 2.267866e-03 9.798801e-01 2.220446e-16 5.220341e-08
#> 48 8.148985e-01 9.519178e-12 1.851015e-01 2.220446e-16 2.220446e-16
#> 49 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16 2.220446e-16

predict(myPLN, trichoptera, "position")

#>          [,1]       [,2]      [,3]       [,4]       [,5]       [,6]       [,7]
#> 1   -6.140754  -6.436193 -2.286924  -5.763906 -0.6135075  -4.943266  -5.694894
#> 2   -8.517587  -9.072904 -2.602323  -7.780329 -0.4768668  -3.508433  -7.637950
#> 3  -16.704199 -16.772233 -3.437150 -16.613979 -0.2719967  -3.461361 -16.596579
#> 4  -17.823752 -17.823752 -3.779915 -17.823752 -0.1258150  -3.540783 -17.823752
#> 5  -15.054901 -15.222925 -3.144732 -14.832692 -0.2089526  -3.402151 -14.789975
#> 6   -9.649470 -10.146723 -2.762187  -8.989167 -0.3743590  -3.383962  -8.861614
#> 7   -6.488056  -7.019021 -2.474783  -5.788370 -0.5807959  -4.043290  -5.654534
#> 8  -11.000790 -11.415699 -3.001999 -10.450426 -0.2823882  -3.404345 -10.344247
#> 9  -15.095131 -15.545081 -2.260280 -14.318814 -0.4950770 -11.650080 -13.983967
#> 10 -11.109925 -11.516223 -2.386379 -10.490342 -0.5900211  -7.808026 -10.284213
#> 11  -9.946837 -10.371739 -2.632798  -9.304054 -0.4779349  -5.337586  -9.101462
#> 12 -11.245238 -10.815137 -3.175134 -10.757522 -0.3861584  -4.474013  -9.672912
#> 13  -6.398742  -7.136320 -2.712242  -5.528525 -0.3018541  -3.469378  -5.351959
#> 14  -6.443604  -7.135643 -2.542542  -5.525574 -0.4808167  -3.447146  -5.302528
#> 15 -17.823753 -17.823753 -4.386193 -17.823753 -0.1971156  -3.409738 -17.823753
#> 16 -14.580541 -14.776506 -3.583488 -14.322920 -0.2338154  -3.145076 -14.273735
#> 17  -6.445497  -7.136644 -2.696919  -5.530194 -0.3448998  -3.524308  -5.353948
#> 18  -6.444982  -7.136395 -2.638278  -5.483082 -0.4871373  -3.386297  -5.352403
#> 19  -6.445356  -7.136583 -2.536568  -5.529824 -0.5616143  -3.308938  -5.353506
#> 20  -6.450598  -7.093175 -2.687831  -5.542751 -0.5628159  -3.448639  -5.323877
#> 21  -6.441765  -7.130612 -2.506930  -5.528212 -0.5107696  -3.549111  -5.352009
#> 22  -5.935709  -5.986885 -2.167314  -5.892488 -0.9721915  -5.808388  -5.890201
#> 23  -5.909513  -5.934674 -2.597566  -5.740467 -0.8823772  -5.905066  -5.906640
#> 24  -9.924688 -10.331818 -2.593046  -9.314540 -0.5234776  -5.629178  -9.125423
#> 25  -5.913677  -5.938152 -2.339955  -5.906804 -1.0204105  -5.751092  -5.751092
#> 26  -5.919572  -5.943907 -2.685228  -5.912738 -1.3266209  -5.917661  -5.917661
#> 27  -5.921354  -5.945647 -2.715803  -5.914532 -1.1299079  -5.919446  -5.919446
#> 28 -25.726582 -25.736137 -2.011206 -25.296256 -0.8013115 -23.509404 -24.775513
#> 29  -6.276703  -6.034393 -2.624043  -6.263533 -0.9059831  -5.930920  -5.931336
#> 30 -16.543444  -8.397871 -6.145589 -16.543444 -0.3121057  -7.285235  -6.459931
#> 31 -16.543448  -8.451358 -5.855280 -16.543448 -0.1128284  -7.360903  -6.501009
#> 32 -17.823689 -17.823689 -3.002083 -17.823689 -0.9583235  -3.172616 -17.823689
#> 33 -17.823751 -17.823751 -2.797815 -17.823751 -0.7926160  -2.999379 -17.823751
#> 34 -17.823752 -17.823752 -3.200619 -17.823752 -0.4179673  -2.894018 -17.823752
#> 35 -17.295292 -17.327381 -2.968022 -17.252820 -0.8135863  -2.850192 -17.244647
#> 36 -17.651938 -17.662359 -3.335746 -17.638167 -0.4534226  -3.298341 -17.635522
#> 37  -6.119950  -6.389827 -1.799190  -5.661966 -0.9949203  -4.829727  -5.716724
#> 38  -5.953731  -6.020523 -2.086921  -5.889138 -1.1973557  -5.752309  -5.882566
#> 39  -6.315766  -6.528660 -2.081888  -6.050926 -0.8938156  -5.172623  -6.004144
#> 40 -11.395979 -11.776860 -2.975475 -10.877836 -0.2793222  -3.451089 -10.766132
#> 41  -6.450909  -7.141345 -2.652788  -5.537277 -0.3963139  -3.581760  -5.361518
#> 42  -6.443863  -7.133574 -2.643826  -5.484143 -0.4738067  -3.566729  -5.354235
#> 43  -6.672885  -7.348986 -2.766310  -5.779693 -0.3407941  -3.569709  -5.608197
#> 44  -6.444375  -7.136072 -2.494911  -5.481534 -0.4463046  -3.504705  -5.259039
#> 45  -6.445801  -7.136806 -2.592564  -5.485209 -0.5017417  -3.444489  -5.309205
#> 46  -6.406461  -7.140034 -2.832397  -5.546667 -0.3953964  -3.629507  -5.328720
#> 47  -6.458645  -7.137866 -2.567389  -5.557031 -0.4306999  -3.554756  -5.382944
#> 48  -5.874006  -6.161252 -2.723170  -5.837651 -0.7028181  -5.472869  -5.808878
#> 49  -5.939538  -5.806908 -2.875796  -5.932836 -1.1579054  -5.937664  -5.937664
#>         [,8]       [,9]      [,10]      [,11]      [,12]      [,13]     [,14]
#> 1  -5.212964  -6.420625  -2.945095  -3.473986  -4.481048  -3.504385 -4.717910
#> 2  -5.479180  -9.070414  -3.390992  -3.714831  -4.061131  -4.289648 -6.025625
#> 3  -6.627966 -16.771974  -4.330956  -5.121806  -3.490198  -3.479606 -7.221077
#> 4  -6.818287 -17.823752  -4.567736  -5.440318  -4.048318  -3.885523 -7.399840
#> 5  -6.406996 -15.222289  -4.230894  -4.888054  -4.094919  -3.807850 -6.991207
#> 6  -5.637121 -10.144836  -3.525200  -3.910327  -4.099116  -4.196421 -6.211736
#> 7  -5.214115  -7.011427  -2.983773  -3.398919  -4.361539  -4.093451 -5.408855
#> 8  -5.829901 -11.414130  -3.703527  -4.159455  -4.185795  -4.083892 -6.407113
#> 9  -4.712650 -15.322244 -11.493587 -11.371018 -12.291623 -23.550022 -5.050626
#> 10 -5.057835 -11.405115  -7.118689  -7.268608  -8.142295 -13.066707 -5.196183
#> 11 -5.357662 -10.316922  -4.971650  -5.242320  -5.860390  -7.856623 -5.686019
#> 12 -5.868060 -10.641981  -4.250075  -4.439103  -5.112796  -5.191033 -5.852376
#> 13 -5.151209  -7.133704  -3.150885  -3.415923  -4.220600  -4.590640 -5.759363
#> 14 -5.192741  -7.133025  -3.163994  -3.392395  -4.255053  -4.582996 -5.756896
#> 15 -6.550164 -17.823753  -3.767815  -5.049700  -4.680928  -4.466172 -7.446307
#> 16 -6.361502 -14.775763  -3.987265  -4.508597  -4.165318  -4.068512 -6.935252
#> 17 -5.199185  -7.087867  -3.084537  -3.301909  -4.271114  -4.594807 -5.760671
#> 18 -5.197391  -7.133779  -3.196598  -3.122196  -4.266663  -4.591579 -5.759659
#> 19 -5.198676  -7.133968  -3.163601  -3.214803  -4.269850  -4.593905 -5.760396
#> 20 -5.171738  -7.136607  -3.165361  -2.852266  -4.312967  -4.625733 -5.725302
#> 21 -5.193494  -7.127893  -2.993584  -3.351999  -4.259023  -4.575702 -5.750047
#> 22 -5.210351  -5.963144  -2.702760  -3.452922  -4.781033  -2.696245 -4.210790
#> 23 -5.214003  -5.910107  -2.863765  -3.475482  -4.808921  -2.661518 -4.001751
#> 24 -5.320893 -10.272832  -5.115495  -5.387527  -6.060325  -8.185978 -5.568303
#> 25 -5.223173  -5.913677  -3.073982  -3.241492  -4.822951  -2.751697 -3.879967
#> 26 -5.234721  -5.919572  -3.028662  -3.310377  -4.686101  -2.519966 -3.778264
#> 27 -5.238188  -5.921354  -2.604163  -3.203478  -4.844938  -2.174782 -4.220350
#> 28 -3.839516 -25.204094 -22.765497 -22.222343 -23.542315 -49.278151 -3.429097
#> 29 -5.292443  -5.957199  -3.033399  -3.291933  -4.914805  -2.656859 -4.131204
#> 30 -7.723461  -6.810371  -5.523305  -3.244467  -8.451183  -4.367732 -3.265448
#> 31 -7.747789  -6.829748  -5.741601  -3.952288  -8.422423  -4.927742 -3.913337
#> 32 -6.795876 -17.823689  -4.552580  -5.356611  -3.781769  -1.339811 -7.190726
#> 33 -6.806257 -17.823751  -4.186868  -5.221847  -2.755137  -2.062256 -7.393024
#> 34 -6.845928 -17.823752  -4.616903  -5.532459  -4.083291  -2.606701 -7.415757
#> 35 -6.717092 -17.327259  -3.972753  -5.248611  -3.871021  -3.168342 -7.309077
#> 36 -6.772278 -17.662320  -4.536790  -5.329683  -4.232558  -3.361205 -7.362925
#> 37 -5.234466  -6.373647  -3.093408  -3.483240  -4.429076  -3.198616 -4.814086
#> 38 -5.212211  -5.997328  -2.471865  -3.461135  -4.770268  -2.632321 -4.254710
#> 39 -5.238421  -6.510789  -2.948894  -3.472379  -4.629133  -3.078853 -4.671185
#> 40 -5.883134 -11.773502  -3.719076  -4.206101  -4.054420  -3.996683 -6.455737
#> 41 -5.202876  -7.138732  -3.147775  -3.446005  -4.234368  -4.599916 -5.763091
#> 42 -5.106286  -7.130906  -3.199852  -3.422154  -4.269372  -4.498033 -5.756245
#> 43 -5.239943  -7.346424  -2.889182  -3.472021  -4.293035  -4.454705 -5.797983
#> 44 -5.149592  -7.133455  -3.097316  -3.407237  -4.216583  -4.587801 -5.758448
#> 45 -5.109222  -7.134192  -2.970558  -3.392103  -4.229232  -4.596679 -5.761277
#> 46 -5.221891  -7.137428  -3.308182  -3.171260  -4.241190  -4.635169 -5.683380
#> 47 -5.143901  -7.134861  -3.128673  -3.325039  -4.253313  -4.537518 -5.729554
#> 48 -5.083152  -6.001630  -2.894460  -3.511850  -4.721859  -2.806773 -4.480142
#> 49 -5.272941  -5.939538  -3.257433  -3.281915  -4.894770  -1.163101 -4.306176
#>        [,15]      [,16]     [,17]
#> 1  -3.183531 -12.795921 -2.600835
#> 2  -4.571035  -5.930488 -2.524206
#> 3  -4.559567  -7.236241 -2.971454
#> 4  -4.510384  -7.458929 -3.460132
#> 5  -4.620533  -6.938759 -2.920475
#> 6  -4.622173  -5.941558 -2.582488
#> 7  -4.053002  -8.228455 -2.482860
#> 8  -4.653127  -6.188681 -2.478042
#> 9  -3.804317 -14.283601 -2.645784
#> 10 -3.742997 -12.156386 -2.556686
#> 11 -4.171431  -8.815678 -2.531375
#> 12 -4.175857  -7.661513 -2.830388
#> 13 -4.603302  -5.309827 -2.582005
#> 14 -4.550199  -5.352860 -2.368082
#> 15 -5.050732  -7.317528 -2.527087
#> 16 -4.852176  -6.864831 -2.676890
#> 17 -4.607414  -5.357690 -2.494250
#> 18 -4.604229  -5.356021 -2.433311
#> 19 -4.606527  -5.357120 -2.338316
#> 20 -4.420700  -5.372492 -2.396316
#> 21 -4.587240  -5.411458 -2.375869
#> 22 -2.153774 -17.392851 -2.520908
#> 23 -1.889988 -17.887297 -2.439106
#> 24 -4.033777  -9.586684 -2.527797
#> 25 -1.758436 -17.898088 -2.668525
#> 26 -1.443045 -17.898088 -2.254354
#> 27 -1.763151 -17.898088 -2.486237
#> 28 -1.926906 -29.668513 -2.850236
#> 29 -1.886789 -17.458951 -2.580893
#> 30 -2.002184  -7.723461 -5.427917
#> 31 -2.859356  -7.747789 -5.929443
#> 32 -4.822972  -7.447064 -2.599271
#> 33 -4.130208  -7.452498 -2.927256
#> 34 -4.453238  -7.473966 -2.355656
#> 35 -4.782158  -7.347560 -1.738378
#> 36 -4.822415  -7.415761 -1.744543
#> 37 -3.446747 -13.276603 -2.127530
#> 38 -2.486254 -17.094865 -2.287873
#> 39 -3.017848 -14.126107 -2.396302
#> 40 -4.638480  -6.273483 -2.655259
#> 41 -4.612961  -5.269836 -2.247118
#> 42 -4.600502  -5.387054 -2.180292
#> 43 -4.628001  -5.407929 -2.537288
#> 44 -4.600491  -5.354505 -2.563531
#> 45 -4.609266  -5.312812 -2.506828
#> 46 -4.517804  -5.287544 -2.286443
#> 47 -4.548233  -5.577035 -2.475820
#> 48 -2.905304 -15.575715 -2.176293
#> 49 -2.096625 -17.898088 -2.770821

predict(myPLN, trichoptera, "response")

#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 
#>  1  3  2  2  2  3  3  3  3  3  3  3  3  3  2  2  3  3  3  3  3  1  1  3  1  1 
#> 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 
#>  1  4  1  5  5  2  2  2  2  2  1  1  1  3  3  3  3  3  3  3  3  1  1 
#> Levels: 1 2 3 4 5