Prediction for a PLNmixturefit object

Predict either posterior probabilities for each group or latent positions based on new samples

# S3 method for PLNmixturefit
predict(
  object,
  newdata,
  type = c("posterior", "response", "position"),
  prior = matrix(rep(1/object$k, object$k), nrow(newdata), object$k, byrow = TRUE),
  control = list(),
  ...
)

Arguments

object

an R6 object with class PLNmixturefit

newdata

A data frame in which to look for variables, offsets and counts with which to predict.

type

The type of prediction required. The default posterior are posterior probabilities for each group , response is the group with maximal posterior probability and latent is the averaged latent in the latent space, with weights equal to the posterior probabilities.

prior

User-specified prior group probabilities in the new data. The default uses a uniform prior.

control

a list for controlling the optimization. See PLN() for details.

...

additional parameters for S3 compatibility. Not used

Value

A matrix of posterior probabilities for each group (if type = "posterior"), a matrix of (average) position in the latent space (if type = "position") or a vector of predicted groups (if type = "response").

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLNmixture(Abundance ~ 1 + offset(log(Offset)),
           data = trichoptera, control_main = list(iterates = 0))  %>% getBestModel()
#> 
#>  Initialization...
#> 
#>  Adjusting 5 PLN mixture models.
#> 	number of cluster = 1 	number of cluster = 2 	number of cluster = 3 	number of cluster = 4 	number of cluster = 5 
#> 
#>  Smoothing PLN mixture models.
#> 
#>  Post-treatments
#>  DONE!
predict(myPLN, trichoptera, "posterior")
#>            [,1]         [,2]         [,3]         [,4]
#> 1  1.470190e-10 9.998423e-01 2.220446e-16 1.576556e-04
#> 2  1.324924e-12 9.868503e-04 2.220446e-16 9.990131e-01
#> 3  3.309535e-15 1.093924e-03 2.220446e-16 9.989061e-01
#> 4  2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 5  2.220446e-16 1.000000e+00 2.220446e-16 6.489916e-12
#> 6  1.215896e-11 2.481084e-02 2.220446e-16 9.751892e-01
#> 7  1.036070e-08 9.323492e-03 2.220446e-16 9.906765e-01
#> 8  1.010093e-12 9.999958e-01 2.220446e-16 4.169819e-06
#> 9  1.969308e-10 3.624631e-03 2.220446e-16 9.963754e-01
#> 10 3.182321e-08 4.260457e-03 2.220446e-16 9.957395e-01
#> 11 3.990643e-07 1.635659e-02 2.220446e-16 9.836430e-01
#> 12 8.641538e-01 1.318448e-01 2.220446e-16 4.001425e-03
#> 13 2.220446e-16 1.000000e+00 2.220446e-16 7.502014e-15
#> 14 1.878600e-14 9.999997e-01 2.220446e-16 3.493327e-07
#> 15 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 16 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 17 1.988509e-15 1.000000e+00 2.220446e-16 1.021895e-13
#> 18 2.220446e-16 1.000000e+00 2.220446e-16 1.222904e-12
#> 19 2.220446e-16 1.000000e+00 2.220446e-16 1.930163e-09
#> 20 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 21 5.016053e-15 9.999689e-01 2.220446e-16 3.109374e-05
#> 22 9.879985e-01 8.658975e-03 2.220446e-16 3.342511e-03
#> 23 9.999962e-01 3.733633e-06 2.220446e-16 5.471593e-08
#> 24 3.772566e-07 1.092046e-02 2.220446e-16 9.890792e-01
#> 25 1.000000e+00 5.283870e-10 2.220446e-16 2.220446e-16
#> 26 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 27 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 28 9.999994e-01 5.428258e-07 2.220446e-16 1.816295e-08
#> 29 9.999779e-01 2.197926e-05 2.220446e-16 1.208260e-07
#> 30 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 31 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
#> 32 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 33 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 34 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 35 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 36 2.220446e-16 1.674795e-15 2.220446e-16 1.000000e+00
#> 37 2.220446e-16 2.220446e-16 2.220446e-16 1.000000e+00
#> 38 6.641026e-14 3.602243e-10 2.220446e-16 1.000000e+00
#> 39 5.884089e-13 2.529953e-09 2.220446e-16 1.000000e+00
#> 40 1.908637e-10 9.999542e-01 2.220446e-16 4.583485e-05
#> 41 2.220446e-16 1.000000e+00 2.220446e-16 1.741387e-13
#> 42 2.220446e-16 4.359273e-13 2.220446e-16 1.000000e+00
#> 43 2.220446e-16 4.930514e-16 2.220446e-16 1.000000e+00
#> 44 2.220446e-16 1.000000e+00 2.220446e-16 3.477536e-12
#> 45 2.220446e-16 1.000000e+00 2.220446e-16 2.757227e-14
#> 46 2.220446e-16 1.000000e+00 2.220446e-16 2.220446e-16
#> 47 3.490925e-09 9.999649e-01 2.220446e-16 3.513933e-05
#> 48 9.999999e-01 8.469402e-16 2.220446e-16 6.358980e-08
#> 49 1.000000e+00 2.220446e-16 2.220446e-16 2.220446e-16
predict(myPLN, trichoptera, "position")
#>          [,1]       [,2]      [,3]      [,4]        [,5]      [,6]       [,7]
#> 1   -7.637612  -7.632259 -2.735553 -6.885417 -0.54967761 -3.649471  -6.163548
#> 2  -13.999792 -13.999786 -2.180961 -6.293225 -0.61643100 -3.448714 -13.998331
#> 3  -13.999113 -13.999107 -2.720620 -6.300694 -0.19521775 -3.568166 -13.997494
#> 4   -7.645516  -7.640209 -3.553626 -6.904305 -0.08970106 -3.710283  -6.200697
#> 5   -7.639348  -7.634009 -3.070559 -6.891312 -0.14248296 -3.600203  -6.174235
#> 6  -13.848009 -13.847876 -2.394271 -6.308583 -0.36767149 -3.342663 -13.811299
#> 7  -13.946678 -13.946628 -2.249439 -6.296644 -0.75997341 -3.423700 -13.932882
#> 8   -7.636909  -7.631556 -3.146424 -6.886091 -0.17475180 -3.565048  -6.163543
#> 9  -13.982986 -13.982967 -2.039492 -6.294474 -0.56624031 -3.443753 -13.977623
#> 10 -13.978934 -13.978911 -2.184617 -6.292479 -0.76240808 -3.402187 -13.972630
#> 11 -13.901867 -13.901779 -2.333171 -6.299628 -0.64060377 -3.403841 -13.877664
#> 12  -8.335090  -7.754580 -4.588543 -8.191952 -0.32390472 -6.488517  -6.392697
#> 13  -7.534201  -7.632357 -3.214196 -6.887836 -0.20965280 -3.512649  -6.167097
#> 14  -7.636830  -7.631477 -2.852888 -6.885977 -0.52089749 -3.462904  -6.060001
#> 15  -7.667491  -7.662301 -4.384159 -6.949943 -0.18404955 -3.487595  -6.291172
#> 16  -7.644427  -7.639115 -3.573714 -6.902017 -0.19872739 -3.251446  -6.196061
#> 17  -7.638198  -7.632853 -3.160152 -6.888880 -0.28378571 -3.635898  -6.169244
#> 18  -7.637815  -7.632467 -3.035873 -6.784707 -0.51567980 -3.326012  -6.167575
#> 19  -7.638089  -7.632743 -2.785618 -6.888648 -0.64533840 -3.152735  -6.168767
#> 20  -7.641973  -7.533333 -3.016171 -6.896851 -0.58780260 -3.461567  -6.083248
#> 21  -7.637135  -7.631783 -2.768380 -6.886191 -0.57734458 -3.667222  -6.163994
#> 22  -8.431127  -7.771555 -2.726002 -8.384005 -0.98643293 -6.923483  -6.443084
#> 23  -8.420677  -7.754489 -4.978538 -7.403800 -0.80658852 -6.970951  -6.430178
#> 24 -13.936501 -13.936442 -2.314243 -6.296123 -0.70019863 -3.397837 -13.920341
#> 25  -8.424207  -7.761288 -3.138460 -8.409294 -1.01319950 -6.020473  -5.516735
#> 26  -8.428753  -7.769983 -4.414053 -8.413907 -1.46042551 -7.003752  -6.484138
#> 27  -8.430143  -7.772626 -4.453164 -8.415318 -1.15070280 -7.009231  -6.492934
#> 28  -8.417482  -7.748285 -4.888743 -8.402467 -1.29167461 -6.957488  -6.407284
#> 29  -8.417822  -7.748976 -4.899538 -8.402796 -0.88253960 -6.958931  -6.409878
#> 30  -8.793537  -7.686821 -6.807978 -8.782496 -0.31064293 -6.883314  -6.399553
#> 31  -8.983325  -8.566025 -5.490280 -8.973443 -0.11026204 -7.652905  -6.640310
#> 32 -14.006089 -14.006089 -2.528018 -6.314740 -0.99263569 -3.294035 -14.006089
#> 33 -14.006095 -14.006095 -2.452180 -6.328324 -0.78604456 -3.067384 -14.006095
#> 34 -14.006121 -14.006121 -3.060249 -6.381871 -0.39954771 -2.853752 -14.006121
#> 35 -14.006086 -14.006086 -2.439202 -6.308313 -0.83812686 -2.955295 -14.006086
#> 36 -14.006089 -14.006089 -2.825185 -6.315322 -0.41653262 -3.418986 -14.006089
#> 37 -14.006086 -14.006086 -1.414083 -6.164915 -1.10931456 -3.423582 -14.006086
#> 38 -14.006079 -14.006079 -2.080031 -6.293243 -1.19305401 -3.458819 -14.006079
#> 39 -14.006079 -14.006079 -1.937274 -6.292945 -0.95081501 -3.453758 -14.006079
#> 40  -7.636406  -7.631049 -3.079127 -6.884437 -0.16771161 -3.622398  -6.160512
#> 41  -7.638856  -7.633514 -3.031210 -6.890270 -0.37129962 -3.765183  -6.172099
#> 42 -14.006084 -14.006084 -2.633273 -6.161776 -0.49221509 -3.633616 -14.006084
#> 43 -14.006090 -14.006090 -2.957481 -6.316775 -0.30127675 -3.675543 -14.006090
#> 44  -7.637376  -7.632026 -2.717899 -6.783737 -0.44718796 -3.592161  -5.959360
#> 45  -7.638417  -7.633073 -2.904876 -6.786039 -0.53175674 -3.456115  -6.067226
#> 46  -7.539910  -7.637854 -3.337363 -6.899379 -0.40720743 -3.869712  -6.088617
#> 47  -7.636118  -7.630760 -2.959144 -6.883978 -0.41528381 -3.610128  -6.159466
#> 48  -7.429197  -7.763981 -5.092194 -8.410718 -0.63748694 -6.991189  -6.463740
#> 49  -8.444497  -6.850872 -4.255383 -8.429879 -1.15998564 -7.063278  -6.576821
#>         [,8]       [,9]     [,10]     [,11]     [,12]      [,13]     [,14]
#> 1  -6.045097  -7.615959 -3.344724 -3.726850 -4.914865 -13.826963 -6.086675
#> 2  -6.303810 -13.999770 -3.622733 -5.902202 -3.684792  -2.739320 -7.040844
#> 3  -6.311103 -13.999089 -3.724074 -5.912857 -3.499103  -2.971175 -7.044262
#> 4  -6.088125  -7.624054 -3.898288 -4.103562 -4.841278 -13.828754 -6.227194
#> 5  -6.058455  -7.617752 -3.697484 -3.856190 -4.953621 -13.828742 -6.201434
#> 6  -6.298753 -13.847471 -3.637194 -5.850307 -3.870774  -3.048741 -7.021012
#> 7  -6.300106 -13.946475 -3.322017 -5.881149 -3.959274  -2.777983 -7.032979
#> 8  -6.046406  -7.615259 -3.559666 -3.740210 -5.021424 -13.828690 -6.191003
#> 9  -6.302803 -13.982907 -3.617843 -5.895787 -3.965681  -2.757814 -7.038435
#> 10 -6.300290 -13.978841 -3.582885 -5.890853 -3.941731  -2.677082 -7.036769
#> 11 -6.297080 -13.901511 -3.580966 -5.863533 -3.954382  -2.811984 -7.026390
#> 12 -6.525388  -6.799350 -4.140016 -3.450955 -7.391085  -4.753051 -3.961757
#> 13 -5.947396  -7.616073 -3.414018 -3.780239 -4.930622 -13.828738 -6.194489
#> 14 -6.046136  -7.615179 -3.457499 -3.737479 -5.020676 -13.828733 -6.190767
#> 15 -5.994417  -7.646502 -3.580242 -4.350409 -5.202032 -13.828800 -6.315388
#> 16 -6.082933  -7.622942 -3.680467 -3.799565 -4.923790 -13.828752 -6.222680
#> 17 -6.052849  -7.513098 -3.254386 -3.513384 -5.039308 -13.828739 -6.196577
#> 18 -6.050973  -7.616185 -3.517026 -3.109448 -5.034124 -13.828739 -6.194953
#> 19 -6.052313  -7.616465 -3.437278 -3.315996 -5.037829 -13.828739 -6.196114
#> 20 -5.969011  -7.620434 -3.387041 -2.459187 -5.089114 -13.828747 -6.110110
#> 21 -6.046682  -7.615487 -3.073174 -3.643522 -5.022144 -13.828387 -6.191259
#> 22 -6.616334  -6.695308 -2.964997 -3.742369 -7.742413  -3.794653 -3.925362
#> 23 -6.630849  -6.670589 -3.713724 -3.823260 -7.782098  -3.799320 -3.296914
#> 24 -6.298234 -13.936263 -3.577443 -5.875378 -3.945620  -2.739950 -7.030916
#> 25 -6.650922  -6.689919 -4.603413 -2.949111 -7.788726  -3.971017 -3.036983
#> 26 -6.675813  -6.713937 -4.067430 -3.195414 -6.817831  -2.860892 -2.940819
#> 27 -6.683219  -6.721091 -2.679647 -2.938601 -7.799767  -2.095704 -4.325112
#> 28 -5.622738  -6.652485 -4.348391 -3.600604 -7.776071  -3.573212 -2.995429
#> 29 -6.614147  -6.654552 -4.365570 -2.898021 -7.776685  -3.604051 -3.800543
#> 30 -7.146418  -6.771593 -5.118699 -3.123445 -8.346719  -4.061094 -3.156936
#> 31 -7.510053  -6.918556 -5.872707 -4.098863 -8.018134  -5.185035 -4.045114
#> 32 -6.325921 -14.006089 -3.900286 -5.936827 -3.787144  -1.027131 -6.909255
#> 33 -6.339355 -14.006095 -3.692442 -5.816664 -2.634406  -1.825074 -7.058733
#> 34 -6.392333 -14.006121 -4.265616 -6.033205 -4.106465  -2.513739 -7.084939
#> 35 -6.319566 -14.006086 -3.446605 -5.927442 -3.851513  -2.831978 -7.049134
#> 36 -6.326497 -14.006089 -3.906697 -5.937676 -4.174968  -3.101000 -7.052484
#> 37 -6.318696 -14.006086 -3.560465 -5.784231 -3.842738  -2.594696 -7.048715
#> 38 -6.304666 -14.006079 -3.076558 -5.905317 -3.971200  -2.615965 -7.041975
#> 39 -6.304372 -14.006079 -3.487591 -5.904878 -3.968166  -2.605094 -7.041834
#> 40 -6.042640  -7.614740 -3.513007 -3.701549 -4.907865 -13.828220 -6.187771
#> 41 -6.056056  -7.617249 -3.388764 -3.833947 -4.946773 -13.828741 -6.199355
#> 42 -6.030430 -14.006084 -3.650203 -5.921852 -4.079840  -2.843434 -7.047315
#> 43 -6.327934 -14.006090 -2.666735 -5.798784 -4.187135  -3.028071 -7.053181
#> 44 -5.945714  -7.615737 -3.296791 -3.764400 -4.925962 -13.828738 -6.193095
#> 45 -5.848190  -7.616801 -2.990213 -3.717021 -4.940649 -13.828740 -6.197504
#> 46 -6.076932  -7.621661 -3.707489 -3.152416 -4.906643 -13.828750 -6.013311
#> 47 -5.938132  -7.614447 -3.398112 -3.588587 -5.007780 -13.828339 -6.186827
#> 48 -5.716974  -5.753398 -3.408299 -4.048189 -7.791347  -2.683584 -4.185144
#> 49 -6.754836  -6.790448 -4.453163 -3.187018 -7.825992  -0.960095 -4.621368
#>        [,15]      [,16]     [,17]
#> 1  -4.375137  -5.797627 -2.782323
#> 2  -4.987803 -13.997967 -2.317927
#> 3  -4.874053 -13.997090 -2.647416
#> 4  -4.470307  -5.851263 -3.476936
#> 5  -4.536336  -5.813486 -2.991056
#> 6  -4.982075 -13.802143 -2.393300
#> 7  -4.978584 -13.929441 -2.238718
#> 8  -4.584999  -5.798096 -2.628594
#> 9  -4.985505 -13.976285 -2.302601
#> 10 -4.976457 -13.971057 -2.166851
#> 11 -4.971157 -13.871628 -2.172082
#> 12 -2.562754  -7.857717 -3.321483
#> 13 -4.602309  -5.700350 -3.000996
#> 14 -4.481906  -5.797719 -2.528741
#> 15 -4.932768  -5.881649 -2.511581
#> 16 -4.734714  -5.844673 -2.762046
#> 17 -4.612563  -5.806316 -2.784949
#> 18 -4.604596  -5.803915 -2.652714
#> 19 -4.610292  -5.805630 -2.426536
#> 20 -4.203277  -5.829696 -2.479475
#> 21 -4.586172  -5.798662 -2.543212
#> 22 -1.974297  -8.149093 -3.055538
#> 23 -1.595622  -8.151854 -2.677520
#> 24 -4.972452 -13.916311 -2.151536
#> 25 -1.462242  -8.156463 -3.413720
#> 26 -1.132708  -8.162377 -2.195705
#> 27 -1.551770  -8.164182 -2.683261
#> 28 -1.298128  -8.147682 -2.862543
#> 29 -1.380241  -8.148099 -2.909101
#> 30 -1.953507  -7.905892 -4.729576
#> 31 -2.913693  -8.184161 -6.397944
#> 32 -5.067183 -14.006089 -2.437486
#> 33 -4.447828 -14.006095 -2.853744
#> 34 -4.669227 -14.006121 -2.305442
#> 35 -5.044706 -14.006086 -1.466301
#> 36 -5.069201 -14.006089 -1.511725
#> 37 -5.041599 -14.006086 -1.857401
#> 38 -4.990448 -14.006079 -1.968270
#> 39 -4.989352 -14.006079 -2.076788
#> 40 -4.568602  -5.793593 -2.817722
#> 41 -4.626085  -5.605358 -2.200850
#> 42 -5.031169 -14.006084 -1.730937
#> 43 -5.074225 -14.006090 -2.667008
#> 44 -4.595418  -5.801165 -2.965459
#> 45 -4.617089  -5.705000 -2.809019
#> 46 -4.421694  -5.634317 -2.221878
#> 47 -4.563862  -5.792117 -2.789910
#> 48 -3.247319  -8.158291 -1.990615
#> 49 -2.067989  -8.182728 -3.191199
predict(myPLN, trichoptera, "response")
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 
#>  2  4  4  2  2  4  4  2  4  4  4  1  2  2  2  2  2  2  2  2  2  1  1  4  1  1 
#> 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 
#>  1  1  1  1  1  4  4  4  4  4  4  4  4  2  2  4  4  2  2  2  2  1  1 
#> Levels: 1 2 4