Prepare data for use in PLN models

Prepare data in proper format for use in PLN model and its variants. The function (i) merges a count table and a covariate data frame in the most comprehensive way and (ii) computes offsets from the count table using one of several normalization schemes (TSS, CSS, RLE, GMPR, etc). The function fails with informative messages when the heuristics used for sample matching fail.

prepare_data(counts, covariates, offset = "TSS", ...)

Arguments

counts
Required. An abundance count table, preferably with dimensions names and species as columns.

covariates
Required. A covariates data frame, preferably with row names.

offset
Optional. Normalisation scheme used to compute scaling factors used as offset during PLN inference. Can be "TSS" (Total Sum Scaling, default), "CSS" (Cumulative Sum Scaling, used in metagenomeSeq), "RLE" (Relative Log Expression, used in DESeq2), "GMPR" (Geometric Mean of Pairwise Ratio, introduced in Chen et al., 2018) or "none".

...
Additional parameters passed on to compute_offset

Value

A data.frame suited for use in PLN and its variants with two specials components: an abundance count matrix (in component "Abundance") and an offset vector/matrix (in component "Offset", only if offset is not set to "none")

References

Chen, L., Reeve, J., Zhang, L., Huang, S., Wang, X. and Chen, J. (2018) GMPR: A robust normalization method for zero-inflated count data with application to microbiome sequencing data. PeerJ, 6, e4600 https://doi.org/10.7717/peerj.4600

Paulson, J. N., Colin Stine, O., Bravo, H. C. and Pop, M. (2013) Differential abundance analysis for microbial marker-gene surveys. Nature Methods, 10, 1200-1202 http://dx.doi.org/10.1038/nmeth.2658

Anders, S. and Huber, W. (2010) Differential expression analysis for sequence count data. Genome Biology, 11, R106 https://doi.org/10.1186/gb-2010-11-10-r106

See also

Examples

data(trichoptera)
counts <- trichoptera$Abundance
covariates <- trichoptera; covariates$Abundance <- NULL
proper_data <- prepare_data(counts, covariates, offset = "TSS")

#> Warning: There are no dimension names for the abundance matrix and/or the covariates data.frame.
#> Function will proceed assuming:
#> - samples are in the same order;
#> - samples are rows of the abundance matrix.

#> Error in if (nrow(counts) != nrow(covariates)) {    stop("Incompatible dimensions")}: argument is of length zero

proper_data$Abundance

#> Error in eval(expr, envir, enclos): object 'proper_data' not found

proper_data$Offset

#> Error in eval(expr, envir, enclos): object 'proper_data' not found