PLN RNG

Random generation for the PLN model with latent mean equal to mu, latent covariance matrix equal to Sigma and average depths (sum of counts in a sample) equal to depths

rPLN(
  n = 10,
  mu = rep(0, ncol(Sigma)),
  Sigma = diag(1, 5, 5),
  depths = rep(10000, n)
)

Arguments

n
the sample size

mu
vectors of means of the latent variable

Sigma
covariance matrix of the latent variable

depths
Numeric vector of target depths. The first is recycled if there are not n values

Value

a n * p count matrix, with row-sums close to depths

Details

The default value for mu and Sigma assume equal abundances and no correlation between the different species.

Examples

## 10 samples of 5 species with equal abundances, no covariance and target depths of 10,000
rPLN()

#>        Y1   Y2   Y3   Y4   Y5
#> S1   1332 3100 1947  682  292
#> S2    634 1500 1798 2241 1515
#> S3   1168 1521  338 9550  102
#> S4    956 6059 2567  233 1262
#> S5   1935 1372 8007 2044 2170
#> S6  19236 1108 1078  202 3775
#> S7   1314  182  517  747  218
#> S8   2161  896 1238 1149  971
#> S9   1393  842  547 2110  985
#> S10   183 3470  276  523  978

## 2 samples of 10 highly correlated species with target depths 1,000 and 100,000
## very different abundances
mu <- rep(c(1, -1), each = 5)
Sigma <- matrix(0.8, 10, 10); diag(Sigma) <- 1
rPLN(n=2, mu = mu, Sigma = Sigma, depths = c(1e3, 1e5))

#>      Y1    Y2   Y3   Y4   Y5  Y6   Y7  Y8  Y9 Y10
#> S1   80    96   27   46   40   9    4   7   8   7
#> S2 4955 12325 6255 4962 4054 395 1473 869 791 498