PLN RNG

Random generation for the PLN model with latent mean equal to mu, latent covariance matrix equal to Sigma and average depths (sum of counts in a sample) equal to depths

rPLN(
  n = 10,
  mu = rep(0, ncol(Sigma)),
  Sigma = diag(1, 5, 5),
  depths = rep(10000, n)
)

Arguments

n
the sample size

mu
vectors of means of the latent variable

Sigma
covariance matrix of the latent variable

depths
Numeric vector of target depths. The first is recycled if there are not n values

Value

a n * p count matrix, with row-sums close to depths

Details

The default value for mu and Sigma assume equal abundances and no correlation between the different species.

Examples

## 10 samples of 5 species with equal abundances, no covariance and target depths of 10,000
rPLN()

#>       Y1    Y2   Y3   Y4   Y5
#> S1  1052   655 2234 1921 3953
#> S2   570  1285 5370 6755 2283
#> S3   752  1389  767  145 2059
#> S4  1016  2884 1634  966  558
#> S5  1401 15999  340 2352 1034
#> S6   450  1342 1903  521 1196
#> S7   345   700 6472  261  805
#> S8  1169  2734 1168 4116  818
#> S9  3831  4158  788 5278 4323
#> S10 4324  7916 2517  362 2828

## 2 samples of 10 highly correlated species with target depths 1,000 and 100,000
## very different abundances
mu <- rep(c(1, -1), each = 5)
Sigma <- matrix(0.8, 10, 10); diag(Sigma) <- 1
rPLN(n=2, mu = mu, Sigma = Sigma, depths = c(1e3, 1e5))

#>       Y1    Y2    Y3    Y4    Y5   Y6   Y7   Y8   Y9  Y10
#> S1    27    27    17    25    23    1    6    3    7    2
#> S2 19743 17751 40569 18207 18057 1207 3245 2272 4442 2109