Extract variance-covariance of residuals 'Sigma'

Extract the variance-covariance matrix of the residuals, usually noted $$\Sigma$$ in PLN models. This captures the correlation between the species in the latent space.

# S3 method for PLNfit
sigma(object, ...)

Arguments

object
an R6 object with class PLNfit

...
additional parameters for S3 compatibility. Not used

Value

A semi definite positive matrix of size p, assuming there are p species in the model.

See also

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLN(Abundance ~ 1 + offset(log(Offset)), data = trichoptera)

#> 
#>  Initialization...
#>  Adjusting a PLN model with full covariance model
#>  Post-treatments...
#>  DONE!

sigma(myPLN) ## Sigma

#>              Che          Hyc        Hym         Hys           Psy         Aga
#> Che  2.527910653 -0.009337094  0.3617320  0.44643957 -0.0180048441  0.30904259
#> Hyc -0.009337094  1.906200016  0.1109625  0.27016045 -0.1464564715 -0.13140626
#> Hym  0.361731955  0.110962478  1.2445919  0.86583510 -0.1186783498  0.51345012
#> Hys  0.446439568  0.270160452  0.8658351  2.66470880 -0.1468048639  0.42563039
#> Psy -0.018004844 -0.146456472 -0.1186783 -0.14680486  0.0762001842  0.00107749
#> Aga  0.309042587 -0.131406257  0.5134501  0.42563039  0.0010774896  1.10073643
#> Glo  0.090108084  0.171722808  0.2102416  0.34371837 -0.0500727294 -0.06368023
#> Ath  0.353026921  0.109727868  0.3859195  0.60842727 -0.0480600549  0.19740722
#> Cea -0.006397199  0.003536292 -0.0140180 -0.01074842  0.0001940469 -0.01457736
#> Ced  0.338525960  0.221806685  0.5084142  0.68193129 -0.0983345247  0.21905443
#> Set -0.059083966  0.477115382 -0.1825192  0.13863005 -0.0912852065 -0.30124598
#> All  0.272852840 -0.263347496  0.6804944  0.36972548 -0.0048905906  0.59215712
#> Han -0.182433934  0.624869766  0.2108334  0.13580567 -0.2882428208 -0.17049472
#> Hfo -0.356234172  0.553846604 -0.3822500 -0.17400482 -0.1445561518 -0.79949024
#> Hsp -0.451000555  0.679088316 -0.3428772 -0.18541858 -0.1926909464 -0.78690649
#> Hve  0.312543358 -0.130211627  0.1616775  0.24307949  0.0449843418  0.18544497
#> Sta  0.364145071  0.256884038  0.4785305  0.61436084 -0.1214981494  0.44005198
#>              Glo         Ath           Cea          Ced          Set
#> Che  0.090108084  0.35302692 -0.0063971985  0.338525960 -0.059083966
#> Hyc  0.171722808  0.10972787  0.0035362915  0.221806685  0.477115382
#> Hym  0.210241640  0.38591951 -0.0140179992  0.508414240 -0.182519173
#> Hys  0.343718374  0.60842727 -0.0107484152  0.681931290  0.138630047
#> Psy -0.050072729 -0.04806005  0.0001940469 -0.098334525 -0.091285207
#> Aga -0.063680227  0.19740722 -0.0145773595  0.219054432 -0.301245976
#> Glo  1.117999646  0.22513362  0.0010388516  0.227165590  0.293501352
#> Ath  0.225133617  1.37476151 -0.0046199297  0.443342367  0.097138290
#> Cea  0.001038852 -0.00461993  0.0266609803 -0.005205607  0.009798642
#> Ced  0.227165590  0.44334237 -0.0052056072  0.958451746  0.194920732
#> Set  0.293501352  0.09713829  0.0097986422  0.194920732  1.585956587
#> All -0.077332250  0.08821420 -0.0160046751  0.080347049 -0.682840815
#> Han -0.187114358 -0.19619249  0.0028170366  0.056713982 -0.042006683
#> Hfo  0.299935318 -0.15264634  0.0174273535 -0.124681102  0.797430085
#> Hsp  0.253287143 -0.22421929  0.0167134679 -0.118070123  0.773423079
#> Hve  0.089173413  0.24256838 -0.0046723984  0.205264830 -0.039023092
#> Sta  0.017506116  0.30863370 -0.0091739142  0.434793577  0.055006676
#>              All          Han         Hfo         Hsp          Hve          Sta
#> Che  0.272852840 -0.182433934 -0.35623417 -0.45100056  0.312543358  0.364145071
#> Hyc -0.263347496  0.624869766  0.55384660  0.67908832 -0.130211627  0.256884038
#> Hym  0.680494396  0.210833355 -0.38224996 -0.34287723  0.161677523  0.478530496
#> Hys  0.369725482  0.135805666 -0.17400482 -0.18541858  0.243079491  0.614360841
#> Psy -0.004890591 -0.288242821 -0.14455615 -0.19269095  0.044984342 -0.121498149
#> Aga  0.592157124 -0.170494719 -0.79949024 -0.78690649  0.185444971  0.440051984
#> Glo -0.077332250 -0.187114358  0.29993532  0.25328714  0.089173413  0.017506116
#> Ath  0.088214197 -0.196192491 -0.15264634 -0.22421929  0.242568380  0.308633700
#> Cea -0.016004675  0.002817037  0.01742735  0.01671347 -0.004672398 -0.009173914
#> Ced  0.080347049  0.056713982 -0.12468110 -0.11807012  0.205264830  0.434793577
#> Set -0.682840815 -0.042006683  0.79743008  0.77342308 -0.039023092  0.055006676
#> All  1.807090795  0.030628224 -0.72717351 -0.66493322  0.145702712  0.199898103
#> Han  0.030628224  3.632497475  0.61210094  0.99306455 -0.595700839  0.463729994
#> Hfo -0.727173510  0.612100942  2.65929035  1.62020675 -0.334213482 -0.250649112
#> Hsp -0.664933218  0.993064549  1.62020675  2.56623654 -0.484698418 -0.196793284
#> Hve  0.145702712 -0.595700839 -0.33421348 -0.48469842  1.347368163  0.144353357
#> Sta  0.199898103  0.463729994 -0.25064911 -0.19679328  0.144353357  1.029454169