Extract variance-covariance of residuals 'Sigma'

Extract the variance-covariance matrix of the residuals, usually noted $$\Sigma$$ in PLN models. This captures the correlation between the species in the latent space.

# S3 method for PLNfit
sigma(object, ...)

Arguments

object
an R6 object with class PLNfit

...
additional parameters for S3 compatibility. Not used

Value

A semi definite positive matrix of size p, assuming there are p species in the model.

See also

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLN(Abundance ~ 1 + offset(log(Offset)), data = trichoptera)

#> 
#>  Initialization...
#>  Adjusting a PLN model with full covariance model
#>  Post-treatments...
#>  DONE!

sigma(myPLN) ## Sigma

#>              Che          Hyc         Hym         Hys           Psy
#> Che  2.523015981 -0.007615322  0.35847302  0.44317168 -0.0180096740
#> Hyc -0.007615322  1.895520115  0.11378738  0.27256210 -0.1464691147
#> Hym  0.358473016  0.113787383  1.24447878  0.86556266 -0.1192100266
#> Hys  0.443171680  0.272562099  0.86556266  2.66110571 -0.1474146187
#> Psy -0.018009674 -0.146469115 -0.11921003 -0.14741462  0.0762988908
#> Aga  0.307475612 -0.133006403  0.51272765  0.42350486  0.0020024040
#> Glo  0.089221905  0.173459979  0.21062446  0.34460143 -0.0506063564
#> Ath  0.351539721  0.112355656  0.38548270  0.60840469 -0.0486631838
#> Cea -0.012272907  0.007913568 -0.02830427 -0.02095567  0.0001760717
#> Ced  0.337032467  0.222767444  0.50821042  0.68126072 -0.0985261541
#> Set -0.056991897  0.475171353 -0.17956163  0.14045286 -0.0916204457
#> All  0.270163413 -0.261681688  0.68026991  0.36911099 -0.0047850461
#> Han -0.177540663  0.619918127  0.21702463  0.14367294 -0.2869909058
#> Hfo -0.355260601  0.555308417 -0.38330716 -0.17279151 -0.1457298442
#> Hsp -0.436760629  0.670025806 -0.32606816 -0.16807219 -0.1920880832
#> Hve  0.311611841 -0.130840119  0.15973380  0.24031060  0.0456673293
#> Sta  0.363579007  0.255607018  0.47903500  0.61416062 -0.1212242111
#>              Aga          Glo          Ath           Cea          Ced
#> Che  0.307475612  0.089221905  0.351539721 -0.0122729075  0.337032467
#> Hyc -0.133006403  0.173459979  0.112355656  0.0079135678  0.222767444
#> Hym  0.512727650  0.210624465  0.385482700 -0.0283042749  0.508210419
#> Hys  0.423504863  0.344601433  0.608404694 -0.0209556688  0.681260719
#> Psy  0.002002404 -0.050606356 -0.048663184  0.0001760717 -0.098526154
#> Aga  1.104306912 -0.066603677  0.195579127 -0.0299562098  0.217691052
#> Glo -0.066603677  1.127559352  0.226814654  0.0025829726  0.227972443
#> Ath  0.195579127  0.226814654  1.377699837 -0.0085966015  0.443840062
#> Cea -0.029956210  0.002582973 -0.008596601  0.0606509876 -0.009587836
#> Ced  0.217691052  0.227972443  0.443840062 -0.0095878364  0.957865434
#> Set -0.303007634  0.295339716  0.099768523  0.0208883537  0.195928095
#> All  0.594304067 -0.079312652  0.086454793 -0.0334298670  0.079899422
#> Han -0.169998978 -0.182026556 -0.190263319  0.0061576545  0.059660937
#> Hfo -0.808489166  0.304465177 -0.149601170  0.0363024027 -0.123928694
#> Hsp -0.780093280  0.257466190 -0.210087679  0.0340089680 -0.107162058
#> Hve  0.186406083  0.088440437  0.241870698 -0.0093224223  0.204403367
#> Sta  0.439979571  0.016860351  0.309061143 -0.0178443881  0.434398404
#>             Set          All          Han        Hfo         Hsp          Hve
#> Che -0.05699190  0.270163413 -0.177540663 -0.3552606 -0.43676063  0.311611841
#> Hyc  0.47517135 -0.261681688  0.619918127  0.5553084  0.67002581 -0.130840119
#> Hym -0.17956163  0.680269907  0.217024631 -0.3833072 -0.32606816  0.159733803
#> Hys  0.14045286  0.369110987  0.143672944 -0.1727915 -0.16807219  0.240310599
#> Psy -0.09162045 -0.004785046 -0.286990906 -0.1457298 -0.19208808  0.045667329
#> Aga -0.30300763  0.594304067 -0.169998978 -0.8084892 -0.78009328  0.186406083
#> Glo  0.29533972 -0.079312652 -0.182026556  0.3044652  0.25746619  0.088440437
#> Ath  0.09976852  0.086454793 -0.190263319 -0.1496012 -0.21008768  0.241870698
#> Cea  0.02088835 -0.033429867  0.006157655  0.0363024  0.03400897 -0.009322422
#> Ced  0.19592809  0.079899422  0.059660937 -0.1239287 -0.10716206  0.204403367
#> Set  1.57909074 -0.681783526 -0.039196017  0.8006738  0.76328325 -0.039736744
#> All -0.68178353  1.807922349  0.035000648 -0.7336931 -0.65499669  0.145484561
#> Han -0.03919602  0.035000648  3.566997088  0.6123879  0.97656745 -0.593208806
#> Hfo  0.80067383 -0.733693131  0.612387852  2.6787445  1.60500535 -0.337477108
#> Hsp  0.76328325 -0.654996693  0.976567450  1.6050054  2.48251216 -0.477436178
#> Hve -0.03973674  0.145484561 -0.593208806 -0.3374771 -0.47743618  1.354264248
#> Sta  0.05501612  0.200968493  0.459165891 -0.2521632 -0.19137475  0.143998188
#>             Sta
#> Che  0.36357901
#> Hyc  0.25560702
#> Hym  0.47903500
#> Hys  0.61416062
#> Psy -0.12122421
#> Aga  0.43997957
#> Glo  0.01686035
#> Ath  0.30906114
#> Cea -0.01784439
#> Ced  0.43439840
#> Set  0.05501612
#> All  0.20096849
#> Han  0.45916589
#> Hfo -0.25216325
#> Hsp -0.19137475
#> Hve  0.14399819
#> Sta  1.02761547