Extract variance-covariance of residuals 'Sigma'

Extract the variance-covariance matrix of the residuals, usually noted $$\Sigma$$ in PLN models. This captures the correlation between the species in the latent space.

# S3 method for PLNfit
sigma(object, ...)

Arguments

object

an R6 object with class PLNfit

...

additional parameters for S3 compatibility. Not used

Value

A semi definite positive matrix of size p, assuming there are p species in the model.

See also

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLN(Abundance ~ 1 + offset(log(Offset)), data = trichoptera)
#> 
#>  Initialization...
#>  Adjusting a PLN model with full covariance model
#>  Post-treatments...
#>  DONE!
sigma(myPLN) ## Sigma
#>              Che          Hyc         Hym         Hys           Psy
#> Che  2.546676813 -0.003596064  0.35479909  0.44314557 -0.0189129187
#> Hyc -0.003596064  1.913142538  0.12132488  0.28435424 -0.1504750120
#> Hym  0.354799095  0.121324884  1.24224962  0.86746658 -0.1218932127
#> Hys  0.443145566  0.284354238  0.86746658  2.67203049 -0.1514194232
#> Psy -0.018912919 -0.150475012 -0.12189321 -0.15141942  0.0777754736
#> Aga  0.305764440 -0.141022365  0.50944307  0.41769593  0.0044670549
#> Glo  0.091286038  0.181539247  0.21321031  0.35166538 -0.0532097562
#> Ath  0.352653072  0.120254662  0.38504589  0.61298431 -0.0511498151
#> Cea -0.017691319  0.016142867 -0.04236166 -0.02899235 -0.0009803407
#> Ced  0.337857639  0.229188243  0.50845293  0.68505315 -0.1007676870
#> Set -0.048649764  0.479448614 -0.17070487  0.15070999 -0.0933394471
#> All  0.266109408 -0.269404448  0.67735620  0.36249380 -0.0028512064
#> Han -0.172847510  0.626184293  0.22782200  0.15696556 -0.2883194824
#> Hfo -0.334629302  0.555682960 -0.35327394 -0.14242249 -0.1476792254
#> Hsp -0.412992392  0.674058938 -0.29266000 -0.13082691 -0.1955588669
#> Hve  0.311581117 -0.136308217  0.15429911  0.23431170  0.0478663868
#> Sta  0.364908671  0.257593128  0.47967912  0.61628308 -0.1220511532
#>              Aga          Glo         Ath           Cea         Ced         Set
#> Che  0.305764440  0.091286038  0.35265307 -0.0176913190  0.33785764 -0.04864976
#> Hyc -0.141022365  0.181539247  0.12025466  0.0161428671  0.22918824  0.47944861
#> Hym  0.509443075  0.213210309  0.38504589 -0.0423616613  0.50845293 -0.17070487
#> Hys  0.417695926  0.351665377  0.61298431 -0.0289923514  0.68505315  0.15070999
#> Psy  0.004467055 -0.053209756 -0.05114982 -0.0009803407 -0.10076769 -0.09333945
#> Aga  1.111875743 -0.071735056  0.19093058 -0.0477897256  0.21389035 -0.30610288
#> Glo -0.071735056  1.133043251  0.23180212  0.0058468143  0.23247266  0.30087566
#> Ath  0.190930578  0.231802118  1.38500401 -0.0111049109  0.44677848  0.10828990
#> Cea -0.047789726  0.005846814 -0.01110491  0.1020742242 -0.01171639  0.03595539
#> Ced  0.213890347  0.232472656  0.44677848 -0.0117163882  0.95822464  0.20020082
#> Set -0.306102883  0.300875665  0.10828990  0.0359553901  0.20020082  1.55873141
#> All  0.602180848 -0.085726181  0.07996397 -0.0545545994  0.07582988 -0.68317080
#> Han -0.174186058 -0.170611444 -0.17992571  0.0139442681  0.06586589 -0.03234971
#> Hfo -0.796614781  0.309116750 -0.12738372  0.0587775943 -0.10261249  0.77926746
#> Hsp -0.774231368  0.268194988 -0.18307854  0.0560552027 -0.08298800  0.74876405
#> Hve  0.189959041  0.085759116  0.23905346 -0.0149304533  0.20045549 -0.04104275
#> Sta  0.438878228  0.018689071  0.31085005 -0.0248991148  0.43491067  0.05660168
#>              All         Han         Hfo        Hsp         Hve         Sta
#> Che  0.266109408 -0.17284751 -0.33462930 -0.4129924  0.31158112  0.36490867
#> Hyc -0.269404448  0.62618429  0.55568296  0.6740589 -0.13630822  0.25759313
#> Hym  0.677356204  0.22782200 -0.35327394 -0.2926600  0.15429911  0.47967912
#> Hys  0.362493801  0.15696556 -0.14242249 -0.1308269  0.23431170  0.61628308
#> Psy -0.002851206 -0.28831948 -0.14767923 -0.1955589  0.04786639 -0.12205115
#> Aga  0.602180848 -0.17418606 -0.79661478 -0.7742314  0.18995904  0.43887823
#> Glo -0.085726181 -0.17061144  0.30911675  0.2681950  0.08575912  0.01868907
#> Ath  0.079963971 -0.17992571 -0.12738372 -0.1830785  0.23905346  0.31085005
#> Cea -0.054554599  0.01394427  0.05877759  0.0560552 -0.01493045 -0.02489911
#> Ced  0.075829883  0.06586589 -0.10261249 -0.0829880  0.20045549  0.43491067
#> Set -0.683170802 -0.03234971  0.77926746  0.7487640 -0.04104275  0.05660168
#> All  1.822442081  0.03297976 -0.72325494 -0.6498372  0.14754365  0.19956423
#> Han  0.032979765  3.47820469  0.60642315  0.9674956 -0.59413106  0.45513928
#> Hfo -0.723254939  0.60642315  2.53243332  1.5378929 -0.33265769 -0.23939553
#> Hsp -0.649837160  0.96749558  1.53789291  2.3724742 -0.47103679 -0.17886628
#> Hve  0.147543653 -0.59413106 -0.33265769 -0.4710368  1.35850640  0.14196037
#> Sta  0.199564229  0.45513928 -0.23939553 -0.1788663  0.14196037  1.02734941