Extract variance-covariance of residuals 'Sigma'

Extract the variance-covariance matrix of the residuals, usually noted $$\Sigma$$ in PLN models. This captures the correlation between the species in the latent space.

# S3 method for PLNfit
sigma(object, ...)

Arguments

object
an R6 object with class PLNfit

...
additional parameters for S3 compatibility. Not used

Value

A semi definite positive matrix of size p, assuming there are p species in the model.

See also

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLN(Abundance ~ 1 + offset(log(Offset)), data = trichoptera)

#> 
#>  Initialization...
#>  Adjusting a PLN model with full covariance model
#>  Post-treatments...
#>  DONE!

sigma(myPLN) ## Sigma

#>              Che          Hyc         Hym         Hys           Psy
#> Che  2.528721747 -0.008614963  0.35918900  0.44447578 -0.0177838001
#> Hyc -0.008614963  1.904478793  0.11312212  0.27172444 -0.1465650739
#> Hym  0.359188999  0.113122119  1.24438461  0.86672480 -0.1188757294
#> Hys  0.444475777  0.271724441  0.86672480  2.66571464 -0.1468499147
#> Psy -0.017783800 -0.146565074 -0.11887573 -0.14684991  0.0761828442
#> Aga  0.308321041 -0.132167102  0.51276845  0.42497794  0.0014529933
#> Glo  0.089626193  0.172614508  0.21117999  0.34515849 -0.0500893092
#> Ath  0.352235614  0.112066775  0.38607257  0.60964580 -0.0484326704
#> Cea -0.010923902  0.007038165 -0.02511974 -0.01867089  0.0001627748
#> Ced  0.337404341  0.223152725  0.50858697  0.68210086 -0.0984397212
#> Set -0.057670504  0.476585689 -0.17926792  0.14046409 -0.0914585920
#> All  0.271368219 -0.263558921  0.68009906  0.36970921 -0.0045743462
#> Han -0.182485620  0.623839658  0.21164315  0.13460583 -0.2882387787
#> Hfo -0.354613821  0.551897130 -0.38073232 -0.17277102 -0.1439938784
#> Hsp -0.439715086  0.672297168 -0.32689003 -0.17061894 -0.1917979637
#> Hve  0.312797500 -0.133107648  0.15940473  0.24053408  0.0463602002
#> Sta  0.363915520  0.256264254  0.47851417  0.61375601 -0.1212481318
#>              Aga          Glo          Ath           Cea         Ced
#> Che  0.308321041  0.089626193  0.352235614 -0.0109239023  0.33740434
#> Hyc -0.132167102  0.172614508  0.112066775  0.0070381654  0.22315272
#> Hym  0.512768451  0.211179986  0.386072571 -0.0251197373  0.50858697
#> Hys  0.424977936  0.345158490  0.609645802 -0.0186708879  0.68210086
#> Psy  0.001452993 -0.050089309 -0.048432670  0.0001627748 -0.09843972
#> Aga  1.101710649 -0.064596502  0.195963114 -0.0263986417  0.21836922
#> Glo -0.064596502  1.119418505  0.227072462  0.0021951211  0.22809360
#> Ath  0.195963114  0.227072462  1.377886034 -0.0076883119  0.44419660
#> Cea -0.026398642  0.002195121 -0.007688312  0.0533332118 -0.00858464
#> Ced  0.218369223  0.228093599  0.444196596 -0.0085846400  0.95829562
#> Set -0.301194802  0.294018313  0.099904305  0.0184909342  0.19583771
#> All  0.593727721 -0.078187655  0.086674054 -0.0295391047  0.07984021
#> Han -0.171461289 -0.188631840 -0.196607813  0.0055480162  0.05673418
#> Hfo -0.799834317  0.298978535 -0.148323152  0.0318842411 -0.12247418
#> Hsp -0.777578342  0.254345194 -0.211302337  0.0300842212 -0.10765564
#> Hve  0.186747116  0.088686750  0.242086236 -0.0082795615  0.20352584
#> Sta  0.440141305  0.016980547  0.308706399 -0.0157645224  0.43447275
#>             Set          All          Han         Hfo         Hsp          Hve
#> Che -0.05767050  0.271368219 -0.182485620 -0.35461382 -0.43971509  0.312797500
#> Hyc  0.47658569 -0.263558921  0.623839658  0.55189713  0.67229717 -0.133107648
#> Hym -0.17926792  0.680099060  0.211643153 -0.38073232 -0.32689003  0.159404728
#> Hys  0.14046409  0.369709214  0.134605830 -0.17277102 -0.17061894  0.240534076
#> Psy -0.09145859 -0.004574346 -0.288238779 -0.14399388 -0.19179796  0.046360200
#> Aga -0.30119480  0.593727721 -0.171461289 -0.79983432 -0.77757834  0.186747116
#> Glo  0.29401831 -0.078187655 -0.188631840  0.29897853  0.25434519  0.088686750
#> Ath  0.09990431  0.086674054 -0.196607813 -0.14832315 -0.21130234  0.242086236
#> Cea  0.01849093 -0.029539105  0.005548016  0.03188424  0.03008422 -0.008279562
#> Ced  0.19583771  0.079840206  0.056734176 -0.12247418 -0.10765564  0.203525844
#> Set  1.57629333 -0.681400643 -0.041343908  0.79327380  0.76306048 -0.040809713
#> All -0.68140064  1.809235149  0.030667177 -0.72926967 -0.65612023  0.146647303
#> Han -0.04134391  0.030667177  3.632478184  0.60998319  0.98317455 -0.603451744
#> Hfo  0.79327380 -0.729269675  0.609983194  2.63644237  1.59189474 -0.337311101
#> Hsp  0.76306048 -0.656120229  0.983174550  1.59189474  2.48941974 -0.481631408
#> Hve -0.04080971  0.146647303 -0.603451744 -0.33731110 -0.48163141  1.358464210
#> Sta  0.05514801  0.200402058  0.462441539 -0.25089001 -0.19140230  0.142776916
#>             Sta
#> Che  0.36391552
#> Hyc  0.25626425
#> Hym  0.47851417
#> Hys  0.61375601
#> Psy -0.12124813
#> Aga  0.44014131
#> Glo  0.01698055
#> Ath  0.30870640
#> Cea -0.01576452
#> Ced  0.43447275
#> Set  0.05514801
#> All  0.20040206
#> Han  0.46244154
#> Hfo -0.25089001
#> Hsp -0.19140230
#> Hve  0.14277692
#> Sta  1.02838340