Extract variance-covariance of residuals 'Sigma'

Extract the variance-covariance matrix of the residuals, usually noted $$\Sigma$$ in PLN models. This captures the correlation between the species in the latent space. or PLNmixture, it is a weighted mean of the variance-covariance matrices of each component.

# S3 method for PLNmixturefit
sigma(object, ...)

Arguments

object
an R6 object with class PLNmixturefit

...
additional parameters for S3 compatibility. Not used

Value

A semi definite positive matrix of size p, assuming there are p species in the model.

See also

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLNmixture(Abundance ~ 1 + offset(log(Offset)), data = trichoptera) %>% getBestModel()

#> 
#>  Initialization...
#> 
#>  Adjusting 5 PLN mixture models.
#> 	number of cluster = 1 	number of cluster = 2 	number of cluster = 3 	number of cluster = 4 	number of cluster = 5 
#> 
#>  Smoothing PLN mixture models.
#>    Going backward ++++                                                                                                       Going forward ++++                                                                                                       Going backward ++++                                                                                                       Going forward ++++                                                                                                    
#>  Post-treatments
#>  DONE!

sigma(myPLN) ## Sigma

#> [[1]]
#>           Che       Hyc       Hym       Hys       Psy       Aga       Glo
#> Che 0.5737386 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Hyc 0.0000000 0.5737386 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Hym 0.0000000 0.0000000 0.5737386 0.0000000 0.0000000 0.0000000 0.0000000
#> Hys 0.0000000 0.0000000 0.0000000 0.5737386 0.0000000 0.0000000 0.0000000
#> Psy 0.0000000 0.0000000 0.0000000 0.0000000 0.5737386 0.0000000 0.0000000
#> Aga 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.5737386 0.0000000
#> Glo 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.5737386
#> Ath 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Cea 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Ced 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Set 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> All 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Han 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Hfo 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Hsp 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Hve 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Sta 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>           Ath       Cea       Ced       Set       All       Han       Hfo
#> Che 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Hyc 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Hym 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Hys 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Psy 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Aga 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Glo 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Ath 0.5737386 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Cea 0.0000000 0.5737386 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Ced 0.0000000 0.0000000 0.5737386 0.0000000 0.0000000 0.0000000 0.0000000
#> Set 0.0000000 0.0000000 0.0000000 0.5737386 0.0000000 0.0000000 0.0000000
#> All 0.0000000 0.0000000 0.0000000 0.0000000 0.5737386 0.0000000 0.0000000
#> Han 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.5737386 0.0000000
#> Hfo 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.5737386
#> Hsp 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Hve 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Sta 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>           Hsp       Hve       Sta
#> Che 0.0000000 0.0000000 0.0000000
#> Hyc 0.0000000 0.0000000 0.0000000
#> Hym 0.0000000 0.0000000 0.0000000
#> Hys 0.0000000 0.0000000 0.0000000
#> Psy 0.0000000 0.0000000 0.0000000
#> Aga 0.0000000 0.0000000 0.0000000
#> Glo 0.0000000 0.0000000 0.0000000
#> Ath 0.0000000 0.0000000 0.0000000
#> Cea 0.0000000 0.0000000 0.0000000
#> Ced 0.0000000 0.0000000 0.0000000
#> Set 0.0000000 0.0000000 0.0000000
#> All 0.0000000 0.0000000 0.0000000
#> Han 0.0000000 0.0000000 0.0000000
#> Hfo 0.0000000 0.0000000 0.0000000
#> Hsp 0.5737386 0.0000000 0.0000000
#> Hve 0.0000000 0.5737386 0.0000000
#> Sta 0.0000000 0.0000000 0.5737386
#> 
#> [[2]]
#>           Che       Hyc       Hym       Hys       Psy       Aga       Glo
#> Che 0.9792721 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Hyc 0.0000000 0.9792721 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Hym 0.0000000 0.0000000 0.9792721 0.0000000 0.0000000 0.0000000 0.0000000
#> Hys 0.0000000 0.0000000 0.0000000 0.9792721 0.0000000 0.0000000 0.0000000
#> Psy 0.0000000 0.0000000 0.0000000 0.0000000 0.9792721 0.0000000 0.0000000
#> Aga 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.9792721 0.0000000
#> Glo 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.9792721
#> Ath 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Cea 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Ced 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Set 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> All 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Han 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Hfo 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Hsp 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Hve 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Sta 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>           Ath       Cea       Ced       Set       All       Han       Hfo
#> Che 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Hyc 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Hym 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Hys 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Psy 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Aga 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Glo 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Ath 0.9792721 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Cea 0.0000000 0.9792721 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Ced 0.0000000 0.0000000 0.9792721 0.0000000 0.0000000 0.0000000 0.0000000
#> Set 0.0000000 0.0000000 0.0000000 0.9792721 0.0000000 0.0000000 0.0000000
#> All 0.0000000 0.0000000 0.0000000 0.0000000 0.9792721 0.0000000 0.0000000
#> Han 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.9792721 0.0000000
#> Hfo 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.9792721
#> Hsp 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Hve 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> Sta 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>           Hsp       Hve       Sta
#> Che 0.0000000 0.0000000 0.0000000
#> Hyc 0.0000000 0.0000000 0.0000000
#> Hym 0.0000000 0.0000000 0.0000000
#> Hys 0.0000000 0.0000000 0.0000000
#> Psy 0.0000000 0.0000000 0.0000000
#> Aga 0.0000000 0.0000000 0.0000000
#> Glo 0.0000000 0.0000000 0.0000000
#> Ath 0.0000000 0.0000000 0.0000000
#> Cea 0.0000000 0.0000000 0.0000000
#> Ced 0.0000000 0.0000000 0.0000000
#> Set 0.0000000 0.0000000 0.0000000
#> All 0.0000000 0.0000000 0.0000000
#> Han 0.0000000 0.0000000 0.0000000
#> Hfo 0.0000000 0.0000000 0.0000000
#> Hsp 0.9792721 0.0000000 0.0000000
#> Hve 0.0000000 0.9792721 0.0000000
#> Sta 0.0000000 0.0000000 0.9792721
#> 
#> [[3]]
#>          Che      Hyc      Hym      Hys      Psy      Aga      Glo      Ath
#> Che 2.820835 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> Hyc 0.000000 2.820835 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> Hym 0.000000 0.000000 2.820835 0.000000 0.000000 0.000000 0.000000 0.000000
#> Hys 0.000000 0.000000 0.000000 2.820835 0.000000 0.000000 0.000000 0.000000
#> Psy 0.000000 0.000000 0.000000 0.000000 2.820835 0.000000 0.000000 0.000000
#> Aga 0.000000 0.000000 0.000000 0.000000 0.000000 2.820835 0.000000 0.000000
#> Glo 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 2.820835 0.000000
#> Ath 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 2.820835
#> Cea 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> Ced 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> Set 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> All 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> Han 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> Hfo 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> Hsp 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> Hve 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> Sta 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>          Cea      Ced      Set      All      Han      Hfo      Hsp      Hve
#> Che 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> Hyc 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> Hym 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> Hys 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> Psy 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> Aga 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> Glo 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> Ath 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> Cea 2.820835 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> Ced 0.000000 2.820835 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> Set 0.000000 0.000000 2.820835 0.000000 0.000000 0.000000 0.000000 0.000000
#> All 0.000000 0.000000 0.000000 2.820835 0.000000 0.000000 0.000000 0.000000
#> Han 0.000000 0.000000 0.000000 0.000000 2.820835 0.000000 0.000000 0.000000
#> Hfo 0.000000 0.000000 0.000000 0.000000 0.000000 2.820835 0.000000 0.000000
#> Hsp 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 2.820835 0.000000
#> Hve 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 2.820835
#> Sta 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#>          Sta
#> Che 0.000000
#> Hyc 0.000000
#> Hym 0.000000
#> Hys 0.000000
#> Psy 0.000000
#> Aga 0.000000
#> Glo 0.000000
#> Ath 0.000000
#> Cea 0.000000
#> Ced 0.000000
#> Set 0.000000
#> All 0.000000
#> Han 0.000000
#> Hfo 0.000000
#> Hsp 0.000000
#> Hve 0.000000
#> Sta 2.820835
#> 
#> [[4]]
#>            Che        Hyc        Hym        Hys        Psy        Aga
#> Che 0.04905692 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Hyc 0.00000000 0.04905692 0.00000000 0.00000000 0.00000000 0.00000000
#> Hym 0.00000000 0.00000000 0.04905692 0.00000000 0.00000000 0.00000000
#> Hys 0.00000000 0.00000000 0.00000000 0.04905692 0.00000000 0.00000000
#> Psy 0.00000000 0.00000000 0.00000000 0.00000000 0.04905692 0.00000000
#> Aga 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.04905692
#> Glo 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Ath 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Cea 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Ced 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Set 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> All 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Han 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Hfo 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Hsp 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Hve 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Sta 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>            Glo        Ath        Cea        Ced        Set        All
#> Che 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Hyc 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Hym 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Hys 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Psy 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Aga 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Glo 0.04905692 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Ath 0.00000000 0.04905692 0.00000000 0.00000000 0.00000000 0.00000000
#> Cea 0.00000000 0.00000000 0.04905692 0.00000000 0.00000000 0.00000000
#> Ced 0.00000000 0.00000000 0.00000000 0.04905692 0.00000000 0.00000000
#> Set 0.00000000 0.00000000 0.00000000 0.00000000 0.04905692 0.00000000
#> All 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.04905692
#> Han 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Hfo 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Hsp 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Hve 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Sta 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>            Han        Hfo        Hsp        Hve        Sta
#> Che 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Hyc 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Hym 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Hys 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Psy 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Aga 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Glo 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Ath 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Cea 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Ced 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Set 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> All 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#> Han 0.04905692 0.00000000 0.00000000 0.00000000 0.00000000
#> Hfo 0.00000000 0.04905692 0.00000000 0.00000000 0.00000000
#> Hsp 0.00000000 0.00000000 0.04905692 0.00000000 0.00000000
#> Hve 0.00000000 0.00000000 0.00000000 0.04905692 0.00000000
#> Sta 0.00000000 0.00000000 0.00000000 0.00000000 0.04905692
#>