Extract variance-covariance of residuals 'Sigma'

Extract the variance-covariance matrix of the residuals, usually noted $$\Sigma$$ in PLN models. This captures the correlation between the species in the latent space. or PLNmixture, it is a weighted mean of the variance-covariance matrices of each component.

# S3 method for PLNmixturefit
sigma(object, ...)

Arguments

object

an R6 object with class PLNmixturefit

...

additional parameters for S3 compatibility. Not used

Value

A semi definite positive matrix of size p, assuming there are p species in the model.

See also

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLNmixture(Abundance ~ 1 + offset(log(Offset)),
           data = trichoptera, control = PLNmixture_param(smoothing = "none"))  %>% getBestModel()
#> 
#>  Initialization...
#> 
#>  Adjusting 5 PLN mixture models.
#> 	number of cluster = 1 	number of cluster = 2 	number of cluster = 3 	number of cluster = 4 	number of cluster = 5 
#>  Post-treatments
#>  DONE!
sigma(myPLN) ## Sigma
#> [[1]]
#> 17 x 17 sparse Matrix of class "dgCMatrix"
#>    [[ suppressing 17 column names ‘Che’, ‘Hyc’, ‘Hym’ ... ]]
#>                                                                          
#> Che 0.4002041 .         .         .         .         .         .        
#> Hyc .         0.4002041 .         .         .         .         .        
#> Hym .         .         0.4002041 .         .         .         .        
#> Hys .         .         .         0.4002041 .         .         .        
#> Psy .         .         .         .         0.4002041 .         .        
#> Aga .         .         .         .         .         0.4002041 .        
#> Glo .         .         .         .         .         .         0.4002041
#> Ath .         .         .         .         .         .         .        
#> Cea .         .         .         .         .         .         .        
#> Ced .         .         .         .         .         .         .        
#> Set .         .         .         .         .         .         .        
#> All .         .         .         .         .         .         .        
#> Han .         .         .         .         .         .         .        
#> Hfo .         .         .         .         .         .         .        
#> Hsp .         .         .         .         .         .         .        
#> Hve .         .         .         .         .         .         .        
#> Sta .         .         .         .         .         .         .        
#>                                                                          
#> Che .         .         .         .         .         .         .        
#> Hyc .         .         .         .         .         .         .        
#> Hym .         .         .         .         .         .         .        
#> Hys .         .         .         .         .         .         .        
#> Psy .         .         .         .         .         .         .        
#> Aga .         .         .         .         .         .         .        
#> Glo .         .         .         .         .         .         .        
#> Ath 0.4002041 .         .         .         .         .         .        
#> Cea .         0.4002041 .         .         .         .         .        
#> Ced .         .         0.4002041 .         .         .         .        
#> Set .         .         .         0.4002041 .         .         .        
#> All .         .         .         .         0.4002041 .         .        
#> Han .         .         .         .         .         0.4002041 .        
#> Hfo .         .         .         .         .         .         0.4002041
#> Hsp .         .         .         .         .         .         .        
#> Hve .         .         .         .         .         .         .        
#> Sta .         .         .         .         .         .         .        
#>                                  
#> Che .         .         .        
#> Hyc .         .         .        
#> Hym .         .         .        
#> Hys .         .         .        
#> Psy .         .         .        
#> Aga .         .         .        
#> Glo .         .         .        
#> Ath .         .         .        
#> Cea .         .         .        
#> Ced .         .         .        
#> Set .         .         .        
#> All .         .         .        
#> Han .         .         .        
#> Hfo .         .         .        
#> Hsp 0.4002041 .         .        
#> Hve .         0.4002041 .        
#> Sta .         .         0.4002041
#> 
#> [[2]]
#> 17 x 17 sparse Matrix of class "dgCMatrix"
#>    [[ suppressing 17 column names ‘Che’, ‘Hyc’, ‘Hym’ ... ]]
#>                                                                          
#> Che 0.2045565 .         .         .         .         .         .        
#> Hyc .         0.2045565 .         .         .         .         .        
#> Hym .         .         0.2045565 .         .         .         .        
#> Hys .         .         .         0.2045565 .         .         .        
#> Psy .         .         .         .         0.2045565 .         .        
#> Aga .         .         .         .         .         0.2045565 .        
#> Glo .         .         .         .         .         .         0.2045565
#> Ath .         .         .         .         .         .         .        
#> Cea .         .         .         .         .         .         .        
#> Ced .         .         .         .         .         .         .        
#> Set .         .         .         .         .         .         .        
#> All .         .         .         .         .         .         .        
#> Han .         .         .         .         .         .         .        
#> Hfo .         .         .         .         .         .         .        
#> Hsp .         .         .         .         .         .         .        
#> Hve .         .         .         .         .         .         .        
#> Sta .         .         .         .         .         .         .        
#>                                                                          
#> Che .         .         .         .         .         .         .        
#> Hyc .         .         .         .         .         .         .        
#> Hym .         .         .         .         .         .         .        
#> Hys .         .         .         .         .         .         .        
#> Psy .         .         .         .         .         .         .        
#> Aga .         .         .         .         .         .         .        
#> Glo .         .         .         .         .         .         .        
#> Ath 0.2045565 .         .         .         .         .         .        
#> Cea .         0.2045565 .         .         .         .         .        
#> Ced .         .         0.2045565 .         .         .         .        
#> Set .         .         .         0.2045565 .         .         .        
#> All .         .         .         .         0.2045565 .         .        
#> Han .         .         .         .         .         0.2045565 .        
#> Hfo .         .         .         .         .         .         0.2045565
#> Hsp .         .         .         .         .         .         .        
#> Hve .         .         .         .         .         .         .        
#> Sta .         .         .         .         .         .         .        
#>                                  
#> Che .         .         .        
#> Hyc .         .         .        
#> Hym .         .         .        
#> Hys .         .         .        
#> Psy .         .         .        
#> Aga .         .         .        
#> Glo .         .         .        
#> Ath .         .         .        
#> Cea .         .         .        
#> Ced .         .         .        
#> Set .         .         .        
#> All .         .         .        
#> Han .         .         .        
#> Hfo .         .         .        
#> Hsp 0.2045565 .         .        
#> Hve .         0.2045565 .        
#> Sta .         .         0.2045565
#> 
#> [[3]]
#> 17 x 17 sparse Matrix of class "dgCMatrix"
#>    [[ suppressing 17 column names ‘Che’, ‘Hyc’, ‘Hym’ ... ]]
#>                                                                          
#> Che 0.2035112 .         .         .         .         .         .        
#> Hyc .         0.2035112 .         .         .         .         .        
#> Hym .         .         0.2035112 .         .         .         .        
#> Hys .         .         .         0.2035112 .         .         .        
#> Psy .         .         .         .         0.2035112 .         .        
#> Aga .         .         .         .         .         0.2035112 .        
#> Glo .         .         .         .         .         .         0.2035112
#> Ath .         .         .         .         .         .         .        
#> Cea .         .         .         .         .         .         .        
#> Ced .         .         .         .         .         .         .        
#> Set .         .         .         .         .         .         .        
#> All .         .         .         .         .         .         .        
#> Han .         .         .         .         .         .         .        
#> Hfo .         .         .         .         .         .         .        
#> Hsp .         .         .         .         .         .         .        
#> Hve .         .         .         .         .         .         .        
#> Sta .         .         .         .         .         .         .        
#>                                                                          
#> Che .         .         .         .         .         .         .        
#> Hyc .         .         .         .         .         .         .        
#> Hym .         .         .         .         .         .         .        
#> Hys .         .         .         .         .         .         .        
#> Psy .         .         .         .         .         .         .        
#> Aga .         .         .         .         .         .         .        
#> Glo .         .         .         .         .         .         .        
#> Ath 0.2035112 .         .         .         .         .         .        
#> Cea .         0.2035112 .         .         .         .         .        
#> Ced .         .         0.2035112 .         .         .         .        
#> Set .         .         .         0.2035112 .         .         .        
#> All .         .         .         .         0.2035112 .         .        
#> Han .         .         .         .         .         0.2035112 .        
#> Hfo .         .         .         .         .         .         0.2035112
#> Hsp .         .         .         .         .         .         .        
#> Hve .         .         .         .         .         .         .        
#> Sta .         .         .         .         .         .         .        
#>                                  
#> Che .         .         .        
#> Hyc .         .         .        
#> Hym .         .         .        
#> Hys .         .         .        
#> Psy .         .         .        
#> Aga .         .         .        
#> Glo .         .         .        
#> Ath .         .         .        
#> Cea .         .         .        
#> Ced .         .         .        
#> Set .         .         .        
#> All .         .         .        
#> Han .         .         .        
#> Hfo .         .         .        
#> Hsp 0.2035112 .         .        
#> Hve .         0.2035112 .        
#> Sta .         .         0.2035112
#> 
#> [[4]]
#> 17 x 17 sparse Matrix of class "dgCMatrix"
#>    [[ suppressing 17 column names ‘Che’, ‘Hyc’, ‘Hym’ ... ]]
#>                                                                          
#> Che 0.1549584 .         .         .         .         .         .        
#> Hyc .         0.1549584 .         .         .         .         .        
#> Hym .         .         0.1549584 .         .         .         .        
#> Hys .         .         .         0.1549584 .         .         .        
#> Psy .         .         .         .         0.1549584 .         .        
#> Aga .         .         .         .         .         0.1549584 .        
#> Glo .         .         .         .         .         .         0.1549584
#> Ath .         .         .         .         .         .         .        
#> Cea .         .         .         .         .         .         .        
#> Ced .         .         .         .         .         .         .        
#> Set .         .         .         .         .         .         .        
#> All .         .         .         .         .         .         .        
#> Han .         .         .         .         .         .         .        
#> Hfo .         .         .         .         .         .         .        
#> Hsp .         .         .         .         .         .         .        
#> Hve .         .         .         .         .         .         .        
#> Sta .         .         .         .         .         .         .        
#>                                                                          
#> Che .         .         .         .         .         .         .        
#> Hyc .         .         .         .         .         .         .        
#> Hym .         .         .         .         .         .         .        
#> Hys .         .         .         .         .         .         .        
#> Psy .         .         .         .         .         .         .        
#> Aga .         .         .         .         .         .         .        
#> Glo .         .         .         .         .         .         .        
#> Ath 0.1549584 .         .         .         .         .         .        
#> Cea .         0.1549584 .         .         .         .         .        
#> Ced .         .         0.1549584 .         .         .         .        
#> Set .         .         .         0.1549584 .         .         .        
#> All .         .         .         .         0.1549584 .         .        
#> Han .         .         .         .         .         0.1549584 .        
#> Hfo .         .         .         .         .         .         0.1549584
#> Hsp .         .         .         .         .         .         .        
#> Hve .         .         .         .         .         .         .        
#> Sta .         .         .         .         .         .         .        
#>                                  
#> Che .         .         .        
#> Hyc .         .         .        
#> Hym .         .         .        
#> Hys .         .         .        
#> Psy .         .         .        
#> Aga .         .         .        
#> Glo .         .         .        
#> Ath .         .         .        
#> Cea .         .         .        
#> Ced .         .         .        
#> Set .         .         .        
#> All .         .         .        
#> Han .         .         .        
#> Hfo .         .         .        
#> Hsp 0.1549584 .         .        
#> Hve .         0.1549584 .        
#> Sta .         .         0.1549584
#> 
#> [[5]]
#> 17 x 17 sparse Matrix of class "dgCMatrix"
#>    [[ suppressing 17 column names ‘Che’, ‘Hyc’, ‘Hym’ ... ]]
#>                                                                          
#> Che 0.2359409 .         .         .         .         .         .        
#> Hyc .         0.2359409 .         .         .         .         .        
#> Hym .         .         0.2359409 .         .         .         .        
#> Hys .         .         .         0.2359409 .         .         .        
#> Psy .         .         .         .         0.2359409 .         .        
#> Aga .         .         .         .         .         0.2359409 .        
#> Glo .         .         .         .         .         .         0.2359409
#> Ath .         .         .         .         .         .         .        
#> Cea .         .         .         .         .         .         .        
#> Ced .         .         .         .         .         .         .        
#> Set .         .         .         .         .         .         .        
#> All .         .         .         .         .         .         .        
#> Han .         .         .         .         .         .         .        
#> Hfo .         .         .         .         .         .         .        
#> Hsp .         .         .         .         .         .         .        
#> Hve .         .         .         .         .         .         .        
#> Sta .         .         .         .         .         .         .        
#>                                                                          
#> Che .         .         .         .         .         .         .        
#> Hyc .         .         .         .         .         .         .        
#> Hym .         .         .         .         .         .         .        
#> Hys .         .         .         .         .         .         .        
#> Psy .         .         .         .         .         .         .        
#> Aga .         .         .         .         .         .         .        
#> Glo .         .         .         .         .         .         .        
#> Ath 0.2359409 .         .         .         .         .         .        
#> Cea .         0.2359409 .         .         .         .         .        
#> Ced .         .         0.2359409 .         .         .         .        
#> Set .         .         .         0.2359409 .         .         .        
#> All .         .         .         .         0.2359409 .         .        
#> Han .         .         .         .         .         0.2359409 .        
#> Hfo .         .         .         .         .         .         0.2359409
#> Hsp .         .         .         .         .         .         .        
#> Hve .         .         .         .         .         .         .        
#> Sta .         .         .         .         .         .         .        
#>                                  
#> Che .         .         .        
#> Hyc .         .         .        
#> Hym .         .         .        
#> Hys .         .         .        
#> Psy .         .         .        
#> Aga .         .         .        
#> Glo .         .         .        
#> Ath .         .         .        
#> Cea .         .         .        
#> Ced .         .         .        
#> Set .         .         .        
#> All .         .         .        
#> Han .         .         .        
#> Hfo .         .         .        
#> Hsp 0.2359409 .         .        
#> Hve .         0.2359409 .        
#> Sta .         .         0.2359409
#>