Extract variance-covariance of residuals 'Sigma'

Extract the variance-covariance matrix of the residuals, usually noted $$\Sigma$$ in PLN models. This captures the correlation between the species in the latent space. or PLNmixture, it is a weighted mean of the variance-covariance matrices of each component.

# S3 method for PLNmixturefit
sigma(object, ...)

Arguments

object
an R6 object with class PLNmixturefit

...
additional parameters for S3 compatibility. Not used

Value

A semi definite positive matrix of size p, assuming there are p species in the model.

See also

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLNmixture(Abundance ~ 1 + offset(log(Offset)),
           data = trichoptera, control_main = list(iterates = 0))  %>% getBestModel()

#> 
#>  Initialization...
#> 
#>  Adjusting 5 PLN mixture models.
#> 	number of cluster = 1 	number of cluster = 2 	number of cluster = 3 	number of cluster = 4 	number of cluster = 5 
#> 
#>  Smoothing PLN mixture models.
#> 
#>  Post-treatments
#>  DONE!

sigma(myPLN) ## Sigma

#> [[1]]
#> 17 x 17 sparse Matrix of class "dgCMatrix"

#>    [[ suppressing 17 column names ‘Che’, ‘Hyc’, ‘Hym’ ... ]]

#>                                                                            
#> Che 0.004292829 .           .           .           .           .          
#> Hyc .           0.004292829 .           .           .           .          
#> Hym .           .           0.004292829 .           .           .          
#> Hys .           .           .           0.004292829 .           .          
#> Psy .           .           .           .           0.004292829 .          
#> Aga .           .           .           .           .           0.004292829
#> Glo .           .           .           .           .           .          
#> Ath .           .           .           .           .           .          
#> Cea .           .           .           .           .           .          
#> Ced .           .           .           .           .           .          
#> Set .           .           .           .           .           .          
#> All .           .           .           .           .           .          
#> Han .           .           .           .           .           .          
#> Hfo .           .           .           .           .           .          
#> Hsp .           .           .           .           .           .          
#> Hve .           .           .           .           .           .          
#> Sta .           .           .           .           .           .          
#>                                                                            
#> Che .           .           .           .           .           .          
#> Hyc .           .           .           .           .           .          
#> Hym .           .           .           .           .           .          
#> Hys .           .           .           .           .           .          
#> Psy .           .           .           .           .           .          
#> Aga .           .           .           .           .           .          
#> Glo 0.004292829 .           .           .           .           .          
#> Ath .           0.004292829 .           .           .           .          
#> Cea .           .           0.004292829 .           .           .          
#> Ced .           .           .           0.004292829 .           .          
#> Set .           .           .           .           0.004292829 .          
#> All .           .           .           .           .           0.004292829
#> Han .           .           .           .           .           .          
#> Hfo .           .           .           .           .           .          
#> Hsp .           .           .           .           .           .          
#> Hve .           .           .           .           .           .          
#> Sta .           .           .           .           .           .          
#>                                                                
#> Che .           .           .           .           .          
#> Hyc .           .           .           .           .          
#> Hym .           .           .           .           .          
#> Hys .           .           .           .           .          
#> Psy .           .           .           .           .          
#> Aga .           .           .           .           .          
#> Glo .           .           .           .           .          
#> Ath .           .           .           .           .          
#> Cea .           .           .           .           .          
#> Ced .           .           .           .           .          
#> Set .           .           .           .           .          
#> All .           .           .           .           .          
#> Han 0.004292829 .           .           .           .          
#> Hfo .           0.004292829 .           .           .          
#> Hsp .           .           0.004292829 .           .          
#> Hve .           .           .           0.004292829 .          
#> Sta .           .           .           .           0.004292829
#> 
#> [[2]]
#> 17 x 17 sparse Matrix of class "dgCMatrix"

#>    [[ suppressing 17 column names ‘Che’, ‘Hyc’, ‘Hym’ ... ]]

#>                                                                      
#> Che 0.04218818 .          .          .          .          .         
#> Hyc .          0.04218818 .          .          .          .         
#> Hym .          .          0.04218818 .          .          .         
#> Hys .          .          .          0.04218818 .          .         
#> Psy .          .          .          .          0.04218818 .         
#> Aga .          .          .          .          .          0.04218818
#> Glo .          .          .          .          .          .         
#> Ath .          .          .          .          .          .         
#> Cea .          .          .          .          .          .         
#> Ced .          .          .          .          .          .         
#> Set .          .          .          .          .          .         
#> All .          .          .          .          .          .         
#> Han .          .          .          .          .          .         
#> Hfo .          .          .          .          .          .         
#> Hsp .          .          .          .          .          .         
#> Hve .          .          .          .          .          .         
#> Sta .          .          .          .          .          .         
#>                                                                      
#> Che .          .          .          .          .          .         
#> Hyc .          .          .          .          .          .         
#> Hym .          .          .          .          .          .         
#> Hys .          .          .          .          .          .         
#> Psy .          .          .          .          .          .         
#> Aga .          .          .          .          .          .         
#> Glo 0.04218818 .          .          .          .          .         
#> Ath .          0.04218818 .          .          .          .         
#> Cea .          .          0.04218818 .          .          .         
#> Ced .          .          .          0.04218818 .          .         
#> Set .          .          .          .          0.04218818 .         
#> All .          .          .          .          .          0.04218818
#> Han .          .          .          .          .          .         
#> Hfo .          .          .          .          .          .         
#> Hsp .          .          .          .          .          .         
#> Hve .          .          .          .          .          .         
#> Sta .          .          .          .          .          .         
#>                                                           
#> Che .          .          .          .          .         
#> Hyc .          .          .          .          .         
#> Hym .          .          .          .          .         
#> Hys .          .          .          .          .         
#> Psy .          .          .          .          .         
#> Aga .          .          .          .          .         
#> Glo .          .          .          .          .         
#> Ath .          .          .          .          .         
#> Cea .          .          .          .          .         
#> Ced .          .          .          .          .         
#> Set .          .          .          .          .         
#> All .          .          .          .          .         
#> Han 0.04218818 .          .          .          .         
#> Hfo .          0.04218818 .          .          .         
#> Hsp .          .          0.04218818 .          .         
#> Hve .          .          .          0.04218818 .         
#> Sta .          .          .          .          0.04218818
#> 
#> [[3]]
#> 17 x 17 sparse Matrix of class "dgCMatrix"

#>    [[ suppressing 17 column names ‘Che’, ‘Hyc’, ‘Hym’ ... ]]

#>                                                                          
#> Che 0.1989391 .         .         .         .         .         .        
#> Hyc .         0.1989391 .         .         .         .         .        
#> Hym .         .         0.1989391 .         .         .         .        
#> Hys .         .         .         0.1989391 .         .         .        
#> Psy .         .         .         .         0.1989391 .         .        
#> Aga .         .         .         .         .         0.1989391 .        
#> Glo .         .         .         .         .         .         0.1989391
#> Ath .         .         .         .         .         .         .        
#> Cea .         .         .         .         .         .         .        
#> Ced .         .         .         .         .         .         .        
#> Set .         .         .         .         .         .         .        
#> All .         .         .         .         .         .         .        
#> Han .         .         .         .         .         .         .        
#> Hfo .         .         .         .         .         .         .        
#> Hsp .         .         .         .         .         .         .        
#> Hve .         .         .         .         .         .         .        
#> Sta .         .         .         .         .         .         .        
#>                                                                          
#> Che .         .         .         .         .         .         .        
#> Hyc .         .         .         .         .         .         .        
#> Hym .         .         .         .         .         .         .        
#> Hys .         .         .         .         .         .         .        
#> Psy .         .         .         .         .         .         .        
#> Aga .         .         .         .         .         .         .        
#> Glo .         .         .         .         .         .         .        
#> Ath 0.1989391 .         .         .         .         .         .        
#> Cea .         0.1989391 .         .         .         .         .        
#> Ced .         .         0.1989391 .         .         .         .        
#> Set .         .         .         0.1989391 .         .         .        
#> All .         .         .         .         0.1989391 .         .        
#> Han .         .         .         .         .         0.1989391 .        
#> Hfo .         .         .         .         .         .         0.1989391
#> Hsp .         .         .         .         .         .         .        
#> Hve .         .         .         .         .         .         .        
#> Sta .         .         .         .         .         .         .        
#>                                  
#> Che .         .         .        
#> Hyc .         .         .        
#> Hym .         .         .        
#> Hys .         .         .        
#> Psy .         .         .        
#> Aga .         .         .        
#> Glo .         .         .        
#> Ath .         .         .        
#> Cea .         .         .        
#> Ced .         .         .        
#> Set .         .         .        
#> All .         .         .        
#> Han .         .         .        
#> Hfo .         .         .        
#> Hsp 0.1989391 .         .        
#> Hve .         0.1989391 .        
#> Sta .         .         0.1989391
#> 
#> [[4]]
#> 17 x 17 sparse Matrix of class "dgCMatrix"

#>    [[ suppressing 17 column names ‘Che’, ‘Hyc’, ‘Hym’ ... ]]

#>                                                                          
#> Che 0.0462633 .         .         .         .         .         .        
#> Hyc .         0.0462633 .         .         .         .         .        
#> Hym .         .         0.0462633 .         .         .         .        
#> Hys .         .         .         0.0462633 .         .         .        
#> Psy .         .         .         .         0.0462633 .         .        
#> Aga .         .         .         .         .         0.0462633 .        
#> Glo .         .         .         .         .         .         0.0462633
#> Ath .         .         .         .         .         .         .        
#> Cea .         .         .         .         .         .         .        
#> Ced .         .         .         .         .         .         .        
#> Set .         .         .         .         .         .         .        
#> All .         .         .         .         .         .         .        
#> Han .         .         .         .         .         .         .        
#> Hfo .         .         .         .         .         .         .        
#> Hsp .         .         .         .         .         .         .        
#> Hve .         .         .         .         .         .         .        
#> Sta .         .         .         .         .         .         .        
#>                                                                          
#> Che .         .         .         .         .         .         .        
#> Hyc .         .         .         .         .         .         .        
#> Hym .         .         .         .         .         .         .        
#> Hys .         .         .         .         .         .         .        
#> Psy .         .         .         .         .         .         .        
#> Aga .         .         .         .         .         .         .        
#> Glo .         .         .         .         .         .         .        
#> Ath 0.0462633 .         .         .         .         .         .        
#> Cea .         0.0462633 .         .         .         .         .        
#> Ced .         .         0.0462633 .         .         .         .        
#> Set .         .         .         0.0462633 .         .         .        
#> All .         .         .         .         0.0462633 .         .        
#> Han .         .         .         .         .         0.0462633 .        
#> Hfo .         .         .         .         .         .         0.0462633
#> Hsp .         .         .         .         .         .         .        
#> Hve .         .         .         .         .         .         .        
#> Sta .         .         .         .         .         .         .        
#>                                  
#> Che .         .         .        
#> Hyc .         .         .        
#> Hym .         .         .        
#> Hys .         .         .        
#> Psy .         .         .        
#> Aga .         .         .        
#> Glo .         .         .        
#> Ath .         .         .        
#> Cea .         .         .        
#> Ced .         .         .        
#> Set .         .         .        
#> All .         .         .        
#> Han .         .         .        
#> Hfo .         .         .        
#> Hsp 0.0462633 .         .        
#> Hve .         0.0462633 .        
#> Sta .         .         0.0462633
#> 
#> [[5]]
#> 17 x 17 sparse Matrix of class "dgCMatrix"

#>    [[ suppressing 17 column names ‘Che’, ‘Hyc’, ‘Hym’ ... ]]

#>                                                                          
#> Che 0.1628787 .         .         .         .         .         .        
#> Hyc .         0.1628787 .         .         .         .         .        
#> Hym .         .         0.1628787 .         .         .         .        
#> Hys .         .         .         0.1628787 .         .         .        
#> Psy .         .         .         .         0.1628787 .         .        
#> Aga .         .         .         .         .         0.1628787 .        
#> Glo .         .         .         .         .         .         0.1628787
#> Ath .         .         .         .         .         .         .        
#> Cea .         .         .         .         .         .         .        
#> Ced .         .         .         .         .         .         .        
#> Set .         .         .         .         .         .         .        
#> All .         .         .         .         .         .         .        
#> Han .         .         .         .         .         .         .        
#> Hfo .         .         .         .         .         .         .        
#> Hsp .         .         .         .         .         .         .        
#> Hve .         .         .         .         .         .         .        
#> Sta .         .         .         .         .         .         .        
#>                                                                          
#> Che .         .         .         .         .         .         .        
#> Hyc .         .         .         .         .         .         .        
#> Hym .         .         .         .         .         .         .        
#> Hys .         .         .         .         .         .         .        
#> Psy .         .         .         .         .         .         .        
#> Aga .         .         .         .         .         .         .        
#> Glo .         .         .         .         .         .         .        
#> Ath 0.1628787 .         .         .         .         .         .        
#> Cea .         0.1628787 .         .         .         .         .        
#> Ced .         .         0.1628787 .         .         .         .        
#> Set .         .         .         0.1628787 .         .         .        
#> All .         .         .         .         0.1628787 .         .        
#> Han .         .         .         .         .         0.1628787 .        
#> Hfo .         .         .         .         .         .         0.1628787
#> Hsp .         .         .         .         .         .         .        
#> Hve .         .         .         .         .         .         .        
#> Sta .         .         .         .         .         .         .        
#>                                  
#> Che .         .         .        
#> Hyc .         .         .        
#> Hym .         .         .        
#> Hys .         .         .        
#> Psy .         .         .        
#> Aga .         .         .        
#> Glo .         .         .        
#> Ath .         .         .        
#> Cea .         .         .        
#> Ced .         .         .        
#> Set .         .         .        
#> All .         .         .        
#> Han .         .         .        
#> Hfo .         .         .        
#> Hsp 0.1628787 .         .        
#> Hve .         0.1628787 .        
#> Sta .         .         0.1628787
#>