Compute the stability path by stability selection

This function computes the StARS stability criteria over a path of penalties. If a path has already been computed, the functions stops with a message unless force = TRUE has been specified.

stability_selection(
  Robject,
  subsamples = NULL,
  control = list(),
  force = FALSE
)

Arguments

Robject

an object with class PLNnetworkfamily, i.e. an output from PLNnetwork()

subsamples

a list of vectors describing the subsamples. The number of vectors (or list length) determines th number of subsamples used in the stability selection. Automatically set to 20 subsamples with size 10*sqrt(n) if n >= 144 and 0.8*n otherwise following Liu et al. (2010) recommendations.

control

a list controlling the main optimization process in each call to PLNnetwork. See PLNnetwork() for details.

force

force computation of the stability path, even if a previous one has been detected.

Value

the list of subsamples. The estimated probabilities of selection of the edges are stored in the fields stability_path of the initial Robject with class PLNnetworkfamily

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
fits <- PLNnetwork(Abundance ~ 1, data = trichoptera)
#> 
#>  Initialization...
#>  Adjusting 30 PLN with sparse inverse covariance estimation
#> 	Joint optimization alternating gradient descent and graphical-lasso
#> 	sparsifying penalty = 7.017338 	sparsifying penalty = 6.48171 	sparsifying penalty = 5.986967 	sparsifying penalty = 5.529987 	sparsifying penalty = 5.107888 	sparsifying penalty = 4.718007 	sparsifying penalty = 4.357886 	sparsifying penalty = 4.025252 	sparsifying penalty = 3.718008 	sparsifying penalty = 3.434215 	sparsifying penalty = 3.172085 	sparsifying penalty = 2.929962 	sparsifying penalty = 2.706321 	sparsifying penalty = 2.49975 	sparsifying penalty = 2.308946 	sparsifying penalty = 2.132706 	sparsifying penalty = 1.969918 	sparsifying penalty = 1.819556 	sparsifying penalty = 1.680671 	sparsifying penalty = 1.552387 	sparsifying penalty = 1.433894 	sparsifying penalty = 1.324447 	sparsifying penalty = 1.223353 	sparsifying penalty = 1.129975 	sparsifying penalty = 1.043725 	sparsifying penalty = 0.9640585 	sparsifying penalty = 0.8904728 	sparsifying penalty = 0.8225038 	sparsifying penalty = 0.7597227 	sparsifying penalty = 0.7017338 
#>  Post-treatments
#>  DONE!
if (FALSE) {
n <- nrow(trichoptera)
subs <- replicate(10, sample.int(n, size = n/2), simplify = FALSE)
stability_selection(nets, subsamples = subs)
}