Component-wise standard errors of Theta

Extracts univariate standard errors for the estimated coefficient of Theta. Standard errors are computed from the (approximate) Fisher information matrix. See fisher.PLNfit() for more details on the approximations.

standard_error(object, type)

# S3 method for PLNfit
standard_error(object, type = c("wald", "louis"))

Arguments

object

an R6 object with class PLNfit

type

Either Wald (default) or Louis. Approximation scheme used to compute the Fisher information matrix

Value

A p * d positive matrix (same size as \(\Theta\)) with standard errors for the coefficients of \(\Theta\)

Methods (by class)

PLNfit: Component-wise standard errors of Theta in PLNfit

See also

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLN(Abundance ~ 1 + offset(log(Offset)), data = trichoptera)
#> 
#>  Initialization...
#>  Adjusting a PLN model with full covariance model
#>  Post-treatments...
#>  DONE!
standard_error(myPLN, "wald")
#>     (Intercept)
#> Che  0.57741398
#> Hyc  0.57706627
#> Hym  0.07392034
#> Hys  0.40818728
#> Psy  0.01292748
#> Aga  0.09578796
#> Glo  0.26724220
#> Ath  0.26722671
#> Cea  0.37869728
#> Ced  0.09284253
#> Set  0.07270957
#> All  0.13868952
#> Han  0.07231423
#> Hfo  0.08660428
#> Hsp  0.04610648
#> Hve  0.33343656
#> Sta  0.05861526