Component-wise standard errors of Theta

Extracts univariate standard errors for the estimated coefficient of Theta. Standard errors are computed from the (approximate) Fisher information matrix. See fisher for more details on the approximations.

standard_error(object, type)

# S3 method for PLNfit
standard_error(object, type = c("wald", "louis"))

Arguments

object
an R6 object with class PLNfit

type
Either `Wald` (default) or `Louis`. Approximation scheme used to compute the Fisher information matrix

Value

A p * d positive matrix (same size as \(\Theta\)) with standard errors for the coefficients of \(\Theta\)

Methods (by class)

PLNfit: Component-wise standard errors of Theta in PLNfit

See also

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLN(Abundance ~ 1 + offset(log(Offset)), data = trichoptera)

#> 
#>  Initialization...
#>  Adjusting a PLN model with full covariance model
#>  Post-treatments...
#>  DONE!

standard_error(myPLN, "wald")

#>     (Intercept)
#> Che  0.57738117
#> Hyc  0.57717051
#> Hym  0.07392176
#> Hys  0.40823393
#> Psy  0.01292380
#> Aga  0.09580570
#> Glo  0.26724765
#> Ath  0.26726899
#> Cea  0.37797519
#> Ced  0.09284341
#> Set  0.07273334
#> All  0.13868125
#> Han  0.07230642
#> Hfo  0.08668585
#> Hsp  0.04611883
#> Hve  0.33342067
#> Sta  0.05861896