Trichoptera data set

Data gathered between 1959 and 1960 during 49 insect trapping nights. For each trapping night, the abundance of 17 Trichoptera species is recorded, which forms the first table. The second table concerns 11 meteorological variables which may influence the abundance of each species. Finally, the observations (that is to say, the trapping nights), have been classified into 12 groups corresponding to contiguous nights between summer 1959 and summer 1960.

trichoptera

Format

A data frame with 49 rows and 3 variables (matrices):

Abundance

a 49 x 17 matrix of abundancies (49 trapping nights and 17 trichoptera species).

T_max

Maximal temperature in Celsius

T_evening

Evening Temperature in Celsius

T_min

Minimal temperature in Celsius

Wind

Wind in m/s

Pressure

Pressure in mm Hg

Pressure_range

Pressure range in mm Hg

Humidity

relative to evening humidity in percent

Cloudiness_night

proportion of sky coverage at 9pm

Precipitation_night

Nighttime precipitation in mm

Cloudiness_average

average proportion of sky coverage

Precipitation_total

Total precipitation in mm

Group

a factor of 12 levels for the definition of the consecutive night groups

This format is convenient for using formula in multivariate analysis (multiple outputs and inputs).

Source

Data from P. Usseglio-Polatera.

References

Usseglio-Polatera, P. and Auda, Y. (1987) Influence des facteurs météorologiques sur les résultats de piégeage lumineux. Annales de Limnologie, 23, 65–79. (code des espèces p. 76) See a data description at http://pbil.univ-lyon1.fr/R/pdf/pps034.pdf (in French)

Examples

data(trichoptera)
str(trichoptera)

#> 'data.frame':	49 obs. of  13 variables:
#>  $ Abundance          : 'AsIs' num [1:49, 1:17] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr  "1" "2" "3" "4" ...
#>   .. ..$ : chr  "Che" "Hyc" "Hym" "Hys" ...
#>  $ T_max              : num  22.2 24.2 27.4 29 28.2 28 20.5 20 18.6 18.1 ...
#>  $ T_evening          : num  18.7 19.8 22 23 22.5 23.9 15 17.2 15.4 14.1 ...
#>  $ T_min              : num  11 8.4 11 12.6 12 13.8 12.6 7.9 10 8.2 ...
#>  $ Wind               : num  -2.3 -2.7 -0.7 2.3 2.3 -2 -4.7 -1 -2.7 -3.7 ...
#>  $ Pressure           : num  998 1000 997 991 990 ...
#>  $ Pressure_range     : num  4.1 -1.1 -3.6 -6.2 1.7 -0.8 6.4 -0.7 3.1 2.4 ...
#>  $ Humidity           : num  60 63 73 71 62 64 93 84 88 75 ...
#>  $ Cloudiness_night   : num  19 0 6 81 50 50 100 19 69 6 ...
#>  $ Precipitation_night: num  0 0 0 0 0 0 1.6 0 1.6 0 ...
#>  $ Cloudiness_average : num  17 27 13 52 34 36 97 48 77 28 ...
#>  $ Precipitation_total: num  0 0 0 0 0 0 1.6 0.1 1.7 0 ...
#>  $ Group              : Factor w/ 12 levels "1","2","3","4",..: 1 1 1 1 1 1 1 1 1 1 ...

## also see the package vignettes