Calculate Variance-Covariance Matrix for a fitted PLN() model object

Returns the variance-covariance matrix of the main parameters of a fitted PLN() model object. The main parameters of the model correspond to $$B$$, as returned by coef.PLNfit(). The function can also be used to return the variance-covariance matrix of the residuals. The latter matrix can also be accessed via sigma.PLNfit()

# S3 method for class 'PLNfit'
vcov(object, type = c("main", "covariance"), ...)

Arguments

object

an R6 object with class PLNfit

type

type of parameter that should be extracted. Either "main" (default) for $$B$$ or "covariance" for $$\Sigma$$

...

additional parameters for S3 compatibility. Not used

Value

A matrix of variance/covariance extracted from the PLNfit model. If type="main" and \(B\) is a matrix of size d * p, the result is a block-diagonal matrix with p (number of species) blocks of size d (number of covariates). if type="main", it is a symmetric matrix of size p. .

See also

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLN(Abundance ~ 1 + offset(log(Offset)), data = trichoptera)
#> 
#>  Initialization...
#>  Adjusting a full covariance PLN model with builtin optimizer
#>  Post-treatments...
#>  DONE!
vcov(myPLN, type = "covariance") ## Sigma
#>              Che          Hyc         Hym         Hys           Psy
#> Che  2.521004879 -0.009621748  0.36202871  0.44639375 -0.0174331701
#> Hyc -0.009621748  1.884847015  0.10623049  0.26240559 -0.1444136471
#> Hym  0.362028711  0.106230486  1.24696664  0.86539400 -0.1168954762
#> Hys  0.446393751  0.262405590  0.86539400  2.65863350 -0.1443170333
#> Psy -0.017433170 -0.144413647 -0.11689548 -0.14431703  0.0757392425
#> Aga  0.307823860 -0.126838378  0.51730324  0.43084002  0.0001376518
#> Glo  0.089844948  0.166211450  0.20809259  0.33855604 -0.0484036391
#> Ath  0.353484462  0.104954266  0.38574934  0.60568792 -0.0463288356
#> Cea -0.017890515  0.012163203 -0.04470148 -0.03254729  0.0003721319
#> Ced  0.338629891  0.218379122  0.50794378  0.67941042 -0.0972694203
#> Set -0.061240053  0.474809922 -0.18965441  0.13067256 -0.0912525748
#> All  0.272227213 -0.260398204  0.68537220  0.37524882 -0.0049566790
#> Han -0.190282757  0.626177994  0.19801950  0.12275677 -0.2909474891
#> Hfo -0.375537379  0.564732976 -0.41675795 -0.20482200 -0.1473791488
#> Hsp -0.460651058  0.675704771 -0.36138697 -0.20672169 -0.1924982413
#> Hve  0.313100965 -0.130219312  0.16492155  0.24505003  0.0453696868
#> Sta  0.362585421  0.256496642  0.47721224  0.61297085 -0.1215073783
#>               Aga          Glo         Ath           Cea         Ced
#> Che  0.3078238599  0.089844948  0.35348446 -0.0178905149  0.33862989
#> Hyc -0.1268383781  0.166211450  0.10495427  0.0121632029  0.21837912
#> Hym  0.5173032352  0.208092591  0.38574934 -0.0447014798  0.50794378
#> Hys  0.4308400156  0.338556038  0.60568792 -0.0325472855  0.67941042
#> Psy  0.0001376518 -0.048403639 -0.04632884  0.0003721319 -0.09726942
#> Aga  1.0969937583 -0.060299881  0.20025725 -0.0462745674  0.22108092
#> Glo -0.0602998814  1.115805410  0.22238118  0.0036140181  0.22419948
#> Ath  0.2002572516  0.222381182  1.37201187 -0.0127252795  0.44180172
#> Cea -0.0462745674  0.003614018 -0.01272528  0.1107330918 -0.01397130
#> Ced  0.2210809197  0.224199475  0.44180172 -0.0139712955  0.95816390
#> Set -0.3011056096  0.290107993  0.09193515  0.0332103341  0.19249831
#> All  0.5918619743 -0.074302561  0.09145003 -0.0530504532  0.08195119
#> Han -0.1752991113 -0.199141863 -0.20907699  0.0102008023  0.05225075
#> Hfo -0.8254721519  0.299208538 -0.17375353  0.0572655297 -0.14377513
#> Hsp -0.7898624952  0.246046886 -0.23974986  0.0526218795 -0.12955968
#> Hve  0.1861438946  0.089892014  0.24432164 -0.0140023858  0.20567692
#> Sta  0.4396727112  0.015858484  0.30713667 -0.0268157180  0.43453646
#>             Set          All         Han         Hfo         Hsp         Hve
#> Che -0.06124005  0.272227213 -0.19028276 -0.37553738 -0.46065106  0.31310097
#> Hyc  0.47480992 -0.260398204  0.62617799  0.56473298  0.67570477 -0.13021931
#> Hym -0.18965441  0.685372201  0.19801950 -0.41675795 -0.36138697  0.16492155
#> Hys  0.13067256  0.375248823  0.12275677 -0.20482200 -0.20672169  0.24505003
#> Psy -0.09125257 -0.004956679 -0.29094749 -0.14737915 -0.19249824  0.04536969
#> Aga -0.30110561  0.591861974 -0.17529911 -0.82547215 -0.78986250  0.18614389
#> Glo  0.29010799 -0.074302561 -0.19914186  0.29920854  0.24604689  0.08989201
#> Ath  0.09193515  0.091450026 -0.20907699 -0.17375353 -0.23974986  0.24432164
#> Cea  0.03321033 -0.053050453  0.01020080  0.05726553  0.05262188 -0.01400239
#> Ced  0.19249831  0.081951192  0.05225075 -0.14377513 -0.12955968  0.20567692
#> Set  1.60083536 -0.686896979 -0.03999138  0.83108168  0.78475434 -0.04267759
#> All -0.68689698  1.805936001  0.02100351 -0.75809328 -0.67247834  0.14837102
#> Han -0.03999138  0.021003507  3.75721785  0.64390222  1.01552123 -0.60939930
#> Hfo  0.83108168 -0.758093278  0.64390222  2.86798568  1.69929471 -0.35294443
#> Hsp  0.78475434 -0.672478343  1.01552123  1.69929471  2.62607097 -0.49539413
#> Hve -0.04267759  0.148371019 -0.60939930 -0.35294443 -0.49539413  1.34599211
#> Sta  0.05587583  0.199051454  0.47126251 -0.26024512 -0.19907362  0.14306509
#>             Sta
#> Che  0.36258542
#> Hyc  0.25649664
#> Hym  0.47721224
#> Hys  0.61297085
#> Psy -0.12150738
#> Aga  0.43967271
#> Glo  0.01585848
#> Ath  0.30713667
#> Cea -0.02681572
#> Ced  0.43453646
#> Set  0.05587583
#> All  0.19905145
#> Han  0.47126251
#> Hfo -0.26024512
#> Hsp -0.19907362
#> Hve  0.14306509
#> Sta  1.03098743