Calculate Variance-Covariance Matrix for a fitted PLN model object

Returns the variance-covariance matrix of the main parameters of a fitted PLN model object. The main parameters of the model correspond to $$\Theta$$, as returned by coef. The function can also be used to return the variance-covariance matrix of the residuals. The latter matrix can also be accessed via sigma

# S3 method for PLNfit
vcov(object, type = c("main", "covariance"), ...)

Arguments

object
an R6 object with class PLNfit

type
type of parameter that should be extracted. Either "main" (default) for $$\Theta$$ or "covariance" for $$\Sigma$$

...
additional parameters for S3 compatibility. Not used

Value

A matrix of variance/covariance extracted from the PLNfit model. If type="main" and \(\Theta\) is a matrix of size d * p, the result is a block-diagonal matrix with p (number of species) blocks of size d (number of covariates). if type="main", it is a symmetric matrix of size p. .

See also

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLN(Abundance ~ 1 + offset(log(Offset)), data = trichoptera)

#> 
#>  Initialization...
#>  Adjusting a PLN model with full covariance model
#>  Post-treatments...
#>  DONE!

vcov(myPLN) ## variance-covariance of Theta

#> 17 x 17 sparse Matrix of class "dgCMatrix"
#>                                                                               
#> Che_(Intercept) 2.999679 .          .      .           .       .       .      
#> Hyc_(Intercept) .        3.001869   .      .           .       .       .      
#> Hym_(Intercept) .        .        183.0018 .           .       .       .      
#> Hys_(Intercept) .        .          .      6.000422    .       .       .      
#> Psy_(Intercept) .        .          .      .        5987.137   .       .      
#> Aga_(Intercept) .        .          .      .           .     108.9475  .      
#> Glo_(Intercept) .        .          .      .           .       .      14.00142
#> Ath_(Intercept) .        .          .      .           .       .       .      
#> Cea_(Intercept) .        .          .      .           .       .       .      
#> Ced_(Intercept) .        .          .      .           .       .       .      
#> Set_(Intercept) .        .          .      .           .       .       .      
#> All_(Intercept) .        .          .      .           .       .       .      
#> Han_(Intercept) .        .          .      .           .       .       .      
#> Hfo_(Intercept) .        .          .      .           .       .       .      
#> Hsp_(Intercept) .        .          .      .           .       .       .      
#> Hve_(Intercept) .        .          .      .           .       .       .      
#> Sta_(Intercept) .        .          .      .           .       .       .      
#>                                                                              
#> Che_(Intercept)  .       .          .        .      .         .        .     
#> Hyc_(Intercept)  .       .          .        .      .         .        .     
#> Hym_(Intercept)  .       .          .        .      .         .        .     
#> Hys_(Intercept)  .       .          .        .      .         .        .     
#> Psy_(Intercept)  .       .          .        .      .         .        .     
#> Aga_(Intercept)  .       .          .        .      .         .        .     
#> Glo_(Intercept)  .       .          .        .      .         .        .     
#> Ath_(Intercept) 13.99919 .          .        .      .         .        .     
#> Cea_(Intercept)  .       6.999603   .        .      .         .        .     
#> Ced_(Intercept)  .       .        116.0106   .      .         .        .     
#> Set_(Intercept)  .       .          .      189.031  .         .        .     
#> All_(Intercept)  .       .          .        .     51.99535   .        .     
#> Han_(Intercept)  .       .          .        .      .       191.2697   .     
#> Hfo_(Intercept)  .       .          .        .      .         .      133.0772
#> Hsp_(Intercept)  .       .          .        .      .         .        .     
#> Hve_(Intercept)  .       .          .        .      .         .        .     
#> Sta_(Intercept)  .       .          .        .      .         .        .     
#>                                           
#> Che_(Intercept)   .      .          .     
#> Hyc_(Intercept)   .      .          .     
#> Hym_(Intercept)   .      .          .     
#> Hys_(Intercept)   .      .          .     
#> Psy_(Intercept)   .      .          .     
#> Aga_(Intercept)   .      .          .     
#> Glo_(Intercept)   .      .          .     
#> Ath_(Intercept)   .      .          .     
#> Cea_(Intercept)   .      .          .     
#> Ced_(Intercept)   .      .          .     
#> Set_(Intercept)   .      .          .     
#> All_(Intercept)   .      .          .     
#> Han_(Intercept)   .      .          .     
#> Hfo_(Intercept)   .      .          .     
#> Hsp_(Intercept) 470.1576 .          .     
#> Hve_(Intercept)   .      8.995285   .     
#> Sta_(Intercept)   .      .        291.0206

vcov(myPLN, type = "covariance") ## Sigma

#>              Che          Hyc         Hym         Hys           Psy
#> Che  2.528781765 -0.009321441  0.36115840  0.44523414 -0.0178562188
#> Hyc -0.009321441  1.905620068  0.11077168  0.26938459 -0.1459207014
#> Hym  0.361158402  0.110771680  1.24461838  0.86526660 -0.1184577858
#> Hys  0.445234144  0.269384590  0.86526660  2.66250773 -0.1464421679
#> Psy -0.017856219 -0.145920701 -0.11845779 -0.14644217  0.0760744067
#> Aga  0.308887610 -0.131135175  0.51329925  0.42558988  0.0011159984
#> Glo  0.089212302  0.170891239  0.20969683  0.34218405 -0.0496706493
#> Ath  0.352527176  0.109054287  0.38593837  0.60735573 -0.0477009923
#> Cea -0.007774911  0.004170301 -0.01740978 -0.01334907  0.0003104627
#> Ced  0.337968442  0.221466684  0.50823585  0.68094771 -0.0981059607
#> Set -0.060652239  0.477423283 -0.18426997  0.13672176 -0.0912743057
#> All  0.273033450 -0.262491701  0.68092426  0.37052261 -0.0049712209
#> Han -0.182872467  0.623163335  0.20974796  0.13513474 -0.2885105601
#> Hfo -0.366433932  0.562253416 -0.39707877 -0.18499710 -0.1463185348
#> Hsp -0.448724077  0.674990120 -0.33999409 -0.18319423 -0.1920106572
#> Hve  0.312639677 -0.131434292  0.16102196  0.24138198  0.0456983444
#> Sta  0.364105311  0.256129639  0.47822405  0.61385913 -0.1213504767
#>              Aga          Glo         Ath           Cea          Ced
#> Che  0.308887610  0.089212302  0.35252718 -0.0077749106  0.337968442
#> Hyc -0.131135175  0.170891239  0.10905429  0.0041703013  0.221466684
#> Hym  0.513299252  0.209696827  0.38593837 -0.0174097811  0.508235849
#> Hys  0.425589881  0.342184054  0.60735573 -0.0133490713  0.680947712
#> Psy  0.001115998 -0.049670649 -0.04770099  0.0003104627 -0.098105961
#> Aga  1.101910616 -0.063544961  0.19784669 -0.0178946528  0.218949159
#> Glo -0.063544961  1.117798373  0.22427496  0.0011841538  0.226213232
#> Ath  0.197846692  0.224274956  1.37358756 -0.0056755800  0.442641753
#> Cea -0.017894653  0.001184154 -0.00567558  0.0341995492 -0.006422698
#> Ced  0.218949159  0.226213232  0.44264175 -0.0064226979  0.958254647
#> Set -0.302123370  0.292835137  0.09532313  0.0119897468  0.194287594
#> All  0.592029083 -0.076641423  0.08949734 -0.0197588039  0.080777696
#> Han -0.171748718 -0.189463999 -0.19833456  0.0031880972  0.056222478
#> Hfo -0.817118690  0.302843502 -0.16049344  0.0216331592 -0.132404217
#> Hsp -0.783944157  0.252083226 -0.22330580  0.0200824436 -0.116839757
#> Hve  0.186124822  0.088581506  0.24244204 -0.0056580906  0.204338024
#> Sta  0.439957507  0.016710729  0.30823764 -0.0113042325  0.434535427
#>             Set          All          Han         Hfo         Hsp          Hve
#> Che -0.06065224  0.273033450 -0.182872467 -0.36643393 -0.44872408  0.312639677
#> Hyc  0.47742328 -0.262491701  0.623163335  0.56225342  0.67499012 -0.131434292
#> Hym -0.18426997  0.680924264  0.209747958 -0.39707877 -0.33999409  0.161021963
#> Hys  0.13672176  0.370522606  0.135134741 -0.18499710 -0.18319423  0.241381976
#> Psy -0.09127431 -0.004971221 -0.288510560 -0.14631853 -0.19201066  0.045698344
#> Aga -0.30212337  0.592029083 -0.171748718 -0.81711869 -0.78394416  0.186124822
#> Glo  0.29283514 -0.076641423 -0.189463999  0.30284350  0.25208323  0.088581506
#> Ath  0.09532313  0.089497335 -0.198334556 -0.16049344 -0.22330580  0.242442043
#> Cea  0.01198975 -0.019758804  0.003188097  0.02163316  0.02008244 -0.005658091
#> Ced  0.19428759  0.080777696  0.056222478 -0.13240422 -0.11683976  0.204338024
#> Set  1.59393898 -0.684794514 -0.042563129  0.81650269  0.77262810 -0.041107782
#> All -0.68479451  1.805809281  0.030345620 -0.74456780 -0.66133353  0.146452819
#> Han -0.04256313  0.030345620  3.652545695  0.62175759  0.99141497 -0.601163498
#> Hfo  0.81650269 -0.744567804  0.621757590  2.77332661  1.64758445 -0.344387358
#> Hsp  0.77262810 -0.661333525  0.991414975  1.64758445  2.54435009 -0.485797413
#> Hve -0.04110778  0.146452819 -0.601163498 -0.34438736 -0.48579741  1.351961466
#> Sta  0.05469796  0.200042056  0.464159581 -0.25744009 -0.19549766  0.143341907
#>             Sta
#> Che  0.36410531
#> Hyc  0.25612964
#> Hym  0.47822405
#> Hys  0.61385913
#> Psy -0.12135048
#> Aga  0.43995751
#> Glo  0.01671073
#> Ath  0.30823764
#> Cea -0.01130423
#> Ced  0.43453543
#> Set  0.05469796
#> All  0.20004206
#> Han  0.46415958
#> Hfo -0.25744009
#> Hsp -0.19549766
#> Hve  0.14334191
#> Sta  1.02936810