Calculate Variance-Covariance Matrix for a fitted PLN model object

Returns the variance-covariance matrix of the main parameters of a fitted PLN model object. The main parameters of the model correspond to $$\Theta$$, as returned by coef. The function can also be used to return the variance-covariance matrix of the residuals. The latter matrix can also be accessed via sigma

# S3 method for PLNfit
vcov(object, type = c("main", "covariance"), ...)

Arguments

object
an R6 object with class PLNfit

type
type of parameter that should be extracted. Either "main" (default) for $$\Theta$$ or "covariance" for $$\Sigma$$

...
additional parameters for S3 compatibility. Not used

Value

A matrix of variance/covariance extracted from the PLNfit model. If type="main" and \(\Theta\) is a matrix of size d * p, the result is a block-diagonal matrix with p (number of species) blocks of size d (number of covariates). if type="main", it is a symmetric matrix of size p. .

See also

Examples

data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLN(Abundance ~ 1 + offset(log(Offset)), data = trichoptera)

#> 
#>  Initialization...
#>  Adjusting a PLN model with full covariance model
#>  Post-treatments...
#>  DONE!

vcov(myPLN) ## variance-covariance of Theta

#> 17 x 17 sparse Matrix of class "dgCMatrix"
#>                                                                              
#> Che_(Intercept) 2.999883 .          .     .           .       .       .      
#> Hyc_(Intercept) .        3.000251   .     .           .       .       .      
#> Hym_(Intercept) .        .        183.006 .           .       .       .      
#> Hys_(Intercept) .        .          .     6.000154    .       .       .      
#> Psy_(Intercept) .        .          .     .        5987.795   .       .      
#> Aga_(Intercept) .        .          .     .           .     108.9309  .      
#> Glo_(Intercept) .        .          .     .           .       .      14.00115
#> Ath_(Intercept) .        .          .     .           .       .       .      
#> Cea_(Intercept) .        .          .     .           .       .       .      
#> Ced_(Intercept) .        .          .     .           .       .       .      
#> Set_(Intercept) .        .          .     .           .       .       .      
#> All_(Intercept) .        .          .     .           .       .       .      
#> Han_(Intercept) .        .          .     .           .       .       .      
#> Hfo_(Intercept) .        .          .     .           .       .       .      
#> Hsp_(Intercept) .        .          .     .           .       .       .      
#> Hve_(Intercept) .        .          .     .           .       .       .      
#> Sta_(Intercept) .        .          .     .           .       .       .      
#>                                                                               
#> Che_(Intercept)  .       .          .        .       .         .        .     
#> Hyc_(Intercept)  .       .          .        .       .         .        .     
#> Hym_(Intercept)  .       .          .        .       .         .        .     
#> Hys_(Intercept)  .       .          .        .       .         .        .     
#> Psy_(Intercept)  .       .          .        .       .         .        .     
#> Aga_(Intercept)  .       .          .        .       .         .        .     
#> Glo_(Intercept)  .       .          .        .       .         .        .     
#> Ath_(Intercept) 13.99934 .          .        .       .         .        .     
#> Cea_(Intercept)  .       6.999747   .        .       .         .        .     
#> Ced_(Intercept)  .       .        116.0018   .       .         .        .     
#> Set_(Intercept)  .       .          .      189.0087  .         .        .     
#> All_(Intercept)  .       .          .        .      51.99403   .        .     
#> Han_(Intercept)  .       .          .        .       .       191.0443   .     
#> Hfo_(Intercept)  .       .          .        .       .         .      133.1213
#> Hsp_(Intercept)  .       .          .        .       .         .        .     
#> Hve_(Intercept)  .       .          .        .       .         .        .     
#> Sta_(Intercept)  .       .          .        .       .         .        .     
#>                                           
#> Che_(Intercept)   .      .          .     
#> Hyc_(Intercept)   .      .          .     
#> Hym_(Intercept)   .      .          .     
#> Hys_(Intercept)   .      .          .     
#> Psy_(Intercept)   .      .          .     
#> Aga_(Intercept)   .      .          .     
#> Glo_(Intercept)   .      .          .     
#> Ath_(Intercept)   .      .          .     
#> Cea_(Intercept)   .      .          .     
#> Ced_(Intercept)   .      .          .     
#> Set_(Intercept)   .      .          .     
#> All_(Intercept)   .      .          .     
#> Han_(Intercept)   .      .          .     
#> Hfo_(Intercept)   .      .          .     
#> Hsp_(Intercept) 470.3516 .          .     
#> Hve_(Intercept)   .      8.998199   .     
#> Sta_(Intercept)   .      .        291.0283

vcov(myPLN, type = "covariance") ## Sigma

#>              Che          Hyc         Hym         Hys           Psy
#> Che  2.527539720 -0.010379195  0.36082220  0.44508083 -0.0175272360
#> Hyc -0.010379195  1.905634239  0.11078984  0.26913191 -0.1460556877
#> Hym  0.360822198  0.110789844  1.24465470  0.86571370 -0.1184892047
#> Hys  0.445080833  0.269131909  0.86571370  2.66323940 -0.1463640232
#> Psy -0.017527236 -0.146055688 -0.11848920 -0.14636402  0.0760420177
#> Aga  0.309242242 -0.131836249  0.51363940  0.42615628  0.0011869070
#> Glo  0.088905334  0.171152180  0.20997518  0.34238082 -0.0497451489
#> Ath  0.352566578  0.109523713  0.38584523  0.60775163 -0.0478314069
#> Cea -0.008600705  0.004829909 -0.01957675 -0.01488447  0.0003187878
#> Ced  0.338095312  0.221717954  0.50827915  0.68111355 -0.0981836103
#> Set -0.061077380  0.477662924 -0.18336023  0.13679517 -0.0913420412
#> All  0.273161537 -0.263754471  0.68093135  0.37073525 -0.0047374653
#> Han -0.185499275  0.625158622  0.20814366  0.13288242 -0.2889685299
#> Hfo -0.365033750  0.558845296 -0.39296795 -0.18385448 -0.1453498794
#> Hsp -0.447862660  0.673662743 -0.33618236 -0.18057257 -0.1915302074
#> Hve  0.313336791 -0.131672331  0.16180597  0.24223423  0.0455642203
#> Sta  0.363654211  0.256322325  0.47827672  0.61394224 -0.1213850095
#>              Aga          Glo          Ath           Cea         Ced
#> Che  0.309242242  0.088905334  0.352566578 -0.0086007050  0.33809531
#> Hyc -0.131836249  0.171152180  0.109523713  0.0048299091  0.22171795
#> Hym  0.513639398  0.209975179  0.385845235 -0.0195767526  0.50827915
#> Hys  0.426156284  0.342380818  0.607751625 -0.0148844674  0.68111355
#> Psy  0.001186907 -0.049745149 -0.047831407  0.0003187878 -0.09818361
#> Aga  1.102805341 -0.063463746  0.197440672 -0.0202238728  0.21885674
#> Glo -0.063463746  1.117679779  0.224748686  0.0014148876  0.22639398
#> Ath  0.197440672  0.224748686  1.374851232 -0.0062178636  0.44298415
#> Cea -0.020223873  0.001414888 -0.006217864  0.0400005056 -0.00702893
#> Ced  0.218856745  0.226393984  0.442984153 -0.0070289303  0.95830305
#> Set -0.302119318  0.292805153  0.096295871  0.0136798138  0.19461104
#> All  0.593106471 -0.077024466  0.088665312 -0.0224455277  0.08036213
#> Han -0.173612612 -0.190749089 -0.199880606  0.0036998080  0.05558420
#> Hfo -0.810672289  0.300138493 -0.158214005  0.0242542960 -0.13062846
#> Hsp -0.781291595  0.252416595 -0.219753003  0.0225514288 -0.11408739
#> Hve  0.186840998  0.088583589  0.242635073 -0.0063295680  0.20465642
#> Sta  0.439927367  0.016943743  0.308268108 -0.0125973905  0.43463781
#>             Set          All          Han        Hfo         Hsp          Hve
#> Che -0.06107738  0.273161537 -0.185499275 -0.3650337 -0.44786266  0.313336791
#> Hyc  0.47766292 -0.263754471  0.625158622  0.5588453  0.67366274 -0.131672331
#> Hym -0.18336023  0.680931351  0.208143656 -0.3929679 -0.33618236  0.161805969
#> Hys  0.13679517  0.370735249  0.132882417 -0.1838545 -0.18057257  0.242234227
#> Psy -0.09134204 -0.004737465 -0.288968530 -0.1453499 -0.19153021  0.045564220
#> Aga -0.30211932  0.593106471 -0.173612612 -0.8106723 -0.78129159  0.186840998
#> Glo  0.29280515 -0.077024466 -0.190749089  0.3001385  0.25241660  0.088583589
#> Ath  0.09629587  0.088665312 -0.199880606 -0.1582140 -0.21975300  0.242635073
#> Cea  0.01367981 -0.022445528  0.003699808  0.0242543  0.02255143 -0.006329568
#> Ced  0.19461104  0.080362130  0.055584195 -0.1306285 -0.11408739  0.204656421
#> Set  1.59076917 -0.684896109 -0.042710690  0.8086657  0.76969030 -0.041445018
#> All -0.68489611  1.808115229  0.028129244 -0.7398110 -0.66008179  0.147492565
#> Han -0.04271069  0.028129244  3.677181720  0.6209594  0.98926626 -0.602842634
#> Hfo  0.80866572 -0.739811023  0.620959356  2.7248708  1.62775723 -0.342806526
#> Hsp  0.76969030 -0.660081790  0.989266262  1.6277572  2.52223701 -0.483422776
#> Hve -0.04144502  0.147492565 -0.602842634 -0.3428065 -0.48342278  1.350778891
#> Sta  0.05508175  0.199837584  0.465157827 -0.2547320 -0.19383569  0.143652823
#>             Sta
#> Che  0.36365421
#> Hyc  0.25632233
#> Hym  0.47827672
#> Hys  0.61394224
#> Psy -0.12138501
#> Aga  0.43992737
#> Glo  0.01694374
#> Ath  0.30826811
#> Cea -0.01259739
#> Ced  0.43463781
#> Set  0.05508175
#> All  0.19983758
#> Han  0.46515783
#> Hfo -0.25473200
#> Hsp -0.19383569
#> Hve  0.14365282
#> Sta  1.02924834